+++ package.json
@@ -17,17 +17,15 @@
   "dependencies": {
     "@babel/core": "^7.8.3",
     "@devexperts/remote-data-ts": "^2.0.3",
-    "@emotion/core": "^10.0.27",
-    "@emotion/styled": "^10.0.27",
     "@octokit/rest": "^16.38.3",
     "@types/node": "^13.5.0",
     "@types/react": "^16.9.19",
     "@types/react-select": "^3.0.10",
+    "@types/styled-components": "^4.4.2",
     "@types/use-persisted-state": "^0.3.0",
     "@zeit/next-css": "^1.0.1",
-    "@zeit/next-sass": "^1.0.1",
     "@zeit/next-typescript": "^1.1.1",
-    "babel-plugin-emotion": "^10.0.27",
+    "babel-plugin-styled-components": "^1.10.7",
     "cssnano": "^4.1.10",
     "dayjs": "^1.8.19",
     "downshift": "^4.0.8",
@@ -38,7 +36,6 @@
     "mq-scss": "^2.1.5",
     "newtype-ts": "^0.3.3",
     "next": "^9.2.1",
-    "next-fonts": "^1.0.3",
     "next-offline": "^4.0.6",
     "njsx": "^3.0.1",
     "njsx-react": "^1.0.1",
@@ -50,14 +47,13 @@
     "postcss-normalize": "^8.0.1",
     "postcss-preset-env": "^6.7.0",
     "postcss-scrollbar": "^0.3.0",
-    "postcss-scss": "^2.0.0",
     "postcss-ts-classnames": "^0.1.2",
     "postcss-will-change-transition": "^1.2.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-feather": "^2.0.3",
     "react-functional-select": "^2.0.4",
-    "react-select": "^3.0.8",
+    "react-is": "^16.12.0",
     "react-window": "^1.8.5",
     "storybook-addon-paddings": "^1.0.0",
     "styled-components": "^5.0.0",
