+++ src/classnames.d.ts
@@ -1,1 +1,1 @@
-type ClassNames = "-m-1" | "-m-10" | "-m-12" | "-m-16" | "-m-2" | "-m-20" | "-m-24" | "-m-3" | "-m-32" | "-m-4" | "-m-40" | "-m-48" | "-m-5" | "-m-56" | "-m-6" | "-m-64" | "-m-8" | "-m-px" | "-mb-1" | "-mb-10" | "-mb-12" | "-mb-16" | "-mb-2" | "-mb-20" | "-mb-24" | "-mb-3" | "-mb-32" | "-mb-4" | "-mb-40" | "-mb-48" | "-mb-5" | "-mb-56" | "-mb-6" | "-mb-64" | "-mb-8" | "-mb-px" | "-ml-1" | "-ml-10" | "-ml-12" | "-ml-16" | "-ml-2" | "-ml-20" | "-ml-24" | "-ml-3" | "-ml-32" | "-ml-4" | "-ml-40" | "-ml-48" | "-ml-5" | "-ml-56" | "-ml-6" | "-ml-64" | "-ml-8" | "-ml-px" | "-mr-1" | "-mr-10" | "-mr-12" | "-mr-16" | "-mr-2" | "-mr-20" | "-mr-24" | "-mr-3" | "-mr-32" | "-mr-4" | "-mr-40" | "-mr-48" | "-mr-5" | "-mr-56" | "-mr-6" | "-mr-64" | "-mr-8" | "-mr-px" | "-mt-1" | "-mt-10" | "-mt-12" | "-mt-16" | "-mt-2" | "-mt-20" | "-mt-24" | "-mt-3" | "-mt-32" | "-mt-4" | "-mt-40" | "-mt-48" | "-mt-5" | "-mt-56" | "-mt-6" | "-mt-64" | "-mt-8" | "-mt-px" | "-mx-1" | "-mx-10" | "-mx-12" | "-mx-16" | "-mx-2" | "-mx-20" | "-mx-24" | "-mx-3" | "-mx-32" | "-mx-4" | "-mx-40" | "-mx-48" | "-mx-5" | "-mx-56" | "-mx-6" | "-mx-64" | "-mx-8" | "-mx-px" | "-my-1" | "-my-10" | "-my-12" | "-my-16" | "-my-2" | "-my-20" | "-my-24" | "-my-3" | "-my-32" | "-my-4" | "-my-40" | "-my-48" | "-my-5" | "-my-56" | "-my-6" | "-my-64" | "-my-8" | "-my-px" | "absolute" | "align-baseline" | "align-bottom" | "align-middle" | "align-text-bottom" | "align-text-top" | "align-top" | "all-small-caps" | "antialiased" | "appearance-none" | "bg-auto" | "bg-black" | "bg-blue-100" | "bg-blue-200" | "bg-blue-300" | "bg-blue-400" | "bg-blue-500" | "bg-blue-600" | "bg-blue-700" | "bg-blue-800" | "bg-blue-900" | "bg-border" | "bg-bottom" | "bg-center" | "bg-contain" | "bg-cover" | "bg-fixed" | "bg-gray-100" | "bg-gray-200" | "bg-gray-300" | "bg-gray-400" | "bg-gray-500" | "bg-gray-600" | "bg-gray-700" | "bg-gray-800" | "bg-gray-900" | "bg-left" | "bg-left-bottom" | "bg-left-top" | "bg-local" | "bg-no-repeat" | "bg-primary" | "bg-repeat" | "bg-repeat-round" | "bg-repeat-space" | "bg-repeat-x" | "bg-repeat-y" | "bg-right" | "bg-right-bottom" | "bg-right-top" | "bg-scroll" | "bg-top" | "bg-white" | "block" | "border" | "border-0" | "border-2" | "border-4" | "border-8" | "border-b" | "border-b-0" | "border-b-2" | "border-b-4" | "border-b-8" | "border-black" | "border-blue-100" | "border-blue-200" | "border-blue-300" | "border-blue-400" | "border-blue-500" | "border-blue-600" | "border-blue-700" | "border-blue-800" | "border-blue-900" | "border-border" | "border-collapse" | "border-dashed" | "border-dotted" | "border-double" | "border-gray-100" | "border-gray-200" | "border-gray-300" | "border-gray-400" | "border-gray-500" | "border-gray-600" | "border-gray-700" | "border-gray-800" | "border-gray-900" | "border-l" | "border-l-0" | "border-l-2" | "border-l-4" | "border-l-8" | "border-none" | "border-primary" | "border-r" | "border-r-0" | "border-r-2" | "border-r-4" | "border-r-8" | "border-separate" | "border-solid" | "border-t" | "border-t-0" | "border-t-2" | "border-t-4" | "border-t-8" | "border-white" | "bottom-0" | "bottom-auto" | "break-all" | "break-normal" | "break-words" | "capitalize" | "clearfix" | "common-ligatures" | "container" | "content-around" | "content-between" | "content-center" | "content-end" | "content-start" | "contextual-ligatures" | "cursor-auto" | "cursor-default" | "cursor-move" | "cursor-not-allowed" | "cursor-pointer" | "cursor-text" | "cursor-wait" | "diagonal-fractions" | "discretionary-ligatures" | "ellipsis" | "fill-current" | "fixed" | "flex" | "flex-1" | "flex-auto" | "flex-col" | "flex-col-reverse" | "flex-grow" | "flex-grow-0" | "flex-initial" | "flex-no-wrap" | "flex-none" | "flex-row" | "flex-row-reverse" | "flex-shrink" | "flex-shrink-0" | "flex-wrap" | "flex-wrap-reverse" | "float-left" | "float-none" | "float-right" | "focus\:bg-black" | "focus\:bg-blue-100" | "focus\:bg-blue-200" | "focus\:bg-blue-300" | "focus\:bg-blue-400" | "focus\:bg-blue-500" | "focus\:bg-blue-600" | "focus\:bg-blue-700" | "focus\:bg-blue-800" | "focus\:bg-blue-900" | "focus\:bg-border" | "focus\:bg-gray-100" | "focus\:bg-gray-200" | "focus\:bg-gray-300" | "focus\:bg-gray-400" | "focus\:bg-gray-500" | "focus\:bg-gray-600" | "focus\:bg-gray-700" | "focus\:bg-gray-800" | "focus\:bg-gray-900" | "focus\:bg-primary" | "focus\:bg-white" | "focus\:border" | "focus\:border-0" | "focus\:border-2" | "focus\:border-4" | "focus\:border-8" | "focus\:border-b" | "focus\:border-b-0" | "focus\:border-b-2" | "focus\:border-b-4" | "focus\:border-b-8" | "focus\:border-black" | "focus\:border-blue-100" | "focus\:border-blue-200" | "focus\:border-blue-300" | "focus\:border-blue-400" | "focus\:border-blue-500" | "focus\:border-blue-600" | "focus\:border-blue-700" | "focus\:border-blue-800" | "focus\:border-blue-900" | "focus\:border-border" | "focus\:border-gray-100" | "focus\:border-gray-200" | "focus\:border-gray-300" | "focus\:border-gray-400" | "focus\:border-gray-500" | "focus\:border-gray-600" | "focus\:border-gray-700" | "focus\:border-gray-800" | "focus\:border-gray-900" | "focus\:border-l" | "focus\:border-l-0" | "focus\:border-l-2" | "focus\:border-l-4" | "focus\:border-l-8" | "focus\:border-primary" | "focus\:border-r" | "focus\:border-r-0" | "focus\:border-r-2" | "focus\:border-r-4" | "focus\:border-r-8" | "focus\:border-t" | "focus\:border-t-0" | "focus\:border-t-2" | "focus\:border-t-4" | "focus\:border-t-8" | "focus\:border-white" | "focus\:font-black" | "focus\:font-bold" | "focus\:font-extrabold" | "focus\:font-hairline" | "focus\:font-light" | "focus\:font-medium" | "focus\:font-normal" | "focus\:font-semibold" | "focus\:font-thin" | "focus\:line-through" | "focus\:no-underline" | "focus\:not-sr-only" | "focus\:opacity-0" | "focus\:opacity-100" | "focus\:opacity-25" | "focus\:opacity-50" | "focus\:opacity-75" | "focus\:outline-none" | "focus\:placeholder-black" | "focus\:placeholder-blue-100" | "focus\:placeholder-blue-200" | "focus\:placeholder-blue-300" | "focus\:placeholder-blue-400" | "focus\:placeholder-blue-500" | "focus\:placeholder-blue-600" | "focus\:placeholder-blue-700" | "focus\:placeholder-blue-800" | "focus\:placeholder-blue-900" | "focus\:placeholder-border" | "focus\:placeholder-gray-100" | "focus\:placeholder-gray-200" | "focus\:placeholder-gray-300" | "focus\:placeholder-gray-400" | "focus\:placeholder-gray-500" | "focus\:placeholder-gray-600" | "focus\:placeholder-gray-700" | "focus\:placeholder-gray-800" | "focus\:placeholder-gray-900" | "focus\:placeholder-primary" | "focus\:placeholder-white" | "focus\:shadow" | "focus\:shadow-2xl" | "focus\:shadow-inner" | "focus\:shadow-lg" | "focus\:shadow-md" | "focus\:shadow-none" | "focus\:shadow-outline" | "focus\:shadow-xl" | "focus\:sr-only" | "focus\:text-black" | "focus\:text-blue-100" | "focus\:text-blue-200" | "focus\:text-blue-300" | "focus\:text-blue-400" | "focus\:text-blue-500" | "focus\:text-blue-600" | "focus\:text-blue-700" | "focus\:text-blue-800" | "focus\:text-blue-900" | "focus\:text-border" | "focus\:text-gray-100" | "focus\:text-gray-200" | "focus\:text-gray-300" | "focus\:text-gray-400" | "focus\:text-gray-500" | "focus\:text-gray-600" | "focus\:text-gray-700" | "focus\:text-gray-800" | "focus\:text-gray-900" | "focus\:text-primary" | "focus\:text-white" | "focus\:underline" | "font-black" | "font-bold" | "font-extrabold" | "font-family-unset" | "font-hairline" | "font-light" | "font-medium" | "font-mono" | "font-normal" | "font-sans" | "font-semibold" | "font-serif" | "font-style-unset" | "font-thin" | "font-weight-unset" | "h-0" | "h-1" | "h-10" | "h-12" | "h-16" | "h-2" | "h-20" | "h-24" | "h-3" | "h-32" | "h-4" | "h-40" | "h-48" | "h-5" | "h-56" | "h-6" | "h-64" | "h-8" | "h-auto" | "h-full" | "h-px" | "h-screen" | "hidden" | "historical-ligatures" | "hover\:bg-black" | "hover\:bg-blue-100" | "hover\:bg-blue-200" | "hover\:bg-blue-300" | "hover\:bg-blue-400" | "hover\:bg-blue-500" | "hover\:bg-blue-600" | "hover\:bg-blue-700" | "hover\:bg-blue-800" | "hover\:bg-blue-900" | "hover\:bg-border" | "hover\:bg-gray-100" | "hover\:bg-gray-200" | "hover\:bg-gray-300" | "hover\:bg-gray-400" | "hover\:bg-gray-500" | "hover\:bg-gray-600" | "hover\:bg-gray-700" | "hover\:bg-gray-800" | "hover\:bg-gray-900" | "hover\:bg-primary" | "hover\:bg-white" | "hover\:border" | "hover\:border-0" | "hover\:border-2" | "hover\:border-4" | "hover\:border-8" | "hover\:border-b" | "hover\:border-b-0" | "hover\:border-b-2" | "hover\:border-b-4" | "hover\:border-b-8" | "hover\:border-black" | "hover\:border-blue-100" | "hover\:border-blue-200" | "hover\:border-blue-300" | "hover\:border-blue-400" | "hover\:border-blue-500" | "hover\:border-blue-600" | "hover\:border-blue-700" | "hover\:border-blue-800" | "hover\:border-blue-900" | "hover\:border-border" | "hover\:border-gray-100" | "hover\:border-gray-200" | "hover\:border-gray-300" | "hover\:border-gray-400" | "hover\:border-gray-500" | "hover\:border-gray-600" | "hover\:border-gray-700" | "hover\:border-gray-800" | "hover\:border-gray-900" | "hover\:border-l" | "hover\:border-l-0" | "hover\:border-l-2" | "hover\:border-l-4" | "hover\:border-l-8" | "hover\:border-primary" | "hover\:border-r" | "hover\:border-r-0" | "hover\:border-r-2" | "hover\:border-r-4" | "hover\:border-r-8" | "hover\:border-t" | "hover\:border-t-0" | "hover\:border-t-2" | "hover\:border-t-4" | "hover\:border-t-8" | "hover\:border-white" | "hover\:font-black" | "hover\:font-bold" | "hover\:font-extrabold" | "hover\:font-hairline" | "hover\:font-light" | "hover\:font-medium" | "hover\:font-normal" | "hover\:font-semibold" | "hover\:font-thin" | "hover\:line-through" | "hover\:no-underline" | "hover\:opacity-0" | "hover\:opacity-100" | "hover\:opacity-25" | "hover\:opacity-50" | "hover\:opacity-75" | "hover\:shadow" | "hover\:shadow-2xl" | "hover\:shadow-inner" | "hover\:shadow-lg" | "hover\:shadow-md" | "hover\:shadow-none" | "hover\:shadow-outline" | "hover\:shadow-xl" | "hover\:text-black" | "hover\:text-blue-100" | "hover\:text-blue-200" | "hover\:text-blue-300" | "hover\:text-blue-400" | "hover\:text-blue-500" | "hover\:text-blue-600" | "hover\:text-blue-700" | "hover\:text-blue-800" | "hover\:text-blue-900" | "hover\:text-border" | "hover\:text-gray-100" | "hover\:text-gray-200" | "hover\:text-gray-300" | "hover\:text-gray-400" | "hover\:text-gray-500" | "hover\:text-gray-600" | "hover\:text-gray-700" | "hover\:text-gray-800" | "hover\:text-gray-900" | "hover\:text-primary" | "hover\:text-white" | "hover\:underline" | "hyphens-auto" | "hyphens-manual" | "hyphens-none" | "inline" | "inline-block" | "inline-flex" | "inset-0" | "inset-auto" | "inset-x-0" | "inset-x-auto" | "inset-y-0" | "inset-y-auto" | "invisible" | "italic" | "items-baseline" | "items-center" | "items-end" | "items-start" | "items-stretch" | "justify-around" | "justify-between" | "justify-center" | "justify-end" | "justify-start" | "language" | "last\:border" | "last\:border-0" | "last\:border-2" | "last\:border-4" | "last\:border-8" | "last\:border-b" | "last\:border-b-0" | "last\:border-b-2" | "last\:border-b-4" | "last\:border-b-8" | "last\:border-l" | "last\:border-l-0" | "last\:border-l-2" | "last\:border-l-4" | "last\:border-l-8" | "last\:border-r" | "last\:border-r-0" | "last\:border-r-2" | "last\:border-r-4" | "last\:border-r-8" | "last\:border-t" | "last\:border-t-0" | "last\:border-t-2" | "last\:border-t-4" | "last\:border-t-8" | "leading-loose" | "leading-none" | "leading-normal" | "leading-relaxed" | "leading-snug" | "leading-tight" | "leading-unset" | "left-0" | "left-auto" | "lg\:-m-1" | "lg\:-m-10" | "lg\:-m-12" | "lg\:-m-16" | "lg\:-m-2" | "lg\:-m-20" | "lg\:-m-24" | "lg\:-m-3" | "lg\:-m-32" | "lg\:-m-4" | "lg\:-m-40" | "lg\:-m-48" | "lg\:-m-5" | "lg\:-m-56" | "lg\:-m-6" | "lg\:-m-64" | "lg\:-m-8" | "lg\:-m-px" | "lg\:-mb-1" | "lg\:-mb-10" | "lg\:-mb-12" | "lg\:-mb-16" | "lg\:-mb-2" | "lg\:-mb-20" | "lg\:-mb-24" | "lg\:-mb-3" | "lg\:-mb-32" | "lg\:-mb-4" | "lg\:-mb-40" | "lg\:-mb-48" | "lg\:-mb-5" | "lg\:-mb-56" | "lg\:-mb-6" | "lg\:-mb-64" | "lg\:-mb-8" | "lg\:-mb-px" | "lg\:-ml-1" | "lg\:-ml-10" | "lg\:-ml-12" | "lg\:-ml-16" | "lg\:-ml-2" | "lg\:-ml-20" | "lg\:-ml-24" | "lg\:-ml-3" | "lg\:-ml-32" | "lg\:-ml-4" | "lg\:-ml-40" | "lg\:-ml-48" | "lg\:-ml-5" | "lg\:-ml-56" | "lg\:-ml-6" | "lg\:-ml-64" | "lg\:-ml-8" | "lg\:-ml-px" | "lg\:-mr-1" | "lg\:-mr-10" | "lg\:-mr-12" | "lg\:-mr-16" | "lg\:-mr-2" | "lg\:-mr-20" | "lg\:-mr-24" | "lg\:-mr-3" | "lg\:-mr-32" | "lg\:-mr-4" | "lg\:-mr-40" | "lg\:-mr-48" | "lg\:-mr-5" | "lg\:-mr-56" | "lg\:-mr-6" | "lg\:-mr-64" | "lg\:-mr-8" | "lg\:-mr-px" | "lg\:-mt-1" | "lg\:-mt-10" | "lg\:-mt-12" | "lg\:-mt-16" | "lg\:-mt-2" | "lg\:-mt-20" | "lg\:-mt-24" | "lg\:-mt-3" | "lg\:-mt-32" | "lg\:-mt-4" | "lg\:-mt-40" | "lg\:-mt-48" | "lg\:-mt-5" | "lg\:-mt-56" | "lg\:-mt-6" | "lg\:-mt-64" | "lg\:-mt-8" | "lg\:-mt-px" | "lg\:-mx-1" | "lg\:-mx-10" | "lg\:-mx-12" | "lg\:-mx-16" | "lg\:-mx-2" | "lg\:-mx-20" | "lg\:-mx-24" | "lg\:-mx-3" | "lg\:-mx-32" | "lg\:-mx-4" | "lg\:-mx-40" | "lg\:-mx-48" | "lg\:-mx-5" | "lg\:-mx-56" | "lg\:-mx-6" | "lg\:-mx-64" | "lg\:-mx-8" | "lg\:-mx-px" | "lg\:-my-1" | "lg\:-my-10" | "lg\:-my-12" | "lg\:-my-16" | "lg\:-my-2" | "lg\:-my-20" | "lg\:-my-24" | "lg\:-my-3" | "lg\:-my-32" | "lg\:-my-4" | "lg\:-my-40" | "lg\:-my-48" | "lg\:-my-5" | "lg\:-my-56" | "lg\:-my-6" | "lg\:-my-64" | "lg\:-my-8" | "lg\:-my-px" | "lg\:absolute" | "lg\:align-baseline" | "lg\:align-bottom" | "lg\:align-middle" | "lg\:align-text-bottom" | "lg\:align-text-top" | "lg\:align-top" | "lg\:all-small-caps" | "lg\:antialiased" | "lg\:appearance-none" | "lg\:bg-auto" | "lg\:bg-black" | "lg\:bg-blue-100" | "lg\:bg-blue-200" | "lg\:bg-blue-300" | "lg\:bg-blue-400" | "lg\:bg-blue-500" | "lg\:bg-blue-600" | "lg\:bg-blue-700" | "lg\:bg-blue-800" | "lg\:bg-blue-900" | "lg\:bg-border" | "lg\:bg-bottom" | "lg\:bg-center" | "lg\:bg-contain" | "lg\:bg-cover" | "lg\:bg-fixed" | "lg\:bg-gray-100" | "lg\:bg-gray-200" | "lg\:bg-gray-300" | "lg\:bg-gray-400" | "lg\:bg-gray-500" | "lg\:bg-gray-600" | "lg\:bg-gray-700" | "lg\:bg-gray-800" | "lg\:bg-gray-900" | "lg\:bg-left" | "lg\:bg-left-bottom" | "lg\:bg-left-top" | "lg\:bg-local" | "lg\:bg-no-repeat" | "lg\:bg-primary" | "lg\:bg-repeat" | "lg\:bg-repeat-round" | "lg\:bg-repeat-space" | "lg\:bg-repeat-x" | "lg\:bg-repeat-y" | "lg\:bg-right" | "lg\:bg-right-bottom" | "lg\:bg-right-top" | "lg\:bg-scroll" | "lg\:bg-top" | "lg\:bg-white" | "lg\:block" | "lg\:border" | "lg\:border-0" | "lg\:border-2" | "lg\:border-4" | "lg\:border-8" | "lg\:border-b" | "lg\:border-b-0" | "lg\:border-b-2" | "lg\:border-b-4" | "lg\:border-b-8" | "lg\:border-black" | "lg\:border-blue-100" | "lg\:border-blue-200" | "lg\:border-blue-300" | "lg\:border-blue-400" | "lg\:border-blue-500" | "lg\:border-blue-600" | "lg\:border-blue-700" | "lg\:border-blue-800" | "lg\:border-blue-900" | "lg\:border-border" | "lg\:border-collapse" | "lg\:border-dashed" | "lg\:border-dotted" | "lg\:border-double" | "lg\:border-gray-100" | "lg\:border-gray-200" | "lg\:border-gray-300" | "lg\:border-gray-400" | "lg\:border-gray-500" | "lg\:border-gray-600" | "lg\:border-gray-700" | "lg\:border-gray-800" | "lg\:border-gray-900" | "lg\:border-l" | "lg\:border-l-0" | "lg\:border-l-2" | "lg\:border-l-4" | "lg\:border-l-8" | "lg\:border-none" | "lg\:border-primary" | "lg\:border-r" | "lg\:border-r-0" | "lg\:border-r-2" | "lg\:border-r-4" | "lg\:border-r-8" | "lg\:border-separate" | "lg\:border-solid" | "lg\:border-t" | "lg\:border-t-0" | "lg\:border-t-2" | "lg\:border-t-4" | "lg\:border-t-8" | "lg\:border-white" | "lg\:bottom-0" | "lg\:bottom-auto" | "lg\:break-all" | "lg\:break-normal" | "lg\:break-words" | "lg\:capitalize" | "lg\:clearfix" | "lg\:common-ligatures" | "lg\:content-around" | "lg\:content-between" | "lg\:content-center" | "lg\:content-end" | "lg\:content-start" | "lg\:contextual-ligatures" | "lg\:cursor-auto" | "lg\:cursor-default" | "lg\:cursor-move" | "lg\:cursor-not-allowed" | "lg\:cursor-pointer" | "lg\:cursor-text" | "lg\:cursor-wait" | "lg\:diagonal-fractions" | "lg\:discretionary-ligatures" | "lg\:ellipsis" | "lg\:fill-current" | "lg\:fixed" | "lg\:flex" | "lg\:flex-1" | "lg\:flex-auto" | "lg\:flex-col" | "lg\:flex-col-reverse" | "lg\:flex-grow" | "lg\:flex-grow-0" | "lg\:flex-initial" | "lg\:flex-no-wrap" | "lg\:flex-none" | "lg\:flex-row" | "lg\:flex-row-reverse" | "lg\:flex-shrink" | "lg\:flex-shrink-0" | "lg\:flex-wrap" | "lg\:flex-wrap-reverse" | "lg\:float-left" | "lg\:float-none" | "lg\:float-right" | "lg\:focus\:bg-black" | "lg\:focus\:bg-blue-100" | "lg\:focus\:bg-blue-200" | "lg\:focus\:bg-blue-300" | "lg\:focus\:bg-blue-400" | "lg\:focus\:bg-blue-500" | "lg\:focus\:bg-blue-600" | "lg\:focus\:bg-blue-700" | "lg\:focus\:bg-blue-800" | "lg\:focus\:bg-blue-900" | "lg\:focus\:bg-border" | "lg\:focus\:bg-gray-100" | "lg\:focus\:bg-gray-200" | "lg\:focus\:bg-gray-300" | "lg\:focus\:bg-gray-400" | "lg\:focus\:bg-gray-500" | "lg\:focus\:bg-gray-600" | "lg\:focus\:bg-gray-700" | "lg\:focus\:bg-gray-800" | "lg\:focus\:bg-gray-900" | "lg\:focus\:bg-primary" | "lg\:focus\:bg-white" | "lg\:focus\:border" | "lg\:focus\:border-0" | "lg\:focus\:border-2" | "lg\:focus\:border-4" | "lg\:focus\:border-8" | "lg\:focus\:border-b" | "lg\:focus\:border-b-0" | "lg\:focus\:border-b-2" | "lg\:focus\:border-b-4" | "lg\:focus\:border-b-8" | "lg\:focus\:border-black" | "lg\:focus\:border-blue-100" | "lg\:focus\:border-blue-200" | "lg\:focus\:border-blue-300" | "lg\:focus\:border-blue-400" | "lg\:focus\:border-blue-500" | "lg\:focus\:border-blue-600" | "lg\:focus\:border-blue-700" | "lg\:focus\:border-blue-800" | "lg\:focus\:border-blue-900" | "lg\:focus\:border-border" | "lg\:focus\:border-gray-100" | "lg\:focus\:border-gray-200" | "lg\:focus\:border-gray-300" | "lg\:focus\:border-gray-400" | "lg\:focus\:border-gray-500" | "lg\:focus\:border-gray-600" | "lg\:focus\:border-gray-700" | "lg\:focus\:border-gray-800" | "lg\:focus\:border-gray-900" | "lg\:focus\:border-l" | "lg\:focus\:border-l-0" | "lg\:focus\:border-l-2" | "lg\:focus\:border-l-4" | "lg\:focus\:border-l-8" | "lg\:focus\:border-primary" | "lg\:focus\:border-r" | "lg\:focus\:border-r-0" | "lg\:focus\:border-r-2" | "lg\:focus\:border-r-4" | "lg\:focus\:border-r-8" | "lg\:focus\:border-t" | "lg\:focus\:border-t-0" | "lg\:focus\:border-t-2" | "lg\:focus\:border-t-4" | "lg\:focus\:border-t-8" | "lg\:focus\:border-white" | "lg\:focus\:font-black" | "lg\:focus\:font-bold" | "lg\:focus\:font-extrabold" | "lg\:focus\:font-hairline" | "lg\:focus\:font-light" | "lg\:focus\:font-medium" | "lg\:focus\:font-normal" | "lg\:focus\:font-semibold" | "lg\:focus\:font-thin" | "lg\:focus\:line-through" | "lg\:focus\:no-underline" | "lg\:focus\:not-sr-only" | "lg\:focus\:opacity-0" | "lg\:focus\:opacity-100" | "lg\:focus\:opacity-25" | "lg\:focus\:opacity-50" | "lg\:focus\:opacity-75" | "lg\:focus\:outline-none" | "lg\:focus\:placeholder-black" | "lg\:focus\:placeholder-blue-100" | "lg\:focus\:placeholder-blue-200" | "lg\:focus\:placeholder-blue-300" | "lg\:focus\:placeholder-blue-400" | "lg\:focus\:placeholder-blue-500" | "lg\:focus\:placeholder-blue-600" | "lg\:focus\:placeholder-blue-700" | "lg\:focus\:placeholder-blue-800" | "lg\:focus\:placeholder-blue-900" | "lg\:focus\:placeholder-border" | "lg\:focus\:placeholder-gray-100" | "lg\:focus\:placeholder-gray-200" | "lg\:focus\:placeholder-gray-300" | "lg\:focus\:placeholder-gray-400" | "lg\:focus\:placeholder-gray-500" | "lg\:focus\:placeholder-gray-600" | "lg\:focus\:placeholder-gray-700" | "lg\:focus\:placeholder-gray-800" | "lg\:focus\:placeholder-gray-900" | "lg\:focus\:placeholder-primary" | "lg\:focus\:placeholder-white" | "lg\:focus\:shadow" | "lg\:focus\:shadow-2xl" | "lg\:focus\:shadow-inner" | "lg\:focus\:shadow-lg" | "lg\:focus\:shadow-md" | "lg\:focus\:shadow-none" | "lg\:focus\:shadow-outline" | "lg\:focus\:shadow-xl" | "lg\:focus\:sr-only" | "lg\:focus\:text-black" | "lg\:focus\:text-blue-100" | "lg\:focus\:text-blue-200" | "lg\:focus\:text-blue-300" | "lg\:focus\:text-blue-400" | "lg\:focus\:text-blue-500" | "lg\:focus\:text-blue-600" | "lg\:focus\:text-blue-700" | "lg\:focus\:text-blue-800" | "lg\:focus\:text-blue-900" | "lg\:focus\:text-border" | "lg\:focus\:text-gray-100" | "lg\:focus\:text-gray-200" | "lg\:focus\:text-gray-300" | "lg\:focus\:text-gray-400" | "lg\:focus\:text-gray-500" | "lg\:focus\:text-gray-600" | "lg\:focus\:text-gray-700" | "lg\:focus\:text-gray-800" | "lg\:focus\:text-gray-900" | "lg\:focus\:text-primary" | "lg\:focus\:text-white" | "lg\:focus\:underline" | "lg\:font-black" | "lg\:font-bold" | "lg\:font-extrabold" | "lg\:font-family-unset" | "lg\:font-hairline" | "lg\:font-light" | "lg\:font-medium" | "lg\:font-mono" | "lg\:font-normal" | "lg\:font-sans" | "lg\:font-semibold" | "lg\:font-serif" | "lg\:font-style-unset" | "lg\:font-thin" | "lg\:font-weight-unset" | "lg\:h-0" | "lg\:h-1" | "lg\:h-10" | "lg\:h-12" | "lg\:h-16" | "lg\:h-2" | "lg\:h-20" | "lg\:h-24" | "lg\:h-3" | "lg\:h-32" | "lg\:h-4" | "lg\:h-40" | "lg\:h-48" | "lg\:h-5" | "lg\:h-56" | "lg\:h-6" | "lg\:h-64" | "lg\:h-8" | "lg\:h-auto" | "lg\:h-full" | "lg\:h-px" | "lg\:h-screen" | "lg\:hidden" | "lg\:historical-ligatures" | "lg\:hover\:bg-black" | "lg\:hover\:bg-blue-100" | "lg\:hover\:bg-blue-200" | "lg\:hover\:bg-blue-300" | "lg\:hover\:bg-blue-400" | "lg\:hover\:bg-blue-500" | "lg\:hover\:bg-blue-600" | "lg\:hover\:bg-blue-700" | "lg\:hover\:bg-blue-800" | "lg\:hover\:bg-blue-900" | "lg\:hover\:bg-border" | "lg\:hover\:bg-gray-100" | "lg\:hover\:bg-gray-200" | "lg\:hover\:bg-gray-300" | "lg\:hover\:bg-gray-400" | "lg\:hover\:bg-gray-500" | "lg\:hover\:bg-gray-600" | "lg\:hover\:bg-gray-700" | "lg\:hover\:bg-gray-800" | "lg\:hover\:bg-gray-900" | "lg\:hover\:bg-primary" | "lg\:hover\:bg-white" | "lg\:hover\:border" | "lg\:hover\:border-0" | "lg\:hover\:border-2" | "lg\:hover\:border-4" | "lg\:hover\:border-8" | "lg\:hover\:border-b" | "lg\:hover\:border-b-0" | "lg\:hover\:border-b-2" | "lg\:hover\:border-b-4" | "lg\:hover\:border-b-8" | "lg\:hover\:border-black" | "lg\:hover\:border-blue-100" | "lg\:hover\:border-blue-200" | "lg\:hover\:border-blue-300" | "lg\:hover\:border-blue-400" | "lg\:hover\:border-blue-500" | "lg\:hover\:border-blue-600" | "lg\:hover\:border-blue-700" | "lg\:hover\:border-blue-800" | "lg\:hover\:border-blue-900" | "lg\:hover\:border-border" | "lg\:hover\:border-gray-100" | "lg\:hover\:border-gray-200" | "lg\:hover\:border-gray-300" | "lg\:hover\:border-gray-400" | "lg\:hover\:border-gray-500" | "lg\:hover\:border-gray-600" | "lg\:hover\:border-gray-700" | "lg\:hover\:border-gray-800" | "lg\:hover\:border-gray-900" | "lg\:hover\:border-l" | "lg\:hover\:border-l-0" | "lg\:hover\:border-l-2" | "lg\:hover\:border-l-4" | "lg\:hover\:border-l-8" | "lg\:hover\:border-primary" | "lg\:hover\:border-r" | "lg\:hover\:border-r-0" | "lg\:hover\:border-r-2" | "lg\:hover\:border-r-4" | "lg\:hover\:border-r-8" | "lg\:hover\:border-t" | "lg\:hover\:border-t-0" | "lg\:hover\:border-t-2" | "lg\:hover\:border-t-4" | "lg\:hover\:border-t-8" | "lg\:hover\:border-white" | "lg\:hover\:font-black" | "lg\:hover\:font-bold" | "lg\:hover\:font-extrabold" | "lg\:hover\:font-hairline" | "lg\:hover\:font-light" | "lg\:hover\:font-medium" | "lg\:hover\:font-normal" | "lg\:hover\:font-semibold" | "lg\:hover\:font-thin" | "lg\:hover\:line-through" | "lg\:hover\:no-underline" | "lg\:hover\:opacity-0" | "lg\:hover\:opacity-100" | "lg\:hover\:opacity-25" | "lg\:hover\:opacity-50" | "lg\:hover\:opacity-75" | "lg\:hover\:shadow" | "lg\:hover\:shadow-2xl" | "lg\:hover\:shadow-inner" | "lg\:hover\:shadow-lg" | "lg\:hover\:shadow-md" | "lg\:hover\:shadow-none" | "lg\:hover\:shadow-outline" | "lg\:hover\:shadow-xl" | "lg\:hover\:text-black" | "lg\:hover\:text-blue-100" | "lg\:hover\:text-blue-200" | "lg\:hover\:text-blue-300" | "lg\:hover\:text-blue-400" | "lg\:hover\:text-blue-500" | "lg\:hover\:text-blue-600" | "lg\:hover\:text-blue-700" | "lg\:hover\:text-blue-800" | "lg\:hover\:text-blue-900" | "lg\:hover\:text-border" | "lg\:hover\:text-gray-100" | "lg\:hover\:text-gray-200" | "lg\:hover\:text-gray-300" | "lg\:hover\:text-gray-400" | "lg\:hover\:text-gray-500" | "lg\:hover\:text-gray-600" | "lg\:hover\:text-gray-700" | "lg\:hover\:text-gray-800" | "lg\:hover\:text-gray-900" | "lg\:hover\:text-primary" | "lg\:hover\:text-white" | "lg\:hover\:underline" | "lg\:hyphens-auto" | "lg\:hyphens-manual" | "lg\:hyphens-none" | "lg\:inline" | "lg\:inline-block" | "lg\:inline-flex" | "lg\:inset-0" | "lg\:inset-auto" | "lg\:inset-x-0" | "lg\:inset-x-auto" | "lg\:inset-y-0" | "lg\:inset-y-auto" | "lg\:invisible" | "lg\:italic" | "lg\:items-baseline" | "lg\:items-center" | "lg\:items-end" | "lg\:items-start" | "lg\:items-stretch" | "lg\:justify-around" | "lg\:justify-between" | "lg\:justify-center" | "lg\:justify-end" | "lg\:justify-start" | "lg\:last\:border" | "lg\:last\:border-0" | "lg\:last\:border-2" | "lg\:last\:border-4" | "lg\:last\:border-8" | "lg\:last\:border-b" | "lg\:last\:border-b-0" | "lg\:last\:border-b-2" | "lg\:last\:border-b-4" | "lg\:last\:border-b-8" | "lg\:last\:border-l" | "lg\:last\:border-l-0" | "lg\:last\:border-l-2" | "lg\:last\:border-l-4" | "lg\:last\:border-l-8" | "lg\:last\:border-r" | "lg\:last\:border-r-0" | "lg\:last\:border-r-2" | "lg\:last\:border-r-4" | "lg\:last\:border-r-8" | "lg\:last\:border-t" | "lg\:last\:border-t-0" | "lg\:last\:border-t-2" | "lg\:last\:border-t-4" | "lg\:last\:border-t-8" | "lg\:leading-loose" | "lg\:leading-none" | "lg\:leading-normal" | "lg\:leading-relaxed" | "lg\:leading-snug" | "lg\:leading-tight" | "lg\:leading-unset" | "lg\:left-0" | "lg\:left-auto" | "lg\:line-through" | "lg\:lining-nums" | "lg\:list-decimal" | "lg\:list-disc" | "lg\:list-inside" | "lg\:list-none" | "lg\:list-outside" | "lg\:lowercase" | "lg\:m-0" | "lg\:m-1" | "lg\:m-10" | "lg\:m-12" | "lg\:m-16" | "lg\:m-2" | "lg\:m-20" | "lg\:m-24" | "lg\:m-3" | "lg\:m-32" | "lg\:m-4" | "lg\:m-40" | "lg\:m-48" | "lg\:m-5" | "lg\:m-56" | "lg\:m-6" | "lg\:m-64" | "lg\:m-8" | "lg\:m-auto" | "lg\:m-px" | "lg\:max-h-full" | "lg\:max-h-screen" | "lg\:max-w-2xl" | "lg\:max-w-3xl" | "lg\:max-w-4xl" | "lg\:max-w-5xl" | "lg\:max-w-6xl" | "lg\:max-w-full" | "lg\:max-w-lg" | "lg\:max-w-md" | "lg\:max-w-sm" | "lg\:max-w-xl" | "lg\:max-w-xs" | "lg\:mb-0" | "lg\:mb-1" | "lg\:mb-10" | "lg\:mb-12" | "lg\:mb-16" | "lg\:mb-2" | "lg\:mb-20" | "lg\:mb-24" | "lg\:mb-3" | "lg\:mb-32" | "lg\:mb-4" | "lg\:mb-40" | "lg\:mb-48" | "lg\:mb-5" | "lg\:mb-56" | "lg\:mb-6" | "lg\:mb-64" | "lg\:mb-8" | "lg\:mb-auto" | "lg\:mb-px" | "lg\:min-h-0" | "lg\:min-h-full" | "lg\:min-h-screen" | "lg\:min-w-0" | "lg\:min-w-full" | "lg\:ml-0" | "lg\:ml-1" | "lg\:ml-10" | "lg\:ml-12" | "lg\:ml-16" | "lg\:ml-2" | "lg\:ml-20" | "lg\:ml-24" | "lg\:ml-3" | "lg\:ml-32" | "lg\:ml-4" | "lg\:ml-40" | "lg\:ml-48" | "lg\:ml-5" | "lg\:ml-56" | "lg\:ml-6" | "lg\:ml-64" | "lg\:ml-8" | "lg\:ml-auto" | "lg\:ml-px" | "lg\:mr-0" | "lg\:mr-1" | "lg\:mr-10" | "lg\:mr-12" | "lg\:mr-16" | "lg\:mr-2" | "lg\:mr-20" | "lg\:mr-24" | "lg\:mr-3" | "lg\:mr-32" | "lg\:mr-4" | "lg\:mr-40" | "lg\:mr-48" | "lg\:mr-5" | "lg\:mr-56" | "lg\:mr-6" | "lg\:mr-64" | "lg\:mr-8" | "lg\:mr-auto" | "lg\:mr-px" | "lg\:mt-0" | "lg\:mt-1" | "lg\:mt-10" | "lg\:mt-12" | "lg\:mt-16" | "lg\:mt-2" | "lg\:mt-20" | "lg\:mt-24" | "lg\:mt-3" | "lg\:mt-32" | "lg\:mt-4" | "lg\:mt-40" | "lg\:mt-48" | "lg\:mt-5" | "lg\:mt-56" | "lg\:mt-6" | "lg\:mt-64" | "lg\:mt-8" | "lg\:mt-auto" | "lg\:mt-px" | "lg\:mx-0" | "lg\:mx-1" | "lg\:mx-10" | "lg\:mx-12" | "lg\:mx-16" | "lg\:mx-2" | "lg\:mx-20" | "lg\:mx-24" | "lg\:mx-3" | "lg\:mx-32" | "lg\:mx-4" | "lg\:mx-40" | "lg\:mx-48" | "lg\:mx-5" | "lg\:mx-56" | "lg\:mx-6" | "lg\:mx-64" | "lg\:mx-8" | "lg\:mx-auto" | "lg\:mx-px" | "lg\:my-0" | "lg\:my-1" | "lg\:my-10" | "lg\:my-12" | "lg\:my-16" | "lg\:my-2" | "lg\:my-20" | "lg\:my-24" | "lg\:my-3" | "lg\:my-32" | "lg\:my-4" | "lg\:my-40" | "lg\:my-48" | "lg\:my-5" | "lg\:my-56" | "lg\:my-6" | "lg\:my-64" | "lg\:my-8" | "lg\:my-auto" | "lg\:my-px" | "lg\:no-common-ligatures" | "lg\:no-contextual-ligatures" | "lg\:no-discretionary-ligatures" | "lg\:no-ellipsis" | "lg\:no-historical-ligatures" | "lg\:no-ligatures" | "lg\:no-underline" | "lg\:normal-caps" | "lg\:normal-case" | "lg\:normal-ligatures" | "lg\:normal-nums" | "lg\:not-italic" | "lg\:not-sr-only" | "lg\:object-bottom" | "lg\:object-center" | "lg\:object-contain" | "lg\:object-cover" | "lg\:object-fill" | "lg\:object-left" | "lg\:object-left-bottom" | "lg\:object-left-top" | "lg\:object-none" | "lg\:object-right" | "lg\:object-right-bottom" | "lg\:object-right-top" | "lg\:object-scale-down" | "lg\:object-top" | "lg\:oldstyle-nums" | "lg\:opacity-0" | "lg\:opacity-100" | "lg\:opacity-25" | "lg\:opacity-50" | "lg\:opacity-75" | "lg\:order-1" | "lg\:order-10" | "lg\:order-11" | "lg\:order-12" | "lg\:order-2" | "lg\:order-3" | "lg\:order-4" | "lg\:order-5" | "lg\:order-6" | "lg\:order-7" | "lg\:order-8" | "lg\:order-9" | "lg\:order-first" | "lg\:order-last" | "lg\:order-none" | "lg\:ordinal-nums" | "lg\:outline-none" | "lg\:overflow-auto" | "lg\:overflow-hidden" | "lg\:overflow-scroll" | "lg\:overflow-visible" | "lg\:overflow-x-auto" | "lg\:overflow-x-hidden" | "lg\:overflow-x-scroll" | "lg\:overflow-x-visible" | "lg\:overflow-y-auto" | "lg\:overflow-y-hidden" | "lg\:overflow-y-scroll" | "lg\:overflow-y-visible" | "lg\:p-0" | "lg\:p-1" | "lg\:p-10" | "lg\:p-12" | "lg\:p-16" | "lg\:p-2" | "lg\:p-20" | "lg\:p-24" | "lg\:p-3" | "lg\:p-32" | "lg\:p-4" | "lg\:p-40" | "lg\:p-48" | "lg\:p-5" | "lg\:p-56" | "lg\:p-6" | "lg\:p-64" | "lg\:p-8" | "lg\:p-px" | "lg\:pb-0" | "lg\:pb-1" | "lg\:pb-10" | "lg\:pb-12" | "lg\:pb-16" | "lg\:pb-2" | "lg\:pb-20" | "lg\:pb-24" | "lg\:pb-3" | "lg\:pb-32" | "lg\:pb-4" | "lg\:pb-40" | "lg\:pb-48" | "lg\:pb-5" | "lg\:pb-56" | "lg\:pb-6" | "lg\:pb-64" | "lg\:pb-8" | "lg\:pb-px" | "lg\:petite-caps" | "lg\:pl-0" | "lg\:pl-1" | "lg\:pl-10" | "lg\:pl-12" | "lg\:pl-16" | "lg\:pl-2" | "lg\:pl-20" | "lg\:pl-24" | "lg\:pl-3" | "lg\:pl-32" | "lg\:pl-4" | "lg\:pl-40" | "lg\:pl-48" | "lg\:pl-5" | "lg\:pl-56" | "lg\:pl-6" | "lg\:pl-64" | "lg\:pl-8" | "lg\:pl-px" | "lg\:placeholder-black" | "lg\:placeholder-blue-100" | "lg\:placeholder-blue-200" | "lg\:placeholder-blue-300" | "lg\:placeholder-blue-400" | "lg\:placeholder-blue-500" | "lg\:placeholder-blue-600" | "lg\:placeholder-blue-700" | "lg\:placeholder-blue-800" | "lg\:placeholder-blue-900" | "lg\:placeholder-border" | "lg\:placeholder-gray-100" | "lg\:placeholder-gray-200" | "lg\:placeholder-gray-300" | "lg\:placeholder-gray-400" | "lg\:placeholder-gray-500" | "lg\:placeholder-gray-600" | "lg\:placeholder-gray-700" | "lg\:placeholder-gray-800" | "lg\:placeholder-gray-900" | "lg\:placeholder-primary" | "lg\:placeholder-white" | "lg\:pointer-events-auto" | "lg\:pointer-events-none" | "lg\:pr-0" | "lg\:pr-1" | "lg\:pr-10" | "lg\:pr-12" | "lg\:pr-16" | "lg\:pr-2" | "lg\:pr-20" | "lg\:pr-24" | "lg\:pr-3" | "lg\:pr-32" | "lg\:pr-4" | "lg\:pr-40" | "lg\:pr-48" | "lg\:pr-5" | "lg\:pr-56" | "lg\:pr-6" | "lg\:pr-64" | "lg\:pr-8" | "lg\:pr-px" | "lg\:proportional-nums" | "lg\:pt-0" | "lg\:pt-1" | "lg\:pt-10" | "lg\:pt-12" | "lg\:pt-16" | "lg\:pt-2" | "lg\:pt-20" | "lg\:pt-24" | "lg\:pt-3" | "lg\:pt-32" | "lg\:pt-4" | "lg\:pt-40" | "lg\:pt-48" | "lg\:pt-5" | "lg\:pt-56" | "lg\:pt-6" | "lg\:pt-64" | "lg\:pt-8" | "lg\:pt-px" | "lg\:px-0" | "lg\:px-1" | "lg\:px-10" | "lg\:px-12" | "lg\:px-16" | "lg\:px-2" | "lg\:px-20" | "lg\:px-24" | "lg\:px-3" | "lg\:px-32" | "lg\:px-4" | "lg\:px-40" | "lg\:px-48" | "lg\:px-5" | "lg\:px-56" | "lg\:px-6" | "lg\:px-64" | "lg\:px-8" | "lg\:px-px" | "lg\:py-0" | "lg\:py-1" | "lg\:py-10" | "lg\:py-12" | "lg\:py-16" | "lg\:py-2" | "lg\:py-20" | "lg\:py-24" | "lg\:py-3" | "lg\:py-32" | "lg\:py-4" | "lg\:py-40" | "lg\:py-48" | "lg\:py-5" | "lg\:py-56" | "lg\:py-6" | "lg\:py-64" | "lg\:py-8" | "lg\:py-px" | "lg\:relative" | "lg\:resize" | "lg\:resize-none" | "lg\:resize-x" | "lg\:resize-y" | "lg\:right-0" | "lg\:right-auto" | "lg\:rotate-180" | "lg\:rotate-270" | "lg\:rotate-90" | "lg\:rounded" | "lg\:rounded-b" | "lg\:rounded-b-full" | "lg\:rounded-b-lg" | "lg\:rounded-b-none" | "lg\:rounded-b-sm" | "lg\:rounded-bl" | "lg\:rounded-bl-full" | "lg\:rounded-bl-lg" | "lg\:rounded-bl-none" | "lg\:rounded-bl-sm" | "lg\:rounded-br" | "lg\:rounded-br-full" | "lg\:rounded-br-lg" | "lg\:rounded-br-none" | "lg\:rounded-br-sm" | "lg\:rounded-full" | "lg\:rounded-l" | "lg\:rounded-l-full" | "lg\:rounded-l-lg" | "lg\:rounded-l-none" | "lg\:rounded-l-sm" | "lg\:rounded-lg" | "lg\:rounded-none" | "lg\:rounded-r" | "lg\:rounded-r-full" | "lg\:rounded-r-lg" | "lg\:rounded-r-none" | "lg\:rounded-r-sm" | "lg\:rounded-sm" | "lg\:rounded-t" | "lg\:rounded-t-full" | "lg\:rounded-t-lg" | "lg\:rounded-t-none" | "lg\:rounded-t-sm" | "lg\:rounded-tl" | "lg\:rounded-tl-full" | "lg\:rounded-tl-lg" | "lg\:rounded-tl-none" | "lg\:rounded-tl-sm" | "lg\:rounded-tr" | "lg\:rounded-tr-full" | "lg\:rounded-tr-lg" | "lg\:rounded-tr-none" | "lg\:rounded-tr-sm" | "lg\:scrolling-auto" | "lg\:scrolling-touch" | "lg\:select-all" | "lg\:select-auto" | "lg\:select-none" | "lg\:select-text" | "lg\:self-auto" | "lg\:self-center" | "lg\:self-end" | "lg\:self-start" | "lg\:self-stretch" | "lg\:shadow" | "lg\:shadow-2xl" | "lg\:shadow-inner" | "lg\:shadow-lg" | "lg\:shadow-md" | "lg\:shadow-none" | "lg\:shadow-outline" | "lg\:shadow-xl" | "lg\:slashed-zeros" | "lg\:small-caps" | "lg\:sr-only" | "lg\:stacked-fractions" | "lg\:static" | "lg\:sticky" | "lg\:stroke-current" | "lg\:subpixel-antialiased" | "lg\:table" | "lg\:table-auto" | "lg\:table-cell" | "lg\:table-fixed" | "lg\:table-row" | "lg\:tabular-nums" | "lg\:text-2xl" | "lg\:text-3xl" | "lg\:text-4xl" | "lg\:text-5xl" | "lg\:text-6xl" | "lg\:text-align-unset" | "lg\:text-base" | "lg\:text-black" | "lg\:text-blue-100" | "lg\:text-blue-200" | "lg\:text-blue-300" | "lg\:text-blue-400" | "lg\:text-blue-500" | "lg\:text-blue-600" | "lg\:text-blue-700" | "lg\:text-blue-800" | "lg\:text-blue-900" | "lg\:text-border" | "lg\:text-center" | "lg\:text-color-unset" | "lg\:text-gray-100" | "lg\:text-gray-200" | "lg\:text-gray-300" | "lg\:text-gray-400" | "lg\:text-gray-500" | "lg\:text-gray-600" | "lg\:text-gray-700" | "lg\:text-gray-800" | "lg\:text-gray-900" | "lg\:text-justify" | "lg\:text-left" | "lg\:text-lg" | "lg\:text-primary" | "lg\:text-right" | "lg\:text-size-unset" | "lg\:text-sm" | "lg\:text-transform-unset" | "lg\:text-white" | "lg\:text-xl" | "lg\:text-xs" | "lg\:titling-caps" | "lg\:top-0" | "lg\:top-auto" | "lg\:tracking-normal" | "lg\:tracking-tight" | "lg\:tracking-tighter" | "lg\:tracking-unset" | "lg\:tracking-wide" | "lg\:tracking-wider" | "lg\:tracking-widest" | "lg\:transform-b" | "lg\:transform-bl" | "lg\:transform-border" | "lg\:transform-br" | "lg\:transform-fill" | "lg\:transform-l" | "lg\:transform-none" | "lg\:transform-r" | "lg\:transform-t" | "lg\:transform-tl" | "lg\:transform-tr" | "lg\:transform-view" | "lg\:transition-0" | "lg\:transition-100" | "lg\:transition-1000" | "lg\:transition-250" | "lg\:transition-500" | "lg\:transition-750" | "lg\:transition-all" | "lg\:transition-bg" | "lg\:transition-border" | "lg\:transition-color" | "lg\:transition-colors" | "lg\:transition-delay-0" | "lg\:transition-delay-100" | "lg\:transition-delay-1000" | "lg\:transition-delay-250" | "lg\:transition-delay-500" | "lg\:transition-delay-750" | "lg\:transition-ease" | "lg\:transition-ease-in" | "lg\:transition-ease-in-out" | "lg\:transition-ease-out" | "lg\:transition-linear" | "lg\:transition-none" | "lg\:transition-opacity" | "lg\:transition-transform" | "lg\:truncate" | "lg\:underline" | "lg\:unicase" | "lg\:uppercase" | "lg\:visible" | "lg\:w-0" | "lg\:w-1" | "lg\:w-10" | "lg\:w-10\/12" | "lg\:w-11\/12" | "lg\:w-12" | "lg\:w-16" | "lg\:w-1\/12" | "lg\:w-1\/2" | "lg\:w-1\/3" | "lg\:w-1\/4" | "lg\:w-1\/5" | "lg\:w-1\/6" | "lg\:w-2" | "lg\:w-20" | "lg\:w-24" | "lg\:w-2\/12" | "lg\:w-2\/3" | "lg\:w-2\/4" | "lg\:w-2\/5" | "lg\:w-2\/6" | "lg\:w-3" | "lg\:w-32" | "lg\:w-3\/12" | "lg\:w-3\/4" | "lg\:w-3\/5" | "lg\:w-3\/6" | "lg\:w-4" | "lg\:w-40" | "lg\:w-48" | "lg\:w-4\/12" | "lg\:w-4\/5" | "lg\:w-4\/6" | "lg\:w-5" | "lg\:w-56" | "lg\:w-5\/12" | "lg\:w-5\/6" | "lg\:w-6" | "lg\:w-64" | "lg\:w-6\/12" | "lg\:w-7\/12" | "lg\:w-8" | "lg\:w-8\/12" | "lg\:w-9\/12" | "lg\:w-auto" | "lg\:w-full" | "lg\:w-px" | "lg\:w-screen" | "lg\:whitespace-no-wrap" | "lg\:whitespace-normal" | "lg\:whitespace-pre" | "lg\:whitespace-pre-line" | "lg\:whitespace-pre-wrap" | "lg\:will-change-auto" | "lg\:will-change-contents" | "lg\:will-change-opacity" | "lg\:will-change-scroll" | "lg\:will-change-transform" | "lg\:z-0" | "lg\:z-10" | "lg\:z-20" | "lg\:z-30" | "lg\:z-40" | "lg\:z-50" | "lg\:z-auto" | "line-through" | "lining-nums" | "list-decimal" | "list-disc" | "list-inside" | "list-none" | "list-outside" | "lowercase" | "m-0" | "m-1" | "m-10" | "m-12" | "m-16" | "m-2" | "m-20" | "m-24" | "m-3" | "m-32" | "m-4" | "m-40" | "m-48" | "m-5" | "m-56" | "m-6" | "m-64" | "m-8" | "m-auto" | "m-px" | "max-h-full" | "max-h-screen" | "max-w-2xl" | "max-w-3xl" | "max-w-4xl" | "max-w-5xl" | "max-w-6xl" | "max-w-full" | "max-w-lg" | "max-w-md" | "max-w-sm" | "max-w-xl" | "max-w-xs" | "mb-0" | "mb-1" | "mb-10" | "mb-12" | "mb-16" | "mb-2" | "mb-20" | "mb-24" | "mb-3" | "mb-32" | "mb-4" | "mb-40" | "mb-48" | "mb-5" | "mb-56" | "mb-6" | "mb-64" | "mb-8" | "mb-auto" | "mb-px" | "md\:-m-1" | "md\:-m-10" | "md\:-m-12" | "md\:-m-16" | "md\:-m-2" | "md\:-m-20" | "md\:-m-24" | "md\:-m-3" | "md\:-m-32" | "md\:-m-4" | "md\:-m-40" | "md\:-m-48" | "md\:-m-5" | "md\:-m-56" | "md\:-m-6" | "md\:-m-64" | "md\:-m-8" | "md\:-m-px" | "md\:-mb-1" | "md\:-mb-10" | "md\:-mb-12" | "md\:-mb-16" | "md\:-mb-2" | "md\:-mb-20" | "md\:-mb-24" | "md\:-mb-3" | "md\:-mb-32" | "md\:-mb-4" | "md\:-mb-40" | "md\:-mb-48" | "md\:-mb-5" | "md\:-mb-56" | "md\:-mb-6" | "md\:-mb-64" | "md\:-mb-8" | "md\:-mb-px" | "md\:-ml-1" | "md\:-ml-10" | "md\:-ml-12" | "md\:-ml-16" | "md\:-ml-2" | "md\:-ml-20" | "md\:-ml-24" | "md\:-ml-3" | "md\:-ml-32" | "md\:-ml-4" | "md\:-ml-40" | "md\:-ml-48" | "md\:-ml-5" | "md\:-ml-56" | "md\:-ml-6" | "md\:-ml-64" | "md\:-ml-8" | "md\:-ml-px" | "md\:-mr-1" | "md\:-mr-10" | "md\:-mr-12" | "md\:-mr-16" | "md\:-mr-2" | "md\:-mr-20" | "md\:-mr-24" | "md\:-mr-3" | "md\:-mr-32" | "md\:-mr-4" | "md\:-mr-40" | "md\:-mr-48" | "md\:-mr-5" | "md\:-mr-56" | "md\:-mr-6" | "md\:-mr-64" | "md\:-mr-8" | "md\:-mr-px" | "md\:-mt-1" | "md\:-mt-10" | "md\:-mt-12" | "md\:-mt-16" | "md\:-mt-2" | "md\:-mt-20" | "md\:-mt-24" | "md\:-mt-3" | "md\:-mt-32" | "md\:-mt-4" | "md\:-mt-40" | "md\:-mt-48" | "md\:-mt-5" | "md\:-mt-56" | "md\:-mt-6" | "md\:-mt-64" | "md\:-mt-8" | "md\:-mt-px" | "md\:-mx-1" | "md\:-mx-10" | "md\:-mx-12" | "md\:-mx-16" | "md\:-mx-2" | "md\:-mx-20" | "md\:-mx-24" | "md\:-mx-3" | "md\:-mx-32" | "md\:-mx-4" | "md\:-mx-40" | "md\:-mx-48" | "md\:-mx-5" | "md\:-mx-56" | "md\:-mx-6" | "md\:-mx-64" | "md\:-mx-8" | "md\:-mx-px" | "md\:-my-1" | "md\:-my-10" | "md\:-my-12" | "md\:-my-16" | "md\:-my-2" | "md\:-my-20" | "md\:-my-24" | "md\:-my-3" | "md\:-my-32" | "md\:-my-4" | "md\:-my-40" | "md\:-my-48" | "md\:-my-5" | "md\:-my-56" | "md\:-my-6" | "md\:-my-64" | "md\:-my-8" | "md\:-my-px" | "md\:absolute" | "md\:align-baseline" | "md\:align-bottom" | "md\:align-middle" | "md\:align-text-bottom" | "md\:align-text-top" | "md\:align-top" | "md\:all-small-caps" | "md\:antialiased" | "md\:appearance-none" | "md\:bg-auto" | "md\:bg-black" | "md\:bg-blue-100" | "md\:bg-blue-200" | "md\:bg-blue-300" | "md\:bg-blue-400" | "md\:bg-blue-500" | "md\:bg-blue-600" | "md\:bg-blue-700" | "md\:bg-blue-800" | "md\:bg-blue-900" | "md\:bg-border" | "md\:bg-bottom" | "md\:bg-center" | "md\:bg-contain" | "md\:bg-cover" | "md\:bg-fixed" | "md\:bg-gray-100" | "md\:bg-gray-200" | "md\:bg-gray-300" | "md\:bg-gray-400" | "md\:bg-gray-500" | "md\:bg-gray-600" | "md\:bg-gray-700" | "md\:bg-gray-800" | "md\:bg-gray-900" | "md\:bg-left" | "md\:bg-left-bottom" | "md\:bg-left-top" | "md\:bg-local" | "md\:bg-no-repeat" | "md\:bg-primary" | "md\:bg-repeat" | "md\:bg-repeat-round" | "md\:bg-repeat-space" | "md\:bg-repeat-x" | "md\:bg-repeat-y" | "md\:bg-right" | "md\:bg-right-bottom" | "md\:bg-right-top" | "md\:bg-scroll" | "md\:bg-top" | "md\:bg-white" | "md\:block" | "md\:border" | "md\:border-0" | "md\:border-2" | "md\:border-4" | "md\:border-8" | "md\:border-b" | "md\:border-b-0" | "md\:border-b-2" | "md\:border-b-4" | "md\:border-b-8" | "md\:border-black" | "md\:border-blue-100" | "md\:border-blue-200" | "md\:border-blue-300" | "md\:border-blue-400" | "md\:border-blue-500" | "md\:border-blue-600" | "md\:border-blue-700" | "md\:border-blue-800" | "md\:border-blue-900" | "md\:border-border" | "md\:border-collapse" | "md\:border-dashed" | "md\:border-dotted" | "md\:border-double" | "md\:border-gray-100" | "md\:border-gray-200" | "md\:border-gray-300" | "md\:border-gray-400" | "md\:border-gray-500" | "md\:border-gray-600" | "md\:border-gray-700" | "md\:border-gray-800" | "md\:border-gray-900" | "md\:border-l" | "md\:border-l-0" | "md\:border-l-2" | "md\:border-l-4" | "md\:border-l-8" | "md\:border-none" | "md\:border-primary" | "md\:border-r" | "md\:border-r-0" | "md\:border-r-2" | "md\:border-r-4" | "md\:border-r-8" | "md\:border-separate" | "md\:border-solid" | "md\:border-t" | "md\:border-t-0" | "md\:border-t-2" | "md\:border-t-4" | "md\:border-t-8" | "md\:border-white" | "md\:bottom-0" | "md\:bottom-auto" | "md\:break-all" | "md\:break-normal" | "md\:break-words" | "md\:capitalize" | "md\:clearfix" | "md\:common-ligatures" | "md\:content-around" | "md\:content-between" | "md\:content-center" | "md\:content-end" | "md\:content-start" | "md\:contextual-ligatures" | "md\:cursor-auto" | "md\:cursor-default" | "md\:cursor-move" | "md\:cursor-not-allowed" | "md\:cursor-pointer" | "md\:cursor-text" | "md\:cursor-wait" | "md\:diagonal-fractions" | "md\:discretionary-ligatures" | "md\:ellipsis" | "md\:fill-current" | "md\:fixed" | "md\:flex" | "md\:flex-1" | "md\:flex-auto" | "md\:flex-col" | "md\:flex-col-reverse" | "md\:flex-grow" | "md\:flex-grow-0" | "md\:flex-initial" | "md\:flex-no-wrap" | "md\:flex-none" | "md\:flex-row" | "md\:flex-row-reverse" | "md\:flex-shrink" | "md\:flex-shrink-0" | "md\:flex-wrap" | "md\:flex-wrap-reverse" | "md\:float-left" | "md\:float-none" | "md\:float-right" | "md\:focus\:bg-black" | "md\:focus\:bg-blue-100" | "md\:focus\:bg-blue-200" | "md\:focus\:bg-blue-300" | "md\:focus\:bg-blue-400" | "md\:focus\:bg-blue-500" | "md\:focus\:bg-blue-600" | "md\:focus\:bg-blue-700" | "md\:focus\:bg-blue-800" | "md\:focus\:bg-blue-900" | "md\:focus\:bg-border" | "md\:focus\:bg-gray-100" | "md\:focus\:bg-gray-200" | "md\:focus\:bg-gray-300" | "md\:focus\:bg-gray-400" | "md\:focus\:bg-gray-500" | "md\:focus\:bg-gray-600" | "md\:focus\:bg-gray-700" | "md\:focus\:bg-gray-800" | "md\:focus\:bg-gray-900" | "md\:focus\:bg-primary" | "md\:focus\:bg-white" | "md\:focus\:border" | "md\:focus\:border-0" | "md\:focus\:border-2" | "md\:focus\:border-4" | "md\:focus\:border-8" | "md\:focus\:border-b" | "md\:focus\:border-b-0" | "md\:focus\:border-b-2" | "md\:focus\:border-b-4" | "md\:focus\:border-b-8" | "md\:focus\:border-black" | "md\:focus\:border-blue-100" | "md\:focus\:border-blue-200" | "md\:focus\:border-blue-300" | "md\:focus\:border-blue-400" | "md\:focus\:border-blue-500" | "md\:focus\:border-blue-600" | "md\:focus\:border-blue-700" | "md\:focus\:border-blue-800" | "md\:focus\:border-blue-900" | "md\:focus\:border-border" | "md\:focus\:border-gray-100" | "md\:focus\:border-gray-200" | "md\:focus\:border-gray-300" | "md\:focus\:border-gray-400" | "md\:focus\:border-gray-500" | "md\:focus\:border-gray-600" | "md\:focus\:border-gray-700" | "md\:focus\:border-gray-800" | "md\:focus\:border-gray-900" | "md\:focus\:border-l" | "md\:focus\:border-l-0" | "md\:focus\:border-l-2" | "md\:focus\:border-l-4" | "md\:focus\:border-l-8" | "md\:focus\:border-primary" | "md\:focus\:border-r" | "md\:focus\:border-r-0" | "md\:focus\:border-r-2" | "md\:focus\:border-r-4" | "md\:focus\:border-r-8" | "md\:focus\:border-t" | "md\:focus\:border-t-0" | "md\:focus\:border-t-2" | "md\:focus\:border-t-4" | "md\:focus\:border-t-8" | "md\:focus\:border-white" | "md\:focus\:font-black" | "md\:focus\:font-bold" | "md\:focus\:font-extrabold" | "md\:focus\:font-hairline" | "md\:focus\:font-light" | "md\:focus\:font-medium" | "md\:focus\:font-normal" | "md\:focus\:font-semibold" | "md\:focus\:font-thin" | "md\:focus\:line-through" | "md\:focus\:no-underline" | "md\:focus\:not-sr-only" | "md\:focus\:opacity-0" | "md\:focus\:opacity-100" | "md\:focus\:opacity-25" | "md\:focus\:opacity-50" | "md\:focus\:opacity-75" | "md\:focus\:outline-none" | "md\:focus\:placeholder-black" | "md\:focus\:placeholder-blue-100" | "md\:focus\:placeholder-blue-200" | "md\:focus\:placeholder-blue-300" | "md\:focus\:placeholder-blue-400" | "md\:focus\:placeholder-blue-500" | "md\:focus\:placeholder-blue-600" | "md\:focus\:placeholder-blue-700" | "md\:focus\:placeholder-blue-800" | "md\:focus\:placeholder-blue-900" | "md\:focus\:placeholder-border" | "md\:focus\:placeholder-gray-100" | "md\:focus\:placeholder-gray-200" | "md\:focus\:placeholder-gray-300" | "md\:focus\:placeholder-gray-400" | "md\:focus\:placeholder-gray-500" | "md\:focus\:placeholder-gray-600" | "md\:focus\:placeholder-gray-700" | "md\:focus\:placeholder-gray-800" | "md\:focus\:placeholder-gray-900" | "md\:focus\:placeholder-primary" | "md\:focus\:placeholder-white" | "md\:focus\:shadow" | "md\:focus\:shadow-2xl" | "md\:focus\:shadow-inner" | "md\:focus\:shadow-lg" | "md\:focus\:shadow-md" | "md\:focus\:shadow-none" | "md\:focus\:shadow-outline" | "md\:focus\:shadow-xl" | "md\:focus\:sr-only" | "md\:focus\:text-black" | "md\:focus\:text-blue-100" | "md\:focus\:text-blue-200" | "md\:focus\:text-blue-300" | "md\:focus\:text-blue-400" | "md\:focus\:text-blue-500" | "md\:focus\:text-blue-600" | "md\:focus\:text-blue-700" | "md\:focus\:text-blue-800" | "md\:focus\:text-blue-900" | "md\:focus\:text-border" | "md\:focus\:text-gray-100" | "md\:focus\:text-gray-200" | "md\:focus\:text-gray-300" | "md\:focus\:text-gray-400" | "md\:focus\:text-gray-500" | "md\:focus\:text-gray-600" | "md\:focus\:text-gray-700" | "md\:focus\:text-gray-800" | "md\:focus\:text-gray-900" | "md\:focus\:text-primary" | "md\:focus\:text-white" | "md\:focus\:underline" | "md\:font-black" | "md\:font-bold" | "md\:font-extrabold" | "md\:font-family-unset" | "md\:font-hairline" | "md\:font-light" | "md\:font-medium" | "md\:font-mono" | "md\:font-normal" | "md\:font-sans" | "md\:font-semibold" | "md\:font-serif" | "md\:font-style-unset" | "md\:font-thin" | "md\:font-weight-unset" | "md\:h-0" | "md\:h-1" | "md\:h-10" | "md\:h-12" | "md\:h-16" | "md\:h-2" | "md\:h-20" | "md\:h-24" | "md\:h-3" | "md\:h-32" | "md\:h-4" | "md\:h-40" | "md\:h-48" | "md\:h-5" | "md\:h-56" | "md\:h-6" | "md\:h-64" | "md\:h-8" | "md\:h-auto" | "md\:h-full" | "md\:h-px" | "md\:h-screen" | "md\:hidden" | "md\:historical-ligatures" | "md\:hover\:bg-black" | "md\:hover\:bg-blue-100" | "md\:hover\:bg-blue-200" | "md\:hover\:bg-blue-300" | "md\:hover\:bg-blue-400" | "md\:hover\:bg-blue-500" | "md\:hover\:bg-blue-600" | "md\:hover\:bg-blue-700" | "md\:hover\:bg-blue-800" | "md\:hover\:bg-blue-900" | "md\:hover\:bg-border" | "md\:hover\:bg-gray-100" | "md\:hover\:bg-gray-200" | "md\:hover\:bg-gray-300" | "md\:hover\:bg-gray-400" | "md\:hover\:bg-gray-500" | "md\:hover\:bg-gray-600" | "md\:hover\:bg-gray-700" | "md\:hover\:bg-gray-800" | "md\:hover\:bg-gray-900" | "md\:hover\:bg-primary" | "md\:hover\:bg-white" | "md\:hover\:border" | "md\:hover\:border-0" | "md\:hover\:border-2" | "md\:hover\:border-4" | "md\:hover\:border-8" | "md\:hover\:border-b" | "md\:hover\:border-b-0" | "md\:hover\:border-b-2" | "md\:hover\:border-b-4" | "md\:hover\:border-b-8" | "md\:hover\:border-black" | "md\:hover\:border-blue-100" | "md\:hover\:border-blue-200" | "md\:hover\:border-blue-300" | "md\:hover\:border-blue-400" | "md\:hover\:border-blue-500" | "md\:hover\:border-blue-600" | "md\:hover\:border-blue-700" | "md\:hover\:border-blue-800" | "md\:hover\:border-blue-900" | "md\:hover\:border-border" | "md\:hover\:border-gray-100" | "md\:hover\:border-gray-200" | "md\:hover\:border-gray-300" | "md\:hover\:border-gray-400" | "md\:hover\:border-gray-500" | "md\:hover\:border-gray-600" | "md\:hover\:border-gray-700" | "md\:hover\:border-gray-800" | "md\:hover\:border-gray-900" | "md\:hover\:border-l" | "md\:hover\:border-l-0" | "md\:hover\:border-l-2" | "md\:hover\:border-l-4" | "md\:hover\:border-l-8" | "md\:hover\:border-primary" | "md\:hover\:border-r" | "md\:hover\:border-r-0" | "md\:hover\:border-r-2" | "md\:hover\:border-r-4" | "md\:hover\:border-r-8" | "md\:hover\:border-t" | "md\:hover\:border-t-0" | "md\:hover\:border-t-2" | "md\:hover\:border-t-4" | "md\:hover\:border-t-8" | "md\:hover\:border-white" | "md\:hover\:font-black" | "md\:hover\:font-bold" | "md\:hover\:font-extrabold" | "md\:hover\:font-hairline" | "md\:hover\:font-light" | "md\:hover\:font-medium" | "md\:hover\:font-normal" | "md\:hover\:font-semibold" | "md\:hover\:font-thin" | "md\:hover\:line-through" | "md\:hover\:no-underline" | "md\:hover\:opacity-0" | "md\:hover\:opacity-100" | "md\:hover\:opacity-25" | "md\:hover\:opacity-50" | "md\:hover\:opacity-75" | "md\:hover\:shadow" | "md\:hover\:shadow-2xl" | "md\:hover\:shadow-inner" | "md\:hover\:shadow-lg" | "md\:hover\:shadow-md" | "md\:hover\:shadow-none" | "md\:hover\:shadow-outline" | "md\:hover\:shadow-xl" | "md\:hover\:text-black" | "md\:hover\:text-blue-100" | "md\:hover\:text-blue-200" | "md\:hover\:text-blue-300" | "md\:hover\:text-blue-400" | "md\:hover\:text-blue-500" | "md\:hover\:text-blue-600" | "md\:hover\:text-blue-700" | "md\:hover\:text-blue-800" | "md\:hover\:text-blue-900" | "md\:hover\:text-border" | "md\:hover\:text-gray-100" | "md\:hover\:text-gray-200" | "md\:hover\:text-gray-300" | "md\:hover\:text-gray-400" | "md\:hover\:text-gray-500" | "md\:hover\:text-gray-600" | "md\:hover\:text-gray-700" | "md\:hover\:text-gray-800" | "md\:hover\:text-gray-900" | "md\:hover\:text-primary" | "md\:hover\:text-white" | "md\:hover\:underline" | "md\:hyphens-auto" | "md\:hyphens-manual" | "md\:hyphens-none" | "md\:inline" | "md\:inline-block" | "md\:inline-flex" | "md\:inset-0" | "md\:inset-auto" | "md\:inset-x-0" | "md\:inset-x-auto" | "md\:inset-y-0" | "md\:inset-y-auto" | "md\:invisible" | "md\:italic" | "md\:items-baseline" | "md\:items-center" | "md\:items-end" | "md\:items-start" | "md\:items-stretch" | "md\:justify-around" | "md\:justify-between" | "md\:justify-center" | "md\:justify-end" | "md\:justify-start" | "md\:last\:border" | "md\:last\:border-0" | "md\:last\:border-2" | "md\:last\:border-4" | "md\:last\:border-8" | "md\:last\:border-b" | "md\:last\:border-b-0" | "md\:last\:border-b-2" | "md\:last\:border-b-4" | "md\:last\:border-b-8" | "md\:last\:border-l" | "md\:last\:border-l-0" | "md\:last\:border-l-2" | "md\:last\:border-l-4" | "md\:last\:border-l-8" | "md\:last\:border-r" | "md\:last\:border-r-0" | "md\:last\:border-r-2" | "md\:last\:border-r-4" | "md\:last\:border-r-8" | "md\:last\:border-t" | "md\:last\:border-t-0" | "md\:last\:border-t-2" | "md\:last\:border-t-4" | "md\:last\:border-t-8" | "md\:leading-loose" | "md\:leading-none" | "md\:leading-normal" | "md\:leading-relaxed" | "md\:leading-snug" | "md\:leading-tight" | "md\:leading-unset" | "md\:left-0" | "md\:left-auto" | "md\:line-through" | "md\:lining-nums" | "md\:list-decimal" | "md\:list-disc" | "md\:list-inside" | "md\:list-none" | "md\:list-outside" | "md\:lowercase" | "md\:m-0" | "md\:m-1" | "md\:m-10" | "md\:m-12" | "md\:m-16" | "md\:m-2" | "md\:m-20" | "md\:m-24" | "md\:m-3" | "md\:m-32" | "md\:m-4" | "md\:m-40" | "md\:m-48" | "md\:m-5" | "md\:m-56" | "md\:m-6" | "md\:m-64" | "md\:m-8" | "md\:m-auto" | "md\:m-px" | "md\:max-h-full" | "md\:max-h-screen" | "md\:max-w-2xl" | "md\:max-w-3xl" | "md\:max-w-4xl" | "md\:max-w-5xl" | "md\:max-w-6xl" | "md\:max-w-full" | "md\:max-w-lg" | "md\:max-w-md" | "md\:max-w-sm" | "md\:max-w-xl" | "md\:max-w-xs" | "md\:mb-0" | "md\:mb-1" | "md\:mb-10" | "md\:mb-12" | "md\:mb-16" | "md\:mb-2" | "md\:mb-20" | "md\:mb-24" | "md\:mb-3" | "md\:mb-32" | "md\:mb-4" | "md\:mb-40" | "md\:mb-48" | "md\:mb-5" | "md\:mb-56" | "md\:mb-6" | "md\:mb-64" | "md\:mb-8" | "md\:mb-auto" | "md\:mb-px" | "md\:min-h-0" | "md\:min-h-full" | "md\:min-h-screen" | "md\:min-w-0" | "md\:min-w-full" | "md\:ml-0" | "md\:ml-1" | "md\:ml-10" | "md\:ml-12" | "md\:ml-16" | "md\:ml-2" | "md\:ml-20" | "md\:ml-24" | "md\:ml-3" | "md\:ml-32" | "md\:ml-4" | "md\:ml-40" | "md\:ml-48" | "md\:ml-5" | "md\:ml-56" | "md\:ml-6" | "md\:ml-64" | "md\:ml-8" | "md\:ml-auto" | "md\:ml-px" | "md\:mr-0" | "md\:mr-1" | "md\:mr-10" | "md\:mr-12" | "md\:mr-16" | "md\:mr-2" | "md\:mr-20" | "md\:mr-24" | "md\:mr-3" | "md\:mr-32" | "md\:mr-4" | "md\:mr-40" | "md\:mr-48" | "md\:mr-5" | "md\:mr-56" | "md\:mr-6" | "md\:mr-64" | "md\:mr-8" | "md\:mr-auto" | "md\:mr-px" | "md\:mt-0" | "md\:mt-1" | "md\:mt-10" | "md\:mt-12" | "md\:mt-16" | "md\:mt-2" | "md\:mt-20" | "md\:mt-24" | "md\:mt-3" | "md\:mt-32" | "md\:mt-4" | "md\:mt-40" | "md\:mt-48" | "md\:mt-5" | "md\:mt-56" | "md\:mt-6" | "md\:mt-64" | "md\:mt-8" | "md\:mt-auto" | "md\:mt-px" | "md\:mx-0" | "md\:mx-1" | "md\:mx-10" | "md\:mx-12" | "md\:mx-16" | "md\:mx-2" | "md\:mx-20" | "md\:mx-24" | "md\:mx-3" | "md\:mx-32" | "md\:mx-4" | "md\:mx-40" | "md\:mx-48" | "md\:mx-5" | "md\:mx-56" | "md\:mx-6" | "md\:mx-64" | "md\:mx-8" | "md\:mx-auto" | "md\:mx-px" | "md\:my-0" | "md\:my-1" | "md\:my-10" | "md\:my-12" | "md\:my-16" | "md\:my-2" | "md\:my-20" | "md\:my-24" | "md\:my-3" | "md\:my-32" | "md\:my-4" | "md\:my-40" | "md\:my-48" | "md\:my-5" | "md\:my-56" | "md\:my-6" | "md\:my-64" | "md\:my-8" | "md\:my-auto" | "md\:my-px" | "md\:no-common-ligatures" | "md\:no-contextual-ligatures" | "md\:no-discretionary-ligatures" | "md\:no-ellipsis" | "md\:no-historical-ligatures" | "md\:no-ligatures" | "md\:no-underline" | "md\:normal-caps" | "md\:normal-case" | "md\:normal-ligatures" | "md\:normal-nums" | "md\:not-italic" | "md\:not-sr-only" | "md\:object-bottom" | "md\:object-center" | "md\:object-contain" | "md\:object-cover" | "md\:object-fill" | "md\:object-left" | "md\:object-left-bottom" | "md\:object-left-top" | "md\:object-none" | "md\:object-right" | "md\:object-right-bottom" | "md\:object-right-top" | "md\:object-scale-down" | "md\:object-top" | "md\:oldstyle-nums" | "md\:opacity-0" | "md\:opacity-100" | "md\:opacity-25" | "md\:opacity-50" | "md\:opacity-75" | "md\:order-1" | "md\:order-10" | "md\:order-11" | "md\:order-12" | "md\:order-2" | "md\:order-3" | "md\:order-4" | "md\:order-5" | "md\:order-6" | "md\:order-7" | "md\:order-8" | "md\:order-9" | "md\:order-first" | "md\:order-last" | "md\:order-none" | "md\:ordinal-nums" | "md\:outline-none" | "md\:overflow-auto" | "md\:overflow-hidden" | "md\:overflow-scroll" | "md\:overflow-visible" | "md\:overflow-x-auto" | "md\:overflow-x-hidden" | "md\:overflow-x-scroll" | "md\:overflow-x-visible" | "md\:overflow-y-auto" | "md\:overflow-y-hidden" | "md\:overflow-y-scroll" | "md\:overflow-y-visible" | "md\:p-0" | "md\:p-1" | "md\:p-10" | "md\:p-12" | "md\:p-16" | "md\:p-2" | "md\:p-20" | "md\:p-24" | "md\:p-3" | "md\:p-32" | "md\:p-4" | "md\:p-40" | "md\:p-48" | "md\:p-5" | "md\:p-56" | "md\:p-6" | "md\:p-64" | "md\:p-8" | "md\:p-px" | "md\:pb-0" | "md\:pb-1" | "md\:pb-10" | "md\:pb-12" | "md\:pb-16" | "md\:pb-2" | "md\:pb-20" | "md\:pb-24" | "md\:pb-3" | "md\:pb-32" | "md\:pb-4" | "md\:pb-40" | "md\:pb-48" | "md\:pb-5" | "md\:pb-56" | "md\:pb-6" | "md\:pb-64" | "md\:pb-8" | "md\:pb-px" | "md\:petite-caps" | "md\:pl-0" | "md\:pl-1" | "md\:pl-10" | "md\:pl-12" | "md\:pl-16" | "md\:pl-2" | "md\:pl-20" | "md\:pl-24" | "md\:pl-3" | "md\:pl-32" | "md\:pl-4" | "md\:pl-40" | "md\:pl-48" | "md\:pl-5" | "md\:pl-56" | "md\:pl-6" | "md\:pl-64" | "md\:pl-8" | "md\:pl-px" | "md\:placeholder-black" | "md\:placeholder-blue-100" | "md\:placeholder-blue-200" | "md\:placeholder-blue-300" | "md\:placeholder-blue-400" | "md\:placeholder-blue-500" | "md\:placeholder-blue-600" | "md\:placeholder-blue-700" | "md\:placeholder-blue-800" | "md\:placeholder-blue-900" | "md\:placeholder-border" | "md\:placeholder-gray-100" | "md\:placeholder-gray-200" | "md\:placeholder-gray-300" | "md\:placeholder-gray-400" | "md\:placeholder-gray-500" | "md\:placeholder-gray-600" | "md\:placeholder-gray-700" | "md\:placeholder-gray-800" | "md\:placeholder-gray-900" | "md\:placeholder-primary" | "md\:placeholder-white" | "md\:pointer-events-auto" | "md\:pointer-events-none" | "md\:pr-0" | "md\:pr-1" | "md\:pr-10" | "md\:pr-12" | "md\:pr-16" | "md\:pr-2" | "md\:pr-20" | "md\:pr-24" | "md\:pr-3" | "md\:pr-32" | "md\:pr-4" | "md\:pr-40" | "md\:pr-48" | "md\:pr-5" | "md\:pr-56" | "md\:pr-6" | "md\:pr-64" | "md\:pr-8" | "md\:pr-px" | "md\:proportional-nums" | "md\:pt-0" | "md\:pt-1" | "md\:pt-10" | "md\:pt-12" | "md\:pt-16" | "md\:pt-2" | "md\:pt-20" | "md\:pt-24" | "md\:pt-3" | "md\:pt-32" | "md\:pt-4" | "md\:pt-40" | "md\:pt-48" | "md\:pt-5" | "md\:pt-56" | "md\:pt-6" | "md\:pt-64" | "md\:pt-8" | "md\:pt-px" | "md\:px-0" | "md\:px-1" | "md\:px-10" | "md\:px-12" | "md\:px-16" | "md\:px-2" | "md\:px-20" | "md\:px-24" | "md\:px-3" | "md\:px-32" | "md\:px-4" | "md\:px-40" | "md\:px-48" | "md\:px-5" | "md\:px-56" | "md\:px-6" | "md\:px-64" | "md\:px-8" | "md\:px-px" | "md\:py-0" | "md\:py-1" | "md\:py-10" | "md\:py-12" | "md\:py-16" | "md\:py-2" | "md\:py-20" | "md\:py-24" | "md\:py-3" | "md\:py-32" | "md\:py-4" | "md\:py-40" | "md\:py-48" | "md\:py-5" | "md\:py-56" | "md\:py-6" | "md\:py-64" | "md\:py-8" | "md\:py-px" | "md\:relative" | "md\:resize" | "md\:resize-none" | "md\:resize-x" | "md\:resize-y" | "md\:right-0" | "md\:right-auto" | "md\:rotate-180" | "md\:rotate-270" | "md\:rotate-90" | "md\:rounded" | "md\:rounded-b" | "md\:rounded-b-full" | "md\:rounded-b-lg" | "md\:rounded-b-none" | "md\:rounded-b-sm" | "md\:rounded-bl" | "md\:rounded-bl-full" | "md\:rounded-bl-lg" | "md\:rounded-bl-none" | "md\:rounded-bl-sm" | "md\:rounded-br" | "md\:rounded-br-full" | "md\:rounded-br-lg" | "md\:rounded-br-none" | "md\:rounded-br-sm" | "md\:rounded-full" | "md\:rounded-l" | "md\:rounded-l-full" | "md\:rounded-l-lg" | "md\:rounded-l-none" | "md\:rounded-l-sm" | "md\:rounded-lg" | "md\:rounded-none" | "md\:rounded-r" | "md\:rounded-r-full" | "md\:rounded-r-lg" | "md\:rounded-r-none" | "md\:rounded-r-sm" | "md\:rounded-sm" | "md\:rounded-t" | "md\:rounded-t-full" | "md\:rounded-t-lg" | "md\:rounded-t-none" | "md\:rounded-t-sm" | "md\:rounded-tl" | "md\:rounded-tl-full" | "md\:rounded-tl-lg" | "md\:rounded-tl-none" | "md\:rounded-tl-sm" | "md\:rounded-tr" | "md\:rounded-tr-full" | "md\:rounded-tr-lg" | "md\:rounded-tr-none" | "md\:rounded-tr-sm" | "md\:scrolling-auto" | "md\:scrolling-touch" | "md\:select-all" | "md\:select-auto" | "md\:select-none" | "md\:select-text" | "md\:self-auto" | "md\:self-center" | "md\:self-end" | "md\:self-start" | "md\:self-stretch" | "md\:shadow" | "md\:shadow-2xl" | "md\:shadow-inner" | "md\:shadow-lg" | "md\:shadow-md" | "md\:shadow-none" | "md\:shadow-outline" | "md\:shadow-xl" | "md\:slashed-zeros" | "md\:small-caps" | "md\:sr-only" | "md\:stacked-fractions" | "md\:static" | "md\:sticky" | "md\:stroke-current" | "md\:subpixel-antialiased" | "md\:table" | "md\:table-auto" | "md\:table-cell" | "md\:table-fixed" | "md\:table-row" | "md\:tabular-nums" | "md\:text-2xl" | "md\:text-3xl" | "md\:text-4xl" | "md\:text-5xl" | "md\:text-6xl" | "md\:text-align-unset" | "md\:text-base" | "md\:text-black" | "md\:text-blue-100" | "md\:text-blue-200" | "md\:text-blue-300" | "md\:text-blue-400" | "md\:text-blue-500" | "md\:text-blue-600" | "md\:text-blue-700" | "md\:text-blue-800" | "md\:text-blue-900" | "md\:text-border" | "md\:text-center" | "md\:text-color-unset" | "md\:text-gray-100" | "md\:text-gray-200" | "md\:text-gray-300" | "md\:text-gray-400" | "md\:text-gray-500" | "md\:text-gray-600" | "md\:text-gray-700" | "md\:text-gray-800" | "md\:text-gray-900" | "md\:text-justify" | "md\:text-left" | "md\:text-lg" | "md\:text-primary" | "md\:text-right" | "md\:text-size-unset" | "md\:text-sm" | "md\:text-transform-unset" | "md\:text-white" | "md\:text-xl" | "md\:text-xs" | "md\:titling-caps" | "md\:top-0" | "md\:top-auto" | "md\:tracking-normal" | "md\:tracking-tight" | "md\:tracking-tighter" | "md\:tracking-unset" | "md\:tracking-wide" | "md\:tracking-wider" | "md\:tracking-widest" | "md\:transform-b" | "md\:transform-bl" | "md\:transform-border" | "md\:transform-br" | "md\:transform-fill" | "md\:transform-l" | "md\:transform-none" | "md\:transform-r" | "md\:transform-t" | "md\:transform-tl" | "md\:transform-tr" | "md\:transform-view" | "md\:transition-0" | "md\:transition-100" | "md\:transition-1000" | "md\:transition-250" | "md\:transition-500" | "md\:transition-750" | "md\:transition-all" | "md\:transition-bg" | "md\:transition-border" | "md\:transition-color" | "md\:transition-colors" | "md\:transition-delay-0" | "md\:transition-delay-100" | "md\:transition-delay-1000" | "md\:transition-delay-250" | "md\:transition-delay-500" | "md\:transition-delay-750" | "md\:transition-ease" | "md\:transition-ease-in" | "md\:transition-ease-in-out" | "md\:transition-ease-out" | "md\:transition-linear" | "md\:transition-none" | "md\:transition-opacity" | "md\:transition-transform" | "md\:truncate" | "md\:underline" | "md\:unicase" | "md\:uppercase" | "md\:visible" | "md\:w-0" | "md\:w-1" | "md\:w-10" | "md\:w-10\/12" | "md\:w-11\/12" | "md\:w-12" | "md\:w-16" | "md\:w-1\/12" | "md\:w-1\/2" | "md\:w-1\/3" | "md\:w-1\/4" | "md\:w-1\/5" | "md\:w-1\/6" | "md\:w-2" | "md\:w-20" | "md\:w-24" | "md\:w-2\/12" | "md\:w-2\/3" | "md\:w-2\/4" | "md\:w-2\/5" | "md\:w-2\/6" | "md\:w-3" | "md\:w-32" | "md\:w-3\/12" | "md\:w-3\/4" | "md\:w-3\/5" | "md\:w-3\/6" | "md\:w-4" | "md\:w-40" | "md\:w-48" | "md\:w-4\/12" | "md\:w-4\/5" | "md\:w-4\/6" | "md\:w-5" | "md\:w-56" | "md\:w-5\/12" | "md\:w-5\/6" | "md\:w-6" | "md\:w-64" | "md\:w-6\/12" | "md\:w-7\/12" | "md\:w-8" | "md\:w-8\/12" | "md\:w-9\/12" | "md\:w-auto" | "md\:w-full" | "md\:w-px" | "md\:w-screen" | "md\:whitespace-no-wrap" | "md\:whitespace-normal" | "md\:whitespace-pre" | "md\:whitespace-pre-line" | "md\:whitespace-pre-wrap" | "md\:will-change-auto" | "md\:will-change-contents" | "md\:will-change-opacity" | "md\:will-change-scroll" | "md\:will-change-transform" | "md\:z-0" | "md\:z-10" | "md\:z-20" | "md\:z-30" | "md\:z-40" | "md\:z-50" | "md\:z-auto" | "min-h-0" | "min-h-full" | "min-h-screen" | "min-w-0" | "min-w-full" | "ml-0" | "ml-1" | "ml-10" | "ml-12" | "ml-16" | "ml-2" | "ml-20" | "ml-24" | "ml-3" | "ml-32" | "ml-4" | "ml-40" | "ml-48" | "ml-5" | "ml-56" | "ml-6" | "ml-64" | "ml-8" | "ml-auto" | "ml-px" | "mr-0" | "mr-1" | "mr-10" | "mr-12" | "mr-16" | "mr-2" | "mr-20" | "mr-24" | "mr-3" | "mr-32" | "mr-4" | "mr-40" | "mr-48" | "mr-5" | "mr-56" | "mr-6" | "mr-64" | "mr-8" | "mr-auto" | "mr-px" | "mt-0" | "mt-1" | "mt-10" | "mt-12" | "mt-16" | "mt-2" | "mt-20" | "mt-24" | "mt-3" | "mt-32" | "mt-4" | "mt-40" | "mt-48" | "mt-5" | "mt-56" | "mt-6" | "mt-64" | "mt-8" | "mt-auto" | "mt-px" | "mx-0" | "mx-1" | "mx-10" | "mx-12" | "mx-16" | "mx-2" | "mx-20" | "mx-24" | "mx-3" | "mx-32" | "mx-4" | "mx-40" | "mx-48" | "mx-5" | "mx-56" | "mx-6" | "mx-64" | "mx-8" | "mx-auto" | "mx-px" | "my-0" | "my-1" | "my-10" | "my-12" | "my-16" | "my-2" | "my-20" | "my-24" | "my-3" | "my-32" | "my-4" | "my-40" | "my-48" | "my-5" | "my-56" | "my-6" | "my-64" | "my-8" | "my-auto" | "my-px" | "no-common-ligatures" | "no-contextual-ligatures" | "no-discretionary-ligatures" | "no-ellipsis" | "no-historical-ligatures" | "no-ligatures" | "no-underline" | "normal-caps" | "normal-case" | "normal-ligatures" | "normal-nums" | "not-italic" | "not-sr-only" | "object-bottom" | "object-center" | "object-contain" | "object-cover" | "object-fill" | "object-left" | "object-left-bottom" | "object-left-top" | "object-none" | "object-right" | "object-right-bottom" | "object-right-top" | "object-scale-down" | "object-top" | "oldstyle-nums" | "opacity-0" | "opacity-100" | "opacity-25" | "opacity-50" | "opacity-75" | "order-1" | "order-10" | "order-11" | "order-12" | "order-2" | "order-3" | "order-4" | "order-5" | "order-6" | "order-7" | "order-8" | "order-9" | "order-first" | "order-last" | "order-none" | "ordinal-nums" | "outline-none" | "overflow-auto" | "overflow-hidden" | "overflow-scroll" | "overflow-visible" | "overflow-x-auto" | "overflow-x-hidden" | "overflow-x-scroll" | "overflow-x-visible" | "overflow-y-auto" | "overflow-y-hidden" | "overflow-y-scroll" | "overflow-y-visible" | "p-0" | "p-1" | "p-10" | "p-12" | "p-16" | "p-2" | "p-20" | "p-24" | "p-3" | "p-32" | "p-4" | "p-40" | "p-48" | "p-5" | "p-56" | "p-6" | "p-64" | "p-8" | "p-px" | "pb-0" | "pb-1" | "pb-10" | "pb-12" | "pb-16" | "pb-2" | "pb-20" | "pb-24" | "pb-3" | "pb-32" | "pb-4" | "pb-40" | "pb-48" | "pb-5" | "pb-56" | "pb-6" | "pb-64" | "pb-8" | "pb-px" | "petite-caps" | "pl-0" | "pl-1" | "pl-10" | "pl-12" | "pl-16" | "pl-2" | "pl-20" | "pl-24" | "pl-3" | "pl-32" | "pl-4" | "pl-40" | "pl-48" | "pl-5" | "pl-56" | "pl-6" | "pl-64" | "pl-8" | "pl-px" | "placeholder-black" | "placeholder-blue-100" | "placeholder-blue-200" | "placeholder-blue-300" | "placeholder-blue-400" | "placeholder-blue-500" | "placeholder-blue-600" | "placeholder-blue-700" | "placeholder-blue-800" | "placeholder-blue-900" | "placeholder-border" | "placeholder-gray-100" | "placeholder-gray-200" | "placeholder-gray-300" | "placeholder-gray-400" | "placeholder-gray-500" | "placeholder-gray-600" | "placeholder-gray-700" | "placeholder-gray-800" | "placeholder-gray-900" | "placeholder-primary" | "placeholder-white" | "pointer-events-auto" | "pointer-events-none" | "pr-0" | "pr-1" | "pr-10" | "pr-12" | "pr-16" | "pr-2" | "pr-20" | "pr-24" | "pr-3" | "pr-32" | "pr-4" | "pr-40" | "pr-48" | "pr-5" | "pr-56" | "pr-6" | "pr-64" | "pr-8" | "pr-px" | "proportional-nums" | "pt-0" | "pt-1" | "pt-10" | "pt-12" | "pt-16" | "pt-2" | "pt-20" | "pt-24" | "pt-3" | "pt-32" | "pt-4" | "pt-40" | "pt-48" | "pt-5" | "pt-56" | "pt-6" | "pt-64" | "pt-8" | "pt-px" | "px-0" | "px-1" | "px-10" | "px-12" | "px-16" | "px-2" | "px-20" | "px-24" | "px-3" | "px-32" | "px-4" | "px-40" | "px-48" | "px-5" | "px-56" | "px-6" | "px-64" | "px-8" | "px-px" | "py-0" | "py-1" | "py-10" | "py-12" | "py-16" | "py-2" | "py-20" | "py-24" | "py-3" | "py-32" | "py-4" | "py-40" | "py-48" | "py-5" | "py-56" | "py-6" | "py-64" | "py-8" | "py-px" | "relative" | "resize" | "resize-none" | "resize-x" | "resize-y" | "rfs-control-container" | "rfs-select-container" | "right-0" | "right-auto" | "rotate-180" | "rotate-270" | "rotate-90" | "rounded" | "rounded-b" | "rounded-b-full" | "rounded-b-lg" | "rounded-b-none" | "rounded-b-sm" | "rounded-bl" | "rounded-bl-full" | "rounded-bl-lg" | "rounded-bl-none" | "rounded-bl-sm" | "rounded-br" | "rounded-br-full" | "rounded-br-lg" | "rounded-br-none" | "rounded-br-sm" | "rounded-full" | "rounded-l" | "rounded-l-full" | "rounded-l-lg" | "rounded-l-none" | "rounded-l-sm" | "rounded-lg" | "rounded-none" | "rounded-r" | "rounded-r-full" | "rounded-r-lg" | "rounded-r-none" | "rounded-r-sm" | "rounded-sm" | "rounded-t" | "rounded-t-full" | "rounded-t-lg" | "rounded-t-none" | "rounded-t-sm" | "rounded-tl" | "rounded-tl-full" | "rounded-tl-lg" | "rounded-tl-none" | "rounded-tl-sm" | "rounded-tr" | "rounded-tr-full" | "rounded-tr-lg" | "rounded-tr-none" | "rounded-tr-sm" | "scrolling-auto" | "scrolling-touch" | "select-all" | "select-auto" | "select-container" | "select-none" | "select-text" | "select__control" | "select__control--menu-is-open" | "select__group-heading" | "select__indicator" | "select__indicator-separator" | "select__input" | "select__menu" | "select__option" | "select__option--is-focused" | "select__option--is-selected" | "select__placeholder" | "select__single-value" | "select__value-container" | "self-auto" | "self-center" | "self-end" | "self-start" | "self-stretch" | "shadow" | "shadow-2xl" | "shadow-inner" | "shadow-lg" | "shadow-md" | "shadow-none" | "shadow-outline" | "shadow-xl" | "slashed-zeros" | "sm\:-m-1" | "sm\:-m-10" | "sm\:-m-12" | "sm\:-m-16" | "sm\:-m-2" | "sm\:-m-20" | "sm\:-m-24" | "sm\:-m-3" | "sm\:-m-32" | "sm\:-m-4" | "sm\:-m-40" | "sm\:-m-48" | "sm\:-m-5" | "sm\:-m-56" | "sm\:-m-6" | "sm\:-m-64" | "sm\:-m-8" | "sm\:-m-px" | "sm\:-mb-1" | "sm\:-mb-10" | "sm\:-mb-12" | "sm\:-mb-16" | "sm\:-mb-2" | "sm\:-mb-20" | "sm\:-mb-24" | "sm\:-mb-3" | "sm\:-mb-32" | "sm\:-mb-4" | "sm\:-mb-40" | "sm\:-mb-48" | "sm\:-mb-5" | "sm\:-mb-56" | "sm\:-mb-6" | "sm\:-mb-64" | "sm\:-mb-8" | "sm\:-mb-px" | "sm\:-ml-1" | "sm\:-ml-10" | "sm\:-ml-12" | "sm\:-ml-16" | "sm\:-ml-2" | "sm\:-ml-20" | "sm\:-ml-24" | "sm\:-ml-3" | "sm\:-ml-32" | "sm\:-ml-4" | "sm\:-ml-40" | "sm\:-ml-48" | "sm\:-ml-5" | "sm\:-ml-56" | "sm\:-ml-6" | "sm\:-ml-64" | "sm\:-ml-8" | "sm\:-ml-px" | "sm\:-mr-1" | "sm\:-mr-10" | "sm\:-mr-12" | "sm\:-mr-16" | "sm\:-mr-2" | "sm\:-mr-20" | "sm\:-mr-24" | "sm\:-mr-3" | "sm\:-mr-32" | "sm\:-mr-4" | "sm\:-mr-40" | "sm\:-mr-48" | "sm\:-mr-5" | "sm\:-mr-56" | "sm\:-mr-6" | "sm\:-mr-64" | "sm\:-mr-8" | "sm\:-mr-px" | "sm\:-mt-1" | "sm\:-mt-10" | "sm\:-mt-12" | "sm\:-mt-16" | "sm\:-mt-2" | "sm\:-mt-20" | "sm\:-mt-24" | "sm\:-mt-3" | "sm\:-mt-32" | "sm\:-mt-4" | "sm\:-mt-40" | "sm\:-mt-48" | "sm\:-mt-5" | "sm\:-mt-56" | "sm\:-mt-6" | "sm\:-mt-64" | "sm\:-mt-8" | "sm\:-mt-px" | "sm\:-mx-1" | "sm\:-mx-10" | "sm\:-mx-12" | "sm\:-mx-16" | "sm\:-mx-2" | "sm\:-mx-20" | "sm\:-mx-24" | "sm\:-mx-3" | "sm\:-mx-32" | "sm\:-mx-4" | "sm\:-mx-40" | "sm\:-mx-48" | "sm\:-mx-5" | "sm\:-mx-56" | "sm\:-mx-6" | "sm\:-mx-64" | "sm\:-mx-8" | "sm\:-mx-px" | "sm\:-my-1" | "sm\:-my-10" | "sm\:-my-12" | "sm\:-my-16" | "sm\:-my-2" | "sm\:-my-20" | "sm\:-my-24" | "sm\:-my-3" | "sm\:-my-32" | "sm\:-my-4" | "sm\:-my-40" | "sm\:-my-48" | "sm\:-my-5" | "sm\:-my-56" | "sm\:-my-6" | "sm\:-my-64" | "sm\:-my-8" | "sm\:-my-px" | "sm\:absolute" | "sm\:align-baseline" | "sm\:align-bottom" | "sm\:align-middle" | "sm\:align-text-bottom" | "sm\:align-text-top" | "sm\:align-top" | "sm\:all-small-caps" | "sm\:antialiased" | "sm\:appearance-none" | "sm\:bg-auto" | "sm\:bg-black" | "sm\:bg-blue-100" | "sm\:bg-blue-200" | "sm\:bg-blue-300" | "sm\:bg-blue-400" | "sm\:bg-blue-500" | "sm\:bg-blue-600" | "sm\:bg-blue-700" | "sm\:bg-blue-800" | "sm\:bg-blue-900" | "sm\:bg-border" | "sm\:bg-bottom" | "sm\:bg-center" | "sm\:bg-contain" | "sm\:bg-cover" | "sm\:bg-fixed" | "sm\:bg-gray-100" | "sm\:bg-gray-200" | "sm\:bg-gray-300" | "sm\:bg-gray-400" | "sm\:bg-gray-500" | "sm\:bg-gray-600" | "sm\:bg-gray-700" | "sm\:bg-gray-800" | "sm\:bg-gray-900" | "sm\:bg-left" | "sm\:bg-left-bottom" | "sm\:bg-left-top" | "sm\:bg-local" | "sm\:bg-no-repeat" | "sm\:bg-primary" | "sm\:bg-repeat" | "sm\:bg-repeat-round" | "sm\:bg-repeat-space" | "sm\:bg-repeat-x" | "sm\:bg-repeat-y" | "sm\:bg-right" | "sm\:bg-right-bottom" | "sm\:bg-right-top" | "sm\:bg-scroll" | "sm\:bg-top" | "sm\:bg-white" | "sm\:block" | "sm\:border" | "sm\:border-0" | "sm\:border-2" | "sm\:border-4" | "sm\:border-8" | "sm\:border-b" | "sm\:border-b-0" | "sm\:border-b-2" | "sm\:border-b-4" | "sm\:border-b-8" | "sm\:border-black" | "sm\:border-blue-100" | "sm\:border-blue-200" | "sm\:border-blue-300" | "sm\:border-blue-400" | "sm\:border-blue-500" | "sm\:border-blue-600" | "sm\:border-blue-700" | "sm\:border-blue-800" | "sm\:border-blue-900" | "sm\:border-border" | "sm\:border-collapse" | "sm\:border-dashed" | "sm\:border-dotted" | "sm\:border-double" | "sm\:border-gray-100" | "sm\:border-gray-200" | "sm\:border-gray-300" | "sm\:border-gray-400" | "sm\:border-gray-500" | "sm\:border-gray-600" | "sm\:border-gray-700" | "sm\:border-gray-800" | "sm\:border-gray-900" | "sm\:border-l" | "sm\:border-l-0" | "sm\:border-l-2" | "sm\:border-l-4" | "sm\:border-l-8" | "sm\:border-none" | "sm\:border-primary" | "sm\:border-r" | "sm\:border-r-0" | "sm\:border-r-2" | "sm\:border-r-4" | "sm\:border-r-8" | "sm\:border-separate" | "sm\:border-solid" | "sm\:border-t" | "sm\:border-t-0" | "sm\:border-t-2" | "sm\:border-t-4" | "sm\:border-t-8" | "sm\:border-white" | "sm\:bottom-0" | "sm\:bottom-auto" | "sm\:break-all" | "sm\:break-normal" | "sm\:break-words" | "sm\:capitalize" | "sm\:clearfix" | "sm\:common-ligatures" | "sm\:content-around" | "sm\:content-between" | "sm\:content-center" | "sm\:content-end" | "sm\:content-start" | "sm\:contextual-ligatures" | "sm\:cursor-auto" | "sm\:cursor-default" | "sm\:cursor-move" | "sm\:cursor-not-allowed" | "sm\:cursor-pointer" | "sm\:cursor-text" | "sm\:cursor-wait" | "sm\:diagonal-fractions" | "sm\:discretionary-ligatures" | "sm\:ellipsis" | "sm\:fill-current" | "sm\:fixed" | "sm\:flex" | "sm\:flex-1" | "sm\:flex-auto" | "sm\:flex-col" | "sm\:flex-col-reverse" | "sm\:flex-grow" | "sm\:flex-grow-0" | "sm\:flex-initial" | "sm\:flex-no-wrap" | "sm\:flex-none" | "sm\:flex-row" | "sm\:flex-row-reverse" | "sm\:flex-shrink" | "sm\:flex-shrink-0" | "sm\:flex-wrap" | "sm\:flex-wrap-reverse" | "sm\:float-left" | "sm\:float-none" | "sm\:float-right" | "sm\:focus\:bg-black" | "sm\:focus\:bg-blue-100" | "sm\:focus\:bg-blue-200" | "sm\:focus\:bg-blue-300" | "sm\:focus\:bg-blue-400" | "sm\:focus\:bg-blue-500" | "sm\:focus\:bg-blue-600" | "sm\:focus\:bg-blue-700" | "sm\:focus\:bg-blue-800" | "sm\:focus\:bg-blue-900" | "sm\:focus\:bg-border" | "sm\:focus\:bg-gray-100" | "sm\:focus\:bg-gray-200" | "sm\:focus\:bg-gray-300" | "sm\:focus\:bg-gray-400" | "sm\:focus\:bg-gray-500" | "sm\:focus\:bg-gray-600" | "sm\:focus\:bg-gray-700" | "sm\:focus\:bg-gray-800" | "sm\:focus\:bg-gray-900" | "sm\:focus\:bg-primary" | "sm\:focus\:bg-white" | "sm\:focus\:border" | "sm\:focus\:border-0" | "sm\:focus\:border-2" | "sm\:focus\:border-4" | "sm\:focus\:border-8" | "sm\:focus\:border-b" | "sm\:focus\:border-b-0" | "sm\:focus\:border-b-2" | "sm\:focus\:border-b-4" | "sm\:focus\:border-b-8" | "sm\:focus\:border-black" | "sm\:focus\:border-blue-100" | "sm\:focus\:border-blue-200" | "sm\:focus\:border-blue-300" | "sm\:focus\:border-blue-400" | "sm\:focus\:border-blue-500" | "sm\:focus\:border-blue-600" | "sm\:focus\:border-blue-700" | "sm\:focus\:border-blue-800" | "sm\:focus\:border-blue-900" | "sm\:focus\:border-border" | "sm\:focus\:border-gray-100" | "sm\:focus\:border-gray-200" | "sm\:focus\:border-gray-300" | "sm\:focus\:border-gray-400" | "sm\:focus\:border-gray-500" | "sm\:focus\:border-gray-600" | "sm\:focus\:border-gray-700" | "sm\:focus\:border-gray-800" | "sm\:focus\:border-gray-900" | "sm\:focus\:border-l" | "sm\:focus\:border-l-0" | "sm\:focus\:border-l-2" | "sm\:focus\:border-l-4" | "sm\:focus\:border-l-8" | "sm\:focus\:border-primary" | "sm\:focus\:border-r" | "sm\:focus\:border-r-0" | "sm\:focus\:border-r-2" | "sm\:focus\:border-r-4" | "sm\:focus\:border-r-8" | "sm\:focus\:border-t" | "sm\:focus\:border-t-0" | "sm\:focus\:border-t-2" | "sm\:focus\:border-t-4" | "sm\:focus\:border-t-8" | "sm\:focus\:border-white" | "sm\:focus\:font-black" | "sm\:focus\:font-bold" | "sm\:focus\:font-extrabold" | "sm\:focus\:font-hairline" | "sm\:focus\:font-light" | "sm\:focus\:font-medium" | "sm\:focus\:font-normal" | "sm\:focus\:font-semibold" | "sm\:focus\:font-thin" | "sm\:focus\:line-through" | "sm\:focus\:no-underline" | "sm\:focus\:not-sr-only" | "sm\:focus\:opacity-0" | "sm\:focus\:opacity-100" | "sm\:focus\:opacity-25" | "sm\:focus\:opacity-50" | "sm\:focus\:opacity-75" | "sm\:focus\:outline-none" | "sm\:focus\:placeholder-black" | "sm\:focus\:placeholder-blue-100" | "sm\:focus\:placeholder-blue-200" | "sm\:focus\:placeholder-blue-300" | "sm\:focus\:placeholder-blue-400" | "sm\:focus\:placeholder-blue-500" | "sm\:focus\:placeholder-blue-600" | "sm\:focus\:placeholder-blue-700" | "sm\:focus\:placeholder-blue-800" | "sm\:focus\:placeholder-blue-900" | "sm\:focus\:placeholder-border" | "sm\:focus\:placeholder-gray-100" | "sm\:focus\:placeholder-gray-200" | "sm\:focus\:placeholder-gray-300" | "sm\:focus\:placeholder-gray-400" | "sm\:focus\:placeholder-gray-500" | "sm\:focus\:placeholder-gray-600" | "sm\:focus\:placeholder-gray-700" | "sm\:focus\:placeholder-gray-800" | "sm\:focus\:placeholder-gray-900" | "sm\:focus\:placeholder-primary" | "sm\:focus\:placeholder-white" | "sm\:focus\:shadow" | "sm\:focus\:shadow-2xl" | "sm\:focus\:shadow-inner" | "sm\:focus\:shadow-lg" | "sm\:focus\:shadow-md" | "sm\:focus\:shadow-none" | "sm\:focus\:shadow-outline" | "sm\:focus\:shadow-xl" | "sm\:focus\:sr-only" | "sm\:focus\:text-black" | "sm\:focus\:text-blue-100" | "sm\:focus\:text-blue-200" | "sm\:focus\:text-blue-300" | "sm\:focus\:text-blue-400" | "sm\:focus\:text-blue-500" | "sm\:focus\:text-blue-600" | "sm\:focus\:text-blue-700" | "sm\:focus\:text-blue-800" | "sm\:focus\:text-blue-900" | "sm\:focus\:text-border" | "sm\:focus\:text-gray-100" | "sm\:focus\:text-gray-200" | "sm\:focus\:text-gray-300" | "sm\:focus\:text-gray-400" | "sm\:focus\:text-gray-500" | "sm\:focus\:text-gray-600" | "sm\:focus\:text-gray-700" | "sm\:focus\:text-gray-800" | "sm\:focus\:text-gray-900" | "sm\:focus\:text-primary" | "sm\:focus\:text-white" | "sm\:focus\:underline" | "sm\:font-black" | "sm\:font-bold" | "sm\:font-extrabold" | "sm\:font-family-unset" | "sm\:font-hairline" | "sm\:font-light" | "sm\:font-medium" | "sm\:font-mono" | "sm\:font-normal" | "sm\:font-sans" | "sm\:font-semibold" | "sm\:font-serif" | "sm\:font-style-unset" | "sm\:font-thin" | "sm\:font-weight-unset" | "sm\:h-0" | "sm\:h-1" | "sm\:h-10" | "sm\:h-12" | "sm\:h-16" | "sm\:h-2" | "sm\:h-20" | "sm\:h-24" | "sm\:h-3" | "sm\:h-32" | "sm\:h-4" | "sm\:h-40" | "sm\:h-48" | "sm\:h-5" | "sm\:h-56" | "sm\:h-6" | "sm\:h-64" | "sm\:h-8" | "sm\:h-auto" | "sm\:h-full" | "sm\:h-px" | "sm\:h-screen" | "sm\:hidden" | "sm\:historical-ligatures" | "sm\:hover\:bg-black" | "sm\:hover\:bg-blue-100" | "sm\:hover\:bg-blue-200" | "sm\:hover\:bg-blue-300" | "sm\:hover\:bg-blue-400" | "sm\:hover\:bg-blue-500" | "sm\:hover\:bg-blue-600" | "sm\:hover\:bg-blue-700" | "sm\:hover\:bg-blue-800" | "sm\:hover\:bg-blue-900" | "sm\:hover\:bg-border" | "sm\:hover\:bg-gray-100" | "sm\:hover\:bg-gray-200" | "sm\:hover\:bg-gray-300" | "sm\:hover\:bg-gray-400" | "sm\:hover\:bg-gray-500" | "sm\:hover\:bg-gray-600" | "sm\:hover\:bg-gray-700" | "sm\:hover\:bg-gray-800" | "sm\:hover\:bg-gray-900" | "sm\:hover\:bg-primary" | "sm\:hover\:bg-white" | "sm\:hover\:border" | "sm\:hover\:border-0" | "sm\:hover\:border-2" | "sm\:hover\:border-4" | "sm\:hover\:border-8" | "sm\:hover\:border-b" | "sm\:hover\:border-b-0" | "sm\:hover\:border-b-2" | "sm\:hover\:border-b-4" | "sm\:hover\:border-b-8" | "sm\:hover\:border-black" | "sm\:hover\:border-blue-100" | "sm\:hover\:border-blue-200" | "sm\:hover\:border-blue-300" | "sm\:hover\:border-blue-400" | "sm\:hover\:border-blue-500" | "sm\:hover\:border-blue-600" | "sm\:hover\:border-blue-700" | "sm\:hover\:border-blue-800" | "sm\:hover\:border-blue-900" | "sm\:hover\:border-border" | "sm\:hover\:border-gray-100" | "sm\:hover\:border-gray-200" | "sm\:hover\:border-gray-300" | "sm\:hover\:border-gray-400" | "sm\:hover\:border-gray-500" | "sm\:hover\:border-gray-600" | "sm\:hover\:border-gray-700" | "sm\:hover\:border-gray-800" | "sm\:hover\:border-gray-900" | "sm\:hover\:border-l" | "sm\:hover\:border-l-0" | "sm\:hover\:border-l-2" | "sm\:hover\:border-l-4" | "sm\:hover\:border-l-8" | "sm\:hover\:border-primary" | "sm\:hover\:border-r" | "sm\:hover\:border-r-0" | "sm\:hover\:border-r-2" | "sm\:hover\:border-r-4" | "sm\:hover\:border-r-8" | "sm\:hover\:border-t" | "sm\:hover\:border-t-0" | "sm\:hover\:border-t-2" | "sm\:hover\:border-t-4" | "sm\:hover\:border-t-8" | "sm\:hover\:border-white" | "sm\:hover\:font-black" | "sm\:hover\:font-bold" | "sm\:hover\:font-extrabold" | "sm\:hover\:font-hairline" | "sm\:hover\:font-light" | "sm\:hover\:font-medium" | "sm\:hover\:font-normal" | "sm\:hover\:font-semibold" | "sm\:hover\:font-thin" | "sm\:hover\:line-through" | "sm\:hover\:no-underline" | "sm\:hover\:opacity-0" | "sm\:hover\:opacity-100" | "sm\:hover\:opacity-25" | "sm\:hover\:opacity-50" | "sm\:hover\:opacity-75" | "sm\:hover\:shadow" | "sm\:hover\:shadow-2xl" | "sm\:hover\:shadow-inner" | "sm\:hover\:shadow-lg" | "sm\:hover\:shadow-md" | "sm\:hover\:shadow-none" | "sm\:hover\:shadow-outline" | "sm\:hover\:shadow-xl" | "sm\:hover\:text-black" | "sm\:hover\:text-blue-100" | "sm\:hover\:text-blue-200" | "sm\:hover\:text-blue-300" | "sm\:hover\:text-blue-400" | "sm\:hover\:text-blue-500" | "sm\:hover\:text-blue-600" | "sm\:hover\:text-blue-700" | "sm\:hover\:text-blue-800" | "sm\:hover\:text-blue-900" | "sm\:hover\:text-border" | "sm\:hover\:text-gray-100" | "sm\:hover\:text-gray-200" | "sm\:hover\:text-gray-300" | "sm\:hover\:text-gray-400" | "sm\:hover\:text-gray-500" | "sm\:hover\:text-gray-600" | "sm\:hover\:text-gray-700" | "sm\:hover\:text-gray-800" | "sm\:hover\:text-gray-900" | "sm\:hover\:text-primary" | "sm\:hover\:text-white" | "sm\:hover\:underline" | "sm\:hyphens-auto" | "sm\:hyphens-manual" | "sm\:hyphens-none" | "sm\:inline" | "sm\:inline-block" | "sm\:inline-flex" | "sm\:inset-0" | "sm\:inset-auto" | "sm\:inset-x-0" | "sm\:inset-x-auto" | "sm\:inset-y-0" | "sm\:inset-y-auto" | "sm\:invisible" | "sm\:italic" | "sm\:items-baseline" | "sm\:items-center" | "sm\:items-end" | "sm\:items-start" | "sm\:items-stretch" | "sm\:justify-around" | "sm\:justify-between" | "sm\:justify-center" | "sm\:justify-end" | "sm\:justify-start" | "sm\:last\:border" | "sm\:last\:border-0" | "sm\:last\:border-2" | "sm\:last\:border-4" | "sm\:last\:border-8" | "sm\:last\:border-b" | "sm\:last\:border-b-0" | "sm\:last\:border-b-2" | "sm\:last\:border-b-4" | "sm\:last\:border-b-8" | "sm\:last\:border-l" | "sm\:last\:border-l-0" | "sm\:last\:border-l-2" | "sm\:last\:border-l-4" | "sm\:last\:border-l-8" | "sm\:last\:border-r" | "sm\:last\:border-r-0" | "sm\:last\:border-r-2" | "sm\:last\:border-r-4" | "sm\:last\:border-r-8" | "sm\:last\:border-t" | "sm\:last\:border-t-0" | "sm\:last\:border-t-2" | "sm\:last\:border-t-4" | "sm\:last\:border-t-8" | "sm\:leading-loose" | "sm\:leading-none" | "sm\:leading-normal" | "sm\:leading-relaxed" | "sm\:leading-snug" | "sm\:leading-tight" | "sm\:leading-unset" | "sm\:left-0" | "sm\:left-auto" | "sm\:line-through" | "sm\:lining-nums" | "sm\:list-decimal" | "sm\:list-disc" | "sm\:list-inside" | "sm\:list-none" | "sm\:list-outside" | "sm\:lowercase" | "sm\:m-0" | "sm\:m-1" | "sm\:m-10" | "sm\:m-12" | "sm\:m-16" | "sm\:m-2" | "sm\:m-20" | "sm\:m-24" | "sm\:m-3" | "sm\:m-32" | "sm\:m-4" | "sm\:m-40" | "sm\:m-48" | "sm\:m-5" | "sm\:m-56" | "sm\:m-6" | "sm\:m-64" | "sm\:m-8" | "sm\:m-auto" | "sm\:m-px" | "sm\:max-h-full" | "sm\:max-h-screen" | "sm\:max-w-2xl" | "sm\:max-w-3xl" | "sm\:max-w-4xl" | "sm\:max-w-5xl" | "sm\:max-w-6xl" | "sm\:max-w-full" | "sm\:max-w-lg" | "sm\:max-w-md" | "sm\:max-w-sm" | "sm\:max-w-xl" | "sm\:max-w-xs" | "sm\:mb-0" | "sm\:mb-1" | "sm\:mb-10" | "sm\:mb-12" | "sm\:mb-16" | "sm\:mb-2" | "sm\:mb-20" | "sm\:mb-24" | "sm\:mb-3" | "sm\:mb-32" | "sm\:mb-4" | "sm\:mb-40" | "sm\:mb-48" | "sm\:mb-5" | "sm\:mb-56" | "sm\:mb-6" | "sm\:mb-64" | "sm\:mb-8" | "sm\:mb-auto" | "sm\:mb-px" | "sm\:min-h-0" | "sm\:min-h-full" | "sm\:min-h-screen" | "sm\:min-w-0" | "sm\:min-w-full" | "sm\:ml-0" | "sm\:ml-1" | "sm\:ml-10" | "sm\:ml-12" | "sm\:ml-16" | "sm\:ml-2" | "sm\:ml-20" | "sm\:ml-24" | "sm\:ml-3" | "sm\:ml-32" | "sm\:ml-4" | "sm\:ml-40" | "sm\:ml-48" | "sm\:ml-5" | "sm\:ml-56" | "sm\:ml-6" | "sm\:ml-64" | "sm\:ml-8" | "sm\:ml-auto" | "sm\:ml-px" | "sm\:mr-0" | "sm\:mr-1" | "sm\:mr-10" | "sm\:mr-12" | "sm\:mr-16" | "sm\:mr-2" | "sm\:mr-20" | "sm\:mr-24" | "sm\:mr-3" | "sm\:mr-32" | "sm\:mr-4" | "sm\:mr-40" | "sm\:mr-48" | "sm\:mr-5" | "sm\:mr-56" | "sm\:mr-6" | "sm\:mr-64" | "sm\:mr-8" | "sm\:mr-auto" | "sm\:mr-px" | "sm\:mt-0" | "sm\:mt-1" | "sm\:mt-10" | "sm\:mt-12" | "sm\:mt-16" | "sm\:mt-2" | "sm\:mt-20" | "sm\:mt-24" | "sm\:mt-3" | "sm\:mt-32" | "sm\:mt-4" | "sm\:mt-40" | "sm\:mt-48" | "sm\:mt-5" | "sm\:mt-56" | "sm\:mt-6" | "sm\:mt-64" | "sm\:mt-8" | "sm\:mt-auto" | "sm\:mt-px" | "sm\:mx-0" | "sm\:mx-1" | "sm\:mx-10" | "sm\:mx-12" | "sm\:mx-16" | "sm\:mx-2" | "sm\:mx-20" | "sm\:mx-24" | "sm\:mx-3" | "sm\:mx-32" | "sm\:mx-4" | "sm\:mx-40" | "sm\:mx-48" | "sm\:mx-5" | "sm\:mx-56" | "sm\:mx-6" | "sm\:mx-64" | "sm\:mx-8" | "sm\:mx-auto" | "sm\:mx-px" | "sm\:my-0" | "sm\:my-1" | "sm\:my-10" | "sm\:my-12" | "sm\:my-16" | "sm\:my-2" | "sm\:my-20" | "sm\:my-24" | "sm\:my-3" | "sm\:my-32" | "sm\:my-4" | "sm\:my-40" | "sm\:my-48" | "sm\:my-5" | "sm\:my-56" | "sm\:my-6" | "sm\:my-64" | "sm\:my-8" | "sm\:my-auto" | "sm\:my-px" | "sm\:no-common-ligatures" | "sm\:no-contextual-ligatures" | "sm\:no-discretionary-ligatures" | "sm\:no-ellipsis" | "sm\:no-historical-ligatures" | "sm\:no-ligatures" | "sm\:no-underline" | "sm\:normal-caps" | "sm\:normal-case" | "sm\:normal-ligatures" | "sm\:normal-nums" | "sm\:not-italic" | "sm\:not-sr-only" | "sm\:object-bottom" | "sm\:object-center" | "sm\:object-contain" | "sm\:object-cover" | "sm\:object-fill" | "sm\:object-left" | "sm\:object-left-bottom" | "sm\:object-left-top" | "sm\:object-none" | "sm\:object-right" | "sm\:object-right-bottom" | "sm\:object-right-top" | "sm\:object-scale-down" | "sm\:object-top" | "sm\:oldstyle-nums" | "sm\:opacity-0" | "sm\:opacity-100" | "sm\:opacity-25" | "sm\:opacity-50" | "sm\:opacity-75" | "sm\:order-1" | "sm\:order-10" | "sm\:order-11" | "sm\:order-12" | "sm\:order-2" | "sm\:order-3" | "sm\:order-4" | "sm\:order-5" | "sm\:order-6" | "sm\:order-7" | "sm\:order-8" | "sm\:order-9" | "sm\:order-first" | "sm\:order-last" | "sm\:order-none" | "sm\:ordinal-nums" | "sm\:outline-none" | "sm\:overflow-auto" | "sm\:overflow-hidden" | "sm\:overflow-scroll" | "sm\:overflow-visible" | "sm\:overflow-x-auto" | "sm\:overflow-x-hidden" | "sm\:overflow-x-scroll" | "sm\:overflow-x-visible" | "sm\:overflow-y-auto" | "sm\:overflow-y-hidden" | "sm\:overflow-y-scroll" | "sm\:overflow-y-visible" | "sm\:p-0" | "sm\:p-1" | "sm\:p-10" | "sm\:p-12" | "sm\:p-16" | "sm\:p-2" | "sm\:p-20" | "sm\:p-24" | "sm\:p-3" | "sm\:p-32" | "sm\:p-4" | "sm\:p-40" | "sm\:p-48" | "sm\:p-5" | "sm\:p-56" | "sm\:p-6" | "sm\:p-64" | "sm\:p-8" | "sm\:p-px" | "sm\:pb-0" | "sm\:pb-1" | "sm\:pb-10" | "sm\:pb-12" | "sm\:pb-16" | "sm\:pb-2" | "sm\:pb-20" | "sm\:pb-24" | "sm\:pb-3" | "sm\:pb-32" | "sm\:pb-4" | "sm\:pb-40" | "sm\:pb-48" | "sm\:pb-5" | "sm\:pb-56" | "sm\:pb-6" | "sm\:pb-64" | "sm\:pb-8" | "sm\:pb-px" | "sm\:petite-caps" | "sm\:pl-0" | "sm\:pl-1" | "sm\:pl-10" | "sm\:pl-12" | "sm\:pl-16" | "sm\:pl-2" | "sm\:pl-20" | "sm\:pl-24" | "sm\:pl-3" | "sm\:pl-32" | "sm\:pl-4" | "sm\:pl-40" | "sm\:pl-48" | "sm\:pl-5" | "sm\:pl-56" | "sm\:pl-6" | "sm\:pl-64" | "sm\:pl-8" | "sm\:pl-px" | "sm\:placeholder-black" | "sm\:placeholder-blue-100" | "sm\:placeholder-blue-200" | "sm\:placeholder-blue-300" | "sm\:placeholder-blue-400" | "sm\:placeholder-blue-500" | "sm\:placeholder-blue-600" | "sm\:placeholder-blue-700" | "sm\:placeholder-blue-800" | "sm\:placeholder-blue-900" | "sm\:placeholder-border" | "sm\:placeholder-gray-100" | "sm\:placeholder-gray-200" | "sm\:placeholder-gray-300" | "sm\:placeholder-gray-400" | "sm\:placeholder-gray-500" | "sm\:placeholder-gray-600" | "sm\:placeholder-gray-700" | "sm\:placeholder-gray-800" | "sm\:placeholder-gray-900" | "sm\:placeholder-primary" | "sm\:placeholder-white" | "sm\:pointer-events-auto" | "sm\:pointer-events-none" | "sm\:pr-0" | "sm\:pr-1" | "sm\:pr-10" | "sm\:pr-12" | "sm\:pr-16" | "sm\:pr-2" | "sm\:pr-20" | "sm\:pr-24" | "sm\:pr-3" | "sm\:pr-32" | "sm\:pr-4" | "sm\:pr-40" | "sm\:pr-48" | "sm\:pr-5" | "sm\:pr-56" | "sm\:pr-6" | "sm\:pr-64" | "sm\:pr-8" | "sm\:pr-px" | "sm\:proportional-nums" | "sm\:pt-0" | "sm\:pt-1" | "sm\:pt-10" | "sm\:pt-12" | "sm\:pt-16" | "sm\:pt-2" | "sm\:pt-20" | "sm\:pt-24" | "sm\:pt-3" | "sm\:pt-32" | "sm\:pt-4" | "sm\:pt-40" | "sm\:pt-48" | "sm\:pt-5" | "sm\:pt-56" | "sm\:pt-6" | "sm\:pt-64" | "sm\:pt-8" | "sm\:pt-px" | "sm\:px-0" | "sm\:px-1" | "sm\:px-10" | "sm\:px-12" | "sm\:px-16" | "sm\:px-2" | "sm\:px-20" | "sm\:px-24" | "sm\:px-3" | "sm\:px-32" | "sm\:px-4" | "sm\:px-40" | "sm\:px-48" | "sm\:px-5" | "sm\:px-56" | "sm\:px-6" | "sm\:px-64" | "sm\:px-8" | "sm\:px-px" | "sm\:py-0" | "sm\:py-1" | "sm\:py-10" | "sm\:py-12" | "sm\:py-16" | "sm\:py-2" | "sm\:py-20" | "sm\:py-24" | "sm\:py-3" | "sm\:py-32" | "sm\:py-4" | "sm\:py-40" | "sm\:py-48" | "sm\:py-5" | "sm\:py-56" | "sm\:py-6" | "sm\:py-64" | "sm\:py-8" | "sm\:py-px" | "sm\:relative" | "sm\:resize" | "sm\:resize-none" | "sm\:resize-x" | "sm\:resize-y" | "sm\:right-0" | "sm\:right-auto" | "sm\:rotate-180" | "sm\:rotate-270" | "sm\:rotate-90" | "sm\:rounded" | "sm\:rounded-b" | "sm\:rounded-b-full" | "sm\:rounded-b-lg" | "sm\:rounded-b-none" | "sm\:rounded-b-sm" | "sm\:rounded-bl" | "sm\:rounded-bl-full" | "sm\:rounded-bl-lg" | "sm\:rounded-bl-none" | "sm\:rounded-bl-sm" | "sm\:rounded-br" | "sm\:rounded-br-full" | "sm\:rounded-br-lg" | "sm\:rounded-br-none" | "sm\:rounded-br-sm" | "sm\:rounded-full" | "sm\:rounded-l" | "sm\:rounded-l-full" | "sm\:rounded-l-lg" | "sm\:rounded-l-none" | "sm\:rounded-l-sm" | "sm\:rounded-lg" | "sm\:rounded-none" | "sm\:rounded-r" | "sm\:rounded-r-full" | "sm\:rounded-r-lg" | "sm\:rounded-r-none" | "sm\:rounded-r-sm" | "sm\:rounded-sm" | "sm\:rounded-t" | "sm\:rounded-t-full" | "sm\:rounded-t-lg" | "sm\:rounded-t-none" | "sm\:rounded-t-sm" | "sm\:rounded-tl" | "sm\:rounded-tl-full" | "sm\:rounded-tl-lg" | "sm\:rounded-tl-none" | "sm\:rounded-tl-sm" | "sm\:rounded-tr" | "sm\:rounded-tr-full" | "sm\:rounded-tr-lg" | "sm\:rounded-tr-none" | "sm\:rounded-tr-sm" | "sm\:scrolling-auto" | "sm\:scrolling-touch" | "sm\:select-all" | "sm\:select-auto" | "sm\:select-none" | "sm\:select-text" | "sm\:self-auto" | "sm\:self-center" | "sm\:self-end" | "sm\:self-start" | "sm\:self-stretch" | "sm\:shadow" | "sm\:shadow-2xl" | "sm\:shadow-inner" | "sm\:shadow-lg" | "sm\:shadow-md" | "sm\:shadow-none" | "sm\:shadow-outline" | "sm\:shadow-xl" | "sm\:slashed-zeros" | "sm\:small-caps" | "sm\:sr-only" | "sm\:stacked-fractions" | "sm\:static" | "sm\:sticky" | "sm\:stroke-current" | "sm\:subpixel-antialiased" | "sm\:table" | "sm\:table-auto" | "sm\:table-cell" | "sm\:table-fixed" | "sm\:table-row" | "sm\:tabular-nums" | "sm\:text-2xl" | "sm\:text-3xl" | "sm\:text-4xl" | "sm\:text-5xl" | "sm\:text-6xl" | "sm\:text-align-unset" | "sm\:text-base" | "sm\:text-black" | "sm\:text-blue-100" | "sm\:text-blue-200" | "sm\:text-blue-300" | "sm\:text-blue-400" | "sm\:text-blue-500" | "sm\:text-blue-600" | "sm\:text-blue-700" | "sm\:text-blue-800" | "sm\:text-blue-900" | "sm\:text-border" | "sm\:text-center" | "sm\:text-color-unset" | "sm\:text-gray-100" | "sm\:text-gray-200" | "sm\:text-gray-300" | "sm\:text-gray-400" | "sm\:text-gray-500" | "sm\:text-gray-600" | "sm\:text-gray-700" | "sm\:text-gray-800" | "sm\:text-gray-900" | "sm\:text-justify" | "sm\:text-left" | "sm\:text-lg" | "sm\:text-primary" | "sm\:text-right" | "sm\:text-size-unset" | "sm\:text-sm" | "sm\:text-transform-unset" | "sm\:text-white" | "sm\:text-xl" | "sm\:text-xs" | "sm\:titling-caps" | "sm\:top-0" | "sm\:top-auto" | "sm\:tracking-normal" | "sm\:tracking-tight" | "sm\:tracking-tighter" | "sm\:tracking-unset" | "sm\:tracking-wide" | "sm\:tracking-wider" | "sm\:tracking-widest" | "sm\:transform-b" | "sm\:transform-bl" | "sm\:transform-border" | "sm\:transform-br" | "sm\:transform-fill" | "sm\:transform-l" | "sm\:transform-none" | "sm\:transform-r" | "sm\:transform-t" | "sm\:transform-tl" | "sm\:transform-tr" | "sm\:transform-view" | "sm\:transition-0" | "sm\:transition-100" | "sm\:transition-1000" | "sm\:transition-250" | "sm\:transition-500" | "sm\:transition-750" | "sm\:transition-all" | "sm\:transition-bg" | "sm\:transition-border" | "sm\:transition-color" | "sm\:transition-colors" | "sm\:transition-delay-0" | "sm\:transition-delay-100" | "sm\:transition-delay-1000" | "sm\:transition-delay-250" | "sm\:transition-delay-500" | "sm\:transition-delay-750" | "sm\:transition-ease" | "sm\:transition-ease-in" | "sm\:transition-ease-in-out" | "sm\:transition-ease-out" | "sm\:transition-linear" | "sm\:transition-none" | "sm\:transition-opacity" | "sm\:transition-transform" | "sm\:truncate" | "sm\:underline" | "sm\:unicase" | "sm\:uppercase" | "sm\:visible" | "sm\:w-0" | "sm\:w-1" | "sm\:w-10" | "sm\:w-10\/12" | "sm\:w-11\/12" | "sm\:w-12" | "sm\:w-16" | "sm\:w-1\/12" | "sm\:w-1\/2" | "sm\:w-1\/3" | "sm\:w-1\/4" | "sm\:w-1\/5" | "sm\:w-1\/6" | "sm\:w-2" | "sm\:w-20" | "sm\:w-24" | "sm\:w-2\/12" | "sm\:w-2\/3" | "sm\:w-2\/4" | "sm\:w-2\/5" | "sm\:w-2\/6" | "sm\:w-3" | "sm\:w-32" | "sm\:w-3\/12" | "sm\:w-3\/4" | "sm\:w-3\/5" | "sm\:w-3\/6" | "sm\:w-4" | "sm\:w-40" | "sm\:w-48" | "sm\:w-4\/12" | "sm\:w-4\/5" | "sm\:w-4\/6" | "sm\:w-5" | "sm\:w-56" | "sm\:w-5\/12" | "sm\:w-5\/6" | "sm\:w-6" | "sm\:w-64" | "sm\:w-6\/12" | "sm\:w-7\/12" | "sm\:w-8" | "sm\:w-8\/12" | "sm\:w-9\/12" | "sm\:w-auto" | "sm\:w-full" | "sm\:w-px" | "sm\:w-screen" | "sm\:whitespace-no-wrap" | "sm\:whitespace-normal" | "sm\:whitespace-pre" | "sm\:whitespace-pre-line" | "sm\:whitespace-pre-wrap" | "sm\:will-change-auto" | "sm\:will-change-contents" | "sm\:will-change-opacity" | "sm\:will-change-scroll" | "sm\:will-change-transform" | "sm\:z-0" | "sm\:z-10" | "sm\:z-20" | "sm\:z-30" | "sm\:z-40" | "sm\:z-50" | "sm\:z-auto" | "small-caps" | "sr-only" | "stacked-fractions" | "static" | "sticky" | "stroke-current" | "subpixel-antialiased" | "table" | "table-auto" | "table-cell" | "table-fixed" | "table-row" | "tabular-nums" | "text-2xl" | "text-3xl" | "text-4xl" | "text-5xl" | "text-6xl" | "text-align-unset" | "text-base" | "text-black" | "text-blue-100" | "text-blue-200" | "text-blue-300" | "text-blue-400" | "text-blue-500" | "text-blue-600" | "text-blue-700" | "text-blue-800" | "text-blue-900" | "text-border" | "text-center" | "text-color-unset" | "text-gray-100" | "text-gray-200" | "text-gray-300" | "text-gray-400" | "text-gray-500" | "text-gray-600" | "text-gray-700" | "text-gray-800" | "text-gray-900" | "text-justify" | "text-left" | "text-lg" | "text-primary" | "text-right" | "text-size-unset" | "text-sm" | "text-transform-unset" | "text-white" | "text-xl" | "text-xs" | "titling-caps" | "top-0" | "top-auto" | "tracking-normal" | "tracking-tight" | "tracking-tighter" | "tracking-unset" | "tracking-wide" | "tracking-wider" | "tracking-widest" | "transform-b" | "transform-bl" | "transform-border" | "transform-br" | "transform-fill" | "transform-l" | "transform-none" | "transform-r" | "transform-t" | "transform-tl" | "transform-tr" | "transform-view" | "transition-0" | "transition-100" | "transition-1000" | "transition-250" | "transition-500" | "transition-750" | "transition-all" | "transition-bg" | "transition-border" | "transition-color" | "transition-colors" | "transition-delay-0" | "transition-delay-100" | "transition-delay-1000" | "transition-delay-250" | "transition-delay-500" | "transition-delay-750" | "transition-ease" | "transition-ease-in" | "transition-ease-in-out" | "transition-ease-out" | "transition-linear" | "transition-none" | "transition-opacity" | "transition-transform" | "truncate" | "underline" | "unicase" | "uppercase" | "visible" | "w-0" | "w-1" | "w-10" | "w-10\/12" | "w-11\/12" | "w-12" | "w-16" | "w-1\/12" | "w-1\/2" | "w-1\/3" | "w-1\/4" | "w-1\/5" | "w-1\/6" | "w-2" | "w-20" | "w-24" | "w-2\/12" | "w-2\/3" | "w-2\/4" | "w-2\/5" | "w-2\/6" | "w-3" | "w-32" | "w-3\/12" | "w-3\/4" | "w-3\/5" | "w-3\/6" | "w-4" | "w-40" | "w-48" | "w-4\/12" | "w-4\/5" | "w-4\/6" | "w-5" | "w-56" | "w-5\/12" | "w-5\/6" | "w-6" | "w-64" | "w-6\/12" | "w-7\/12" | "w-8" | "w-8\/12" | "w-9\/12" | "w-auto" | "w-full" | "w-px" | "w-screen" | "whitespace-no-wrap" | "whitespace-normal" | "whitespace-pre" | "whitespace-pre-line" | "whitespace-pre-wrap" | "will-change-auto" | "will-change-contents" | "will-change-opacity" | "will-change-scroll" | "will-change-transform" | "xl\:-m-1" | "xl\:-m-10" | "xl\:-m-12" | "xl\:-m-16" | "xl\:-m-2" | "xl\:-m-20" | "xl\:-m-24" | "xl\:-m-3" | "xl\:-m-32" | "xl\:-m-4" | "xl\:-m-40" | "xl\:-m-48" | "xl\:-m-5" | "xl\:-m-56" | "xl\:-m-6" | "xl\:-m-64" | "xl\:-m-8" | "xl\:-m-px" | "xl\:-mb-1" | "xl\:-mb-10" | "xl\:-mb-12" | "xl\:-mb-16" | "xl\:-mb-2" | "xl\:-mb-20" | "xl\:-mb-24" | "xl\:-mb-3" | "xl\:-mb-32" | "xl\:-mb-4" | "xl\:-mb-40" | "xl\:-mb-48" | "xl\:-mb-5" | "xl\:-mb-56" | "xl\:-mb-6" | "xl\:-mb-64" | "xl\:-mb-8" | "xl\:-mb-px" | "xl\:-ml-1" | "xl\:-ml-10" | "xl\:-ml-12" | "xl\:-ml-16" | "xl\:-ml-2" | "xl\:-ml-20" | "xl\:-ml-24" | "xl\:-ml-3" | "xl\:-ml-32" | "xl\:-ml-4" | "xl\:-ml-40" | "xl\:-ml-48" | "xl\:-ml-5" | "xl\:-ml-56" | "xl\:-ml-6" | "xl\:-ml-64" | "xl\:-ml-8" | "xl\:-ml-px" | "xl\:-mr-1" | "xl\:-mr-10" | "xl\:-mr-12" | "xl\:-mr-16" | "xl\:-mr-2" | "xl\:-mr-20" | "xl\:-mr-24" | "xl\:-mr-3" | "xl\:-mr-32" | "xl\:-mr-4" | "xl\:-mr-40" | "xl\:-mr-48" | "xl\:-mr-5" | "xl\:-mr-56" | "xl\:-mr-6" | "xl\:-mr-64" | "xl\:-mr-8" | "xl\:-mr-px" | "xl\:-mt-1" | "xl\:-mt-10" | "xl\:-mt-12" | "xl\:-mt-16" | "xl\:-mt-2" | "xl\:-mt-20" | "xl\:-mt-24" | "xl\:-mt-3" | "xl\:-mt-32" | "xl\:-mt-4" | "xl\:-mt-40" | "xl\:-mt-48" | "xl\:-mt-5" | "xl\:-mt-56" | "xl\:-mt-6" | "xl\:-mt-64" | "xl\:-mt-8" | "xl\:-mt-px" | "xl\:-mx-1" | "xl\:-mx-10" | "xl\:-mx-12" | "xl\:-mx-16" | "xl\:-mx-2" | "xl\:-mx-20" | "xl\:-mx-24" | "xl\:-mx-3" | "xl\:-mx-32" | "xl\:-mx-4" | "xl\:-mx-40" | "xl\:-mx-48" | "xl\:-mx-5" | "xl\:-mx-56" | "xl\:-mx-6" | "xl\:-mx-64" | "xl\:-mx-8" | "xl\:-mx-px" | "xl\:-my-1" | "xl\:-my-10" | "xl\:-my-12" | "xl\:-my-16" | "xl\:-my-2" | "xl\:-my-20" | "xl\:-my-24" | "xl\:-my-3" | "xl\:-my-32" | "xl\:-my-4" | "xl\:-my-40" | "xl\:-my-48" | "xl\:-my-5" | "xl\:-my-56" | "xl\:-my-6" | "xl\:-my-64" | "xl\:-my-8" | "xl\:-my-px" | "xl\:absolute" | "xl\:align-baseline" | "xl\:align-bottom" | "xl\:align-middle" | "xl\:align-text-bottom" | "xl\:align-text-top" | "xl\:align-top" | "xl\:all-small-caps" | "xl\:antialiased" | "xl\:appearance-none" | "xl\:bg-auto" | "xl\:bg-black" | "xl\:bg-blue-100" | "xl\:bg-blue-200" | "xl\:bg-blue-300" | "xl\:bg-blue-400" | "xl\:bg-blue-500" | "xl\:bg-blue-600" | "xl\:bg-blue-700" | "xl\:bg-blue-800" | "xl\:bg-blue-900" | "xl\:bg-border" | "xl\:bg-bottom" | "xl\:bg-center" | "xl\:bg-contain" | "xl\:bg-cover" | "xl\:bg-fixed" | "xl\:bg-gray-100" | "xl\:bg-gray-200" | "xl\:bg-gray-300" | "xl\:bg-gray-400" | "xl\:bg-gray-500" | "xl\:bg-gray-600" | "xl\:bg-gray-700" | "xl\:bg-gray-800" | "xl\:bg-gray-900" | "xl\:bg-left" | "xl\:bg-left-bottom" | "xl\:bg-left-top" | "xl\:bg-local" | "xl\:bg-no-repeat" | "xl\:bg-primary" | "xl\:bg-repeat" | "xl\:bg-repeat-round" | "xl\:bg-repeat-space" | "xl\:bg-repeat-x" | "xl\:bg-repeat-y" | "xl\:bg-right" | "xl\:bg-right-bottom" | "xl\:bg-right-top" | "xl\:bg-scroll" | "xl\:bg-top" | "xl\:bg-white" | "xl\:block" | "xl\:border" | "xl\:border-0" | "xl\:border-2" | "xl\:border-4" | "xl\:border-8" | "xl\:border-b" | "xl\:border-b-0" | "xl\:border-b-2" | "xl\:border-b-4" | "xl\:border-b-8" | "xl\:border-black" | "xl\:border-blue-100" | "xl\:border-blue-200" | "xl\:border-blue-300" | "xl\:border-blue-400" | "xl\:border-blue-500" | "xl\:border-blue-600" | "xl\:border-blue-700" | "xl\:border-blue-800" | "xl\:border-blue-900" | "xl\:border-border" | "xl\:border-collapse" | "xl\:border-dashed" | "xl\:border-dotted" | "xl\:border-double" | "xl\:border-gray-100" | "xl\:border-gray-200" | "xl\:border-gray-300" | "xl\:border-gray-400" | "xl\:border-gray-500" | "xl\:border-gray-600" | "xl\:border-gray-700" | "xl\:border-gray-800" | "xl\:border-gray-900" | "xl\:border-l" | "xl\:border-l-0" | "xl\:border-l-2" | "xl\:border-l-4" | "xl\:border-l-8" | "xl\:border-none" | "xl\:border-primary" | "xl\:border-r" | "xl\:border-r-0" | "xl\:border-r-2" | "xl\:border-r-4" | "xl\:border-r-8" | "xl\:border-separate" | "xl\:border-solid" | "xl\:border-t" | "xl\:border-t-0" | "xl\:border-t-2" | "xl\:border-t-4" | "xl\:border-t-8" | "xl\:border-white" | "xl\:bottom-0" | "xl\:bottom-auto" | "xl\:break-all" | "xl\:break-normal" | "xl\:break-words" | "xl\:capitalize" | "xl\:clearfix" | "xl\:common-ligatures" | "xl\:content-around" | "xl\:content-between" | "xl\:content-center" | "xl\:content-end" | "xl\:content-start" | "xl\:contextual-ligatures" | "xl\:cursor-auto" | "xl\:cursor-default" | "xl\:cursor-move" | "xl\:cursor-not-allowed" | "xl\:cursor-pointer" | "xl\:cursor-text" | "xl\:cursor-wait" | "xl\:diagonal-fractions" | "xl\:discretionary-ligatures" | "xl\:ellipsis" | "xl\:fill-current" | "xl\:fixed" | "xl\:flex" | "xl\:flex-1" | "xl\:flex-auto" | "xl\:flex-col" | "xl\:flex-col-reverse" | "xl\:flex-grow" | "xl\:flex-grow-0" | "xl\:flex-initial" | "xl\:flex-no-wrap" | "xl\:flex-none" | "xl\:flex-row" | "xl\:flex-row-reverse" | "xl\:flex-shrink" | "xl\:flex-shrink-0" | "xl\:flex-wrap" | "xl\:flex-wrap-reverse" | "xl\:float-left" | "xl\:float-none" | "xl\:float-right" | "xl\:focus\:bg-black" | "xl\:focus\:bg-blue-100" | "xl\:focus\:bg-blue-200" | "xl\:focus\:bg-blue-300" | "xl\:focus\:bg-blue-400" | "xl\:focus\:bg-blue-500" | "xl\:focus\:bg-blue-600" | "xl\:focus\:bg-blue-700" | "xl\:focus\:bg-blue-800" | "xl\:focus\:bg-blue-900" | "xl\:focus\:bg-border" | "xl\:focus\:bg-gray-100" | "xl\:focus\:bg-gray-200" | "xl\:focus\:bg-gray-300" | "xl\:focus\:bg-gray-400" | "xl\:focus\:bg-gray-500" | "xl\:focus\:bg-gray-600" | "xl\:focus\:bg-gray-700" | "xl\:focus\:bg-gray-800" | "xl\:focus\:bg-gray-900" | "xl\:focus\:bg-primary" | "xl\:focus\:bg-white" | "xl\:focus\:border" | "xl\:focus\:border-0" | "xl\:focus\:border-2" | "xl\:focus\:border-4" | "xl\:focus\:border-8" | "xl\:focus\:border-b" | "xl\:focus\:border-b-0" | "xl\:focus\:border-b-2" | "xl\:focus\:border-b-4" | "xl\:focus\:border-b-8" | "xl\:focus\:border-black" | "xl\:focus\:border-blue-100" | "xl\:focus\:border-blue-200" | "xl\:focus\:border-blue-300" | "xl\:focus\:border-blue-400" | "xl\:focus\:border-blue-500" | "xl\:focus\:border-blue-600" | "xl\:focus\:border-blue-700" | "xl\:focus\:border-blue-800" | "xl\:focus\:border-blue-900" | "xl\:focus\:border-border" | "xl\:focus\:border-gray-100" | "xl\:focus\:border-gray-200" | "xl\:focus\:border-gray-300" | "xl\:focus\:border-gray-400" | "xl\:focus\:border-gray-500" | "xl\:focus\:border-gray-600" | "xl\:focus\:border-gray-700" | "xl\:focus\:border-gray-800" | "xl\:focus\:border-gray-900" | "xl\:focus\:border-l" | "xl\:focus\:border-l-0" | "xl\:focus\:border-l-2" | "xl\:focus\:border-l-4" | "xl\:focus\:border-l-8" | "xl\:focus\:border-primary" | "xl\:focus\:border-r" | "xl\:focus\:border-r-0" | "xl\:focus\:border-r-2" | "xl\:focus\:border-r-4" | "xl\:focus\:border-r-8" | "xl\:focus\:border-t" | "xl\:focus\:border-t-0" | "xl\:focus\:border-t-2" | "xl\:focus\:border-t-4" | "xl\:focus\:border-t-8" | "xl\:focus\:border-white" | "xl\:focus\:font-black" | "xl\:focus\:font-bold" | "xl\:focus\:font-extrabold" | "xl\:focus\:font-hairline" | "xl\:focus\:font-light" | "xl\:focus\:font-medium" | "xl\:focus\:font-normal" | "xl\:focus\:font-semibold" | "xl\:focus\:font-thin" | "xl\:focus\:line-through" | "xl\:focus\:no-underline" | "xl\:focus\:not-sr-only" | "xl\:focus\:opacity-0" | "xl\:focus\:opacity-100" | "xl\:focus\:opacity-25" | "xl\:focus\:opacity-50" | "xl\:focus\:opacity-75" | "xl\:focus\:outline-none" | "xl\:focus\:placeholder-black" | "xl\:focus\:placeholder-blue-100" | "xl\:focus\:placeholder-blue-200" | "xl\:focus\:placeholder-blue-300" | "xl\:focus\:placeholder-blue-400" | "xl\:focus\:placeholder-blue-500" | "xl\:focus\:placeholder-blue-600" | "xl\:focus\:placeholder-blue-700" | "xl\:focus\:placeholder-blue-800" | "xl\:focus\:placeholder-blue-900" | "xl\:focus\:placeholder-border" | "xl\:focus\:placeholder-gray-100" | "xl\:focus\:placeholder-gray-200" | "xl\:focus\:placeholder-gray-300" | "xl\:focus\:placeholder-gray-400" | "xl\:focus\:placeholder-gray-500" | "xl\:focus\:placeholder-gray-600" | "xl\:focus\:placeholder-gray-700" | "xl\:focus\:placeholder-gray-800" | "xl\:focus\:placeholder-gray-900" | "xl\:focus\:placeholder-primary" | "xl\:focus\:placeholder-white" | "xl\:focus\:shadow" | "xl\:focus\:shadow-2xl" | "xl\:focus\:shadow-inner" | "xl\:focus\:shadow-lg" | "xl\:focus\:shadow-md" | "xl\:focus\:shadow-none" | "xl\:focus\:shadow-outline" | "xl\:focus\:shadow-xl" | "xl\:focus\:sr-only" | "xl\:focus\:text-black" | "xl\:focus\:text-blue-100" | "xl\:focus\:text-blue-200" | "xl\:focus\:text-blue-300" | "xl\:focus\:text-blue-400" | "xl\:focus\:text-blue-500" | "xl\:focus\:text-blue-600" | "xl\:focus\:text-blue-700" | "xl\:focus\:text-blue-800" | "xl\:focus\:text-blue-900" | "xl\:focus\:text-border" | "xl\:focus\:text-gray-100" | "xl\:focus\:text-gray-200" | "xl\:focus\:text-gray-300" | "xl\:focus\:text-gray-400" | "xl\:focus\:text-gray-500" | "xl\:focus\:text-gray-600" | "xl\:focus\:text-gray-700" | "xl\:focus\:text-gray-800" | "xl\:focus\:text-gray-900" | "xl\:focus\:text-primary" | "xl\:focus\:text-white" | "xl\:focus\:underline" | "xl\:font-black" | "xl\:font-bold" | "xl\:font-extrabold" | "xl\:font-family-unset" | "xl\:font-hairline" | "xl\:font-light" | "xl\:font-medium" | "xl\:font-mono" | "xl\:font-normal" | "xl\:font-sans" | "xl\:font-semibold" | "xl\:font-serif" | "xl\:font-style-unset" | "xl\:font-thin" | "xl\:font-weight-unset" | "xl\:h-0" | "xl\:h-1" | "xl\:h-10" | "xl\:h-12" | "xl\:h-16" | "xl\:h-2" | "xl\:h-20" | "xl\:h-24" | "xl\:h-3" | "xl\:h-32" | "xl\:h-4" | "xl\:h-40" | "xl\:h-48" | "xl\:h-5" | "xl\:h-56" | "xl\:h-6" | "xl\:h-64" | "xl\:h-8" | "xl\:h-auto" | "xl\:h-full" | "xl\:h-px" | "xl\:h-screen" | "xl\:hidden" | "xl\:historical-ligatures" | "xl\:hover\:bg-black" | "xl\:hover\:bg-blue-100" | "xl\:hover\:bg-blue-200" | "xl\:hover\:bg-blue-300" | "xl\:hover\:bg-blue-400" | "xl\:hover\:bg-blue-500" | "xl\:hover\:bg-blue-600" | "xl\:hover\:bg-blue-700" | "xl\:hover\:bg-blue-800" | "xl\:hover\:bg-blue-900" | "xl\:hover\:bg-border" | "xl\:hover\:bg-gray-100" | "xl\:hover\:bg-gray-200" | "xl\:hover\:bg-gray-300" | "xl\:hover\:bg-gray-400" | "xl\:hover\:bg-gray-500" | "xl\:hover\:bg-gray-600" | "xl\:hover\:bg-gray-700" | "xl\:hover\:bg-gray-800" | "xl\:hover\:bg-gray-900" | "xl\:hover\:bg-primary" | "xl\:hover\:bg-white" | "xl\:hover\:border" | "xl\:hover\:border-0" | "xl\:hover\:border-2" | "xl\:hover\:border-4" | "xl\:hover\:border-8" | "xl\:hover\:border-b" | "xl\:hover\:border-b-0" | "xl\:hover\:border-b-2" | "xl\:hover\:border-b-4" | "xl\:hover\:border-b-8" | "xl\:hover\:border-black" | "xl\:hover\:border-blue-100" | "xl\:hover\:border-blue-200" | "xl\:hover\:border-blue-300" | "xl\:hover\:border-blue-400" | "xl\:hover\:border-blue-500" | "xl\:hover\:border-blue-600" | "xl\:hover\:border-blue-700" | "xl\:hover\:border-blue-800" | "xl\:hover\:border-blue-900" | "xl\:hover\:border-border" | "xl\:hover\:border-gray-100" | "xl\:hover\:border-gray-200" | "xl\:hover\:border-gray-300" | "xl\:hover\:border-gray-400" | "xl\:hover\:border-gray-500" | "xl\:hover\:border-gray-600" | "xl\:hover\:border-gray-700" | "xl\:hover\:border-gray-800" | "xl\:hover\:border-gray-900" | "xl\:hover\:border-l" | "xl\:hover\:border-l-0" | "xl\:hover\:border-l-2" | "xl\:hover\:border-l-4" | "xl\:hover\:border-l-8" | "xl\:hover\:border-primary" | "xl\:hover\:border-r" | "xl\:hover\:border-r-0" | "xl\:hover\:border-r-2" | "xl\:hover\:border-r-4" | "xl\:hover\:border-r-8" | "xl\:hover\:border-t" | "xl\:hover\:border-t-0" | "xl\:hover\:border-t-2" | "xl\:hover\:border-t-4" | "xl\:hover\:border-t-8" | "xl\:hover\:border-white" | "xl\:hover\:font-black" | "xl\:hover\:font-bold" | "xl\:hover\:font-extrabold" | "xl\:hover\:font-hairline" | "xl\:hover\:font-light" | "xl\:hover\:font-medium" | "xl\:hover\:font-normal" | "xl\:hover\:font-semibold" | "xl\:hover\:font-thin" | "xl\:hover\:line-through" | "xl\:hover\:no-underline" | "xl\:hover\:opacity-0" | "xl\:hover\:opacity-100" | "xl\:hover\:opacity-25" | "xl\:hover\:opacity-50" | "xl\:hover\:opacity-75" | "xl\:hover\:shadow" | "xl\:hover\:shadow-2xl" | "xl\:hover\:shadow-inner" | "xl\:hover\:shadow-lg" | "xl\:hover\:shadow-md" | "xl\:hover\:shadow-none" | "xl\:hover\:shadow-outline" | "xl\:hover\:shadow-xl" | "xl\:hover\:text-black" | "xl\:hover\:text-blue-100" | "xl\:hover\:text-blue-200" | "xl\:hover\:text-blue-300" | "xl\:hover\:text-blue-400" | "xl\:hover\:text-blue-500" | "xl\:hover\:text-blue-600" | "xl\:hover\:text-blue-700" | "xl\:hover\:text-blue-800" | "xl\:hover\:text-blue-900" | "xl\:hover\:text-border" | "xl\:hover\:text-gray-100" | "xl\:hover\:text-gray-200" | "xl\:hover\:text-gray-300" | "xl\:hover\:text-gray-400" | "xl\:hover\:text-gray-500" | "xl\:hover\:text-gray-600" | "xl\:hover\:text-gray-700" | "xl\:hover\:text-gray-800" | "xl\:hover\:text-gray-900" | "xl\:hover\:text-primary" | "xl\:hover\:text-white" | "xl\:hover\:underline" | "xl\:hyphens-auto" | "xl\:hyphens-manual" | "xl\:hyphens-none" | "xl\:inline" | "xl\:inline-block" | "xl\:inline-flex" | "xl\:inset-0" | "xl\:inset-auto" | "xl\:inset-x-0" | "xl\:inset-x-auto" | "xl\:inset-y-0" | "xl\:inset-y-auto" | "xl\:invisible" | "xl\:italic" | "xl\:items-baseline" | "xl\:items-center" | "xl\:items-end" | "xl\:items-start" | "xl\:items-stretch" | "xl\:justify-around" | "xl\:justify-between" | "xl\:justify-center" | "xl\:justify-end" | "xl\:justify-start" | "xl\:last\:border" | "xl\:last\:border-0" | "xl\:last\:border-2" | "xl\:last\:border-4" | "xl\:last\:border-8" | "xl\:last\:border-b" | "xl\:last\:border-b-0" | "xl\:last\:border-b-2" | "xl\:last\:border-b-4" | "xl\:last\:border-b-8" | "xl\:last\:border-l" | "xl\:last\:border-l-0" | "xl\:last\:border-l-2" | "xl\:last\:border-l-4" | "xl\:last\:border-l-8" | "xl\:last\:border-r" | "xl\:last\:border-r-0" | "xl\:last\:border-r-2" | "xl\:last\:border-r-4" | "xl\:last\:border-r-8" | "xl\:last\:border-t" | "xl\:last\:border-t-0" | "xl\:last\:border-t-2" | "xl\:last\:border-t-4" | "xl\:last\:border-t-8" | "xl\:leading-loose" | "xl\:leading-none" | "xl\:leading-normal" | "xl\:leading-relaxed" | "xl\:leading-snug" | "xl\:leading-tight" | "xl\:leading-unset" | "xl\:left-0" | "xl\:left-auto" | "xl\:line-through" | "xl\:lining-nums" | "xl\:list-decimal" | "xl\:list-disc" | "xl\:list-inside" | "xl\:list-none" | "xl\:list-outside" | "xl\:lowercase" | "xl\:m-0" | "xl\:m-1" | "xl\:m-10" | "xl\:m-12" | "xl\:m-16" | "xl\:m-2" | "xl\:m-20" | "xl\:m-24" | "xl\:m-3" | "xl\:m-32" | "xl\:m-4" | "xl\:m-40" | "xl\:m-48" | "xl\:m-5" | "xl\:m-56" | "xl\:m-6" | "xl\:m-64" | "xl\:m-8" | "xl\:m-auto" | "xl\:m-px" | "xl\:max-h-full" | "xl\:max-h-screen" | "xl\:max-w-2xl" | "xl\:max-w-3xl" | "xl\:max-w-4xl" | "xl\:max-w-5xl" | "xl\:max-w-6xl" | "xl\:max-w-full" | "xl\:max-w-lg" | "xl\:max-w-md" | "xl\:max-w-sm" | "xl\:max-w-xl" | "xl\:max-w-xs" | "xl\:mb-0" | "xl\:mb-1" | "xl\:mb-10" | "xl\:mb-12" | "xl\:mb-16" | "xl\:mb-2" | "xl\:mb-20" | "xl\:mb-24" | "xl\:mb-3" | "xl\:mb-32" | "xl\:mb-4" | "xl\:mb-40" | "xl\:mb-48" | "xl\:mb-5" | "xl\:mb-56" | "xl\:mb-6" | "xl\:mb-64" | "xl\:mb-8" | "xl\:mb-auto" | "xl\:mb-px" | "xl\:min-h-0" | "xl\:min-h-full" | "xl\:min-h-screen" | "xl\:min-w-0" | "xl\:min-w-full" | "xl\:ml-0" | "xl\:ml-1" | "xl\:ml-10" | "xl\:ml-12" | "xl\:ml-16" | "xl\:ml-2" | "xl\:ml-20" | "xl\:ml-24" | "xl\:ml-3" | "xl\:ml-32" | "xl\:ml-4" | "xl\:ml-40" | "xl\:ml-48" | "xl\:ml-5" | "xl\:ml-56" | "xl\:ml-6" | "xl\:ml-64" | "xl\:ml-8" | "xl\:ml-auto" | "xl\:ml-px" | "xl\:mr-0" | "xl\:mr-1" | "xl\:mr-10" | "xl\:mr-12" | "xl\:mr-16" | "xl\:mr-2" | "xl\:mr-20" | "xl\:mr-24" | "xl\:mr-3" | "xl\:mr-32" | "xl\:mr-4" | "xl\:mr-40" | "xl\:mr-48" | "xl\:mr-5" | "xl\:mr-56" | "xl\:mr-6" | "xl\:mr-64" | "xl\:mr-8" | "xl\:mr-auto" | "xl\:mr-px" | "xl\:mt-0" | "xl\:mt-1" | "xl\:mt-10" | "xl\:mt-12" | "xl\:mt-16" | "xl\:mt-2" | "xl\:mt-20" | "xl\:mt-24" | "xl\:mt-3" | "xl\:mt-32" | "xl\:mt-4" | "xl\:mt-40" | "xl\:mt-48" | "xl\:mt-5" | "xl\:mt-56" | "xl\:mt-6" | "xl\:mt-64" | "xl\:mt-8" | "xl\:mt-auto" | "xl\:mt-px" | "xl\:mx-0" | "xl\:mx-1" | "xl\:mx-10" | "xl\:mx-12" | "xl\:mx-16" | "xl\:mx-2" | "xl\:mx-20" | "xl\:mx-24" | "xl\:mx-3" | "xl\:mx-32" | "xl\:mx-4" | "xl\:mx-40" | "xl\:mx-48" | "xl\:mx-5" | "xl\:mx-56" | "xl\:mx-6" | "xl\:mx-64" | "xl\:mx-8" | "xl\:mx-auto" | "xl\:mx-px" | "xl\:my-0" | "xl\:my-1" | "xl\:my-10" | "xl\:my-12" | "xl\:my-16" | "xl\:my-2" | "xl\:my-20" | "xl\:my-24" | "xl\:my-3" | "xl\:my-32" | "xl\:my-4" | "xl\:my-40" | "xl\:my-48" | "xl\:my-5" | "xl\:my-56" | "xl\:my-6" | "xl\:my-64" | "xl\:my-8" | "xl\:my-auto" | "xl\:my-px" | "xl\:no-common-ligatures" | "xl\:no-contextual-ligatures" | "xl\:no-discretionary-ligatures" | "xl\:no-ellipsis" | "xl\:no-historical-ligatures" | "xl\:no-ligatures" | "xl\:no-underline" | "xl\:normal-caps" | "xl\:normal-case" | "xl\:normal-ligatures" | "xl\:normal-nums" | "xl\:not-italic" | "xl\:not-sr-only" | "xl\:object-bottom" | "xl\:object-center" | "xl\:object-contain" | "xl\:object-cover" | "xl\:object-fill" | "xl\:object-left" | "xl\:object-left-bottom" | "xl\:object-left-top" | "xl\:object-none" | "xl\:object-right" | "xl\:object-right-bottom" | "xl\:object-right-top" | "xl\:object-scale-down" | "xl\:object-top" | "xl\:oldstyle-nums" | "xl\:opacity-0" | "xl\:opacity-100" | "xl\:opacity-25" | "xl\:opacity-50" | "xl\:opacity-75" | "xl\:order-1" | "xl\:order-10" | "xl\:order-11" | "xl\:order-12" | "xl\:order-2" | "xl\:order-3" | "xl\:order-4" | "xl\:order-5" | "xl\:order-6" | "xl\:order-7" | "xl\:order-8" | "xl\:order-9" | "xl\:order-first" | "xl\:order-last" | "xl\:order-none" | "xl\:ordinal-nums" | "xl\:outline-none" | "xl\:overflow-auto" | "xl\:overflow-hidden" | "xl\:overflow-scroll" | "xl\:overflow-visible" | "xl\:overflow-x-auto" | "xl\:overflow-x-hidden" | "xl\:overflow-x-scroll" | "xl\:overflow-x-visible" | "xl\:overflow-y-auto" | "xl\:overflow-y-hidden" | "xl\:overflow-y-scroll" | "xl\:overflow-y-visible" | "xl\:p-0" | "xl\:p-1" | "xl\:p-10" | "xl\:p-12" | "xl\:p-16" | "xl\:p-2" | "xl\:p-20" | "xl\:p-24" | "xl\:p-3" | "xl\:p-32" | "xl\:p-4" | "xl\:p-40" | "xl\:p-48" | "xl\:p-5" | "xl\:p-56" | "xl\:p-6" | "xl\:p-64" | "xl\:p-8" | "xl\:p-px" | "xl\:pb-0" | "xl\:pb-1" | "xl\:pb-10" | "xl\:pb-12" | "xl\:pb-16" | "xl\:pb-2" | "xl\:pb-20" | "xl\:pb-24" | "xl\:pb-3" | "xl\:pb-32" | "xl\:pb-4" | "xl\:pb-40" | "xl\:pb-48" | "xl\:pb-5" | "xl\:pb-56" | "xl\:pb-6" | "xl\:pb-64" | "xl\:pb-8" | "xl\:pb-px" | "xl\:petite-caps" | "xl\:pl-0" | "xl\:pl-1" | "xl\:pl-10" | "xl\:pl-12" | "xl\:pl-16" | "xl\:pl-2" | "xl\:pl-20" | "xl\:pl-24" | "xl\:pl-3" | "xl\:pl-32" | "xl\:pl-4" | "xl\:pl-40" | "xl\:pl-48" | "xl\:pl-5" | "xl\:pl-56" | "xl\:pl-6" | "xl\:pl-64" | "xl\:pl-8" | "xl\:pl-px" | "xl\:placeholder-black" | "xl\:placeholder-blue-100" | "xl\:placeholder-blue-200" | "xl\:placeholder-blue-300" | "xl\:placeholder-blue-400" | "xl\:placeholder-blue-500" | "xl\:placeholder-blue-600" | "xl\:placeholder-blue-700" | "xl\:placeholder-blue-800" | "xl\:placeholder-blue-900" | "xl\:placeholder-border" | "xl\:placeholder-gray-100" | "xl\:placeholder-gray-200" | "xl\:placeholder-gray-300" | "xl\:placeholder-gray-400" | "xl\:placeholder-gray-500" | "xl\:placeholder-gray-600" | "xl\:placeholder-gray-700" | "xl\:placeholder-gray-800" | "xl\:placeholder-gray-900" | "xl\:placeholder-primary" | "xl\:placeholder-white" | "xl\:pointer-events-auto" | "xl\:pointer-events-none" | "xl\:pr-0" | "xl\:pr-1" | "xl\:pr-10" | "xl\:pr-12" | "xl\:pr-16" | "xl\:pr-2" | "xl\:pr-20" | "xl\:pr-24" | "xl\:pr-3" | "xl\:pr-32" | "xl\:pr-4" | "xl\:pr-40" | "xl\:pr-48" | "xl\:pr-5" | "xl\:pr-56" | "xl\:pr-6" | "xl\:pr-64" | "xl\:pr-8" | "xl\:pr-px" | "xl\:proportional-nums" | "xl\:pt-0" | "xl\:pt-1" | "xl\:pt-10" | "xl\:pt-12" | "xl\:pt-16" | "xl\:pt-2" | "xl\:pt-20" | "xl\:pt-24" | "xl\:pt-3" | "xl\:pt-32" | "xl\:pt-4" | "xl\:pt-40" | "xl\:pt-48" | "xl\:pt-5" | "xl\:pt-56" | "xl\:pt-6" | "xl\:pt-64" | "xl\:pt-8" | "xl\:pt-px" | "xl\:px-0" | "xl\:px-1" | "xl\:px-10" | "xl\:px-12" | "xl\:px-16" | "xl\:px-2" | "xl\:px-20" | "xl\:px-24" | "xl\:px-3" | "xl\:px-32" | "xl\:px-4" | "xl\:px-40" | "xl\:px-48" | "xl\:px-5" | "xl\:px-56" | "xl\:px-6" | "xl\:px-64" | "xl\:px-8" | "xl\:px-px" | "xl\:py-0" | "xl\:py-1" | "xl\:py-10" | "xl\:py-12" | "xl\:py-16" | "xl\:py-2" | "xl\:py-20" | "xl\:py-24" | "xl\:py-3" | "xl\:py-32" | "xl\:py-4" | "xl\:py-40" | "xl\:py-48" | "xl\:py-5" | "xl\:py-56" | "xl\:py-6" | "xl\:py-64" | "xl\:py-8" | "xl\:py-px" | "xl\:relative" | "xl\:resize" | "xl\:resize-none" | "xl\:resize-x" | "xl\:resize-y" | "xl\:right-0" | "xl\:right-auto" | "xl\:rotate-180" | "xl\:rotate-270" | "xl\:rotate-90" | "xl\:rounded" | "xl\:rounded-b" | "xl\:rounded-b-full" | "xl\:rounded-b-lg" | "xl\:rounded-b-none" | "xl\:rounded-b-sm" | "xl\:rounded-bl" | "xl\:rounded-bl-full" | "xl\:rounded-bl-lg" | "xl\:rounded-bl-none" | "xl\:rounded-bl-sm" | "xl\:rounded-br" | "xl\:rounded-br-full" | "xl\:rounded-br-lg" | "xl\:rounded-br-none" | "xl\:rounded-br-sm" | "xl\:rounded-full" | "xl\:rounded-l" | "xl\:rounded-l-full" | "xl\:rounded-l-lg" | "xl\:rounded-l-none" | "xl\:rounded-l-sm" | "xl\:rounded-lg" | "xl\:rounded-none" | "xl\:rounded-r" | "xl\:rounded-r-full" | "xl\:rounded-r-lg" | "xl\:rounded-r-none" | "xl\:rounded-r-sm" | "xl\:rounded-sm" | "xl\:rounded-t" | "xl\:rounded-t-full" | "xl\:rounded-t-lg" | "xl\:rounded-t-none" | "xl\:rounded-t-sm" | "xl\:rounded-tl" | "xl\:rounded-tl-full" | "xl\:rounded-tl-lg" | "xl\:rounded-tl-none" | "xl\:rounded-tl-sm" | "xl\:rounded-tr" | "xl\:rounded-tr-full" | "xl\:rounded-tr-lg" | "xl\:rounded-tr-none" | "xl\:rounded-tr-sm" | "xl\:scrolling-auto" | "xl\:scrolling-touch" | "xl\:select-all" | "xl\:select-auto" | "xl\:select-none" | "xl\:select-text" | "xl\:self-auto" | "xl\:self-center" | "xl\:self-end" | "xl\:self-start" | "xl\:self-stretch" | "xl\:shadow" | "xl\:shadow-2xl" | "xl\:shadow-inner" | "xl\:shadow-lg" | "xl\:shadow-md" | "xl\:shadow-none" | "xl\:shadow-outline" | "xl\:shadow-xl" | "xl\:slashed-zeros" | "xl\:small-caps" | "xl\:sr-only" | "xl\:stacked-fractions" | "xl\:static" | "xl\:sticky" | "xl\:stroke-current" | "xl\:subpixel-antialiased" | "xl\:table" | "xl\:table-auto" | "xl\:table-cell" | "xl\:table-fixed" | "xl\:table-row" | "xl\:tabular-nums" | "xl\:text-2xl" | "xl\:text-3xl" | "xl\:text-4xl" | "xl\:text-5xl" | "xl\:text-6xl" | "xl\:text-align-unset" | "xl\:text-base" | "xl\:text-black" | "xl\:text-blue-100" | "xl\:text-blue-200" | "xl\:text-blue-300" | "xl\:text-blue-400" | "xl\:text-blue-500" | "xl\:text-blue-600" | "xl\:text-blue-700" | "xl\:text-blue-800" | "xl\:text-blue-900" | "xl\:text-border" | "xl\:text-center" | "xl\:text-color-unset" | "xl\:text-gray-100" | "xl\:text-gray-200" | "xl\:text-gray-300" | "xl\:text-gray-400" | "xl\:text-gray-500" | "xl\:text-gray-600" | "xl\:text-gray-700" | "xl\:text-gray-800" | "xl\:text-gray-900" | "xl\:text-justify" | "xl\:text-left" | "xl\:text-lg" | "xl\:text-primary" | "xl\:text-right" | "xl\:text-size-unset" | "xl\:text-sm" | "xl\:text-transform-unset" | "xl\:text-white" | "xl\:text-xl" | "xl\:text-xs" | "xl\:titling-caps" | "xl\:top-0" | "xl\:top-auto" | "xl\:tracking-normal" | "xl\:tracking-tight" | "xl\:tracking-tighter" | "xl\:tracking-unset" | "xl\:tracking-wide" | "xl\:tracking-wider" | "xl\:tracking-widest" | "xl\:transform-b" | "xl\:transform-bl" | "xl\:transform-border" | "xl\:transform-br" | "xl\:transform-fill" | "xl\:transform-l" | "xl\:transform-none" | "xl\:transform-r" | "xl\:transform-t" | "xl\:transform-tl" | "xl\:transform-tr" | "xl\:transform-view" | "xl\:transition-0" | "xl\:transition-100" | "xl\:transition-1000" | "xl\:transition-250" | "xl\:transition-500" | "xl\:transition-750" | "xl\:transition-all" | "xl\:transition-bg" | "xl\:transition-border" | "xl\:transition-color" | "xl\:transition-colors" | "xl\:transition-delay-0" | "xl\:transition-delay-100" | "xl\:transition-delay-1000" | "xl\:transition-delay-250" | "xl\:transition-delay-500" | "xl\:transition-delay-750" | "xl\:transition-ease" | "xl\:transition-ease-in" | "xl\:transition-ease-in-out" | "xl\:transition-ease-out" | "xl\:transition-linear" | "xl\:transition-none" | "xl\:transition-opacity" | "xl\:transition-transform" | "xl\:truncate" | "xl\:underline" | "xl\:unicase" | "xl\:uppercase" | "xl\:visible" | "xl\:w-0" | "xl\:w-1" | "xl\:w-10" | "xl\:w-10\/12" | "xl\:w-11\/12" | "xl\:w-12" | "xl\:w-16" | "xl\:w-1\/12" | "xl\:w-1\/2" | "xl\:w-1\/3" | "xl\:w-1\/4" | "xl\:w-1\/5" | "xl\:w-1\/6" | "xl\:w-2" | "xl\:w-20" | "xl\:w-24" | "xl\:w-2\/12" | "xl\:w-2\/3" | "xl\:w-2\/4" | "xl\:w-2\/5" | "xl\:w-2\/6" | "xl\:w-3" | "xl\:w-32" | "xl\:w-3\/12" | "xl\:w-3\/4" | "xl\:w-3\/5" | "xl\:w-3\/6" | "xl\:w-4" | "xl\:w-40" | "xl\:w-48" | "xl\:w-4\/12" | "xl\:w-4\/5" | "xl\:w-4\/6" | "xl\:w-5" | "xl\:w-56" | "xl\:w-5\/12" | "xl\:w-5\/6" | "xl\:w-6" | "xl\:w-64" | "xl\:w-6\/12" | "xl\:w-7\/12" | "xl\:w-8" | "xl\:w-8\/12" | "xl\:w-9\/12" | "xl\:w-auto" | "xl\:w-full" | "xl\:w-px" | "xl\:w-screen" | "xl\:whitespace-no-wrap" | "xl\:whitespace-normal" | "xl\:whitespace-pre" | "xl\:whitespace-pre-line" | "xl\:whitespace-pre-wrap" | "xl\:will-change-auto" | "xl\:will-change-contents" | "xl\:will-change-opacity" | "xl\:will-change-scroll" | "xl\:will-change-transform" | "xl\:z-0" | "xl\:z-10" | "xl\:z-20" | "xl\:z-30" | "xl\:z-40" | "xl\:z-50" | "xl\:z-auto" | "xs\:-m-1" | "xs\:-m-10" | "xs\:-m-12" | "xs\:-m-16" | "xs\:-m-2" | "xs\:-m-20" | "xs\:-m-24" | "xs\:-m-3" | "xs\:-m-32" | "xs\:-m-4" | "xs\:-m-40" | "xs\:-m-48" | "xs\:-m-5" | "xs\:-m-56" | "xs\:-m-6" | "xs\:-m-64" | "xs\:-m-8" | "xs\:-m-px" | "xs\:-mb-1" | "xs\:-mb-10" | "xs\:-mb-12" | "xs\:-mb-16" | "xs\:-mb-2" | "xs\:-mb-20" | "xs\:-mb-24" | "xs\:-mb-3" | "xs\:-mb-32" | "xs\:-mb-4" | "xs\:-mb-40" | "xs\:-mb-48" | "xs\:-mb-5" | "xs\:-mb-56" | "xs\:-mb-6" | "xs\:-mb-64" | "xs\:-mb-8" | "xs\:-mb-px" | "xs\:-ml-1" | "xs\:-ml-10" | "xs\:-ml-12" | "xs\:-ml-16" | "xs\:-ml-2" | "xs\:-ml-20" | "xs\:-ml-24" | "xs\:-ml-3" | "xs\:-ml-32" | "xs\:-ml-4" | "xs\:-ml-40" | "xs\:-ml-48" | "xs\:-ml-5" | "xs\:-ml-56" | "xs\:-ml-6" | "xs\:-ml-64" | "xs\:-ml-8" | "xs\:-ml-px" | "xs\:-mr-1" | "xs\:-mr-10" | "xs\:-mr-12" | "xs\:-mr-16" | "xs\:-mr-2" | "xs\:-mr-20" | "xs\:-mr-24" | "xs\:-mr-3" | "xs\:-mr-32" | "xs\:-mr-4" | "xs\:-mr-40" | "xs\:-mr-48" | "xs\:-mr-5" | "xs\:-mr-56" | "xs\:-mr-6" | "xs\:-mr-64" | "xs\:-mr-8" | "xs\:-mr-px" | "xs\:-mt-1" | "xs\:-mt-10" | "xs\:-mt-12" | "xs\:-mt-16" | "xs\:-mt-2" | "xs\:-mt-20" | "xs\:-mt-24" | "xs\:-mt-3" | "xs\:-mt-32" | "xs\:-mt-4" | "xs\:-mt-40" | "xs\:-mt-48" | "xs\:-mt-5" | "xs\:-mt-56" | "xs\:-mt-6" | "xs\:-mt-64" | "xs\:-mt-8" | "xs\:-mt-px" | "xs\:-mx-1" | "xs\:-mx-10" | "xs\:-mx-12" | "xs\:-mx-16" | "xs\:-mx-2" | "xs\:-mx-20" | "xs\:-mx-24" | "xs\:-mx-3" | "xs\:-mx-32" | "xs\:-mx-4" | "xs\:-mx-40" | "xs\:-mx-48" | "xs\:-mx-5" | "xs\:-mx-56" | "xs\:-mx-6" | "xs\:-mx-64" | "xs\:-mx-8" | "xs\:-mx-px" | "xs\:-my-1" | "xs\:-my-10" | "xs\:-my-12" | "xs\:-my-16" | "xs\:-my-2" | "xs\:-my-20" | "xs\:-my-24" | "xs\:-my-3" | "xs\:-my-32" | "xs\:-my-4" | "xs\:-my-40" | "xs\:-my-48" | "xs\:-my-5" | "xs\:-my-56" | "xs\:-my-6" | "xs\:-my-64" | "xs\:-my-8" | "xs\:-my-px" | "xs\:absolute" | "xs\:align-baseline" | "xs\:align-bottom" | "xs\:align-middle" | "xs\:align-text-bottom" | "xs\:align-text-top" | "xs\:align-top" | "xs\:all-small-caps" | "xs\:antialiased" | "xs\:appearance-none" | "xs\:bg-auto" | "xs\:bg-black" | "xs\:bg-blue-100" | "xs\:bg-blue-200" | "xs\:bg-blue-300" | "xs\:bg-blue-400" | "xs\:bg-blue-500" | "xs\:bg-blue-600" | "xs\:bg-blue-700" | "xs\:bg-blue-800" | "xs\:bg-blue-900" | "xs\:bg-border" | "xs\:bg-bottom" | "xs\:bg-center" | "xs\:bg-contain" | "xs\:bg-cover" | "xs\:bg-fixed" | "xs\:bg-gray-100" | "xs\:bg-gray-200" | "xs\:bg-gray-300" | "xs\:bg-gray-400" | "xs\:bg-gray-500" | "xs\:bg-gray-600" | "xs\:bg-gray-700" | "xs\:bg-gray-800" | "xs\:bg-gray-900" | "xs\:bg-left" | "xs\:bg-left-bottom" | "xs\:bg-left-top" | "xs\:bg-local" | "xs\:bg-no-repeat" | "xs\:bg-primary" | "xs\:bg-repeat" | "xs\:bg-repeat-round" | "xs\:bg-repeat-space" | "xs\:bg-repeat-x" | "xs\:bg-repeat-y" | "xs\:bg-right" | "xs\:bg-right-bottom" | "xs\:bg-right-top" | "xs\:bg-scroll" | "xs\:bg-top" | "xs\:bg-white" | "xs\:block" | "xs\:border" | "xs\:border-0" | "xs\:border-2" | "xs\:border-4" | "xs\:border-8" | "xs\:border-b" | "xs\:border-b-0" | "xs\:border-b-2" | "xs\:border-b-4" | "xs\:border-b-8" | "xs\:border-black" | "xs\:border-blue-100" | "xs\:border-blue-200" | "xs\:border-blue-300" | "xs\:border-blue-400" | "xs\:border-blue-500" | "xs\:border-blue-600" | "xs\:border-blue-700" | "xs\:border-blue-800" | "xs\:border-blue-900" | "xs\:border-border" | "xs\:border-collapse" | "xs\:border-dashed" | "xs\:border-dotted" | "xs\:border-double" | "xs\:border-gray-100" | "xs\:border-gray-200" | "xs\:border-gray-300" | "xs\:border-gray-400" | "xs\:border-gray-500" | "xs\:border-gray-600" | "xs\:border-gray-700" | "xs\:border-gray-800" | "xs\:border-gray-900" | "xs\:border-l" | "xs\:border-l-0" | "xs\:border-l-2" | "xs\:border-l-4" | "xs\:border-l-8" | "xs\:border-none" | "xs\:border-primary" | "xs\:border-r" | "xs\:border-r-0" | "xs\:border-r-2" | "xs\:border-r-4" | "xs\:border-r-8" | "xs\:border-separate" | "xs\:border-solid" | "xs\:border-t" | "xs\:border-t-0" | "xs\:border-t-2" | "xs\:border-t-4" | "xs\:border-t-8" | "xs\:border-white" | "xs\:bottom-0" | "xs\:bottom-auto" | "xs\:break-all" | "xs\:break-normal" | "xs\:break-words" | "xs\:capitalize" | "xs\:clearfix" | "xs\:common-ligatures" | "xs\:content-around" | "xs\:content-between" | "xs\:content-center" | "xs\:content-end" | "xs\:content-start" | "xs\:contextual-ligatures" | "xs\:cursor-auto" | "xs\:cursor-default" | "xs\:cursor-move" | "xs\:cursor-not-allowed" | "xs\:cursor-pointer" | "xs\:cursor-text" | "xs\:cursor-wait" | "xs\:diagonal-fractions" | "xs\:discretionary-ligatures" | "xs\:ellipsis" | "xs\:fill-current" | "xs\:fixed" | "xs\:flex" | "xs\:flex-1" | "xs\:flex-auto" | "xs\:flex-col" | "xs\:flex-col-reverse" | "xs\:flex-grow" | "xs\:flex-grow-0" | "xs\:flex-initial" | "xs\:flex-no-wrap" | "xs\:flex-none" | "xs\:flex-row" | "xs\:flex-row-reverse" | "xs\:flex-shrink" | "xs\:flex-shrink-0" | "xs\:flex-wrap" | "xs\:flex-wrap-reverse" | "xs\:float-left" | "xs\:float-none" | "xs\:float-right" | "xs\:focus\:bg-black" | "xs\:focus\:bg-blue-100" | "xs\:focus\:bg-blue-200" | "xs\:focus\:bg-blue-300" | "xs\:focus\:bg-blue-400" | "xs\:focus\:bg-blue-500" | "xs\:focus\:bg-blue-600" | "xs\:focus\:bg-blue-700" | "xs\:focus\:bg-blue-800" | "xs\:focus\:bg-blue-900" | "xs\:focus\:bg-border" | "xs\:focus\:bg-gray-100" | "xs\:focus\:bg-gray-200" | "xs\:focus\:bg-gray-300" | "xs\:focus\:bg-gray-400" | "xs\:focus\:bg-gray-500" | "xs\:focus\:bg-gray-600" | "xs\:focus\:bg-gray-700" | "xs\:focus\:bg-gray-800" | "xs\:focus\:bg-gray-900" | "xs\:focus\:bg-primary" | "xs\:focus\:bg-white" | "xs\:focus\:border" | "xs\:focus\:border-0" | "xs\:focus\:border-2" | "xs\:focus\:border-4" | "xs\:focus\:border-8" | "xs\:focus\:border-b" | "xs\:focus\:border-b-0" | "xs\:focus\:border-b-2" | "xs\:focus\:border-b-4" | "xs\:focus\:border-b-8" | "xs\:focus\:border-black" | "xs\:focus\:border-blue-100" | "xs\:focus\:border-blue-200" | "xs\:focus\:border-blue-300" | "xs\:focus\:border-blue-400" | "xs\:focus\:border-blue-500" | "xs\:focus\:border-blue-600" | "xs\:focus\:border-blue-700" | "xs\:focus\:border-blue-800" | "xs\:focus\:border-blue-900" | "xs\:focus\:border-border" | "xs\:focus\:border-gray-100" | "xs\:focus\:border-gray-200" | "xs\:focus\:border-gray-300" | "xs\:focus\:border-gray-400" | "xs\:focus\:border-gray-500" | "xs\:focus\:border-gray-600" | "xs\:focus\:border-gray-700" | "xs\:focus\:border-gray-800" | "xs\:focus\:border-gray-900" | "xs\:focus\:border-l" | "xs\:focus\:border-l-0" | "xs\:focus\:border-l-2" | "xs\:focus\:border-l-4" | "xs\:focus\:border-l-8" | "xs\:focus\:border-primary" | "xs\:focus\:border-r" | "xs\:focus\:border-r-0" | "xs\:focus\:border-r-2" | "xs\:focus\:border-r-4" | "xs\:focus\:border-r-8" | "xs\:focus\:border-t" | "xs\:focus\:border-t-0" | "xs\:focus\:border-t-2" | "xs\:focus\:border-t-4" | "xs\:focus\:border-t-8" | "xs\:focus\:border-white" | "xs\:focus\:font-black" | "xs\:focus\:font-bold" | "xs\:focus\:font-extrabold" | "xs\:focus\:font-hairline" | "xs\:focus\:font-light" | "xs\:focus\:font-medium" | "xs\:focus\:font-normal" | "xs\:focus\:font-semibold" | "xs\:focus\:font-thin" | "xs\:focus\:line-through" | "xs\:focus\:no-underline" | "xs\:focus\:not-sr-only" | "xs\:focus\:opacity-0" | "xs\:focus\:opacity-100" | "xs\:focus\:opacity-25" | "xs\:focus\:opacity-50" | "xs\:focus\:opacity-75" | "xs\:focus\:outline-none" | "xs\:focus\:placeholder-black" | "xs\:focus\:placeholder-blue-100" | "xs\:focus\:placeholder-blue-200" | "xs\:focus\:placeholder-blue-300" | "xs\:focus\:placeholder-blue-400" | "xs\:focus\:placeholder-blue-500" | "xs\:focus\:placeholder-blue-600" | "xs\:focus\:placeholder-blue-700" | "xs\:focus\:placeholder-blue-800" | "xs\:focus\:placeholder-blue-900" | "xs\:focus\:placeholder-border" | "xs\:focus\:placeholder-gray-100" | "xs\:focus\:placeholder-gray-200" | "xs\:focus\:placeholder-gray-300" | "xs\:focus\:placeholder-gray-400" | "xs\:focus\:placeholder-gray-500" | "xs\:focus\:placeholder-gray-600" | "xs\:focus\:placeholder-gray-700" | "xs\:focus\:placeholder-gray-800" | "xs\:focus\:placeholder-gray-900" | "xs\:focus\:placeholder-primary" | "xs\:focus\:placeholder-white" | "xs\:focus\:shadow" | "xs\:focus\:shadow-2xl" | "xs\:focus\:shadow-inner" | "xs\:focus\:shadow-lg" | "xs\:focus\:shadow-md" | "xs\:focus\:shadow-none" | "xs\:focus\:shadow-outline" | "xs\:focus\:shadow-xl" | "xs\:focus\:sr-only" | "xs\:focus\:text-black" | "xs\:focus\:text-blue-100" | "xs\:focus\:text-blue-200" | "xs\:focus\:text-blue-300" | "xs\:focus\:text-blue-400" | "xs\:focus\:text-blue-500" | "xs\:focus\:text-blue-600" | "xs\:focus\:text-blue-700" | "xs\:focus\:text-blue-800" | "xs\:focus\:text-blue-900" | "xs\:focus\:text-border" | "xs\:focus\:text-gray-100" | "xs\:focus\:text-gray-200" | "xs\:focus\:text-gray-300" | "xs\:focus\:text-gray-400" | "xs\:focus\:text-gray-500" | "xs\:focus\:text-gray-600" | "xs\:focus\:text-gray-700" | "xs\:focus\:text-gray-800" | "xs\:focus\:text-gray-900" | "xs\:focus\:text-primary" | "xs\:focus\:text-white" | "xs\:focus\:underline" | "xs\:font-black" | "xs\:font-bold" | "xs\:font-extrabold" | "xs\:font-family-unset" | "xs\:font-hairline" | "xs\:font-light" | "xs\:font-medium" | "xs\:font-mono" | "xs\:font-normal" | "xs\:font-sans" | "xs\:font-semibold" | "xs\:font-serif" | "xs\:font-style-unset" | "xs\:font-thin" | "xs\:font-weight-unset" | "xs\:h-0" | "xs\:h-1" | "xs\:h-10" | "xs\:h-12" | "xs\:h-16" | "xs\:h-2" | "xs\:h-20" | "xs\:h-24" | "xs\:h-3" | "xs\:h-32" | "xs\:h-4" | "xs\:h-40" | "xs\:h-48" | "xs\:h-5" | "xs\:h-56" | "xs\:h-6" | "xs\:h-64" | "xs\:h-8" | "xs\:h-auto" | "xs\:h-full" | "xs\:h-px" | "xs\:h-screen" | "xs\:hidden" | "xs\:historical-ligatures" | "xs\:hover\:bg-black" | "xs\:hover\:bg-blue-100" | "xs\:hover\:bg-blue-200" | "xs\:hover\:bg-blue-300" | "xs\:hover\:bg-blue-400" | "xs\:hover\:bg-blue-500" | "xs\:hover\:bg-blue-600" | "xs\:hover\:bg-blue-700" | "xs\:hover\:bg-blue-800" | "xs\:hover\:bg-blue-900" | "xs\:hover\:bg-border" | "xs\:hover\:bg-gray-100" | "xs\:hover\:bg-gray-200" | "xs\:hover\:bg-gray-300" | "xs\:hover\:bg-gray-400" | "xs\:hover\:bg-gray-500" | "xs\:hover\:bg-gray-600" | "xs\:hover\:bg-gray-700" | "xs\:hover\:bg-gray-800" | "xs\:hover\:bg-gray-900" | "xs\:hover\:bg-primary" | "xs\:hover\:bg-white" | "xs\:hover\:border" | "xs\:hover\:border-0" | "xs\:hover\:border-2" | "xs\:hover\:border-4" | "xs\:hover\:border-8" | "xs\:hover\:border-b" | "xs\:hover\:border-b-0" | "xs\:hover\:border-b-2" | "xs\:hover\:border-b-4" | "xs\:hover\:border-b-8" | "xs\:hover\:border-black" | "xs\:hover\:border-blue-100" | "xs\:hover\:border-blue-200" | "xs\:hover\:border-blue-300" | "xs\:hover\:border-blue-400" | "xs\:hover\:border-blue-500" | "xs\:hover\:border-blue-600" | "xs\:hover\:border-blue-700" | "xs\:hover\:border-blue-800" | "xs\:hover\:border-blue-900" | "xs\:hover\:border-border" | "xs\:hover\:border-gray-100" | "xs\:hover\:border-gray-200" | "xs\:hover\:border-gray-300" | "xs\:hover\:border-gray-400" | "xs\:hover\:border-gray-500" | "xs\:hover\:border-gray-600" | "xs\:hover\:border-gray-700" | "xs\:hover\:border-gray-800" | "xs\:hover\:border-gray-900" | "xs\:hover\:border-l" | "xs\:hover\:border-l-0" | "xs\:hover\:border-l-2" | "xs\:hover\:border-l-4" | "xs\:hover\:border-l-8" | "xs\:hover\:border-primary" | "xs\:hover\:border-r" | "xs\:hover\:border-r-0" | "xs\:hover\:border-r-2" | "xs\:hover\:border-r-4" | "xs\:hover\:border-r-8" | "xs\:hover\:border-t" | "xs\:hover\:border-t-0" | "xs\:hover\:border-t-2" | "xs\:hover\:border-t-4" | "xs\:hover\:border-t-8" | "xs\:hover\:border-white" | "xs\:hover\:font-black" | "xs\:hover\:font-bold" | "xs\:hover\:font-extrabold" | "xs\:hover\:font-hairline" | "xs\:hover\:font-light" | "xs\:hover\:font-medium" | "xs\:hover\:font-normal" | "xs\:hover\:font-semibold" | "xs\:hover\:font-thin" | "xs\:hover\:line-through" | "xs\:hover\:no-underline" | "xs\:hover\:opacity-0" | "xs\:hover\:opacity-100" | "xs\:hover\:opacity-25" | "xs\:hover\:opacity-50" | "xs\:hover\:opacity-75" | "xs\:hover\:shadow" | "xs\:hover\:shadow-2xl" | "xs\:hover\:shadow-inner" | "xs\:hover\:shadow-lg" | "xs\:hover\:shadow-md" | "xs\:hover\:shadow-none" | "xs\:hover\:shadow-outline" | "xs\:hover\:shadow-xl" | "xs\:hover\:text-black" | "xs\:hover\:text-blue-100" | "xs\:hover\:text-blue-200" | "xs\:hover\:text-blue-300" | "xs\:hover\:text-blue-400" | "xs\:hover\:text-blue-500" | "xs\:hover\:text-blue-600" | "xs\:hover\:text-blue-700" | "xs\:hover\:text-blue-800" | "xs\:hover\:text-blue-900" | "xs\:hover\:text-border" | "xs\:hover\:text-gray-100" | "xs\:hover\:text-gray-200" | "xs\:hover\:text-gray-300" | "xs\:hover\:text-gray-400" | "xs\:hover\:text-gray-500" | "xs\:hover\:text-gray-600" | "xs\:hover\:text-gray-700" | "xs\:hover\:text-gray-800" | "xs\:hover\:text-gray-900" | "xs\:hover\:text-primary" | "xs\:hover\:text-white" | "xs\:hover\:underline" | "xs\:hyphens-auto" | "xs\:hyphens-manual" | "xs\:hyphens-none" | "xs\:inline" | "xs\:inline-block" | "xs\:inline-flex" | "xs\:inset-0" | "xs\:inset-auto" | "xs\:inset-x-0" | "xs\:inset-x-auto" | "xs\:inset-y-0" | "xs\:inset-y-auto" | "xs\:invisible" | "xs\:italic" | "xs\:items-baseline" | "xs\:items-center" | "xs\:items-end" | "xs\:items-start" | "xs\:items-stretch" | "xs\:justify-around" | "xs\:justify-between" | "xs\:justify-center" | "xs\:justify-end" | "xs\:justify-start" | "xs\:last\:border" | "xs\:last\:border-0" | "xs\:last\:border-2" | "xs\:last\:border-4" | "xs\:last\:border-8" | "xs\:last\:border-b" | "xs\:last\:border-b-0" | "xs\:last\:border-b-2" | "xs\:last\:border-b-4" | "xs\:last\:border-b-8" | "xs\:last\:border-l" | "xs\:last\:border-l-0" | "xs\:last\:border-l-2" | "xs\:last\:border-l-4" | "xs\:last\:border-l-8" | "xs\:last\:border-r" | "xs\:last\:border-r-0" | "xs\:last\:border-r-2" | "xs\:last\:border-r-4" | "xs\:last\:border-r-8" | "xs\:last\:border-t" | "xs\:last\:border-t-0" | "xs\:last\:border-t-2" | "xs\:last\:border-t-4" | "xs\:last\:border-t-8" | "xs\:leading-loose" | "xs\:leading-none" | "xs\:leading-normal" | "xs\:leading-relaxed" | "xs\:leading-snug" | "xs\:leading-tight" | "xs\:leading-unset" | "xs\:left-0" | "xs\:left-auto" | "xs\:line-through" | "xs\:lining-nums" | "xs\:list-decimal" | "xs\:list-disc" | "xs\:list-inside" | "xs\:list-none" | "xs\:list-outside" | "xs\:lowercase" | "xs\:m-0" | "xs\:m-1" | "xs\:m-10" | "xs\:m-12" | "xs\:m-16" | "xs\:m-2" | "xs\:m-20" | "xs\:m-24" | "xs\:m-3" | "xs\:m-32" | "xs\:m-4" | "xs\:m-40" | "xs\:m-48" | "xs\:m-5" | "xs\:m-56" | "xs\:m-6" | "xs\:m-64" | "xs\:m-8" | "xs\:m-auto" | "xs\:m-px" | "xs\:max-h-full" | "xs\:max-h-screen" | "xs\:max-w-2xl" | "xs\:max-w-3xl" | "xs\:max-w-4xl" | "xs\:max-w-5xl" | "xs\:max-w-6xl" | "xs\:max-w-full" | "xs\:max-w-lg" | "xs\:max-w-md" | "xs\:max-w-sm" | "xs\:max-w-xl" | "xs\:max-w-xs" | "xs\:mb-0" | "xs\:mb-1" | "xs\:mb-10" | "xs\:mb-12" | "xs\:mb-16" | "xs\:mb-2" | "xs\:mb-20" | "xs\:mb-24" | "xs\:mb-3" | "xs\:mb-32" | "xs\:mb-4" | "xs\:mb-40" | "xs\:mb-48" | "xs\:mb-5" | "xs\:mb-56" | "xs\:mb-6" | "xs\:mb-64" | "xs\:mb-8" | "xs\:mb-auto" | "xs\:mb-px" | "xs\:min-h-0" | "xs\:min-h-full" | "xs\:min-h-screen" | "xs\:min-w-0" | "xs\:min-w-full" | "xs\:ml-0" | "xs\:ml-1" | "xs\:ml-10" | "xs\:ml-12" | "xs\:ml-16" | "xs\:ml-2" | "xs\:ml-20" | "xs\:ml-24" | "xs\:ml-3" | "xs\:ml-32" | "xs\:ml-4" | "xs\:ml-40" | "xs\:ml-48" | "xs\:ml-5" | "xs\:ml-56" | "xs\:ml-6" | "xs\:ml-64" | "xs\:ml-8" | "xs\:ml-auto" | "xs\:ml-px" | "xs\:mr-0" | "xs\:mr-1" | "xs\:mr-10" | "xs\:mr-12" | "xs\:mr-16" | "xs\:mr-2" | "xs\:mr-20" | "xs\:mr-24" | "xs\:mr-3" | "xs\:mr-32" | "xs\:mr-4" | "xs\:mr-40" | "xs\:mr-48" | "xs\:mr-5" | "xs\:mr-56" | "xs\:mr-6" | "xs\:mr-64" | "xs\:mr-8" | "xs\:mr-auto" | "xs\:mr-px" | "xs\:mt-0" | "xs\:mt-1" | "xs\:mt-10" | "xs\:mt-12" | "xs\:mt-16" | "xs\:mt-2" | "xs\:mt-20" | "xs\:mt-24" | "xs\:mt-3" | "xs\:mt-32" | "xs\:mt-4" | "xs\:mt-40" | "xs\:mt-48" | "xs\:mt-5" | "xs\:mt-56" | "xs\:mt-6" | "xs\:mt-64" | "xs\:mt-8" | "xs\:mt-auto" | "xs\:mt-px" | "xs\:mx-0" | "xs\:mx-1" | "xs\:mx-10" | "xs\:mx-12" | "xs\:mx-16" | "xs\:mx-2" | "xs\:mx-20" | "xs\:mx-24" | "xs\:mx-3" | "xs\:mx-32" | "xs\:mx-4" | "xs\:mx-40" | "xs\:mx-48" | "xs\:mx-5" | "xs\:mx-56" | "xs\:mx-6" | "xs\:mx-64" | "xs\:mx-8" | "xs\:mx-auto" | "xs\:mx-px" | "xs\:my-0" | "xs\:my-1" | "xs\:my-10" | "xs\:my-12" | "xs\:my-16" | "xs\:my-2" | "xs\:my-20" | "xs\:my-24" | "xs\:my-3" | "xs\:my-32" | "xs\:my-4" | "xs\:my-40" | "xs\:my-48" | "xs\:my-5" | "xs\:my-56" | "xs\:my-6" | "xs\:my-64" | "xs\:my-8" | "xs\:my-auto" | "xs\:my-px" | "xs\:no-common-ligatures" | "xs\:no-contextual-ligatures" | "xs\:no-discretionary-ligatures" | "xs\:no-ellipsis" | "xs\:no-historical-ligatures" | "xs\:no-ligatures" | "xs\:no-underline" | "xs\:normal-caps" | "xs\:normal-case" | "xs\:normal-ligatures" | "xs\:normal-nums" | "xs\:not-italic" | "xs\:not-sr-only" | "xs\:object-bottom" | "xs\:object-center" | "xs\:object-contain" | "xs\:object-cover" | "xs\:object-fill" | "xs\:object-left" | "xs\:object-left-bottom" | "xs\:object-left-top" | "xs\:object-none" | "xs\:object-right" | "xs\:object-right-bottom" | "xs\:object-right-top" | "xs\:object-scale-down" | "xs\:object-top" | "xs\:oldstyle-nums" | "xs\:opacity-0" | "xs\:opacity-100" | "xs\:opacity-25" | "xs\:opacity-50" | "xs\:opacity-75" | "xs\:order-1" | "xs\:order-10" | "xs\:order-11" | "xs\:order-12" | "xs\:order-2" | "xs\:order-3" | "xs\:order-4" | "xs\:order-5" | "xs\:order-6" | "xs\:order-7" | "xs\:order-8" | "xs\:order-9" | "xs\:order-first" | "xs\:order-last" | "xs\:order-none" | "xs\:ordinal-nums" | "xs\:outline-none" | "xs\:overflow-auto" | "xs\:overflow-hidden" | "xs\:overflow-scroll" | "xs\:overflow-visible" | "xs\:overflow-x-auto" | "xs\:overflow-x-hidden" | "xs\:overflow-x-scroll" | "xs\:overflow-x-visible" | "xs\:overflow-y-auto" | "xs\:overflow-y-hidden" | "xs\:overflow-y-scroll" | "xs\:overflow-y-visible" | "xs\:p-0" | "xs\:p-1" | "xs\:p-10" | "xs\:p-12" | "xs\:p-16" | "xs\:p-2" | "xs\:p-20" | "xs\:p-24" | "xs\:p-3" | "xs\:p-32" | "xs\:p-4" | "xs\:p-40" | "xs\:p-48" | "xs\:p-5" | "xs\:p-56" | "xs\:p-6" | "xs\:p-64" | "xs\:p-8" | "xs\:p-px" | "xs\:pb-0" | "xs\:pb-1" | "xs\:pb-10" | "xs\:pb-12" | "xs\:pb-16" | "xs\:pb-2" | "xs\:pb-20" | "xs\:pb-24" | "xs\:pb-3" | "xs\:pb-32" | "xs\:pb-4" | "xs\:pb-40" | "xs\:pb-48" | "xs\:pb-5" | "xs\:pb-56" | "xs\:pb-6" | "xs\:pb-64" | "xs\:pb-8" | "xs\:pb-px" | "xs\:petite-caps" | "xs\:pl-0" | "xs\:pl-1" | "xs\:pl-10" | "xs\:pl-12" | "xs\:pl-16" | "xs\:pl-2" | "xs\:pl-20" | "xs\:pl-24" | "xs\:pl-3" | "xs\:pl-32" | "xs\:pl-4" | "xs\:pl-40" | "xs\:pl-48" | "xs\:pl-5" | "xs\:pl-56" | "xs\:pl-6" | "xs\:pl-64" | "xs\:pl-8" | "xs\:pl-px" | "xs\:placeholder-black" | "xs\:placeholder-blue-100" | "xs\:placeholder-blue-200" | "xs\:placeholder-blue-300" | "xs\:placeholder-blue-400" | "xs\:placeholder-blue-500" | "xs\:placeholder-blue-600" | "xs\:placeholder-blue-700" | "xs\:placeholder-blue-800" | "xs\:placeholder-blue-900" | "xs\:placeholder-border" | "xs\:placeholder-gray-100" | "xs\:placeholder-gray-200" | "xs\:placeholder-gray-300" | "xs\:placeholder-gray-400" | "xs\:placeholder-gray-500" | "xs\:placeholder-gray-600" | "xs\:placeholder-gray-700" | "xs\:placeholder-gray-800" | "xs\:placeholder-gray-900" | "xs\:placeholder-primary" | "xs\:placeholder-white" | "xs\:pointer-events-auto" | "xs\:pointer-events-none" | "xs\:pr-0" | "xs\:pr-1" | "xs\:pr-10" | "xs\:pr-12" | "xs\:pr-16" | "xs\:pr-2" | "xs\:pr-20" | "xs\:pr-24" | "xs\:pr-3" | "xs\:pr-32" | "xs\:pr-4" | "xs\:pr-40" | "xs\:pr-48" | "xs\:pr-5" | "xs\:pr-56" | "xs\:pr-6" | "xs\:pr-64" | "xs\:pr-8" | "xs\:pr-px" | "xs\:proportional-nums" | "xs\:pt-0" | "xs\:pt-1" | "xs\:pt-10" | "xs\:pt-12" | "xs\:pt-16" | "xs\:pt-2" | "xs\:pt-20" | "xs\:pt-24" | "xs\:pt-3" | "xs\:pt-32" | "xs\:pt-4" | "xs\:pt-40" | "xs\:pt-48" | "xs\:pt-5" | "xs\:pt-56" | "xs\:pt-6" | "xs\:pt-64" | "xs\:pt-8" | "xs\:pt-px" | "xs\:px-0" | "xs\:px-1" | "xs\:px-10" | "xs\:px-12" | "xs\:px-16" | "xs\:px-2" | "xs\:px-20" | "xs\:px-24" | "xs\:px-3" | "xs\:px-32" | "xs\:px-4" | "xs\:px-40" | "xs\:px-48" | "xs\:px-5" | "xs\:px-56" | "xs\:px-6" | "xs\:px-64" | "xs\:px-8" | "xs\:px-px" | "xs\:py-0" | "xs\:py-1" | "xs\:py-10" | "xs\:py-12" | "xs\:py-16" | "xs\:py-2" | "xs\:py-20" | "xs\:py-24" | "xs\:py-3" | "xs\:py-32" | "xs\:py-4" | "xs\:py-40" | "xs\:py-48" | "xs\:py-5" | "xs\:py-56" | "xs\:py-6" | "xs\:py-64" | "xs\:py-8" | "xs\:py-px" | "xs\:relative" | "xs\:resize" | "xs\:resize-none" | "xs\:resize-x" | "xs\:resize-y" | "xs\:right-0" | "xs\:right-auto" | "xs\:rotate-180" | "xs\:rotate-270" | "xs\:rotate-90" | "xs\:rounded" | "xs\:rounded-b" | "xs\:rounded-b-full" | "xs\:rounded-b-lg" | "xs\:rounded-b-none" | "xs\:rounded-b-sm" | "xs\:rounded-bl" | "xs\:rounded-bl-full" | "xs\:rounded-bl-lg" | "xs\:rounded-bl-none" | "xs\:rounded-bl-sm" | "xs\:rounded-br" | "xs\:rounded-br-full" | "xs\:rounded-br-lg" | "xs\:rounded-br-none" | "xs\:rounded-br-sm" | "xs\:rounded-full" | "xs\:rounded-l" | "xs\:rounded-l-full" | "xs\:rounded-l-lg" | "xs\:rounded-l-none" | "xs\:rounded-l-sm" | "xs\:rounded-lg" | "xs\:rounded-none" | "xs\:rounded-r" | "xs\:rounded-r-full" | "xs\:rounded-r-lg" | "xs\:rounded-r-none" | "xs\:rounded-r-sm" | "xs\:rounded-sm" | "xs\:rounded-t" | "xs\:rounded-t-full" | "xs\:rounded-t-lg" | "xs\:rounded-t-none" | "xs\:rounded-t-sm" | "xs\:rounded-tl" | "xs\:rounded-tl-full" | "xs\:rounded-tl-lg" | "xs\:rounded-tl-none" | "xs\:rounded-tl-sm" | "xs\:rounded-tr" | "xs\:rounded-tr-full" | "xs\:rounded-tr-lg" | "xs\:rounded-tr-none" | "xs\:rounded-tr-sm" | "xs\:scrolling-auto" | "xs\:scrolling-touch" | "xs\:select-all" | "xs\:select-auto" | "xs\:select-none" | "xs\:select-text" | "xs\:self-auto" | "xs\:self-center" | "xs\:self-end" | "xs\:self-start" | "xs\:self-stretch" | "xs\:shadow" | "xs\:shadow-2xl" | "xs\:shadow-inner" | "xs\:shadow-lg" | "xs\:shadow-md" | "xs\:shadow-none" | "xs\:shadow-outline" | "xs\:shadow-xl" | "xs\:slashed-zeros" | "xs\:small-caps" | "xs\:sr-only" | "xs\:stacked-fractions" | "xs\:static" | "xs\:sticky" | "xs\:stroke-current" | "xs\:subpixel-antialiased" | "xs\:table" | "xs\:table-auto" | "xs\:table-cell" | "xs\:table-fixed" | "xs\:table-row" | "xs\:tabular-nums" | "xs\:text-2xl" | "xs\:text-3xl" | "xs\:text-4xl" | "xs\:text-5xl" | "xs\:text-6xl" | "xs\:text-align-unset" | "xs\:text-base" | "xs\:text-black" | "xs\:text-blue-100" | "xs\:text-blue-200" | "xs\:text-blue-300" | "xs\:text-blue-400" | "xs\:text-blue-500" | "xs\:text-blue-600" | "xs\:text-blue-700" | "xs\:text-blue-800" | "xs\:text-blue-900" | "xs\:text-border" | "xs\:text-center" | "xs\:text-color-unset" | "xs\:text-gray-100" | "xs\:text-gray-200" | "xs\:text-gray-300" | "xs\:text-gray-400" | "xs\:text-gray-500" | "xs\:text-gray-600" | "xs\:text-gray-700" | "xs\:text-gray-800" | "xs\:text-gray-900" | "xs\:text-justify" | "xs\:text-left" | "xs\:text-lg" | "xs\:text-primary" | "xs\:text-right" | "xs\:text-size-unset" | "xs\:text-sm" | "xs\:text-transform-unset" | "xs\:text-white" | "xs\:text-xl" | "xs\:text-xs" | "xs\:titling-caps" | "xs\:top-0" | "xs\:top-auto" | "xs\:tracking-normal" | "xs\:tracking-tight" | "xs\:tracking-tighter" | "xs\:tracking-unset" | "xs\:tracking-wide" | "xs\:tracking-wider" | "xs\:tracking-widest" | "xs\:transform-b" | "xs\:transform-bl" | "xs\:transform-border" | "xs\:transform-br" | "xs\:transform-fill" | "xs\:transform-l" | "xs\:transform-none" | "xs\:transform-r" | "xs\:transform-t" | "xs\:transform-tl" | "xs\:transform-tr" | "xs\:transform-view" | "xs\:transition-0" | "xs\:transition-100" | "xs\:transition-1000" | "xs\:transition-250" | "xs\:transition-500" | "xs\:transition-750" | "xs\:transition-all" | "xs\:transition-bg" | "xs\:transition-border" | "xs\:transition-color" | "xs\:transition-colors" | "xs\:transition-delay-0" | "xs\:transition-delay-100" | "xs\:transition-delay-1000" | "xs\:transition-delay-250" | "xs\:transition-delay-500" | "xs\:transition-delay-750" | "xs\:transition-ease" | "xs\:transition-ease-in" | "xs\:transition-ease-in-out" | "xs\:transition-ease-out" | "xs\:transition-linear" | "xs\:transition-none" | "xs\:transition-opacity" | "xs\:transition-transform" | "xs\:truncate" | "xs\:underline" | "xs\:unicase" | "xs\:uppercase" | "xs\:visible" | "xs\:w-0" | "xs\:w-1" | "xs\:w-10" | "xs\:w-10\/12" | "xs\:w-11\/12" | "xs\:w-12" | "xs\:w-16" | "xs\:w-1\/12" | "xs\:w-1\/2" | "xs\:w-1\/3" | "xs\:w-1\/4" | "xs\:w-1\/5" | "xs\:w-1\/6" | "xs\:w-2" | "xs\:w-20" | "xs\:w-24" | "xs\:w-2\/12" | "xs\:w-2\/3" | "xs\:w-2\/4" | "xs\:w-2\/5" | "xs\:w-2\/6" | "xs\:w-3" | "xs\:w-32" | "xs\:w-3\/12" | "xs\:w-3\/4" | "xs\:w-3\/5" | "xs\:w-3\/6" | "xs\:w-4" | "xs\:w-40" | "xs\:w-48" | "xs\:w-4\/12" | "xs\:w-4\/5" | "xs\:w-4\/6" | "xs\:w-5" | "xs\:w-56" | "xs\:w-5\/12" | "xs\:w-5\/6" | "xs\:w-6" | "xs\:w-64" | "xs\:w-6\/12" | "xs\:w-7\/12" | "xs\:w-8" | "xs\:w-8\/12" | "xs\:w-9\/12" | "xs\:w-auto" | "xs\:w-full" | "xs\:w-px" | "xs\:w-screen" | "xs\:whitespace-no-wrap" | "xs\:whitespace-normal" | "xs\:whitespace-pre" | "xs\:whitespace-pre-line" | "xs\:whitespace-pre-wrap" | "xs\:will-change-auto" | "xs\:will-change-contents" | "xs\:will-change-opacity" | "xs\:will-change-scroll" | "xs\:will-change-transform" | "xs\:z-0" | "xs\:z-10" | "xs\:z-20" | "xs\:z-30" | "xs\:z-40" | "xs\:z-50" | "xs\:z-auto" | "z-0" | "z-10" | "z-20" | "z-30" | "z-40" | "z-50" | "z-auto";
+type ClassNames = "-m-1" | "-m-10" | "-m-12" | "-m-16" | "-m-2" | "-m-20" | "-m-24" | "-m-3" | "-m-32" | "-m-4" | "-m-40" | "-m-48" | "-m-5" | "-m-56" | "-m-6" | "-m-64" | "-m-8" | "-m-px" | "-mb-1" | "-mb-10" | "-mb-12" | "-mb-16" | "-mb-2" | "-mb-20" | "-mb-24" | "-mb-3" | "-mb-32" | "-mb-4" | "-mb-40" | "-mb-48" | "-mb-5" | "-mb-56" | "-mb-6" | "-mb-64" | "-mb-8" | "-mb-px" | "-ml-1" | "-ml-10" | "-ml-12" | "-ml-16" | "-ml-2" | "-ml-20" | "-ml-24" | "-ml-3" | "-ml-32" | "-ml-4" | "-ml-40" | "-ml-48" | "-ml-5" | "-ml-56" | "-ml-6" | "-ml-64" | "-ml-8" | "-ml-px" | "-mr-1" | "-mr-10" | "-mr-12" | "-mr-16" | "-mr-2" | "-mr-20" | "-mr-24" | "-mr-3" | "-mr-32" | "-mr-4" | "-mr-40" | "-mr-48" | "-mr-5" | "-mr-56" | "-mr-6" | "-mr-64" | "-mr-8" | "-mr-px" | "-mt-1" | "-mt-10" | "-mt-12" | "-mt-16" | "-mt-2" | "-mt-20" | "-mt-24" | "-mt-3" | "-mt-32" | "-mt-4" | "-mt-40" | "-mt-48" | "-mt-5" | "-mt-56" | "-mt-6" | "-mt-64" | "-mt-8" | "-mt-px" | "-mx-1" | "-mx-10" | "-mx-12" | "-mx-16" | "-mx-2" | "-mx-20" | "-mx-24" | "-mx-3" | "-mx-32" | "-mx-4" | "-mx-40" | "-mx-48" | "-mx-5" | "-mx-56" | "-mx-6" | "-mx-64" | "-mx-8" | "-mx-px" | "-my-1" | "-my-10" | "-my-12" | "-my-16" | "-my-2" | "-my-20" | "-my-24" | "-my-3" | "-my-32" | "-my-4" | "-my-40" | "-my-48" | "-my-5" | "-my-56" | "-my-6" | "-my-64" | "-my-8" | "-my-px" | "absolute" | "align-baseline" | "align-bottom" | "align-middle" | "align-text-bottom" | "align-text-top" | "align-top" | "all-small-caps" | "antialiased" | "appearance-none" | "bg-auto" | "bg-black" | "bg-blue-100" | "bg-blue-200" | "bg-blue-300" | "bg-blue-400" | "bg-blue-500" | "bg-blue-600" | "bg-blue-700" | "bg-blue-800" | "bg-blue-900" | "bg-border" | "bg-bottom" | "bg-center" | "bg-contain" | "bg-cover" | "bg-fixed" | "bg-gray-100" | "bg-gray-200" | "bg-gray-300" | "bg-gray-400" | "bg-gray-500" | "bg-gray-600" | "bg-gray-700" | "bg-gray-800" | "bg-gray-900" | "bg-left" | "bg-left-bottom" | "bg-left-top" | "bg-local" | "bg-no-repeat" | "bg-primary" | "bg-repeat" | "bg-repeat-round" | "bg-repeat-space" | "bg-repeat-x" | "bg-repeat-y" | "bg-right" | "bg-right-bottom" | "bg-right-top" | "bg-scroll" | "bg-top" | "bg-white" | "block" | "border" | "border-0" | "border-2" | "border-4" | "border-8" | "border-b" | "border-b-0" | "border-b-2" | "border-b-4" | "border-b-8" | "border-black" | "border-blue-100" | "border-blue-200" | "border-blue-300" | "border-blue-400" | "border-blue-500" | "border-blue-600" | "border-blue-700" | "border-blue-800" | "border-blue-900" | "border-border" | "border-collapse" | "border-dashed" | "border-dotted" | "border-double" | "border-gray-100" | "border-gray-200" | "border-gray-300" | "border-gray-400" | "border-gray-500" | "border-gray-600" | "border-gray-700" | "border-gray-800" | "border-gray-900" | "border-l" | "border-l-0" | "border-l-2" | "border-l-4" | "border-l-8" | "border-none" | "border-primary" | "border-r" | "border-r-0" | "border-r-2" | "border-r-4" | "border-r-8" | "border-separate" | "border-solid" | "border-t" | "border-t-0" | "border-t-2" | "border-t-4" | "border-t-8" | "border-white" | "bottom-0" | "bottom-auto" | "break-all" | "break-normal" | "break-words" | "capitalize" | "clearfix" | "common-ligatures" | "container" | "content-around" | "content-between" | "content-center" | "content-end" | "content-start" | "contextual-ligatures" | "cursor-auto" | "cursor-default" | "cursor-move" | "cursor-not-allowed" | "cursor-pointer" | "cursor-text" | "cursor-wait" | "diagonal-fractions" | "discretionary-ligatures" | "ellipsis" | "fill-current" | "fixed" | "flex" | "flex-1" | "flex-auto" | "flex-col" | "flex-col-reverse" | "flex-grow" | "flex-grow-0" | "flex-initial" | "flex-no-wrap" | "flex-none" | "flex-row" | "flex-row-reverse" | "flex-shrink" | "flex-shrink-0" | "flex-wrap" | "flex-wrap-reverse" | "float-left" | "float-none" | "float-right" | "focus\:bg-black" | "focus\:bg-blue-100" | "focus\:bg-blue-200" | "focus\:bg-blue-300" | "focus\:bg-blue-400" | "focus\:bg-blue-500" | "focus\:bg-blue-600" | "focus\:bg-blue-700" | "focus\:bg-blue-800" | "focus\:bg-blue-900" | "focus\:bg-border" | "focus\:bg-gray-100" | "focus\:bg-gray-200" | "focus\:bg-gray-300" | "focus\:bg-gray-400" | "focus\:bg-gray-500" | "focus\:bg-gray-600" | "focus\:bg-gray-700" | "focus\:bg-gray-800" | "focus\:bg-gray-900" | "focus\:bg-primary" | "focus\:bg-white" | "focus\:border" | "focus\:border-0" | "focus\:border-2" | "focus\:border-4" | "focus\:border-8" | "focus\:border-b" | "focus\:border-b-0" | "focus\:border-b-2" | "focus\:border-b-4" | "focus\:border-b-8" | "focus\:border-black" | "focus\:border-blue-100" | "focus\:border-blue-200" | "focus\:border-blue-300" | "focus\:border-blue-400" | "focus\:border-blue-500" | "focus\:border-blue-600" | "focus\:border-blue-700" | "focus\:border-blue-800" | "focus\:border-blue-900" | "focus\:border-border" | "focus\:border-gray-100" | "focus\:border-gray-200" | "focus\:border-gray-300" | "focus\:border-gray-400" | "focus\:border-gray-500" | "focus\:border-gray-600" | "focus\:border-gray-700" | "focus\:border-gray-800" | "focus\:border-gray-900" | "focus\:border-l" | "focus\:border-l-0" | "focus\:border-l-2" | "focus\:border-l-4" | "focus\:border-l-8" | "focus\:border-primary" | "focus\:border-r" | "focus\:border-r-0" | "focus\:border-r-2" | "focus\:border-r-4" | "focus\:border-r-8" | "focus\:border-t" | "focus\:border-t-0" | "focus\:border-t-2" | "focus\:border-t-4" | "focus\:border-t-8" | "focus\:border-white" | "focus\:font-black" | "focus\:font-bold" | "focus\:font-extrabold" | "focus\:font-hairline" | "focus\:font-light" | "focus\:font-medium" | "focus\:font-normal" | "focus\:font-semibold" | "focus\:font-thin" | "focus\:line-through" | "focus\:no-underline" | "focus\:not-sr-only" | "focus\:opacity-0" | "focus\:opacity-100" | "focus\:opacity-25" | "focus\:opacity-50" | "focus\:opacity-75" | "focus\:outline-none" | "focus\:placeholder-black" | "focus\:placeholder-blue-100" | "focus\:placeholder-blue-200" | "focus\:placeholder-blue-300" | "focus\:placeholder-blue-400" | "focus\:placeholder-blue-500" | "focus\:placeholder-blue-600" | "focus\:placeholder-blue-700" | "focus\:placeholder-blue-800" | "focus\:placeholder-blue-900" | "focus\:placeholder-border" | "focus\:placeholder-gray-100" | "focus\:placeholder-gray-200" | "focus\:placeholder-gray-300" | "focus\:placeholder-gray-400" | "focus\:placeholder-gray-500" | "focus\:placeholder-gray-600" | "focus\:placeholder-gray-700" | "focus\:placeholder-gray-800" | "focus\:placeholder-gray-900" | "focus\:placeholder-primary" | "focus\:placeholder-white" | "focus\:shadow" | "focus\:shadow-2xl" | "focus\:shadow-inner" | "focus\:shadow-lg" | "focus\:shadow-md" | "focus\:shadow-none" | "focus\:shadow-outline" | "focus\:shadow-xl" | "focus\:sr-only" | "focus\:text-black" | "focus\:text-blue-100" | "focus\:text-blue-200" | "focus\:text-blue-300" | "focus\:text-blue-400" | "focus\:text-blue-500" | "focus\:text-blue-600" | "focus\:text-blue-700" | "focus\:text-blue-800" | "focus\:text-blue-900" | "focus\:text-border" | "focus\:text-gray-100" | "focus\:text-gray-200" | "focus\:text-gray-300" | "focus\:text-gray-400" | "focus\:text-gray-500" | "focus\:text-gray-600" | "focus\:text-gray-700" | "focus\:text-gray-800" | "focus\:text-gray-900" | "focus\:text-primary" | "focus\:text-white" | "focus\:underline" | "font-black" | "font-bold" | "font-extrabold" | "font-family-unset" | "font-hairline" | "font-light" | "font-medium" | "font-mono" | "font-normal" | "font-sans" | "font-semibold" | "font-serif" | "font-style-unset" | "font-thin" | "font-weight-unset" | "h-0" | "h-1" | "h-10" | "h-12" | "h-16" | "h-2" | "h-20" | "h-24" | "h-3" | "h-32" | "h-4" | "h-40" | "h-48" | "h-5" | "h-56" | "h-6" | "h-64" | "h-8" | "h-auto" | "h-full" | "h-px" | "h-screen" | "hidden" | "historical-ligatures" | "hover\:bg-black" | "hover\:bg-blue-100" | "hover\:bg-blue-200" | "hover\:bg-blue-300" | "hover\:bg-blue-400" | "hover\:bg-blue-500" | "hover\:bg-blue-600" | "hover\:bg-blue-700" | "hover\:bg-blue-800" | "hover\:bg-blue-900" | "hover\:bg-border" | "hover\:bg-gray-100" | "hover\:bg-gray-200" | "hover\:bg-gray-300" | "hover\:bg-gray-400" | "hover\:bg-gray-500" | "hover\:bg-gray-600" | "hover\:bg-gray-700" | "hover\:bg-gray-800" | "hover\:bg-gray-900" | "hover\:bg-primary" | "hover\:bg-white" | "hover\:border" | "hover\:border-0" | "hover\:border-2" | "hover\:border-4" | "hover\:border-8" | "hover\:border-b" | "hover\:border-b-0" | "hover\:border-b-2" | "hover\:border-b-4" | "hover\:border-b-8" | "hover\:border-black" | "hover\:border-blue-100" | "hover\:border-blue-200" | "hover\:border-blue-300" | "hover\:border-blue-400" | "hover\:border-blue-500" | "hover\:border-blue-600" | "hover\:border-blue-700" | "hover\:border-blue-800" | "hover\:border-blue-900" | "hover\:border-border" | "hover\:border-gray-100" | "hover\:border-gray-200" | "hover\:border-gray-300" | "hover\:border-gray-400" | "hover\:border-gray-500" | "hover\:border-gray-600" | "hover\:border-gray-700" | "hover\:border-gray-800" | "hover\:border-gray-900" | "hover\:border-l" | "hover\:border-l-0" | "hover\:border-l-2" | "hover\:border-l-4" | "hover\:border-l-8" | "hover\:border-primary" | "hover\:border-r" | "hover\:border-r-0" | "hover\:border-r-2" | "hover\:border-r-4" | "hover\:border-r-8" | "hover\:border-t" | "hover\:border-t-0" | "hover\:border-t-2" | "hover\:border-t-4" | "hover\:border-t-8" | "hover\:border-white" | "hover\:font-black" | "hover\:font-bold" | "hover\:font-extrabold" | "hover\:font-hairline" | "hover\:font-light" | "hover\:font-medium" | "hover\:font-normal" | "hover\:font-semibold" | "hover\:font-thin" | "hover\:line-through" | "hover\:no-underline" | "hover\:opacity-0" | "hover\:opacity-100" | "hover\:opacity-25" | "hover\:opacity-50" | "hover\:opacity-75" | "hover\:shadow" | "hover\:shadow-2xl" | "hover\:shadow-inner" | "hover\:shadow-lg" | "hover\:shadow-md" | "hover\:shadow-none" | "hover\:shadow-outline" | "hover\:shadow-xl" | "hover\:text-black" | "hover\:text-blue-100" | "hover\:text-blue-200" | "hover\:text-blue-300" | "hover\:text-blue-400" | "hover\:text-blue-500" | "hover\:text-blue-600" | "hover\:text-blue-700" | "hover\:text-blue-800" | "hover\:text-blue-900" | "hover\:text-border" | "hover\:text-gray-100" | "hover\:text-gray-200" | "hover\:text-gray-300" | "hover\:text-gray-400" | "hover\:text-gray-500" | "hover\:text-gray-600" | "hover\:text-gray-700" | "hover\:text-gray-800" | "hover\:text-gray-900" | "hover\:text-primary" | "hover\:text-white" | "hover\:underline" | "hyphens-auto" | "hyphens-manual" | "hyphens-none" | "inline" | "inline-block" | "inline-flex" | "inset-0" | "inset-auto" | "inset-x-0" | "inset-x-auto" | "inset-y-0" | "inset-y-auto" | "invisible" | "italic" | "items-baseline" | "items-center" | "items-end" | "items-start" | "items-stretch" | "justify-around" | "justify-between" | "justify-center" | "justify-end" | "justify-start" | "last\:border" | "last\:border-0" | "last\:border-2" | "last\:border-4" | "last\:border-8" | "last\:border-b" | "last\:border-b-0" | "last\:border-b-2" | "last\:border-b-4" | "last\:border-b-8" | "last\:border-l" | "last\:border-l-0" | "last\:border-l-2" | "last\:border-l-4" | "last\:border-l-8" | "last\:border-r" | "last\:border-r-0" | "last\:border-r-2" | "last\:border-r-4" | "last\:border-r-8" | "last\:border-t" | "last\:border-t-0" | "last\:border-t-2" | "last\:border-t-4" | "last\:border-t-8" | "leading-loose" | "leading-none" | "leading-normal" | "leading-relaxed" | "leading-snug" | "leading-tight" | "leading-unset" | "left-0" | "left-auto" | "lg\:-m-1" | "lg\:-m-10" | "lg\:-m-12" | "lg\:-m-16" | "lg\:-m-2" | "lg\:-m-20" | "lg\:-m-24" | "lg\:-m-3" | "lg\:-m-32" | "lg\:-m-4" | "lg\:-m-40" | "lg\:-m-48" | "lg\:-m-5" | "lg\:-m-56" | "lg\:-m-6" | "lg\:-m-64" | "lg\:-m-8" | "lg\:-m-px" | "lg\:-mb-1" | "lg\:-mb-10" | "lg\:-mb-12" | "lg\:-mb-16" | "lg\:-mb-2" | "lg\:-mb-20" | "lg\:-mb-24" | "lg\:-mb-3" | "lg\:-mb-32" | "lg\:-mb-4" | "lg\:-mb-40" | "lg\:-mb-48" | "lg\:-mb-5" | "lg\:-mb-56" | "lg\:-mb-6" | "lg\:-mb-64" | "lg\:-mb-8" | "lg\:-mb-px" | "lg\:-ml-1" | "lg\:-ml-10" | "lg\:-ml-12" | "lg\:-ml-16" | "lg\:-ml-2" | "lg\:-ml-20" | "lg\:-ml-24" | "lg\:-ml-3" | "lg\:-ml-32" | "lg\:-ml-4" | "lg\:-ml-40" | "lg\:-ml-48" | "lg\:-ml-5" | "lg\:-ml-56" | "lg\:-ml-6" | "lg\:-ml-64" | "lg\:-ml-8" | "lg\:-ml-px" | "lg\:-mr-1" | "lg\:-mr-10" | "lg\:-mr-12" | "lg\:-mr-16" | "lg\:-mr-2" | "lg\:-mr-20" | "lg\:-mr-24" | "lg\:-mr-3" | "lg\:-mr-32" | "lg\:-mr-4" | "lg\:-mr-40" | "lg\:-mr-48" | "lg\:-mr-5" | "lg\:-mr-56" | "lg\:-mr-6" | "lg\:-mr-64" | "lg\:-mr-8" | "lg\:-mr-px" | "lg\:-mt-1" | "lg\:-mt-10" | "lg\:-mt-12" | "lg\:-mt-16" | "lg\:-mt-2" | "lg\:-mt-20" | "lg\:-mt-24" | "lg\:-mt-3" | "lg\:-mt-32" | "lg\:-mt-4" | "lg\:-mt-40" | "lg\:-mt-48" | "lg\:-mt-5" | "lg\:-mt-56" | "lg\:-mt-6" | "lg\:-mt-64" | "lg\:-mt-8" | "lg\:-mt-px" | "lg\:-mx-1" | "lg\:-mx-10" | "lg\:-mx-12" | "lg\:-mx-16" | "lg\:-mx-2" | "lg\:-mx-20" | "lg\:-mx-24" | "lg\:-mx-3" | "lg\:-mx-32" | "lg\:-mx-4" | "lg\:-mx-40" | "lg\:-mx-48" | "lg\:-mx-5" | "lg\:-mx-56" | "lg\:-mx-6" | "lg\:-mx-64" | "lg\:-mx-8" | "lg\:-mx-px" | "lg\:-my-1" | "lg\:-my-10" | "lg\:-my-12" | "lg\:-my-16" | "lg\:-my-2" | "lg\:-my-20" | "lg\:-my-24" | "lg\:-my-3" | "lg\:-my-32" | "lg\:-my-4" | "lg\:-my-40" | "lg\:-my-48" | "lg\:-my-5" | "lg\:-my-56" | "lg\:-my-6" | "lg\:-my-64" | "lg\:-my-8" | "lg\:-my-px" | "lg\:absolute" | "lg\:align-baseline" | "lg\:align-bottom" | "lg\:align-middle" | "lg\:align-text-bottom" | "lg\:align-text-top" | "lg\:align-top" | "lg\:all-small-caps" | "lg\:antialiased" | "lg\:appearance-none" | "lg\:bg-auto" | "lg\:bg-black" | "lg\:bg-blue-100" | "lg\:bg-blue-200" | "lg\:bg-blue-300" | "lg\:bg-blue-400" | "lg\:bg-blue-500" | "lg\:bg-blue-600" | "lg\:bg-blue-700" | "lg\:bg-blue-800" | "lg\:bg-blue-900" | "lg\:bg-border" | "lg\:bg-bottom" | "lg\:bg-center" | "lg\:bg-contain" | "lg\:bg-cover" | "lg\:bg-fixed" | "lg\:bg-gray-100" | "lg\:bg-gray-200" | "lg\:bg-gray-300" | "lg\:bg-gray-400" | "lg\:bg-gray-500" | "lg\:bg-gray-600" | "lg\:bg-gray-700" | "lg\:bg-gray-800" | "lg\:bg-gray-900" | "lg\:bg-left" | "lg\:bg-left-bottom" | "lg\:bg-left-top" | "lg\:bg-local" | "lg\:bg-no-repeat" | "lg\:bg-primary" | "lg\:bg-repeat" | "lg\:bg-repeat-round" | "lg\:bg-repeat-space" | "lg\:bg-repeat-x" | "lg\:bg-repeat-y" | "lg\:bg-right" | "lg\:bg-right-bottom" | "lg\:bg-right-top" | "lg\:bg-scroll" | "lg\:bg-top" | "lg\:bg-white" | "lg\:block" | "lg\:border" | "lg\:border-0" | "lg\:border-2" | "lg\:border-4" | "lg\:border-8" | "lg\:border-b" | "lg\:border-b-0" | "lg\:border-b-2" | "lg\:border-b-4" | "lg\:border-b-8" | "lg\:border-black" | "lg\:border-blue-100" | "lg\:border-blue-200" | "lg\:border-blue-300" | "lg\:border-blue-400" | "lg\:border-blue-500" | "lg\:border-blue-600" | "lg\:border-blue-700" | "lg\:border-blue-800" | "lg\:border-blue-900" | "lg\:border-border" | "lg\:border-collapse" | "lg\:border-dashed" | "lg\:border-dotted" | "lg\:border-double" | "lg\:border-gray-100" | "lg\:border-gray-200" | "lg\:border-gray-300" | "lg\:border-gray-400" | "lg\:border-gray-500" | "lg\:border-gray-600" | "lg\:border-gray-700" | "lg\:border-gray-800" | "lg\:border-gray-900" | "lg\:border-l" | "lg\:border-l-0" | "lg\:border-l-2" | "lg\:border-l-4" | "lg\:border-l-8" | "lg\:border-none" | "lg\:border-primary" | "lg\:border-r" | "lg\:border-r-0" | "lg\:border-r-2" | "lg\:border-r-4" | "lg\:border-r-8" | "lg\:border-separate" | "lg\:border-solid" | "lg\:border-t" | "lg\:border-t-0" | "lg\:border-t-2" | "lg\:border-t-4" | "lg\:border-t-8" | "lg\:border-white" | "lg\:bottom-0" | "lg\:bottom-auto" | "lg\:break-all" | "lg\:break-normal" | "lg\:break-words" | "lg\:capitalize" | "lg\:clearfix" | "lg\:common-ligatures" | "lg\:content-around" | "lg\:content-between" | "lg\:content-center" | "lg\:content-end" | "lg\:content-start" | "lg\:contextual-ligatures" | "lg\:cursor-auto" | "lg\:cursor-default" | "lg\:cursor-move" | "lg\:cursor-not-allowed" | "lg\:cursor-pointer" | "lg\:cursor-text" | "lg\:cursor-wait" | "lg\:diagonal-fractions" | "lg\:discretionary-ligatures" | "lg\:ellipsis" | "lg\:fill-current" | "lg\:fixed" | "lg\:flex" | "lg\:flex-1" | "lg\:flex-auto" | "lg\:flex-col" | "lg\:flex-col-reverse" | "lg\:flex-grow" | "lg\:flex-grow-0" | "lg\:flex-initial" | "lg\:flex-no-wrap" | "lg\:flex-none" | "lg\:flex-row" | "lg\:flex-row-reverse" | "lg\:flex-shrink" | "lg\:flex-shrink-0" | "lg\:flex-wrap" | "lg\:flex-wrap-reverse" | "lg\:float-left" | "lg\:float-none" | "lg\:float-right" | "lg\:focus\:bg-black" | "lg\:focus\:bg-blue-100" | "lg\:focus\:bg-blue-200" | "lg\:focus\:bg-blue-300" | "lg\:focus\:bg-blue-400" | "lg\:focus\:bg-blue-500" | "lg\:focus\:bg-blue-600" | "lg\:focus\:bg-blue-700" | "lg\:focus\:bg-blue-800" | "lg\:focus\:bg-blue-900" | "lg\:focus\:bg-border" | "lg\:focus\:bg-gray-100" | "lg\:focus\:bg-gray-200" | "lg\:focus\:bg-gray-300" | "lg\:focus\:bg-gray-400" | "lg\:focus\:bg-gray-500" | "lg\:focus\:bg-gray-600" | "lg\:focus\:bg-gray-700" | "lg\:focus\:bg-gray-800" | "lg\:focus\:bg-gray-900" | "lg\:focus\:bg-primary" | "lg\:focus\:bg-white" | "lg\:focus\:border" | "lg\:focus\:border-0" | "lg\:focus\:border-2" | "lg\:focus\:border-4" | "lg\:focus\:border-8" | "lg\:focus\:border-b" | "lg\:focus\:border-b-0" | "lg\:focus\:border-b-2" | "lg\:focus\:border-b-4" | "lg\:focus\:border-b-8" | "lg\:focus\:border-black" | "lg\:focus\:border-blue-100" | "lg\:focus\:border-blue-200" | "lg\:focus\:border-blue-300" | "lg\:focus\:border-blue-400" | "lg\:focus\:border-blue-500" | "lg\:focus\:border-blue-600" | "lg\:focus\:border-blue-700" | "lg\:focus\:border-blue-800" | "lg\:focus\:border-blue-900" | "lg\:focus\:border-border" | "lg\:focus\:border-gray-100" | "lg\:focus\:border-gray-200" | "lg\:focus\:border-gray-300" | "lg\:focus\:border-gray-400" | "lg\:focus\:border-gray-500" | "lg\:focus\:border-gray-600" | "lg\:focus\:border-gray-700" | "lg\:focus\:border-gray-800" | "lg\:focus\:border-gray-900" | "lg\:focus\:border-l" | "lg\:focus\:border-l-0" | "lg\:focus\:border-l-2" | "lg\:focus\:border-l-4" | "lg\:focus\:border-l-8" | "lg\:focus\:border-primary" | "lg\:focus\:border-r" | "lg\:focus\:border-r-0" | "lg\:focus\:border-r-2" | "lg\:focus\:border-r-4" | "lg\:focus\:border-r-8" | "lg\:focus\:border-t" | "lg\:focus\:border-t-0" | "lg\:focus\:border-t-2" | "lg\:focus\:border-t-4" | "lg\:focus\:border-t-8" | "lg\:focus\:border-white" | "lg\:focus\:font-black" | "lg\:focus\:font-bold" | "lg\:focus\:font-extrabold" | "lg\:focus\:font-hairline" | "lg\:focus\:font-light" | "lg\:focus\:font-medium" | "lg\:focus\:font-normal" | "lg\:focus\:font-semibold" | "lg\:focus\:font-thin" | "lg\:focus\:line-through" | "lg\:focus\:no-underline" | "lg\:focus\:not-sr-only" | "lg\:focus\:opacity-0" | "lg\:focus\:opacity-100" | "lg\:focus\:opacity-25" | "lg\:focus\:opacity-50" | "lg\:focus\:opacity-75" | "lg\:focus\:outline-none" | "lg\:focus\:placeholder-black" | "lg\:focus\:placeholder-blue-100" | "lg\:focus\:placeholder-blue-200" | "lg\:focus\:placeholder-blue-300" | "lg\:focus\:placeholder-blue-400" | "lg\:focus\:placeholder-blue-500" | "lg\:focus\:placeholder-blue-600" | "lg\:focus\:placeholder-blue-700" | "lg\:focus\:placeholder-blue-800" | "lg\:focus\:placeholder-blue-900" | "lg\:focus\:placeholder-border" | "lg\:focus\:placeholder-gray-100" | "lg\:focus\:placeholder-gray-200" | "lg\:focus\:placeholder-gray-300" | "lg\:focus\:placeholder-gray-400" | "lg\:focus\:placeholder-gray-500" | "lg\:focus\:placeholder-gray-600" | "lg\:focus\:placeholder-gray-700" | "lg\:focus\:placeholder-gray-800" | "lg\:focus\:placeholder-gray-900" | "lg\:focus\:placeholder-primary" | "lg\:focus\:placeholder-white" | "lg\:focus\:shadow" | "lg\:focus\:shadow-2xl" | "lg\:focus\:shadow-inner" | "lg\:focus\:shadow-lg" | "lg\:focus\:shadow-md" | "lg\:focus\:shadow-none" | "lg\:focus\:shadow-outline" | "lg\:focus\:shadow-xl" | "lg\:focus\:sr-only" | "lg\:focus\:text-black" | "lg\:focus\:text-blue-100" | "lg\:focus\:text-blue-200" | "lg\:focus\:text-blue-300" | "lg\:focus\:text-blue-400" | "lg\:focus\:text-blue-500" | "lg\:focus\:text-blue-600" | "lg\:focus\:text-blue-700" | "lg\:focus\:text-blue-800" | "lg\:focus\:text-blue-900" | "lg\:focus\:text-border" | "lg\:focus\:text-gray-100" | "lg\:focus\:text-gray-200" | "lg\:focus\:text-gray-300" | "lg\:focus\:text-gray-400" | "lg\:focus\:text-gray-500" | "lg\:focus\:text-gray-600" | "lg\:focus\:text-gray-700" | "lg\:focus\:text-gray-800" | "lg\:focus\:text-gray-900" | "lg\:focus\:text-primary" | "lg\:focus\:text-white" | "lg\:focus\:underline" | "lg\:font-black" | "lg\:font-bold" | "lg\:font-extrabold" | "lg\:font-family-unset" | "lg\:font-hairline" | "lg\:font-light" | "lg\:font-medium" | "lg\:font-mono" | "lg\:font-normal" | "lg\:font-sans" | "lg\:font-semibold" | "lg\:font-serif" | "lg\:font-style-unset" | "lg\:font-thin" | "lg\:font-weight-unset" | "lg\:h-0" | "lg\:h-1" | "lg\:h-10" | "lg\:h-12" | "lg\:h-16" | "lg\:h-2" | "lg\:h-20" | "lg\:h-24" | "lg\:h-3" | "lg\:h-32" | "lg\:h-4" | "lg\:h-40" | "lg\:h-48" | "lg\:h-5" | "lg\:h-56" | "lg\:h-6" | "lg\:h-64" | "lg\:h-8" | "lg\:h-auto" | "lg\:h-full" | "lg\:h-px" | "lg\:h-screen" | "lg\:hidden" | "lg\:historical-ligatures" | "lg\:hover\:bg-black" | "lg\:hover\:bg-blue-100" | "lg\:hover\:bg-blue-200" | "lg\:hover\:bg-blue-300" | "lg\:hover\:bg-blue-400" | "lg\:hover\:bg-blue-500" | "lg\:hover\:bg-blue-600" | "lg\:hover\:bg-blue-700" | "lg\:hover\:bg-blue-800" | "lg\:hover\:bg-blue-900" | "lg\:hover\:bg-border" | "lg\:hover\:bg-gray-100" | "lg\:hover\:bg-gray-200" | "lg\:hover\:bg-gray-300" | "lg\:hover\:bg-gray-400" | "lg\:hover\:bg-gray-500" | "lg\:hover\:bg-gray-600" | "lg\:hover\:bg-gray-700" | "lg\:hover\:bg-gray-800" | "lg\:hover\:bg-gray-900" | "lg\:hover\:bg-primary" | "lg\:hover\:bg-white" | "lg\:hover\:border" | "lg\:hover\:border-0" | "lg\:hover\:border-2" | "lg\:hover\:border-4" | "lg\:hover\:border-8" | "lg\:hover\:border-b" | "lg\:hover\:border-b-0" | "lg\:hover\:border-b-2" | "lg\:hover\:border-b-4" | "lg\:hover\:border-b-8" | "lg\:hover\:border-black" | "lg\:hover\:border-blue-100" | "lg\:hover\:border-blue-200" | "lg\:hover\:border-blue-300" | "lg\:hover\:border-blue-400" | "lg\:hover\:border-blue-500" | "lg\:hover\:border-blue-600" | "lg\:hover\:border-blue-700" | "lg\:hover\:border-blue-800" | "lg\:hover\:border-blue-900" | "lg\:hover\:border-border" | "lg\:hover\:border-gray-100" | "lg\:hover\:border-gray-200" | "lg\:hover\:border-gray-300" | "lg\:hover\:border-gray-400" | "lg\:hover\:border-gray-500" | "lg\:hover\:border-gray-600" | "lg\:hover\:border-gray-700" | "lg\:hover\:border-gray-800" | "lg\:hover\:border-gray-900" | "lg\:hover\:border-l" | "lg\:hover\:border-l-0" | "lg\:hover\:border-l-2" | "lg\:hover\:border-l-4" | "lg\:hover\:border-l-8" | "lg\:hover\:border-primary" | "lg\:hover\:border-r" | "lg\:hover\:border-r-0" | "lg\:hover\:border-r-2" | "lg\:hover\:border-r-4" | "lg\:hover\:border-r-8" | "lg\:hover\:border-t" | "lg\:hover\:border-t-0" | "lg\:hover\:border-t-2" | "lg\:hover\:border-t-4" | "lg\:hover\:border-t-8" | "lg\:hover\:border-white" | "lg\:hover\:font-black" | "lg\:hover\:font-bold" | "lg\:hover\:font-extrabold" | "lg\:hover\:font-hairline" | "lg\:hover\:font-light" | "lg\:hover\:font-medium" | "lg\:hover\:font-normal" | "lg\:hover\:font-semibold" | "lg\:hover\:font-thin" | "lg\:hover\:line-through" | "lg\:hover\:no-underline" | "lg\:hover\:opacity-0" | "lg\:hover\:opacity-100" | "lg\:hover\:opacity-25" | "lg\:hover\:opacity-50" | "lg\:hover\:opacity-75" | "lg\:hover\:shadow" | "lg\:hover\:shadow-2xl" | "lg\:hover\:shadow-inner" | "lg\:hover\:shadow-lg" | "lg\:hover\:shadow-md" | "lg\:hover\:shadow-none" | "lg\:hover\:shadow-outline" | "lg\:hover\:shadow-xl" | "lg\:hover\:text-black" | "lg\:hover\:text-blue-100" | "lg\:hover\:text-blue-200" | "lg\:hover\:text-blue-300" | "lg\:hover\:text-blue-400" | "lg\:hover\:text-blue-500" | "lg\:hover\:text-blue-600" | "lg\:hover\:text-blue-700" | "lg\:hover\:text-blue-800" | "lg\:hover\:text-blue-900" | "lg\:hover\:text-border" | "lg\:hover\:text-gray-100" | "lg\:hover\:text-gray-200" | "lg\:hover\:text-gray-300" | "lg\:hover\:text-gray-400" | "lg\:hover\:text-gray-500" | "lg\:hover\:text-gray-600" | "lg\:hover\:text-gray-700" | "lg\:hover\:text-gray-800" | "lg\:hover\:text-gray-900" | "lg\:hover\:text-primary" | "lg\:hover\:text-white" | "lg\:hover\:underline" | "lg\:hyphens-auto" | "lg\:hyphens-manual" | "lg\:hyphens-none" | "lg\:inline" | "lg\:inline-block" | "lg\:inline-flex" | "lg\:inset-0" | "lg\:inset-auto" | "lg\:inset-x-0" | "lg\:inset-x-auto" | "lg\:inset-y-0" | "lg\:inset-y-auto" | "lg\:invisible" | "lg\:italic" | "lg\:items-baseline" | "lg\:items-center" | "lg\:items-end" | "lg\:items-start" | "lg\:items-stretch" | "lg\:justify-around" | "lg\:justify-between" | "lg\:justify-center" | "lg\:justify-end" | "lg\:justify-start" | "lg\:last\:border" | "lg\:last\:border-0" | "lg\:last\:border-2" | "lg\:last\:border-4" | "lg\:last\:border-8" | "lg\:last\:border-b" | "lg\:last\:border-b-0" | "lg\:last\:border-b-2" | "lg\:last\:border-b-4" | "lg\:last\:border-b-8" | "lg\:last\:border-l" | "lg\:last\:border-l-0" | "lg\:last\:border-l-2" | "lg\:last\:border-l-4" | "lg\:last\:border-l-8" | "lg\:last\:border-r" | "lg\:last\:border-r-0" | "lg\:last\:border-r-2" | "lg\:last\:border-r-4" | "lg\:last\:border-r-8" | "lg\:last\:border-t" | "lg\:last\:border-t-0" | "lg\:last\:border-t-2" | "lg\:last\:border-t-4" | "lg\:last\:border-t-8" | "lg\:leading-loose" | "lg\:leading-none" | "lg\:leading-normal" | "lg\:leading-relaxed" | "lg\:leading-snug" | "lg\:leading-tight" | "lg\:leading-unset" | "lg\:left-0" | "lg\:left-auto" | "lg\:line-through" | "lg\:lining-nums" | "lg\:list-decimal" | "lg\:list-disc" | "lg\:list-inside" | "lg\:list-none" | "lg\:list-outside" | "lg\:lowercase" | "lg\:m-0" | "lg\:m-1" | "lg\:m-10" | "lg\:m-12" | "lg\:m-16" | "lg\:m-2" | "lg\:m-20" | "lg\:m-24" | "lg\:m-3" | "lg\:m-32" | "lg\:m-4" | "lg\:m-40" | "lg\:m-48" | "lg\:m-5" | "lg\:m-56" | "lg\:m-6" | "lg\:m-64" | "lg\:m-8" | "lg\:m-auto" | "lg\:m-px" | "lg\:max-h-full" | "lg\:max-h-screen" | "lg\:max-w-2xl" | "lg\:max-w-3xl" | "lg\:max-w-4xl" | "lg\:max-w-5xl" | "lg\:max-w-6xl" | "lg\:max-w-full" | "lg\:max-w-lg" | "lg\:max-w-md" | "lg\:max-w-sm" | "lg\:max-w-xl" | "lg\:max-w-xs" | "lg\:mb-0" | "lg\:mb-1" | "lg\:mb-10" | "lg\:mb-12" | "lg\:mb-16" | "lg\:mb-2" | "lg\:mb-20" | "lg\:mb-24" | "lg\:mb-3" | "lg\:mb-32" | "lg\:mb-4" | "lg\:mb-40" | "lg\:mb-48" | "lg\:mb-5" | "lg\:mb-56" | "lg\:mb-6" | "lg\:mb-64" | "lg\:mb-8" | "lg\:mb-auto" | "lg\:mb-px" | "lg\:min-h-0" | "lg\:min-h-full" | "lg\:min-h-screen" | "lg\:min-w-0" | "lg\:min-w-full" | "lg\:ml-0" | "lg\:ml-1" | "lg\:ml-10" | "lg\:ml-12" | "lg\:ml-16" | "lg\:ml-2" | "lg\:ml-20" | "lg\:ml-24" | "lg\:ml-3" | "lg\:ml-32" | "lg\:ml-4" | "lg\:ml-40" | "lg\:ml-48" | "lg\:ml-5" | "lg\:ml-56" | "lg\:ml-6" | "lg\:ml-64" | "lg\:ml-8" | "lg\:ml-auto" | "lg\:ml-px" | "lg\:mr-0" | "lg\:mr-1" | "lg\:mr-10" | "lg\:mr-12" | "lg\:mr-16" | "lg\:mr-2" | "lg\:mr-20" | "lg\:mr-24" | "lg\:mr-3" | "lg\:mr-32" | "lg\:mr-4" | "lg\:mr-40" | "lg\:mr-48" | "lg\:mr-5" | "lg\:mr-56" | "lg\:mr-6" | "lg\:mr-64" | "lg\:mr-8" | "lg\:mr-auto" | "lg\:mr-px" | "lg\:mt-0" | "lg\:mt-1" | "lg\:mt-10" | "lg\:mt-12" | "lg\:mt-16" | "lg\:mt-2" | "lg\:mt-20" | "lg\:mt-24" | "lg\:mt-3" | "lg\:mt-32" | "lg\:mt-4" | "lg\:mt-40" | "lg\:mt-48" | "lg\:mt-5" | "lg\:mt-56" | "lg\:mt-6" | "lg\:mt-64" | "lg\:mt-8" | "lg\:mt-auto" | "lg\:mt-px" | "lg\:mx-0" | "lg\:mx-1" | "lg\:mx-10" | "lg\:mx-12" | "lg\:mx-16" | "lg\:mx-2" | "lg\:mx-20" | "lg\:mx-24" | "lg\:mx-3" | "lg\:mx-32" | "lg\:mx-4" | "lg\:mx-40" | "lg\:mx-48" | "lg\:mx-5" | "lg\:mx-56" | "lg\:mx-6" | "lg\:mx-64" | "lg\:mx-8" | "lg\:mx-auto" | "lg\:mx-px" | "lg\:my-0" | "lg\:my-1" | "lg\:my-10" | "lg\:my-12" | "lg\:my-16" | "lg\:my-2" | "lg\:my-20" | "lg\:my-24" | "lg\:my-3" | "lg\:my-32" | "lg\:my-4" | "lg\:my-40" | "lg\:my-48" | "lg\:my-5" | "lg\:my-56" | "lg\:my-6" | "lg\:my-64" | "lg\:my-8" | "lg\:my-auto" | "lg\:my-px" | "lg\:no-common-ligatures" | "lg\:no-contextual-ligatures" | "lg\:no-discretionary-ligatures" | "lg\:no-ellipsis" | "lg\:no-historical-ligatures" | "lg\:no-ligatures" | "lg\:no-underline" | "lg\:normal-caps" | "lg\:normal-case" | "lg\:normal-ligatures" | "lg\:normal-nums" | "lg\:not-italic" | "lg\:not-sr-only" | "lg\:object-bottom" | "lg\:object-center" | "lg\:object-contain" | "lg\:object-cover" | "lg\:object-fill" | "lg\:object-left" | "lg\:object-left-bottom" | "lg\:object-left-top" | "lg\:object-none" | "lg\:object-right" | "lg\:object-right-bottom" | "lg\:object-right-top" | "lg\:object-scale-down" | "lg\:object-top" | "lg\:oldstyle-nums" | "lg\:opacity-0" | "lg\:opacity-100" | "lg\:opacity-25" | "lg\:opacity-50" | "lg\:opacity-75" | "lg\:order-1" | "lg\:order-10" | "lg\:order-11" | "lg\:order-12" | "lg\:order-2" | "lg\:order-3" | "lg\:order-4" | "lg\:order-5" | "lg\:order-6" | "lg\:order-7" | "lg\:order-8" | "lg\:order-9" | "lg\:order-first" | "lg\:order-last" | "lg\:order-none" | "lg\:ordinal-nums" | "lg\:outline-none" | "lg\:overflow-auto" | "lg\:overflow-hidden" | "lg\:overflow-scroll" | "lg\:overflow-visible" | "lg\:overflow-x-auto" | "lg\:overflow-x-hidden" | "lg\:overflow-x-scroll" | "lg\:overflow-x-visible" | "lg\:overflow-y-auto" | "lg\:overflow-y-hidden" | "lg\:overflow-y-scroll" | "lg\:overflow-y-visible" | "lg\:p-0" | "lg\:p-1" | "lg\:p-10" | "lg\:p-12" | "lg\:p-16" | "lg\:p-2" | "lg\:p-20" | "lg\:p-24" | "lg\:p-3" | "lg\:p-32" | "lg\:p-4" | "lg\:p-40" | "lg\:p-48" | "lg\:p-5" | "lg\:p-56" | "lg\:p-6" | "lg\:p-64" | "lg\:p-8" | "lg\:p-px" | "lg\:pb-0" | "lg\:pb-1" | "lg\:pb-10" | "lg\:pb-12" | "lg\:pb-16" | "lg\:pb-2" | "lg\:pb-20" | "lg\:pb-24" | "lg\:pb-3" | "lg\:pb-32" | "lg\:pb-4" | "lg\:pb-40" | "lg\:pb-48" | "lg\:pb-5" | "lg\:pb-56" | "lg\:pb-6" | "lg\:pb-64" | "lg\:pb-8" | "lg\:pb-px" | "lg\:petite-caps" | "lg\:pl-0" | "lg\:pl-1" | "lg\:pl-10" | "lg\:pl-12" | "lg\:pl-16" | "lg\:pl-2" | "lg\:pl-20" | "lg\:pl-24" | "lg\:pl-3" | "lg\:pl-32" | "lg\:pl-4" | "lg\:pl-40" | "lg\:pl-48" | "lg\:pl-5" | "lg\:pl-56" | "lg\:pl-6" | "lg\:pl-64" | "lg\:pl-8" | "lg\:pl-px" | "lg\:placeholder-black" | "lg\:placeholder-blue-100" | "lg\:placeholder-blue-200" | "lg\:placeholder-blue-300" | "lg\:placeholder-blue-400" | "lg\:placeholder-blue-500" | "lg\:placeholder-blue-600" | "lg\:placeholder-blue-700" | "lg\:placeholder-blue-800" | "lg\:placeholder-blue-900" | "lg\:placeholder-border" | "lg\:placeholder-gray-100" | "lg\:placeholder-gray-200" | "lg\:placeholder-gray-300" | "lg\:placeholder-gray-400" | "lg\:placeholder-gray-500" | "lg\:placeholder-gray-600" | "lg\:placeholder-gray-700" | "lg\:placeholder-gray-800" | "lg\:placeholder-gray-900" | "lg\:placeholder-primary" | "lg\:placeholder-white" | "lg\:pointer-events-auto" | "lg\:pointer-events-none" | "lg\:pr-0" | "lg\:pr-1" | "lg\:pr-10" | "lg\:pr-12" | "lg\:pr-16" | "lg\:pr-2" | "lg\:pr-20" | "lg\:pr-24" | "lg\:pr-3" | "lg\:pr-32" | "lg\:pr-4" | "lg\:pr-40" | "lg\:pr-48" | "lg\:pr-5" | "lg\:pr-56" | "lg\:pr-6" | "lg\:pr-64" | "lg\:pr-8" | "lg\:pr-px" | "lg\:proportional-nums" | "lg\:pt-0" | "lg\:pt-1" | "lg\:pt-10" | "lg\:pt-12" | "lg\:pt-16" | "lg\:pt-2" | "lg\:pt-20" | "lg\:pt-24" | "lg\:pt-3" | "lg\:pt-32" | "lg\:pt-4" | "lg\:pt-40" | "lg\:pt-48" | "lg\:pt-5" | "lg\:pt-56" | "lg\:pt-6" | "lg\:pt-64" | "lg\:pt-8" | "lg\:pt-px" | "lg\:px-0" | "lg\:px-1" | "lg\:px-10" | "lg\:px-12" | "lg\:px-16" | "lg\:px-2" | "lg\:px-20" | "lg\:px-24" | "lg\:px-3" | "lg\:px-32" | "lg\:px-4" | "lg\:px-40" | "lg\:px-48" | "lg\:px-5" | "lg\:px-56" | "lg\:px-6" | "lg\:px-64" | "lg\:px-8" | "lg\:px-px" | "lg\:py-0" | "lg\:py-1" | "lg\:py-10" | "lg\:py-12" | "lg\:py-16" | "lg\:py-2" | "lg\:py-20" | "lg\:py-24" | "lg\:py-3" | "lg\:py-32" | "lg\:py-4" | "lg\:py-40" | "lg\:py-48" | "lg\:py-5" | "lg\:py-56" | "lg\:py-6" | "lg\:py-64" | "lg\:py-8" | "lg\:py-px" | "lg\:relative" | "lg\:resize" | "lg\:resize-none" | "lg\:resize-x" | "lg\:resize-y" | "lg\:right-0" | "lg\:right-auto" | "lg\:rotate-180" | "lg\:rotate-270" | "lg\:rotate-90" | "lg\:rounded" | "lg\:rounded-b" | "lg\:rounded-b-full" | "lg\:rounded-b-lg" | "lg\:rounded-b-none" | "lg\:rounded-b-sm" | "lg\:rounded-bl" | "lg\:rounded-bl-full" | "lg\:rounded-bl-lg" | "lg\:rounded-bl-none" | "lg\:rounded-bl-sm" | "lg\:rounded-br" | "lg\:rounded-br-full" | "lg\:rounded-br-lg" | "lg\:rounded-br-none" | "lg\:rounded-br-sm" | "lg\:rounded-full" | "lg\:rounded-l" | "lg\:rounded-l-full" | "lg\:rounded-l-lg" | "lg\:rounded-l-none" | "lg\:rounded-l-sm" | "lg\:rounded-lg" | "lg\:rounded-none" | "lg\:rounded-r" | "lg\:rounded-r-full" | "lg\:rounded-r-lg" | "lg\:rounded-r-none" | "lg\:rounded-r-sm" | "lg\:rounded-sm" | "lg\:rounded-t" | "lg\:rounded-t-full" | "lg\:rounded-t-lg" | "lg\:rounded-t-none" | "lg\:rounded-t-sm" | "lg\:rounded-tl" | "lg\:rounded-tl-full" | "lg\:rounded-tl-lg" | "lg\:rounded-tl-none" | "lg\:rounded-tl-sm" | "lg\:rounded-tr" | "lg\:rounded-tr-full" | "lg\:rounded-tr-lg" | "lg\:rounded-tr-none" | "lg\:rounded-tr-sm" | "lg\:scrolling-auto" | "lg\:scrolling-touch" | "lg\:select-all" | "lg\:select-auto" | "lg\:select-none" | "lg\:select-text" | "lg\:self-auto" | "lg\:self-center" | "lg\:self-end" | "lg\:self-start" | "lg\:self-stretch" | "lg\:shadow" | "lg\:shadow-2xl" | "lg\:shadow-inner" | "lg\:shadow-lg" | "lg\:shadow-md" | "lg\:shadow-none" | "lg\:shadow-outline" | "lg\:shadow-xl" | "lg\:slashed-zeros" | "lg\:small-caps" | "lg\:sr-only" | "lg\:stacked-fractions" | "lg\:static" | "lg\:sticky" | "lg\:stroke-current" | "lg\:subpixel-antialiased" | "lg\:table" | "lg\:table-auto" | "lg\:table-cell" | "lg\:table-fixed" | "lg\:table-row" | "lg\:tabular-nums" | "lg\:text-2xl" | "lg\:text-3xl" | "lg\:text-4xl" | "lg\:text-5xl" | "lg\:text-6xl" | "lg\:text-align-unset" | "lg\:text-base" | "lg\:text-black" | "lg\:text-blue-100" | "lg\:text-blue-200" | "lg\:text-blue-300" | "lg\:text-blue-400" | "lg\:text-blue-500" | "lg\:text-blue-600" | "lg\:text-blue-700" | "lg\:text-blue-800" | "lg\:text-blue-900" | "lg\:text-border" | "lg\:text-center" | "lg\:text-color-unset" | "lg\:text-gray-100" | "lg\:text-gray-200" | "lg\:text-gray-300" | "lg\:text-gray-400" | "lg\:text-gray-500" | "lg\:text-gray-600" | "lg\:text-gray-700" | "lg\:text-gray-800" | "lg\:text-gray-900" | "lg\:text-justify" | "lg\:text-left" | "lg\:text-lg" | "lg\:text-primary" | "lg\:text-right" | "lg\:text-size-unset" | "lg\:text-sm" | "lg\:text-transform-unset" | "lg\:text-white" | "lg\:text-xl" | "lg\:text-xs" | "lg\:titling-caps" | "lg\:top-0" | "lg\:top-auto" | "lg\:tracking-normal" | "lg\:tracking-tight" | "lg\:tracking-tighter" | "lg\:tracking-unset" | "lg\:tracking-wide" | "lg\:tracking-wider" | "lg\:tracking-widest" | "lg\:transform-b" | "lg\:transform-bl" | "lg\:transform-border" | "lg\:transform-br" | "lg\:transform-fill" | "lg\:transform-l" | "lg\:transform-none" | "lg\:transform-r" | "lg\:transform-t" | "lg\:transform-tl" | "lg\:transform-tr" | "lg\:transform-view" | "lg\:transition-0" | "lg\:transition-100" | "lg\:transition-1000" | "lg\:transition-250" | "lg\:transition-500" | "lg\:transition-750" | "lg\:transition-all" | "lg\:transition-bg" | "lg\:transition-border" | "lg\:transition-color" | "lg\:transition-colors" | "lg\:transition-delay-0" | "lg\:transition-delay-100" | "lg\:transition-delay-1000" | "lg\:transition-delay-250" | "lg\:transition-delay-500" | "lg\:transition-delay-750" | "lg\:transition-ease" | "lg\:transition-ease-in" | "lg\:transition-ease-in-out" | "lg\:transition-ease-out" | "lg\:transition-linear" | "lg\:transition-none" | "lg\:transition-opacity" | "lg\:transition-transform" | "lg\:truncate" | "lg\:underline" | "lg\:unicase" | "lg\:uppercase" | "lg\:visible" | "lg\:w-0" | "lg\:w-1" | "lg\:w-10" | "lg\:w-10\/12" | "lg\:w-11\/12" | "lg\:w-12" | "lg\:w-16" | "lg\:w-1\/12" | "lg\:w-1\/2" | "lg\:w-1\/3" | "lg\:w-1\/4" | "lg\:w-1\/5" | "lg\:w-1\/6" | "lg\:w-2" | "lg\:w-20" | "lg\:w-24" | "lg\:w-2\/12" | "lg\:w-2\/3" | "lg\:w-2\/4" | "lg\:w-2\/5" | "lg\:w-2\/6" | "lg\:w-3" | "lg\:w-32" | "lg\:w-3\/12" | "lg\:w-3\/4" | "lg\:w-3\/5" | "lg\:w-3\/6" | "lg\:w-4" | "lg\:w-40" | "lg\:w-48" | "lg\:w-4\/12" | "lg\:w-4\/5" | "lg\:w-4\/6" | "lg\:w-5" | "lg\:w-56" | "lg\:w-5\/12" | "lg\:w-5\/6" | "lg\:w-6" | "lg\:w-64" | "lg\:w-6\/12" | "lg\:w-7\/12" | "lg\:w-8" | "lg\:w-8\/12" | "lg\:w-9\/12" | "lg\:w-auto" | "lg\:w-full" | "lg\:w-px" | "lg\:w-screen" | "lg\:whitespace-no-wrap" | "lg\:whitespace-normal" | "lg\:whitespace-pre" | "lg\:whitespace-pre-line" | "lg\:whitespace-pre-wrap" | "lg\:will-change-auto" | "lg\:will-change-contents" | "lg\:will-change-opacity" | "lg\:will-change-scroll" | "lg\:will-change-transform" | "lg\:z-0" | "lg\:z-10" | "lg\:z-20" | "lg\:z-30" | "lg\:z-40" | "lg\:z-50" | "lg\:z-auto" | "line-through" | "lining-nums" | "list-decimal" | "list-disc" | "list-inside" | "list-none" | "list-outside" | "lowercase" | "m-0" | "m-1" | "m-10" | "m-12" | "m-16" | "m-2" | "m-20" | "m-24" | "m-3" | "m-32" | "m-4" | "m-40" | "m-48" | "m-5" | "m-56" | "m-6" | "m-64" | "m-8" | "m-auto" | "m-px" | "max-h-full" | "max-h-screen" | "max-w-2xl" | "max-w-3xl" | "max-w-4xl" | "max-w-5xl" | "max-w-6xl" | "max-w-full" | "max-w-lg" | "max-w-md" | "max-w-sm" | "max-w-xl" | "max-w-xs" | "mb-0" | "mb-1" | "mb-10" | "mb-12" | "mb-16" | "mb-2" | "mb-20" | "mb-24" | "mb-3" | "mb-32" | "mb-4" | "mb-40" | "mb-48" | "mb-5" | "mb-56" | "mb-6" | "mb-64" | "mb-8" | "mb-auto" | "mb-px" | "md\:-m-1" | "md\:-m-10" | "md\:-m-12" | "md\:-m-16" | "md\:-m-2" | "md\:-m-20" | "md\:-m-24" | "md\:-m-3" | "md\:-m-32" | "md\:-m-4" | "md\:-m-40" | "md\:-m-48" | "md\:-m-5" | "md\:-m-56" | "md\:-m-6" | "md\:-m-64" | "md\:-m-8" | "md\:-m-px" | "md\:-mb-1" | "md\:-mb-10" | "md\:-mb-12" | "md\:-mb-16" | "md\:-mb-2" | "md\:-mb-20" | "md\:-mb-24" | "md\:-mb-3" | "md\:-mb-32" | "md\:-mb-4" | "md\:-mb-40" | "md\:-mb-48" | "md\:-mb-5" | "md\:-mb-56" | "md\:-mb-6" | "md\:-mb-64" | "md\:-mb-8" | "md\:-mb-px" | "md\:-ml-1" | "md\:-ml-10" | "md\:-ml-12" | "md\:-ml-16" | "md\:-ml-2" | "md\:-ml-20" | "md\:-ml-24" | "md\:-ml-3" | "md\:-ml-32" | "md\:-ml-4" | "md\:-ml-40" | "md\:-ml-48" | "md\:-ml-5" | "md\:-ml-56" | "md\:-ml-6" | "md\:-ml-64" | "md\:-ml-8" | "md\:-ml-px" | "md\:-mr-1" | "md\:-mr-10" | "md\:-mr-12" | "md\:-mr-16" | "md\:-mr-2" | "md\:-mr-20" | "md\:-mr-24" | "md\:-mr-3" | "md\:-mr-32" | "md\:-mr-4" | "md\:-mr-40" | "md\:-mr-48" | "md\:-mr-5" | "md\:-mr-56" | "md\:-mr-6" | "md\:-mr-64" | "md\:-mr-8" | "md\:-mr-px" | "md\:-mt-1" | "md\:-mt-10" | "md\:-mt-12" | "md\:-mt-16" | "md\:-mt-2" | "md\:-mt-20" | "md\:-mt-24" | "md\:-mt-3" | "md\:-mt-32" | "md\:-mt-4" | "md\:-mt-40" | "md\:-mt-48" | "md\:-mt-5" | "md\:-mt-56" | "md\:-mt-6" | "md\:-mt-64" | "md\:-mt-8" | "md\:-mt-px" | "md\:-mx-1" | "md\:-mx-10" | "md\:-mx-12" | "md\:-mx-16" | "md\:-mx-2" | "md\:-mx-20" | "md\:-mx-24" | "md\:-mx-3" | "md\:-mx-32" | "md\:-mx-4" | "md\:-mx-40" | "md\:-mx-48" | "md\:-mx-5" | "md\:-mx-56" | "md\:-mx-6" | "md\:-mx-64" | "md\:-mx-8" | "md\:-mx-px" | "md\:-my-1" | "md\:-my-10" | "md\:-my-12" | "md\:-my-16" | "md\:-my-2" | "md\:-my-20" | "md\:-my-24" | "md\:-my-3" | "md\:-my-32" | "md\:-my-4" | "md\:-my-40" | "md\:-my-48" | "md\:-my-5" | "md\:-my-56" | "md\:-my-6" | "md\:-my-64" | "md\:-my-8" | "md\:-my-px" | "md\:absolute" | "md\:align-baseline" | "md\:align-bottom" | "md\:align-middle" | "md\:align-text-bottom" | "md\:align-text-top" | "md\:align-top" | "md\:all-small-caps" | "md\:antialiased" | "md\:appearance-none" | "md\:bg-auto" | "md\:bg-black" | "md\:bg-blue-100" | "md\:bg-blue-200" | "md\:bg-blue-300" | "md\:bg-blue-400" | "md\:bg-blue-500" | "md\:bg-blue-600" | "md\:bg-blue-700" | "md\:bg-blue-800" | "md\:bg-blue-900" | "md\:bg-border" | "md\:bg-bottom" | "md\:bg-center" | "md\:bg-contain" | "md\:bg-cover" | "md\:bg-fixed" | "md\:bg-gray-100" | "md\:bg-gray-200" | "md\:bg-gray-300" | "md\:bg-gray-400" | "md\:bg-gray-500" | "md\:bg-gray-600" | "md\:bg-gray-700" | "md\:bg-gray-800" | "md\:bg-gray-900" | "md\:bg-left" | "md\:bg-left-bottom" | "md\:bg-left-top" | "md\:bg-local" | "md\:bg-no-repeat" | "md\:bg-primary" | "md\:bg-repeat" | "md\:bg-repeat-round" | "md\:bg-repeat-space" | "md\:bg-repeat-x" | "md\:bg-repeat-y" | "md\:bg-right" | "md\:bg-right-bottom" | "md\:bg-right-top" | "md\:bg-scroll" | "md\:bg-top" | "md\:bg-white" | "md\:block" | "md\:border" | "md\:border-0" | "md\:border-2" | "md\:border-4" | "md\:border-8" | "md\:border-b" | "md\:border-b-0" | "md\:border-b-2" | "md\:border-b-4" | "md\:border-b-8" | "md\:border-black" | "md\:border-blue-100" | "md\:border-blue-200" | "md\:border-blue-300" | "md\:border-blue-400" | "md\:border-blue-500" | "md\:border-blue-600" | "md\:border-blue-700" | "md\:border-blue-800" | "md\:border-blue-900" | "md\:border-border" | "md\:border-collapse" | "md\:border-dashed" | "md\:border-dotted" | "md\:border-double" | "md\:border-gray-100" | "md\:border-gray-200" | "md\:border-gray-300" | "md\:border-gray-400" | "md\:border-gray-500" | "md\:border-gray-600" | "md\:border-gray-700" | "md\:border-gray-800" | "md\:border-gray-900" | "md\:border-l" | "md\:border-l-0" | "md\:border-l-2" | "md\:border-l-4" | "md\:border-l-8" | "md\:border-none" | "md\:border-primary" | "md\:border-r" | "md\:border-r-0" | "md\:border-r-2" | "md\:border-r-4" | "md\:border-r-8" | "md\:border-separate" | "md\:border-solid" | "md\:border-t" | "md\:border-t-0" | "md\:border-t-2" | "md\:border-t-4" | "md\:border-t-8" | "md\:border-white" | "md\:bottom-0" | "md\:bottom-auto" | "md\:break-all" | "md\:break-normal" | "md\:break-words" | "md\:capitalize" | "md\:clearfix" | "md\:common-ligatures" | "md\:content-around" | "md\:content-between" | "md\:content-center" | "md\:content-end" | "md\:content-start" | "md\:contextual-ligatures" | "md\:cursor-auto" | "md\:cursor-default" | "md\:cursor-move" | "md\:cursor-not-allowed" | "md\:cursor-pointer" | "md\:cursor-text" | "md\:cursor-wait" | "md\:diagonal-fractions" | "md\:discretionary-ligatures" | "md\:ellipsis" | "md\:fill-current" | "md\:fixed" | "md\:flex" | "md\:flex-1" | "md\:flex-auto" | "md\:flex-col" | "md\:flex-col-reverse" | "md\:flex-grow" | "md\:flex-grow-0" | "md\:flex-initial" | "md\:flex-no-wrap" | "md\:flex-none" | "md\:flex-row" | "md\:flex-row-reverse" | "md\:flex-shrink" | "md\:flex-shrink-0" | "md\:flex-wrap" | "md\:flex-wrap-reverse" | "md\:float-left" | "md\:float-none" | "md\:float-right" | "md\:focus\:bg-black" | "md\:focus\:bg-blue-100" | "md\:focus\:bg-blue-200" | "md\:focus\:bg-blue-300" | "md\:focus\:bg-blue-400" | "md\:focus\:bg-blue-500" | "md\:focus\:bg-blue-600" | "md\:focus\:bg-blue-700" | "md\:focus\:bg-blue-800" | "md\:focus\:bg-blue-900" | "md\:focus\:bg-border" | "md\:focus\:bg-gray-100" | "md\:focus\:bg-gray-200" | "md\:focus\:bg-gray-300" | "md\:focus\:bg-gray-400" | "md\:focus\:bg-gray-500" | "md\:focus\:bg-gray-600" | "md\:focus\:bg-gray-700" | "md\:focus\:bg-gray-800" | "md\:focus\:bg-gray-900" | "md\:focus\:bg-primary" | "md\:focus\:bg-white" | "md\:focus\:border" | "md\:focus\:border-0" | "md\:focus\:border-2" | "md\:focus\:border-4" | "md\:focus\:border-8" | "md\:focus\:border-b" | "md\:focus\:border-b-0" | "md\:focus\:border-b-2" | "md\:focus\:border-b-4" | "md\:focus\:border-b-8" | "md\:focus\:border-black" | "md\:focus\:border-blue-100" | "md\:focus\:border-blue-200" | "md\:focus\:border-blue-300" | "md\:focus\:border-blue-400" | "md\:focus\:border-blue-500" | "md\:focus\:border-blue-600" | "md\:focus\:border-blue-700" | "md\:focus\:border-blue-800" | "md\:focus\:border-blue-900" | "md\:focus\:border-border" | "md\:focus\:border-gray-100" | "md\:focus\:border-gray-200" | "md\:focus\:border-gray-300" | "md\:focus\:border-gray-400" | "md\:focus\:border-gray-500" | "md\:focus\:border-gray-600" | "md\:focus\:border-gray-700" | "md\:focus\:border-gray-800" | "md\:focus\:border-gray-900" | "md\:focus\:border-l" | "md\:focus\:border-l-0" | "md\:focus\:border-l-2" | "md\:focus\:border-l-4" | "md\:focus\:border-l-8" | "md\:focus\:border-primary" | "md\:focus\:border-r" | "md\:focus\:border-r-0" | "md\:focus\:border-r-2" | "md\:focus\:border-r-4" | "md\:focus\:border-r-8" | "md\:focus\:border-t" | "md\:focus\:border-t-0" | "md\:focus\:border-t-2" | "md\:focus\:border-t-4" | "md\:focus\:border-t-8" | "md\:focus\:border-white" | "md\:focus\:font-black" | "md\:focus\:font-bold" | "md\:focus\:font-extrabold" | "md\:focus\:font-hairline" | "md\:focus\:font-light" | "md\:focus\:font-medium" | "md\:focus\:font-normal" | "md\:focus\:font-semibold" | "md\:focus\:font-thin" | "md\:focus\:line-through" | "md\:focus\:no-underline" | "md\:focus\:not-sr-only" | "md\:focus\:opacity-0" | "md\:focus\:opacity-100" | "md\:focus\:opacity-25" | "md\:focus\:opacity-50" | "md\:focus\:opacity-75" | "md\:focus\:outline-none" | "md\:focus\:placeholder-black" | "md\:focus\:placeholder-blue-100" | "md\:focus\:placeholder-blue-200" | "md\:focus\:placeholder-blue-300" | "md\:focus\:placeholder-blue-400" | "md\:focus\:placeholder-blue-500" | "md\:focus\:placeholder-blue-600" | "md\:focus\:placeholder-blue-700" | "md\:focus\:placeholder-blue-800" | "md\:focus\:placeholder-blue-900" | "md\:focus\:placeholder-border" | "md\:focus\:placeholder-gray-100" | "md\:focus\:placeholder-gray-200" | "md\:focus\:placeholder-gray-300" | "md\:focus\:placeholder-gray-400" | "md\:focus\:placeholder-gray-500" | "md\:focus\:placeholder-gray-600" | "md\:focus\:placeholder-gray-700" | "md\:focus\:placeholder-gray-800" | "md\:focus\:placeholder-gray-900" | "md\:focus\:placeholder-primary" | "md\:focus\:placeholder-white" | "md\:focus\:shadow" | "md\:focus\:shadow-2xl" | "md\:focus\:shadow-inner" | "md\:focus\:shadow-lg" | "md\:focus\:shadow-md" | "md\:focus\:shadow-none" | "md\:focus\:shadow-outline" | "md\:focus\:shadow-xl" | "md\:focus\:sr-only" | "md\:focus\:text-black" | "md\:focus\:text-blue-100" | "md\:focus\:text-blue-200" | "md\:focus\:text-blue-300" | "md\:focus\:text-blue-400" | "md\:focus\:text-blue-500" | "md\:focus\:text-blue-600" | "md\:focus\:text-blue-700" | "md\:focus\:text-blue-800" | "md\:focus\:text-blue-900" | "md\:focus\:text-border" | "md\:focus\:text-gray-100" | "md\:focus\:text-gray-200" | "md\:focus\:text-gray-300" | "md\:focus\:text-gray-400" | "md\:focus\:text-gray-500" | "md\:focus\:text-gray-600" | "md\:focus\:text-gray-700" | "md\:focus\:text-gray-800" | "md\:focus\:text-gray-900" | "md\:focus\:text-primary" | "md\:focus\:text-white" | "md\:focus\:underline" | "md\:font-black" | "md\:font-bold" | "md\:font-extrabold" | "md\:font-family-unset" | "md\:font-hairline" | "md\:font-light" | "md\:font-medium" | "md\:font-mono" | "md\:font-normal" | "md\:font-sans" | "md\:font-semibold" | "md\:font-serif" | "md\:font-style-unset" | "md\:font-thin" | "md\:font-weight-unset" | "md\:h-0" | "md\:h-1" | "md\:h-10" | "md\:h-12" | "md\:h-16" | "md\:h-2" | "md\:h-20" | "md\:h-24" | "md\:h-3" | "md\:h-32" | "md\:h-4" | "md\:h-40" | "md\:h-48" | "md\:h-5" | "md\:h-56" | "md\:h-6" | "md\:h-64" | "md\:h-8" | "md\:h-auto" | "md\:h-full" | "md\:h-px" | "md\:h-screen" | "md\:hidden" | "md\:historical-ligatures" | "md\:hover\:bg-black" | "md\:hover\:bg-blue-100" | "md\:hover\:bg-blue-200" | "md\:hover\:bg-blue-300" | "md\:hover\:bg-blue-400" | "md\:hover\:bg-blue-500" | "md\:hover\:bg-blue-600" | "md\:hover\:bg-blue-700" | "md\:hover\:bg-blue-800" | "md\:hover\:bg-blue-900" | "md\:hover\:bg-border" | "md\:hover\:bg-gray-100" | "md\:hover\:bg-gray-200" | "md\:hover\:bg-gray-300" | "md\:hover\:bg-gray-400" | "md\:hover\:bg-gray-500" | "md\:hover\:bg-gray-600" | "md\:hover\:bg-gray-700" | "md\:hover\:bg-gray-800" | "md\:hover\:bg-gray-900" | "md\:hover\:bg-primary" | "md\:hover\:bg-white" | "md\:hover\:border" | "md\:hover\:border-0" | "md\:hover\:border-2" | "md\:hover\:border-4" | "md\:hover\:border-8" | "md\:hover\:border-b" | "md\:hover\:border-b-0" | "md\:hover\:border-b-2" | "md\:hover\:border-b-4" | "md\:hover\:border-b-8" | "md\:hover\:border-black" | "md\:hover\:border-blue-100" | "md\:hover\:border-blue-200" | "md\:hover\:border-blue-300" | "md\:hover\:border-blue-400" | "md\:hover\:border-blue-500" | "md\:hover\:border-blue-600" | "md\:hover\:border-blue-700" | "md\:hover\:border-blue-800" | "md\:hover\:border-blue-900" | "md\:hover\:border-border" | "md\:hover\:border-gray-100" | "md\:hover\:border-gray-200" | "md\:hover\:border-gray-300" | "md\:hover\:border-gray-400" | "md\:hover\:border-gray-500" | "md\:hover\:border-gray-600" | "md\:hover\:border-gray-700" | "md\:hover\:border-gray-800" | "md\:hover\:border-gray-900" | "md\:hover\:border-l" | "md\:hover\:border-l-0" | "md\:hover\:border-l-2" | "md\:hover\:border-l-4" | "md\:hover\:border-l-8" | "md\:hover\:border-primary" | "md\:hover\:border-r" | "md\:hover\:border-r-0" | "md\:hover\:border-r-2" | "md\:hover\:border-r-4" | "md\:hover\:border-r-8" | "md\:hover\:border-t" | "md\:hover\:border-t-0" | "md\:hover\:border-t-2" | "md\:hover\:border-t-4" | "md\:hover\:border-t-8" | "md\:hover\:border-white" | "md\:hover\:font-black" | "md\:hover\:font-bold" | "md\:hover\:font-extrabold" | "md\:hover\:font-hairline" | "md\:hover\:font-light" | "md\:hover\:font-medium" | "md\:hover\:font-normal" | "md\:hover\:font-semibold" | "md\:hover\:font-thin" | "md\:hover\:line-through" | "md\:hover\:no-underline" | "md\:hover\:opacity-0" | "md\:hover\:opacity-100" | "md\:hover\:opacity-25" | "md\:hover\:opacity-50" | "md\:hover\:opacity-75" | "md\:hover\:shadow" | "md\:hover\:shadow-2xl" | "md\:hover\:shadow-inner" | "md\:hover\:shadow-lg" | "md\:hover\:shadow-md" | "md\:hover\:shadow-none" | "md\:hover\:shadow-outline" | "md\:hover\:shadow-xl" | "md\:hover\:text-black" | "md\:hover\:text-blue-100" | "md\:hover\:text-blue-200" | "md\:hover\:text-blue-300" | "md\:hover\:text-blue-400" | "md\:hover\:text-blue-500" | "md\:hover\:text-blue-600" | "md\:hover\:text-blue-700" | "md\:hover\:text-blue-800" | "md\:hover\:text-blue-900" | "md\:hover\:text-border" | "md\:hover\:text-gray-100" | "md\:hover\:text-gray-200" | "md\:hover\:text-gray-300" | "md\:hover\:text-gray-400" | "md\:hover\:text-gray-500" | "md\:hover\:text-gray-600" | "md\:hover\:text-gray-700" | "md\:hover\:text-gray-800" | "md\:hover\:text-gray-900" | "md\:hover\:text-primary" | "md\:hover\:text-white" | "md\:hover\:underline" | "md\:hyphens-auto" | "md\:hyphens-manual" | "md\:hyphens-none" | "md\:inline" | "md\:inline-block" | "md\:inline-flex" | "md\:inset-0" | "md\:inset-auto" | "md\:inset-x-0" | "md\:inset-x-auto" | "md\:inset-y-0" | "md\:inset-y-auto" | "md\:invisible" | "md\:italic" | "md\:items-baseline" | "md\:items-center" | "md\:items-end" | "md\:items-start" | "md\:items-stretch" | "md\:justify-around" | "md\:justify-between" | "md\:justify-center" | "md\:justify-end" | "md\:justify-start" | "md\:last\:border" | "md\:last\:border-0" | "md\:last\:border-2" | "md\:last\:border-4" | "md\:last\:border-8" | "md\:last\:border-b" | "md\:last\:border-b-0" | "md\:last\:border-b-2" | "md\:last\:border-b-4" | "md\:last\:border-b-8" | "md\:last\:border-l" | "md\:last\:border-l-0" | "md\:last\:border-l-2" | "md\:last\:border-l-4" | "md\:last\:border-l-8" | "md\:last\:border-r" | "md\:last\:border-r-0" | "md\:last\:border-r-2" | "md\:last\:border-r-4" | "md\:last\:border-r-8" | "md\:last\:border-t" | "md\:last\:border-t-0" | "md\:last\:border-t-2" | "md\:last\:border-t-4" | "md\:last\:border-t-8" | "md\:leading-loose" | "md\:leading-none" | "md\:leading-normal" | "md\:leading-relaxed" | "md\:leading-snug" | "md\:leading-tight" | "md\:leading-unset" | "md\:left-0" | "md\:left-auto" | "md\:line-through" | "md\:lining-nums" | "md\:list-decimal" | "md\:list-disc" | "md\:list-inside" | "md\:list-none" | "md\:list-outside" | "md\:lowercase" | "md\:m-0" | "md\:m-1" | "md\:m-10" | "md\:m-12" | "md\:m-16" | "md\:m-2" | "md\:m-20" | "md\:m-24" | "md\:m-3" | "md\:m-32" | "md\:m-4" | "md\:m-40" | "md\:m-48" | "md\:m-5" | "md\:m-56" | "md\:m-6" | "md\:m-64" | "md\:m-8" | "md\:m-auto" | "md\:m-px" | "md\:max-h-full" | "md\:max-h-screen" | "md\:max-w-2xl" | "md\:max-w-3xl" | "md\:max-w-4xl" | "md\:max-w-5xl" | "md\:max-w-6xl" | "md\:max-w-full" | "md\:max-w-lg" | "md\:max-w-md" | "md\:max-w-sm" | "md\:max-w-xl" | "md\:max-w-xs" | "md\:mb-0" | "md\:mb-1" | "md\:mb-10" | "md\:mb-12" | "md\:mb-16" | "md\:mb-2" | "md\:mb-20" | "md\:mb-24" | "md\:mb-3" | "md\:mb-32" | "md\:mb-4" | "md\:mb-40" | "md\:mb-48" | "md\:mb-5" | "md\:mb-56" | "md\:mb-6" | "md\:mb-64" | "md\:mb-8" | "md\:mb-auto" | "md\:mb-px" | "md\:min-h-0" | "md\:min-h-full" | "md\:min-h-screen" | "md\:min-w-0" | "md\:min-w-full" | "md\:ml-0" | "md\:ml-1" | "md\:ml-10" | "md\:ml-12" | "md\:ml-16" | "md\:ml-2" | "md\:ml-20" | "md\:ml-24" | "md\:ml-3" | "md\:ml-32" | "md\:ml-4" | "md\:ml-40" | "md\:ml-48" | "md\:ml-5" | "md\:ml-56" | "md\:ml-6" | "md\:ml-64" | "md\:ml-8" | "md\:ml-auto" | "md\:ml-px" | "md\:mr-0" | "md\:mr-1" | "md\:mr-10" | "md\:mr-12" | "md\:mr-16" | "md\:mr-2" | "md\:mr-20" | "md\:mr-24" | "md\:mr-3" | "md\:mr-32" | "md\:mr-4" | "md\:mr-40" | "md\:mr-48" | "md\:mr-5" | "md\:mr-56" | "md\:mr-6" | "md\:mr-64" | "md\:mr-8" | "md\:mr-auto" | "md\:mr-px" | "md\:mt-0" | "md\:mt-1" | "md\:mt-10" | "md\:mt-12" | "md\:mt-16" | "md\:mt-2" | "md\:mt-20" | "md\:mt-24" | "md\:mt-3" | "md\:mt-32" | "md\:mt-4" | "md\:mt-40" | "md\:mt-48" | "md\:mt-5" | "md\:mt-56" | "md\:mt-6" | "md\:mt-64" | "md\:mt-8" | "md\:mt-auto" | "md\:mt-px" | "md\:mx-0" | "md\:mx-1" | "md\:mx-10" | "md\:mx-12" | "md\:mx-16" | "md\:mx-2" | "md\:mx-20" | "md\:mx-24" | "md\:mx-3" | "md\:mx-32" | "md\:mx-4" | "md\:mx-40" | "md\:mx-48" | "md\:mx-5" | "md\:mx-56" | "md\:mx-6" | "md\:mx-64" | "md\:mx-8" | "md\:mx-auto" | "md\:mx-px" | "md\:my-0" | "md\:my-1" | "md\:my-10" | "md\:my-12" | "md\:my-16" | "md\:my-2" | "md\:my-20" | "md\:my-24" | "md\:my-3" | "md\:my-32" | "md\:my-4" | "md\:my-40" | "md\:my-48" | "md\:my-5" | "md\:my-56" | "md\:my-6" | "md\:my-64" | "md\:my-8" | "md\:my-auto" | "md\:my-px" | "md\:no-common-ligatures" | "md\:no-contextual-ligatures" | "md\:no-discretionary-ligatures" | "md\:no-ellipsis" | "md\:no-historical-ligatures" | "md\:no-ligatures" | "md\:no-underline" | "md\:normal-caps" | "md\:normal-case" | "md\:normal-ligatures" | "md\:normal-nums" | "md\:not-italic" | "md\:not-sr-only" | "md\:object-bottom" | "md\:object-center" | "md\:object-contain" | "md\:object-cover" | "md\:object-fill" | "md\:object-left" | "md\:object-left-bottom" | "md\:object-left-top" | "md\:object-none" | "md\:object-right" | "md\:object-right-bottom" | "md\:object-right-top" | "md\:object-scale-down" | "md\:object-top" | "md\:oldstyle-nums" | "md\:opacity-0" | "md\:opacity-100" | "md\:opacity-25" | "md\:opacity-50" | "md\:opacity-75" | "md\:order-1" | "md\:order-10" | "md\:order-11" | "md\:order-12" | "md\:order-2" | "md\:order-3" | "md\:order-4" | "md\:order-5" | "md\:order-6" | "md\:order-7" | "md\:order-8" | "md\:order-9" | "md\:order-first" | "md\:order-last" | "md\:order-none" | "md\:ordinal-nums" | "md\:outline-none" | "md\:overflow-auto" | "md\:overflow-hidden" | "md\:overflow-scroll" | "md\:overflow-visible" | "md\:overflow-x-auto" | "md\:overflow-x-hidden" | "md\:overflow-x-scroll" | "md\:overflow-x-visible" | "md\:overflow-y-auto" | "md\:overflow-y-hidden" | "md\:overflow-y-scroll" | "md\:overflow-y-visible" | "md\:p-0" | "md\:p-1" | "md\:p-10" | "md\:p-12" | "md\:p-16" | "md\:p-2" | "md\:p-20" | "md\:p-24" | "md\:p-3" | "md\:p-32" | "md\:p-4" | "md\:p-40" | "md\:p-48" | "md\:p-5" | "md\:p-56" | "md\:p-6" | "md\:p-64" | "md\:p-8" | "md\:p-px" | "md\:pb-0" | "md\:pb-1" | "md\:pb-10" | "md\:pb-12" | "md\:pb-16" | "md\:pb-2" | "md\:pb-20" | "md\:pb-24" | "md\:pb-3" | "md\:pb-32" | "md\:pb-4" | "md\:pb-40" | "md\:pb-48" | "md\:pb-5" | "md\:pb-56" | "md\:pb-6" | "md\:pb-64" | "md\:pb-8" | "md\:pb-px" | "md\:petite-caps" | "md\:pl-0" | "md\:pl-1" | "md\:pl-10" | "md\:pl-12" | "md\:pl-16" | "md\:pl-2" | "md\:pl-20" | "md\:pl-24" | "md\:pl-3" | "md\:pl-32" | "md\:pl-4" | "md\:pl-40" | "md\:pl-48" | "md\:pl-5" | "md\:pl-56" | "md\:pl-6" | "md\:pl-64" | "md\:pl-8" | "md\:pl-px" | "md\:placeholder-black" | "md\:placeholder-blue-100" | "md\:placeholder-blue-200" | "md\:placeholder-blue-300" | "md\:placeholder-blue-400" | "md\:placeholder-blue-500" | "md\:placeholder-blue-600" | "md\:placeholder-blue-700" | "md\:placeholder-blue-800" | "md\:placeholder-blue-900" | "md\:placeholder-border" | "md\:placeholder-gray-100" | "md\:placeholder-gray-200" | "md\:placeholder-gray-300" | "md\:placeholder-gray-400" | "md\:placeholder-gray-500" | "md\:placeholder-gray-600" | "md\:placeholder-gray-700" | "md\:placeholder-gray-800" | "md\:placeholder-gray-900" | "md\:placeholder-primary" | "md\:placeholder-white" | "md\:pointer-events-auto" | "md\:pointer-events-none" | "md\:pr-0" | "md\:pr-1" | "md\:pr-10" | "md\:pr-12" | "md\:pr-16" | "md\:pr-2" | "md\:pr-20" | "md\:pr-24" | "md\:pr-3" | "md\:pr-32" | "md\:pr-4" | "md\:pr-40" | "md\:pr-48" | "md\:pr-5" | "md\:pr-56" | "md\:pr-6" | "md\:pr-64" | "md\:pr-8" | "md\:pr-px" | "md\:proportional-nums" | "md\:pt-0" | "md\:pt-1" | "md\:pt-10" | "md\:pt-12" | "md\:pt-16" | "md\:pt-2" | "md\:pt-20" | "md\:pt-24" | "md\:pt-3" | "md\:pt-32" | "md\:pt-4" | "md\:pt-40" | "md\:pt-48" | "md\:pt-5" | "md\:pt-56" | "md\:pt-6" | "md\:pt-64" | "md\:pt-8" | "md\:pt-px" | "md\:px-0" | "md\:px-1" | "md\:px-10" | "md\:px-12" | "md\:px-16" | "md\:px-2" | "md\:px-20" | "md\:px-24" | "md\:px-3" | "md\:px-32" | "md\:px-4" | "md\:px-40" | "md\:px-48" | "md\:px-5" | "md\:px-56" | "md\:px-6" | "md\:px-64" | "md\:px-8" | "md\:px-px" | "md\:py-0" | "md\:py-1" | "md\:py-10" | "md\:py-12" | "md\:py-16" | "md\:py-2" | "md\:py-20" | "md\:py-24" | "md\:py-3" | "md\:py-32" | "md\:py-4" | "md\:py-40" | "md\:py-48" | "md\:py-5" | "md\:py-56" | "md\:py-6" | "md\:py-64" | "md\:py-8" | "md\:py-px" | "md\:relative" | "md\:resize" | "md\:resize-none" | "md\:resize-x" | "md\:resize-y" | "md\:right-0" | "md\:right-auto" | "md\:rotate-180" | "md\:rotate-270" | "md\:rotate-90" | "md\:rounded" | "md\:rounded-b" | "md\:rounded-b-full" | "md\:rounded-b-lg" | "md\:rounded-b-none" | "md\:rounded-b-sm" | "md\:rounded-bl" | "md\:rounded-bl-full" | "md\:rounded-bl-lg" | "md\:rounded-bl-none" | "md\:rounded-bl-sm" | "md\:rounded-br" | "md\:rounded-br-full" | "md\:rounded-br-lg" | "md\:rounded-br-none" | "md\:rounded-br-sm" | "md\:rounded-full" | "md\:rounded-l" | "md\:rounded-l-full" | "md\:rounded-l-lg" | "md\:rounded-l-none" | "md\:rounded-l-sm" | "md\:rounded-lg" | "md\:rounded-none" | "md\:rounded-r" | "md\:rounded-r-full" | "md\:rounded-r-lg" | "md\:rounded-r-none" | "md\:rounded-r-sm" | "md\:rounded-sm" | "md\:rounded-t" | "md\:rounded-t-full" | "md\:rounded-t-lg" | "md\:rounded-t-none" | "md\:rounded-t-sm" | "md\:rounded-tl" | "md\:rounded-tl-full" | "md\:rounded-tl-lg" | "md\:rounded-tl-none" | "md\:rounded-tl-sm" | "md\:rounded-tr" | "md\:rounded-tr-full" | "md\:rounded-tr-lg" | "md\:rounded-tr-none" | "md\:rounded-tr-sm" | "md\:scrolling-auto" | "md\:scrolling-touch" | "md\:select-all" | "md\:select-auto" | "md\:select-none" | "md\:select-text" | "md\:self-auto" | "md\:self-center" | "md\:self-end" | "md\:self-start" | "md\:self-stretch" | "md\:shadow" | "md\:shadow-2xl" | "md\:shadow-inner" | "md\:shadow-lg" | "md\:shadow-md" | "md\:shadow-none" | "md\:shadow-outline" | "md\:shadow-xl" | "md\:slashed-zeros" | "md\:small-caps" | "md\:sr-only" | "md\:stacked-fractions" | "md\:static" | "md\:sticky" | "md\:stroke-current" | "md\:subpixel-antialiased" | "md\:table" | "md\:table-auto" | "md\:table-cell" | "md\:table-fixed" | "md\:table-row" | "md\:tabular-nums" | "md\:text-2xl" | "md\:text-3xl" | "md\:text-4xl" | "md\:text-5xl" | "md\:text-6xl" | "md\:text-align-unset" | "md\:text-base" | "md\:text-black" | "md\:text-blue-100" | "md\:text-blue-200" | "md\:text-blue-300" | "md\:text-blue-400" | "md\:text-blue-500" | "md\:text-blue-600" | "md\:text-blue-700" | "md\:text-blue-800" | "md\:text-blue-900" | "md\:text-border" | "md\:text-center" | "md\:text-color-unset" | "md\:text-gray-100" | "md\:text-gray-200" | "md\:text-gray-300" | "md\:text-gray-400" | "md\:text-gray-500" | "md\:text-gray-600" | "md\:text-gray-700" | "md\:text-gray-800" | "md\:text-gray-900" | "md\:text-justify" | "md\:text-left" | "md\:text-lg" | "md\:text-primary" | "md\:text-right" | "md\:text-size-unset" | "md\:text-sm" | "md\:text-transform-unset" | "md\:text-white" | "md\:text-xl" | "md\:text-xs" | "md\:titling-caps" | "md\:top-0" | "md\:top-auto" | "md\:tracking-normal" | "md\:tracking-tight" | "md\:tracking-tighter" | "md\:tracking-unset" | "md\:tracking-wide" | "md\:tracking-wider" | "md\:tracking-widest" | "md\:transform-b" | "md\:transform-bl" | "md\:transform-border" | "md\:transform-br" | "md\:transform-fill" | "md\:transform-l" | "md\:transform-none" | "md\:transform-r" | "md\:transform-t" | "md\:transform-tl" | "md\:transform-tr" | "md\:transform-view" | "md\:transition-0" | "md\:transition-100" | "md\:transition-1000" | "md\:transition-250" | "md\:transition-500" | "md\:transition-750" | "md\:transition-all" | "md\:transition-bg" | "md\:transition-border" | "md\:transition-color" | "md\:transition-colors" | "md\:transition-delay-0" | "md\:transition-delay-100" | "md\:transition-delay-1000" | "md\:transition-delay-250" | "md\:transition-delay-500" | "md\:transition-delay-750" | "md\:transition-ease" | "md\:transition-ease-in" | "md\:transition-ease-in-out" | "md\:transition-ease-out" | "md\:transition-linear" | "md\:transition-none" | "md\:transition-opacity" | "md\:transition-transform" | "md\:truncate" | "md\:underline" | "md\:unicase" | "md\:uppercase" | "md\:visible" | "md\:w-0" | "md\:w-1" | "md\:w-10" | "md\:w-10\/12" | "md\:w-11\/12" | "md\:w-12" | "md\:w-16" | "md\:w-1\/12" | "md\:w-1\/2" | "md\:w-1\/3" | "md\:w-1\/4" | "md\:w-1\/5" | "md\:w-1\/6" | "md\:w-2" | "md\:w-20" | "md\:w-24" | "md\:w-2\/12" | "md\:w-2\/3" | "md\:w-2\/4" | "md\:w-2\/5" | "md\:w-2\/6" | "md\:w-3" | "md\:w-32" | "md\:w-3\/12" | "md\:w-3\/4" | "md\:w-3\/5" | "md\:w-3\/6" | "md\:w-4" | "md\:w-40" | "md\:w-48" | "md\:w-4\/12" | "md\:w-4\/5" | "md\:w-4\/6" | "md\:w-5" | "md\:w-56" | "md\:w-5\/12" | "md\:w-5\/6" | "md\:w-6" | "md\:w-64" | "md\:w-6\/12" | "md\:w-7\/12" | "md\:w-8" | "md\:w-8\/12" | "md\:w-9\/12" | "md\:w-auto" | "md\:w-full" | "md\:w-px" | "md\:w-screen" | "md\:whitespace-no-wrap" | "md\:whitespace-normal" | "md\:whitespace-pre" | "md\:whitespace-pre-line" | "md\:whitespace-pre-wrap" | "md\:will-change-auto" | "md\:will-change-contents" | "md\:will-change-opacity" | "md\:will-change-scroll" | "md\:will-change-transform" | "md\:z-0" | "md\:z-10" | "md\:z-20" | "md\:z-30" | "md\:z-40" | "md\:z-50" | "md\:z-auto" | "min-h-0" | "min-h-full" | "min-h-screen" | "min-w-0" | "min-w-full" | "ml-0" | "ml-1" | "ml-10" | "ml-12" | "ml-16" | "ml-2" | "ml-20" | "ml-24" | "ml-3" | "ml-32" | "ml-4" | "ml-40" | "ml-48" | "ml-5" | "ml-56" | "ml-6" | "ml-64" | "ml-8" | "ml-auto" | "ml-px" | "mr-0" | "mr-1" | "mr-10" | "mr-12" | "mr-16" | "mr-2" | "mr-20" | "mr-24" | "mr-3" | "mr-32" | "mr-4" | "mr-40" | "mr-48" | "mr-5" | "mr-56" | "mr-6" | "mr-64" | "mr-8" | "mr-auto" | "mr-px" | "mt-0" | "mt-1" | "mt-10" | "mt-12" | "mt-16" | "mt-2" | "mt-20" | "mt-24" | "mt-3" | "mt-32" | "mt-4" | "mt-40" | "mt-48" | "mt-5" | "mt-56" | "mt-6" | "mt-64" | "mt-8" | "mt-auto" | "mt-px" | "mx-0" | "mx-1" | "mx-10" | "mx-12" | "mx-16" | "mx-2" | "mx-20" | "mx-24" | "mx-3" | "mx-32" | "mx-4" | "mx-40" | "mx-48" | "mx-5" | "mx-56" | "mx-6" | "mx-64" | "mx-8" | "mx-auto" | "mx-px" | "my-0" | "my-1" | "my-10" | "my-12" | "my-16" | "my-2" | "my-20" | "my-24" | "my-3" | "my-32" | "my-4" | "my-40" | "my-48" | "my-5" | "my-56" | "my-6" | "my-64" | "my-8" | "my-auto" | "my-px" | "no-common-ligatures" | "no-contextual-ligatures" | "no-discretionary-ligatures" | "no-ellipsis" | "no-historical-ligatures" | "no-ligatures" | "no-underline" | "normal-caps" | "normal-case" | "normal-ligatures" | "normal-nums" | "not-italic" | "not-sr-only" | "object-bottom" | "object-center" | "object-contain" | "object-cover" | "object-fill" | "object-left" | "object-left-bottom" | "object-left-top" | "object-none" | "object-right" | "object-right-bottom" | "object-right-top" | "object-scale-down" | "object-top" | "oldstyle-nums" | "opacity-0" | "opacity-100" | "opacity-25" | "opacity-50" | "opacity-75" | "order-1" | "order-10" | "order-11" | "order-12" | "order-2" | "order-3" | "order-4" | "order-5" | "order-6" | "order-7" | "order-8" | "order-9" | "order-first" | "order-last" | "order-none" | "ordinal-nums" | "outline-none" | "overflow-auto" | "overflow-hidden" | "overflow-scroll" | "overflow-visible" | "overflow-x-auto" | "overflow-x-hidden" | "overflow-x-scroll" | "overflow-x-visible" | "overflow-y-auto" | "overflow-y-hidden" | "overflow-y-scroll" | "overflow-y-visible" | "p-0" | "p-1" | "p-10" | "p-12" | "p-16" | "p-2" | "p-20" | "p-24" | "p-3" | "p-32" | "p-4" | "p-40" | "p-48" | "p-5" | "p-56" | "p-6" | "p-64" | "p-8" | "p-px" | "pb-0" | "pb-1" | "pb-10" | "pb-12" | "pb-16" | "pb-2" | "pb-20" | "pb-24" | "pb-3" | "pb-32" | "pb-4" | "pb-40" | "pb-48" | "pb-5" | "pb-56" | "pb-6" | "pb-64" | "pb-8" | "pb-px" | "petite-caps" | "pl-0" | "pl-1" | "pl-10" | "pl-12" | "pl-16" | "pl-2" | "pl-20" | "pl-24" | "pl-3" | "pl-32" | "pl-4" | "pl-40" | "pl-48" | "pl-5" | "pl-56" | "pl-6" | "pl-64" | "pl-8" | "pl-px" | "placeholder-black" | "placeholder-blue-100" | "placeholder-blue-200" | "placeholder-blue-300" | "placeholder-blue-400" | "placeholder-blue-500" | "placeholder-blue-600" | "placeholder-blue-700" | "placeholder-blue-800" | "placeholder-blue-900" | "placeholder-border" | "placeholder-gray-100" | "placeholder-gray-200" | "placeholder-gray-300" | "placeholder-gray-400" | "placeholder-gray-500" | "placeholder-gray-600" | "placeholder-gray-700" | "placeholder-gray-800" | "placeholder-gray-900" | "placeholder-primary" | "placeholder-white" | "pointer-events-auto" | "pointer-events-none" | "pr-0" | "pr-1" | "pr-10" | "pr-12" | "pr-16" | "pr-2" | "pr-20" | "pr-24" | "pr-3" | "pr-32" | "pr-4" | "pr-40" | "pr-48" | "pr-5" | "pr-56" | "pr-6" | "pr-64" | "pr-8" | "pr-px" | "proportional-nums" | "pt-0" | "pt-1" | "pt-10" | "pt-12" | "pt-16" | "pt-2" | "pt-20" | "pt-24" | "pt-3" | "pt-32" | "pt-4" | "pt-40" | "pt-48" | "pt-5" | "pt-56" | "pt-6" | "pt-64" | "pt-8" | "pt-px" | "px-0" | "px-1" | "px-10" | "px-12" | "px-16" | "px-2" | "px-20" | "px-24" | "px-3" | "px-32" | "px-4" | "px-40" | "px-48" | "px-5" | "px-56" | "px-6" | "px-64" | "px-8" | "px-px" | "py-0" | "py-1" | "py-10" | "py-12" | "py-16" | "py-2" | "py-20" | "py-24" | "py-3" | "py-32" | "py-4" | "py-40" | "py-48" | "py-5" | "py-56" | "py-6" | "py-64" | "py-8" | "py-px" | "relative" | "resize" | "resize-none" | "resize-x" | "resize-y" | "rfs-control-container" | "rfs-select-container" | "right-0" | "right-auto" | "rotate-180" | "rotate-270" | "rotate-90" | "rounded" | "rounded-b" | "rounded-b-full" | "rounded-b-lg" | "rounded-b-none" | "rounded-b-sm" | "rounded-bl" | "rounded-bl-full" | "rounded-bl-lg" | "rounded-bl-none" | "rounded-bl-sm" | "rounded-br" | "rounded-br-full" | "rounded-br-lg" | "rounded-br-none" | "rounded-br-sm" | "rounded-full" | "rounded-l" | "rounded-l-full" | "rounded-l-lg" | "rounded-l-none" | "rounded-l-sm" | "rounded-lg" | "rounded-none" | "rounded-r" | "rounded-r-full" | "rounded-r-lg" | "rounded-r-none" | "rounded-r-sm" | "rounded-sm" | "rounded-t" | "rounded-t-full" | "rounded-t-lg" | "rounded-t-none" | "rounded-t-sm" | "rounded-tl" | "rounded-tl-full" | "rounded-tl-lg" | "rounded-tl-none" | "rounded-tl-sm" | "rounded-tr" | "rounded-tr-full" | "rounded-tr-lg" | "rounded-tr-none" | "rounded-tr-sm" | "scrolling-auto" | "scrolling-touch" | "select-all" | "select-auto" | "select-none" | "select-text" | "self-auto" | "self-center" | "self-end" | "self-start" | "self-stretch" | "shadow" | "shadow-2xl" | "shadow-inner" | "shadow-lg" | "shadow-md" | "shadow-none" | "shadow-outline" | "shadow-xl" | "slashed-zeros" | "sm\:-m-1" | "sm\:-m-10" | "sm\:-m-12" | "sm\:-m-16" | "sm\:-m-2" | "sm\:-m-20" | "sm\:-m-24" | "sm\:-m-3" | "sm\:-m-32" | "sm\:-m-4" | "sm\:-m-40" | "sm\:-m-48" | "sm\:-m-5" | "sm\:-m-56" | "sm\:-m-6" | "sm\:-m-64" | "sm\:-m-8" | "sm\:-m-px" | "sm\:-mb-1" | "sm\:-mb-10" | "sm\:-mb-12" | "sm\:-mb-16" | "sm\:-mb-2" | "sm\:-mb-20" | "sm\:-mb-24" | "sm\:-mb-3" | "sm\:-mb-32" | "sm\:-mb-4" | "sm\:-mb-40" | "sm\:-mb-48" | "sm\:-mb-5" | "sm\:-mb-56" | "sm\:-mb-6" | "sm\:-mb-64" | "sm\:-mb-8" | "sm\:-mb-px" | "sm\:-ml-1" | "sm\:-ml-10" | "sm\:-ml-12" | "sm\:-ml-16" | "sm\:-ml-2" | "sm\:-ml-20" | "sm\:-ml-24" | "sm\:-ml-3" | "sm\:-ml-32" | "sm\:-ml-4" | "sm\:-ml-40" | "sm\:-ml-48" | "sm\:-ml-5" | "sm\:-ml-56" | "sm\:-ml-6" | "sm\:-ml-64" | "sm\:-ml-8" | "sm\:-ml-px" | "sm\:-mr-1" | "sm\:-mr-10" | "sm\:-mr-12" | "sm\:-mr-16" | "sm\:-mr-2" | "sm\:-mr-20" | "sm\:-mr-24" | "sm\:-mr-3" | "sm\:-mr-32" | "sm\:-mr-4" | "sm\:-mr-40" | "sm\:-mr-48" | "sm\:-mr-5" | "sm\:-mr-56" | "sm\:-mr-6" | "sm\:-mr-64" | "sm\:-mr-8" | "sm\:-mr-px" | "sm\:-mt-1" | "sm\:-mt-10" | "sm\:-mt-12" | "sm\:-mt-16" | "sm\:-mt-2" | "sm\:-mt-20" | "sm\:-mt-24" | "sm\:-mt-3" | "sm\:-mt-32" | "sm\:-mt-4" | "sm\:-mt-40" | "sm\:-mt-48" | "sm\:-mt-5" | "sm\:-mt-56" | "sm\:-mt-6" | "sm\:-mt-64" | "sm\:-mt-8" | "sm\:-mt-px" | "sm\:-mx-1" | "sm\:-mx-10" | "sm\:-mx-12" | "sm\:-mx-16" | "sm\:-mx-2" | "sm\:-mx-20" | "sm\:-mx-24" | "sm\:-mx-3" | "sm\:-mx-32" | "sm\:-mx-4" | "sm\:-mx-40" | "sm\:-mx-48" | "sm\:-mx-5" | "sm\:-mx-56" | "sm\:-mx-6" | "sm\:-mx-64" | "sm\:-mx-8" | "sm\:-mx-px" | "sm\:-my-1" | "sm\:-my-10" | "sm\:-my-12" | "sm\:-my-16" | "sm\:-my-2" | "sm\:-my-20" | "sm\:-my-24" | "sm\:-my-3" | "sm\:-my-32" | "sm\:-my-4" | "sm\:-my-40" | "sm\:-my-48" | "sm\:-my-5" | "sm\:-my-56" | "sm\:-my-6" | "sm\:-my-64" | "sm\:-my-8" | "sm\:-my-px" | "sm\:absolute" | "sm\:align-baseline" | "sm\:align-bottom" | "sm\:align-middle" | "sm\:align-text-bottom" | "sm\:align-text-top" | "sm\:align-top" | "sm\:all-small-caps" | "sm\:antialiased" | "sm\:appearance-none" | "sm\:bg-auto" | "sm\:bg-black" | "sm\:bg-blue-100" | "sm\:bg-blue-200" | "sm\:bg-blue-300" | "sm\:bg-blue-400" | "sm\:bg-blue-500" | "sm\:bg-blue-600" | "sm\:bg-blue-700" | "sm\:bg-blue-800" | "sm\:bg-blue-900" | "sm\:bg-border" | "sm\:bg-bottom" | "sm\:bg-center" | "sm\:bg-contain" | "sm\:bg-cover" | "sm\:bg-fixed" | "sm\:bg-gray-100" | "sm\:bg-gray-200" | "sm\:bg-gray-300" | "sm\:bg-gray-400" | "sm\:bg-gray-500" | "sm\:bg-gray-600" | "sm\:bg-gray-700" | "sm\:bg-gray-800" | "sm\:bg-gray-900" | "sm\:bg-left" | "sm\:bg-left-bottom" | "sm\:bg-left-top" | "sm\:bg-local" | "sm\:bg-no-repeat" | "sm\:bg-primary" | "sm\:bg-repeat" | "sm\:bg-repeat-round" | "sm\:bg-repeat-space" | "sm\:bg-repeat-x" | "sm\:bg-repeat-y" | "sm\:bg-right" | "sm\:bg-right-bottom" | "sm\:bg-right-top" | "sm\:bg-scroll" | "sm\:bg-top" | "sm\:bg-white" | "sm\:block" | "sm\:border" | "sm\:border-0" | "sm\:border-2" | "sm\:border-4" | "sm\:border-8" | "sm\:border-b" | "sm\:border-b-0" | "sm\:border-b-2" | "sm\:border-b-4" | "sm\:border-b-8" | "sm\:border-black" | "sm\:border-blue-100" | "sm\:border-blue-200" | "sm\:border-blue-300" | "sm\:border-blue-400" | "sm\:border-blue-500" | "sm\:border-blue-600" | "sm\:border-blue-700" | "sm\:border-blue-800" | "sm\:border-blue-900" | "sm\:border-border" | "sm\:border-collapse" | "sm\:border-dashed" | "sm\:border-dotted" | "sm\:border-double" | "sm\:border-gray-100" | "sm\:border-gray-200" | "sm\:border-gray-300" | "sm\:border-gray-400" | "sm\:border-gray-500" | "sm\:border-gray-600" | "sm\:border-gray-700" | "sm\:border-gray-800" | "sm\:border-gray-900" | "sm\:border-l" | "sm\:border-l-0" | "sm\:border-l-2" | "sm\:border-l-4" | "sm\:border-l-8" | "sm\:border-none" | "sm\:border-primary" | "sm\:border-r" | "sm\:border-r-0" | "sm\:border-r-2" | "sm\:border-r-4" | "sm\:border-r-8" | "sm\:border-separate" | "sm\:border-solid" | "sm\:border-t" | "sm\:border-t-0" | "sm\:border-t-2" | "sm\:border-t-4" | "sm\:border-t-8" | "sm\:border-white" | "sm\:bottom-0" | "sm\:bottom-auto" | "sm\:break-all" | "sm\:break-normal" | "sm\:break-words" | "sm\:capitalize" | "sm\:clearfix" | "sm\:common-ligatures" | "sm\:content-around" | "sm\:content-between" | "sm\:content-center" | "sm\:content-end" | "sm\:content-start" | "sm\:contextual-ligatures" | "sm\:cursor-auto" | "sm\:cursor-default" | "sm\:cursor-move" | "sm\:cursor-not-allowed" | "sm\:cursor-pointer" | "sm\:cursor-text" | "sm\:cursor-wait" | "sm\:diagonal-fractions" | "sm\:discretionary-ligatures" | "sm\:ellipsis" | "sm\:fill-current" | "sm\:fixed" | "sm\:flex" | "sm\:flex-1" | "sm\:flex-auto" | "sm\:flex-col" | "sm\:flex-col-reverse" | "sm\:flex-grow" | "sm\:flex-grow-0" | "sm\:flex-initial" | "sm\:flex-no-wrap" | "sm\:flex-none" | "sm\:flex-row" | "sm\:flex-row-reverse" | "sm\:flex-shrink" | "sm\:flex-shrink-0" | "sm\:flex-wrap" | "sm\:flex-wrap-reverse" | "sm\:float-left" | "sm\:float-none" | "sm\:float-right" | "sm\:focus\:bg-black" | "sm\:focus\:bg-blue-100" | "sm\:focus\:bg-blue-200" | "sm\:focus\:bg-blue-300" | "sm\:focus\:bg-blue-400" | "sm\:focus\:bg-blue-500" | "sm\:focus\:bg-blue-600" | "sm\:focus\:bg-blue-700" | "sm\:focus\:bg-blue-800" | "sm\:focus\:bg-blue-900" | "sm\:focus\:bg-border" | "sm\:focus\:bg-gray-100" | "sm\:focus\:bg-gray-200" | "sm\:focus\:bg-gray-300" | "sm\:focus\:bg-gray-400" | "sm\:focus\:bg-gray-500" | "sm\:focus\:bg-gray-600" | "sm\:focus\:bg-gray-700" | "sm\:focus\:bg-gray-800" | "sm\:focus\:bg-gray-900" | "sm\:focus\:bg-primary" | "sm\:focus\:bg-white" | "sm\:focus\:border" | "sm\:focus\:border-0" | "sm\:focus\:border-2" | "sm\:focus\:border-4" | "sm\:focus\:border-8" | "sm\:focus\:border-b" | "sm\:focus\:border-b-0" | "sm\:focus\:border-b-2" | "sm\:focus\:border-b-4" | "sm\:focus\:border-b-8" | "sm\:focus\:border-black" | "sm\:focus\:border-blue-100" | "sm\:focus\:border-blue-200" | "sm\:focus\:border-blue-300" | "sm\:focus\:border-blue-400" | "sm\:focus\:border-blue-500" | "sm\:focus\:border-blue-600" | "sm\:focus\:border-blue-700" | "sm\:focus\:border-blue-800" | "sm\:focus\:border-blue-900" | "sm\:focus\:border-border" | "sm\:focus\:border-gray-100" | "sm\:focus\:border-gray-200" | "sm\:focus\:border-gray-300" | "sm\:focus\:border-gray-400" | "sm\:focus\:border-gray-500" | "sm\:focus\:border-gray-600" | "sm\:focus\:border-gray-700" | "sm\:focus\:border-gray-800" | "sm\:focus\:border-gray-900" | "sm\:focus\:border-l" | "sm\:focus\:border-l-0" | "sm\:focus\:border-l-2" | "sm\:focus\:border-l-4" | "sm\:focus\:border-l-8" | "sm\:focus\:border-primary" | "sm\:focus\:border-r" | "sm\:focus\:border-r-0" | "sm\:focus\:border-r-2" | "sm\:focus\:border-r-4" | "sm\:focus\:border-r-8" | "sm\:focus\:border-t" | "sm\:focus\:border-t-0" | "sm\:focus\:border-t-2" | "sm\:focus\:border-t-4" | "sm\:focus\:border-t-8" | "sm\:focus\:border-white" | "sm\:focus\:font-black" | "sm\:focus\:font-bold" | "sm\:focus\:font-extrabold" | "sm\:focus\:font-hairline" | "sm\:focus\:font-light" | "sm\:focus\:font-medium" | "sm\:focus\:font-normal" | "sm\:focus\:font-semibold" | "sm\:focus\:font-thin" | "sm\:focus\:line-through" | "sm\:focus\:no-underline" | "sm\:focus\:not-sr-only" | "sm\:focus\:opacity-0" | "sm\:focus\:opacity-100" | "sm\:focus\:opacity-25" | "sm\:focus\:opacity-50" | "sm\:focus\:opacity-75" | "sm\:focus\:outline-none" | "sm\:focus\:placeholder-black" | "sm\:focus\:placeholder-blue-100" | "sm\:focus\:placeholder-blue-200" | "sm\:focus\:placeholder-blue-300" | "sm\:focus\:placeholder-blue-400" | "sm\:focus\:placeholder-blue-500" | "sm\:focus\:placeholder-blue-600" | "sm\:focus\:placeholder-blue-700" | "sm\:focus\:placeholder-blue-800" | "sm\:focus\:placeholder-blue-900" | "sm\:focus\:placeholder-border" | "sm\:focus\:placeholder-gray-100" | "sm\:focus\:placeholder-gray-200" | "sm\:focus\:placeholder-gray-300" | "sm\:focus\:placeholder-gray-400" | "sm\:focus\:placeholder-gray-500" | "sm\:focus\:placeholder-gray-600" | "sm\:focus\:placeholder-gray-700" | "sm\:focus\:placeholder-gray-800" | "sm\:focus\:placeholder-gray-900" | "sm\:focus\:placeholder-primary" | "sm\:focus\:placeholder-white" | "sm\:focus\:shadow" | "sm\:focus\:shadow-2xl" | "sm\:focus\:shadow-inner" | "sm\:focus\:shadow-lg" | "sm\:focus\:shadow-md" | "sm\:focus\:shadow-none" | "sm\:focus\:shadow-outline" | "sm\:focus\:shadow-xl" | "sm\:focus\:sr-only" | "sm\:focus\:text-black" | "sm\:focus\:text-blue-100" | "sm\:focus\:text-blue-200" | "sm\:focus\:text-blue-300" | "sm\:focus\:text-blue-400" | "sm\:focus\:text-blue-500" | "sm\:focus\:text-blue-600" | "sm\:focus\:text-blue-700" | "sm\:focus\:text-blue-800" | "sm\:focus\:text-blue-900" | "sm\:focus\:text-border" | "sm\:focus\:text-gray-100" | "sm\:focus\:text-gray-200" | "sm\:focus\:text-gray-300" | "sm\:focus\:text-gray-400" | "sm\:focus\:text-gray-500" | "sm\:focus\:text-gray-600" | "sm\:focus\:text-gray-700" | "sm\:focus\:text-gray-800" | "sm\:focus\:text-gray-900" | "sm\:focus\:text-primary" | "sm\:focus\:text-white" | "sm\:focus\:underline" | "sm\:font-black" | "sm\:font-bold" | "sm\:font-extrabold" | "sm\:font-family-unset" | "sm\:font-hairline" | "sm\:font-light" | "sm\:font-medium" | "sm\:font-mono" | "sm\:font-normal" | "sm\:font-sans" | "sm\:font-semibold" | "sm\:font-serif" | "sm\:font-style-unset" | "sm\:font-thin" | "sm\:font-weight-unset" | "sm\:h-0" | "sm\:h-1" | "sm\:h-10" | "sm\:h-12" | "sm\:h-16" | "sm\:h-2" | "sm\:h-20" | "sm\:h-24" | "sm\:h-3" | "sm\:h-32" | "sm\:h-4" | "sm\:h-40" | "sm\:h-48" | "sm\:h-5" | "sm\:h-56" | "sm\:h-6" | "sm\:h-64" | "sm\:h-8" | "sm\:h-auto" | "sm\:h-full" | "sm\:h-px" | "sm\:h-screen" | "sm\:hidden" | "sm\:historical-ligatures" | "sm\:hover\:bg-black" | "sm\:hover\:bg-blue-100" | "sm\:hover\:bg-blue-200" | "sm\:hover\:bg-blue-300" | "sm\:hover\:bg-blue-400" | "sm\:hover\:bg-blue-500" | "sm\:hover\:bg-blue-600" | "sm\:hover\:bg-blue-700" | "sm\:hover\:bg-blue-800" | "sm\:hover\:bg-blue-900" | "sm\:hover\:bg-border" | "sm\:hover\:bg-gray-100" | "sm\:hover\:bg-gray-200" | "sm\:hover\:bg-gray-300" | "sm\:hover\:bg-gray-400" | "sm\:hover\:bg-gray-500" | "sm\:hover\:bg-gray-600" | "sm\:hover\:bg-gray-700" | "sm\:hover\:bg-gray-800" | "sm\:hover\:bg-gray-900" | "sm\:hover\:bg-primary" | "sm\:hover\:bg-white" | "sm\:hover\:border" | "sm\:hover\:border-0" | "sm\:hover\:border-2" | "sm\:hover\:border-4" | "sm\:hover\:border-8" | "sm\:hover\:border-b" | "sm\:hover\:border-b-0" | "sm\:hover\:border-b-2" | "sm\:hover\:border-b-4" | "sm\:hover\:border-b-8" | "sm\:hover\:border-black" | "sm\:hover\:border-blue-100" | "sm\:hover\:border-blue-200" | "sm\:hover\:border-blue-300" | "sm\:hover\:border-blue-400" | "sm\:hover\:border-blue-500" | "sm\:hover\:border-blue-600" | "sm\:hover\:border-blue-700" | "sm\:hover\:border-blue-800" | "sm\:hover\:border-blue-900" | "sm\:hover\:border-border" | "sm\:hover\:border-gray-100" | "sm\:hover\:border-gray-200" | "sm\:hover\:border-gray-300" | "sm\:hover\:border-gray-400" | "sm\:hover\:border-gray-500" | "sm\:hover\:border-gray-600" | "sm\:hover\:border-gray-700" | "sm\:hover\:border-gray-800" | "sm\:hover\:border-gray-900" | "sm\:hover\:border-l" | "sm\:hover\:border-l-0" | "sm\:hover\:border-l-2" | "sm\:hover\:border-l-4" | "sm\:hover\:border-l-8" | "sm\:hover\:border-primary" | "sm\:hover\:border-r" | "sm\:hover\:border-r-0" | "sm\:hover\:border-r-2" | "sm\:hover\:border-r-4" | "sm\:hover\:border-r-8" | "sm\:hover\:border-t" | "sm\:hover\:border-t-0" | "sm\:hover\:border-t-2" | "sm\:hover\:border-t-4" | "sm\:hover\:border-t-8" | "sm\:hover\:border-white" | "sm\:hover\:font-black" | "sm\:hover\:font-bold" | "sm\:hover\:font-extrabold" | "sm\:hover\:font-hairline" | "sm\:hover\:font-light" | "sm\:hover\:font-medium" | "sm\:hover\:font-normal" | "sm\:hover\:font-semibold" | "sm\:hover\:font-thin" | "sm\:hover\:line-through" | "sm\:hover\:no-underline" | "sm\:hover\:opacity-0" | "sm\:hover\:opacity-100" | "sm\:hover\:opacity-25" | "sm\:hover\:opacity-50" | "sm\:hover\:opacity-75" | "sm\:hover\:shadow" | "sm\:hover\:shadow-2xl" | "sm\:hover\:shadow-inner" | "sm\:hover\:shadow-lg" | "sm\:hover\:shadow-md" | "sm\:hover\:shadow-none" | "sm\:hover\:shadow-outline" | "sm\:hover\:shadow-xl" | "sm\:hover\:text-black" | "sm\:hover\:text-blue-100" | "sm\:hover\:text-blue-200" | "sm\:hover\:text-blue-300" | "sm\:hover\:text-blue-400" | "sm\:hover\:text-blue-500" | "sm\:hover\:text-blue-600" | "sm\:hover\:text-blue-700" | "sm\:hover\:text-blue-800" | "sm\:hover\:text-blue-900" | "sm\:hover\:text-border" | "sm\:hover\:text-gray-100" | "sm\:hover\:text-gray-200" | "sm\:hover\:text-gray-300" | "sm\:hover\:text-gray-400" | "sm\:hover\:text-gray-500" | "sm\:hover\:text-gray-600" | "sm\:hover\:text-gray-700" | "sm\:hover\:text-gray-800" | "sm\:hover\:text-gray-900" | "sm\:hover\:text-primary" | "sm\:hover\:text-white" | "sm\:hover\:underline" | "sm\:hyphens-auto" | "sm\:hyphens-manual" | "sm\:hyphens-none" | "sm\:inline" | "sm\:inline-block" | "sm\:inline-flex" | "sm\:inset-0" | "sm\:inset-auto" | "sm\:inset-x-0" | "sm\:inset-x-auto" | "sm\:inset-y-0" | "sm\:inset-y-auto" | "sm\:invisible" | "sm\:italic" | "sm\:items-baseline" | "sm\:items-center" | "sm\:items-end" | "sm\:items-start" | "sm\:items-stretch" | "sm\:justify-around" | "sm\:justify-between" | "sm\:justify-center" | "sm\:justify-end" | "sm\:justify-start" | "sm\:last\:border" | "sm\:last\:border-0" | "sm\:last\:border-2" | "sm\:last\:border-4" | "sm\:last\:border-8" | "sm\:last\:border-b" | "sm\:last\:border-b-0" | "sm\:last\:border-b-2" | "sm\:last\:border-b-4" | "sm\:last\:border-b-8" | "sm\:last\:border-l" | "sm\:last\:border-l-0" | "sm\:last\:border-l-2" | "sm\:last\:border-l-4" | "sm\:last\:border-l-8" | "sm\:last\:border-r" | "sm\:last\:border-r-0" | "sm\:last\:border-r-2" | "sm\:last\:border-r-4" | "sm\:last\:border-r-8" | "sm\:last\:border-t" | "sm\:last\:border-t-0" | "sm\:last\:border-t-2" | "sm\:last\:border-t-4" | "sm\:last\:border-t-8" | "sm\:leading-loose" | "sm\:leading-none" | "sm\:leading-normal" | "sm\:leading-relaxed" | "sm\:leading-snug" | "sm\:leading-tight" | "sm\:leading-unset" | "sm\:left-0" | "sm\:left-auto" | "sm\:line-through" | "sm\:lining-nums" | "sm\:list-decimal" | "sm\:list-disc" | "sm\:list-inside" | "sm\:list-none" | "sm\:list-outside" | "sm\:lowercase" | "sm\:m-0" | "sm\:m-1" | "sm\:m-10" | "sm\:m-12" | "sm\:m-16" | "sm\:m-2" | "sm\:m-20" | "sm\:m-24" | "sm\:m-3" | "sm\:m-32" | "sm\:m-4" | "sm\:m-40" | "sm\:m-48" | "sm\:m-5" | "sm\:m-56" | "sm\:m-6" | "sm\:m-64" | "sm\:m-8" | "sm\:m-auto" | "sm\:m-px" | "sm\:max-h-full" | "sm\:max-h-screen" | "sm\:max-w-2xl" | "sm\:max-w-3xl" | "sm\:max-w-4xl" | "sm\:max-w-5xl" | "sm\:max-w-6xl" | "sm\:max-w-full" | "sm\:max-w-lg" | "sm\:max-w-md" | "sm\:max-w-sm" | "sm\:max-w-xl" | "sm\:max-w-xs" | "sm\:mb-0" | "sm\:mb-1" | "sm\:mb-10" | "sm\:mb-12" | "sm\:mb-16" | "sm\:mb-2" | "sm\:mb-20" | "sm\:mb-24" | "sm\:mb-3" | "sm\:mb-32" | "sm\:mb-4" | "sm\:mb-40" | "sm\:mb-48" | "sm\:mb-5" | "sm\:mb-56" | "sm\:mb-6" | "sm\:mb-64" | "sm\:mb-8" | "sm\:mb-auto" | "sm\:mb-px" | "sm\:min-h-0" | "sm\:min-h-full" | "sm\:min-h-screen" | "sm\:min-w-0" | "sm\:min-w-full" | "sm\:ml-0" | "sm\:ml-1" | "sm\:ml-10" | "sm\:ml-12" | "sm\:ml-16" | "sm\:ml-2" | "sm\:ml-20" | "sm\:ml-24" | "sm\:ml-3" | "sm\:ml-32" | "sm\:ml-4" | "sm\:ml-40" | "sm\:ml-48" | "sm\:ml-5" | "sm\:ml-56" | "sm\:ml-6" | "sm\:ml-64" | "sm\:ml-8" | "sm\:ml-auto" | "sm\:ml-px" | "sm\:mr-0" | "sm\:mr-1" | "sm\:mr-10" | "sm\:mr-12" | "sm\:mr-16" | "sm\:mr-2" | "sm\:mr-20" | "sm\:mr-24" | "sm\:mr-3" | "sm\:mr-32" | "sm\:mr-4" | "sm\:mr-40" | "sm\:mr-48" | "sm\:mr-5" | "sm\:mr-56" | "sm\:mr-6" | "sm\:mr-64" | "sm\:mr-8" | "sm\:mr-auto" | "sm\:mr-px" | "sm\:mt-0" | "sm\:mt-1" | "sm\:mt-10" | "sm\:mt-12" | "sm\:mt-16" | "sm\:mt-2" | "sm\:mt-20" | "sm\:mt-24" | "sm\:mt-3" | "sm\:mt-32" | "sm\:mt-4" | "sm\:mt-40" | "sm\:mt-48" | "sm\:mt-5" | "sm\:mt-56" | "sm\:mt-6" | "sm\:mt-64" | "sm\:mt-8" | "sm\:mt-auto" | "sm\:mt-px" | "sm\:mx-0" | "sm\:mx-1" | "sm\:mx-10" | "sm\:mx-12" | "sm\:mx-16" | "sm\:mx-2" | "sm\:mx-20" | "sm\:mx-24" | "sm\:mx-3" | "sm\:mx-32" | "sm\:mx-4" | "sm\:mx-40" | "sm\:mx-48" | "sm\:mx-5" | "sm\:mx-56" | "sm\:mx-6" | "sm\:mx-64" | "sm\:mx-8" | "sm\:mx-auto" | "sm\:mx-px" | "sm\:my-0" | "sm\:my-1" | "sm\:my-10" | "sm\:my-12" | "sm\:my-16" | "sm\:my-2" | "sm\:my-20" | "sm\:my-24" | "sm\:my-3" | "sm\:my-32" | "sm\:my-4" | "sm\:my-40" | "sm\:my-48" | "sm\:my-5" | "sm\:my-56" | "sm\:my-6" | "sm\:my-64" | "sm\:my-8" | "sm\:my-auto" | "sm\:my-px" | "sm\:no-common-ligatures" | "sm\:no-contextual-ligatures" | "sm\:no-discretionary-ligatures" | "sm\:no-ellipsis" | "sm\:no-historical-ligatures" | "sm\:no-ligatures" | "sm\:no-underline" | "sm\:normal-caps" | "sm\:normal-case" | "sm\:normal-ligatures" | "sm\:normal-nums" | "sm\:not-italic" | "sm\:not-sr-only" | "sm\:object-bottom" | "sm\:object-center" | "sm\:object-contain" | "sm\:object-cover" | "sm\:object-fill" | "sm\:object-left" | "sm\:object-left-bottom" | "sm\:object-left-top" | "sm\:object-none" | "sm\:object-right" | "sm\:object-right-bottom" | "sm\:object-right-top" | "sm\:object-scale-down" | "sm\:object-top" | "sm\:oldstyle-nums" | "sm\:opacity-0" | "sm\:opacity-100" | "sm\:opacity-25" | "sm\:opacity-50" | "sm\:opacity-75" | "sm\:order-1" | "sm\:order-10" | "sm\:order-11" | "sm\:order-12" | "sm\:order-2" | "sm\:order-3" | "sm\:order-4" | "sm\:order-5" | "sm\:order-6" | "sm\:order-7" | "sm\:order-8" | "sm\:order-9" | "sm\:order-first" | "sm\:order-last" | "sm\:order-none" | "sm\:ordinal-nums" | "sm\:outline-none" | "sm\:overflow-auto" | "sm\:overflow-hidden" | "sm\:overflow-scroll" | "sm\:overflow-visible" | "sm\:overflow-x-auto" | "sm\:overflow-x-hidden" | "sm\:overflow-x-scroll" | "sm\:overflow-x-visible" | "sm\:overflow-y-auto" | "sm\:overflow-y-hidden" | "sm\:overflow-y-scroll" | "sm\:overflow-y-visible" | "sm\:p-0" | "sm\:p-1" | "sm\:p-10" | "sm\:p-12" | "sm\:p-16" | "sm\:p-2" | "sm\:p-20" | "sm\:p-24" | "sm\:p-3" | "sm\:p-32" | "sm\:p-4" | "sm\:p-40" | "sm\:p-48" | "sm\:p-5" | "sm\:p-56" | "sm\:p-6" | "sm\:p-64" | "sm\:p-8" | "sm\:p-px" | "sm\:pb-0" | "sm\:pb-1" | "sm\:pb-10" | "sm\:pb-12" | "sm\:pb-16" | "sm\:pb-2" | "sm\:pb-20" | "sm\:pb-24" | "sm\:pb-3" | "sm\:pb-32" | "sm\:pb-4" | "sm\:pb-40" | "sm\:pb-48" | "sm\:pb-5" | "sm\:pb-56" | "sm\:pb-6" | "sm\:pb-64" | "sm\:pb-8" | "sm\:pb-px" | "sm\:petite-caps" | "sm\:pl-0" | "sm\:pl-1" | "sm\:pl-10" | "sm\:pl-12" | "sm\:pl-16" | "sm\:pl-2" | "sm\:pl-20" | "sm\:pl-24" | "sm\:pl-3" | "sm\:pl-32" | "sm\:pl-4" | "sm\:pl-40" | "sm\:pl-48" | "sm\:pl-5" | "sm\:pl-56" | "sm\:pl-6" | "sm\:pl-64" | "sm\:pl-8" | "sm\:pl-px" | "sm\:placeholder-black" | "sm\:placeholder-blue-100" | "sm\:placeholder-blue-200" | "sm\:placeholder-blue-300" | "sm\:placeholder-blue-400" | "sm\:placeholder-blue-500" | "sm\:placeholder-blue-600" | "sm\:placeholder-blue-700" | "sm\:placeholder-blue-800" | "sm\:placeholder-blue-900" | "sm\:placeholder-border" | "sm\:placeholder-gray-100" | "sm\:placeholder-gray-200" | "sm\:placeholder-gray-300" | "sm\:placeholder-gray-400" | "sm\:placeholder-gray-500" | "sm\:placeholder-gray-600" | "sm\:placeholder-gray-700" | "sm\:placeholder-gray-800" | "sm\:placeholder-gray-900" | "sm\:placeholder-primary" | "sm\:placeholder-white" | "sm\:pointer-events-auto" | "sm\:pointer-events-none" | "sm\:pr-0" | "sm\:pr-1" | "sm\:pr-10" | "sm\:pr-12" | "sm\:pr-16" | "sm\:pr-2" | "sm\:pr-20" | "sm\:pr-24" | "sm\:pr-3" | "sm\:pr-32" | "sm\:pr-4" | "sm\:pr-40" | "sm\:pr-48" | "sm\:pr-5" | "sm\:pr-56" | "sm\:pr-6" | "sm\:pr-64" | "sm\:pr-8" | "sm\:pr-px" | "sm\:proportional-nums" | "sm\:pt-0" | "sm\:pt-1" | "sm\:pt-10" | "sm\:pt-12" | "sm\:pt-16" | "sm\:pt-2" | "sm\:pt-20" | "sm\:pt-24" | "sm\:pt-3" | "sm\:pt-32" | "sm\:pt-4" | "sm\:pt-40" | "sm\:pt-48" | "sm\:pt-5" | "sm\:pt-56" | "sm\:pt-6" | "sm\:pt-64" | "sm\:pt-8" | "sm\:pt-px" | "sm\:px-0" | "sm\:px-1" | "sm\:px-10" | "sm\:px-12" | "sm\:px-16" | "sm\:px-2" | "sm\:px-20" | "sm\:px-24" | "sm\:px-3" | "sm\:px-32" | "sm\:px-4" | "sm\:px-40" | "sm\:px-48" | "sm\:px-5" | "sm\:px-56" | "sm\:px-6" | "sm\:px-64" | "sm\:px-8" | "sm\:px-px" | "sm\:py-0" | "sm\:py-1" | "sm\:py-10" | "sm\:py-12" | "sm\:py-16" | "sm\:py-2" | "sm\:py-20" | "sm\:py-24" | "sm\:py-3" | "sm\:py-32" | "sm\:py-4" | "sm\:py-40" | "sm\:py-48" | "sm\:py-5" | "sm\:py-56" | "sm\:py-6" | "sm\:py-64" | "sm\:py-8" | "sm\:py-px" | "sm\:relative" | "sm\:resize" | "sm\:resize-none" | "sm\:resize-x" | "sm\:resize-y" | "sm\:right-0" | "sm\:right-auto" | "sm\:rotate-180" | "sm\:rotate-270" | "sm\:rotate-90" | "sm\:rounded" | "sm\:rounded-b" | "sm\:rounded-b-full" | "sm\:rounded-b-lg" | "sm\:rounded-b-none" | "sm\:rounded-b-sm" | "sm\:rounded-bl" | "sm\:rounded-bl-full" | "sm\:rounded-bl-lg" | "sm\:rounded-bl-none" | "sm\:rounded-bl-sm" | "sm\:rounded-br" | "sm\:rounded-br-full" | "sm\:rounded-br-lg" | "sm\:rounded-br-none" | "sm\:rounded-br-sm" | "sm\:rounded-full" | "sm\:rounded-l" | "sm\:rounded-l-full" | "sm\:rounded-l-lg" | "sm\:rounded-l-none" | "sm\:rounded-l-sm" | "sm\:rounded-lg" | "sm\:rounded-none" | "sm\:rounded-r" | "sm\:rounded-r-full" | "sm\:rounded-r-lg" | "sm\:rounded-r-none" | "sm\:rounded-r-sm" | "sm\:rounded-sm" | "sm\:rounded-t" | "sm\:rounded-t-full" | "sm\:rounded-t-lg" | "sm\:rounded-t-none" | "sm\:rounded-t-sm" | "sm\:rounded-tl" | "sm\:rounded-tl-full" | "sm\:rounded-tl-lg" | "sm\:rounded-tl-none" | "sm\:rounded-tl-sm" | "sm\:rounded-tr" | "sm\:rounded-tr-full" | "sm\:rounded-tr-lg" | "sm\:rounded-tr-none" | "sm\:rounded-tr-sm" | "sm\:scrolling-auto" | "sm\:scrolling-touch" | "sm\:select-all" | "sm\:select-auto" | "sm\:select-none" | "sm\:select-text" | "sm\:self-auto" | "sm\:self-center" | "sm\:self-end" | "sm\:self-start" | "sm\:self-stretch" | "sm\:shadow" | "sm\:shadow-2xl" | "sm\:shadow-inner" | "sm\:shadow-lg" | "sm\:shadow-md" | "sm\:shadow-none" | "sm\:shadow-outline" | "sm\:shadow-xl" | "sm\:slashed-zeros" | "sm\:small-caps" | "sm\:sr-only" | "sm\:stacked-fractions" | "sm\:static" | "sm\:sticky" | "sm\:stroke-current" | "sm\:subpixel-antialiased" | "sm\:table" | "sm\:table-auto" | "sm\:table-cell" | "sm\:table-fixed" | "sm\:table-row" | "sm\:tabular-nums" | "sm\:text-2xl" | "sm\:text-3xl" | "sm\:text-4xl" | "sm\:text-5xl" | "sm\:text-6xl" | "sm\:text-align-unset" | "sm\:text-base" | "sm\:text-black" | "sm\:text-blue-100" | "sm\:text-blue-200" | "sm\:text-blue-300" | "sm\:text-blue-400" | "sm\:text-blue-500" | "sm\:text-blue-600" | "sm\:text-blue-700" | "sm\:text-blue-800" | "sm\:text-blue-900" | "sm\:text-border" | "sm\:text-center" | "sm\:text-color-unset" | "sm\:text-gray-100" | "sm\:text-gray-200" | "sm\:text-gray-300" | "sm\:text-gray-400" | "sm\:text-gray-500" | "sm\:text-gray-600" | "sm\:text-gray-700" | "sm\:text-gray-800" | "sm\:text-gray-900" | "sm\:text-justify" | "sm\:text-left" | "sm\:text-lg" | "sm\:text-primary" | "sm\:text-right" | "sm\:text-size-unset" | "sm\:text-sm" | "sm\:text-transform-unset" | "sm\:text-white" | "sm\:text-xl" | "sm\:text-xs" | "sm\:titling-caps" | "sm\:top-0" | "sm\:top-auto" | "sm\:tracking-normal" | "sm\:tracking-tight" | "sm\:tracking-tighter" | "sm\:tracking-unset" | "sm\:tracking-wide" | "sm\:tracking-wider" | "sm\:tracking-widest" | "sm\:transform-b" | "sm\:transform-bl" | "sm\:transform-border" | "sm\:transform-br" | "sm\:transform-fill" | "sm\:transform-l" | "sm\:transform-none" | "sm\:transform-r" | "sm\:transform-t" | "sm\:transform-tl" | "sm\:transform-tr" | "sm\:transform-view" | "sm\:transition-0" | "sm\:transition-100" | "sm\:transition-1000" | "sm\:transition-250" | "sm\:transition-500" | "sm\:transition-750" | "sm\:transition-all" | "sm\:transition-bg" | "sm\:transition-border" | "sm\:transition-color" | "sm\:transition-colors" | "sm\:transition-delay-0" | "sm\:transition-delay-100" | "sm\:transition-delay-1000" | "sm\:transition-delay-250" | "sm\:transition-delay-500" | "sm\:transition-delay-750" | "sm\:transition-ease" | "sm\:transition-ease-in" | "sm\:transition-ease-in-out" | "sm\:transition-ease-out" | "sm\:transition-linear" | "sm\:transition-none" | "sm\:transition-opacity" | "sm\:transition-transform" | "sm\:truncate" | "sm\:underline" | "sm\:unicase" | "sm\:uppercase" | "sm\:visible" | "sm\:w-0" | "sm\:w-1" | "sm\:w-10" | "sm\:w-10\/12" | "sm\:w-11\/12" | "sm\:w-12" | "sm\:w-16" | "sm\:w-1\/12" | "sm\:w-1\/2" | "sm\:w-1\/3" | "sm\:w-1\/4" | "sm\:w-1\/5" | "sm\:w-1\/6" | "sm\:w-2" | "sm\:w-20" | "sm\:w-24" | "sm\:w-2\/12" | "sm\:w-2\/3" | "sm\:w-2\/4" | "sm\:w-2\/5" | "sm\:w-2\/6" | "sm\:w-3" | "sm\:w-32" | "sm\:w-3\/12" | "sm\:w-3\/4" | "sm\:w-3\/5" | "sm\:w-3\/6" | "sm\:w-4" | "sm\:w-40" | "sm\:w-48" | "sm\:w-4\/12" | "sm\:w-4\/5" | "sm\:w-4\/6" | "sm\:w-5" | "sm\:w-56" | "sm\:w-5\/12" | "sm\:w-5\/6" | "sm\:w-6" | "sm\:w-64" | "sm\:w-6\/12" | "sm\:w-7\/12" | "sm\:w-8" | "sm\:w-8\/12" | "sm\:w-9\/12" | "sm\:w-auto" | "sm\:w-full" | "sm\:w-px" | "sm\:w-screen" | "sm\:whitespace-no-wrap" | "sm\:whitespace-normal" | "sm\:whitespace-pre" | "sm\:whitespace-pre-line" | "sm\:whitespace-pre-wrap" | "sm\:will-change-auto" | "sm\:will-change-contents" | "sm\:will-change-opacity" | "sm\:will-change-scroll" | "sm\:will-change-transform" | "sm\:z-0" | "sm\:z-10" | "sm\:z-20" | "sm\:z-30" | "sm\:z-40" | "sm\:z-50" | "sm\:z-auto" | "small-caps" | "sr-only" | "stacked-fractions" | "static" | "sticky" | "stroke-current" | "subpixel-antialiased" | "table" | "table-auto" | "table-cell" | "table-fixed" | "table-row" | "tabular-nums" | "text-2xl" | "text-3xl" | "text-4xl" | "text-5xl" | "text-6xl" | "text-align-unset" | "text-base" | "text-black" | "text-blue-100" | "text-blue-200" | "text-blue-300" | "text-blue-400" | "text-blue-500" | "text-blue-600" | "text-blue-700" | "text-blue-800" | "text-blue-900" | "text-border" | "text-center" | "text-color-unset" | "text-gray-100" | "text-gray-200" | "text-gray-300" | "text-gray-400" | "text-gray-500" | "text-gray-600" | "text-gray-700" | "text-gray-800" | "text-gray-900" | "text-justify" | "text-left" | "text-lg" | "text-primary" | "text-right" | "text-size-unset" | "text-sm" | "text-transform-unset" | "text-white" | "text-xl" | "text-xs" | "titling-caps" | "top-0" | "top-auto" | "tracking-normal" | "tracking-tight" | "tracking-tighter" | "tracking-unset" | "tracking-wide" | "tracking-wider" | "tracking-widest" | "transform-b" | "transform-bl" | "transform-border" | "transform-br" | "transform-fill" | "transform-l" | "transform-none" | "transform-r" | "transform-t" | "transform-tl" | "transform-tr" | "transform-view" | "transition-0" | "transition-100" | "transition-1000" | "transition-250" | "transition-500" | "transition-750" | "transition-all" | "transition-bg" | "transition-border" | "transition-color" | "transition-colors" | "transition-delay-0" | "transition-delay-100" | "transition-delay-1000" | "transition-delay-250" | "transition-delay-500" | "transition-delay-750" | "transition-ease" | "transition-ease-in" | "transition-ease-in-out" | "transition-ease-out" | "transition-linear" | "transition-none" | "transition-opacity" | "transition-transform" | "truncate" | "underline" | "unicase" | "uppercase" | "visible" | "w-0" | "w-1" | "w-10" | "w-10\/12" | "w-11\/12" | "w-12" | "w-16" | "w-1\/12" | "w-1\/2" | "w-1\/3" | "w-1\/4" | "w-1\/5" | "w-1\/6" | "w-2" | "w-20" | "w-24" | "w-2\/12" | "w-2\/3" | "w-2\/4" | "w-2\/5" | "w-2\/6" | "w-3" | "w-32" | "w-3\/12" | "w-3\/4" | "w-3\/5" | "w-3\/6" | "w-4" | "w-40" | "w-48" | "w-4\/12" | "w-4\/5" | "w-4\/6" | "w-5" | "w-56" | "w-5\/12" | "w-5\/6" | "w-6" | "w-64" | "w-6\/12" | "w-7\/12" | "w-8" | "w-8\/12" | "w-9\/12" | "w-auto" | "w-full" | "w-px" | "w-screen" | "whitespace-no-wrap" | "whitespace-normal" | "whitespace-pre" | "whitespace-pre-line" | "whitespace-pre-wrap" | "will-change-auto" | "will-change-contents" | "will-change-opacity" | "will-change-scroll" | "will-change-transform" | "xl\:-m-1" | "xl\:-m-10" | "xl\:-m-12" | "xl\:-m-16" | "xl\:-m-2" | "xl\:-m-20" | "xl\:-m-24" | "xl\:-m-3" | "xl\:-m-32" | "xl\:-m-4" | "xl\:-m-40" | "xl\:-m-48" | "xl\:-m-5" | "xl\:-m-56" | "xl\:-m-6" | "xl\:-m-64" | "xl\:-m-8" | "xl\:-m-px" | "xl\:-mb-1" | "xl\:-mb-10" | "xl\:-mb-12" | "xl\:-mb-16" | "xl\:-mb-2" | "xl\:-mb-20" | "xl\:-mb-24" | "xl\:-mb-3" | "xl\:-mb-32" | "xl\:-mb-4" | "xl\:-mb-40" | "xl\:-mb-48" | "xl\:-mb-5" | "xl\:-mb-56" | "xl\:-mb-6" | "xl\:-mb-64" | "xl\:-mb-8" | "xl\:-mb-px" | "xl\:-ml-1" | "xl\:-ml-10" | "xl\:-ml-12" | "xl\:-ml-16" | "xl\:-ml-2" | "xl\:-ml-20" | "xl\:-ml-24" | "xl\:-ml-3" | "xl\:-ml-32" | "xl\:-ml-4" | "xl\:-ml-40" | "xl\:-ml-48" | "xl\:-ml-5" | "xl\:-ml-56" | "xl\:-ml-6" | "xl\:-ml-64" | "xl\:-ml-8" | "xl\:-ml-px" | "xl\:-mr-1" | "xl\:-mr-10" | "xl\:-mr-12" | "xl\:-mr-16" | "xl\:-mr-2" | "xl\:-mr-20" | "xl\:-mr-24" | "xl\:-mr-3" | "xl\:-mr-32" | "xl\:-mr-4" | "xl\:-mr-40" | "xl\:-mr-48" | "xl\:-mr-5" | "xl\:-mr-56" | "xl\:-mr-6" | "xl\:-mr-64" | "xl\:-mr-8" | "xl\:-mr-px" | "xl\:-mt-1" | "xl\:-mt-10" | "xl\:-mt-12" | "xl\:-mt-16" | "xl\:-mt-2" | "xl\:-mt-20" | "xl\:-mt-24" | "xl\:-mt-3" | "xl\:-mt-32" | "xl\:-mt-4" | "xl\:-mt-40" | "xl\:-mt-48" | "xl\:-mt-5" | "xl\:-mt-56" | "xl\:-mt-6" | "xl\:-mt-64" | "xl\:-mt-8" | "xl\:-mt-px" | "xl\:-mx-1" | "xl\:-mx-10" | "xl\:-mx-12" | "xl\:-mx-16" | "xl\:-mx-2" | "xl\:-mx-20" | "xl\:-mx-24" | "xl\:-mx-3" | "xl\:-mx-32" | "xl\:-mx-4" | "xl\:-mx-40" | "xl\:-mx-48" | "xl\:-mx-5" | "xl\:-mx-56" | "xl\:-mx-6" | "xl\:-mx-64" | "xl\:-mx-8" | "xl\:-mx-px" | "xl\:-my-1" | "xl\:-my-10" | "xl\:-my-12" | "xl\:-my-16" | "xl\:-my-2" | "xl\:-my-20" | "xl\:-my-24" | "xl\:-my-3" | "xl\:-my-32" | "xl\:-my-4" | "xl\:-my-40" | "xl\:-my-48" | "xl\:-my-5" | "xl\:-my-56" | "xl\:-my-6" | "xl\:-my-64" | "xl\:-my-8" | "xl\:-my-px" | "xl\:absolute" | "xl\:align-baseline" | "xl\:align-bottom" | "xl\:align-middle" | "xl\:align-text-bottom" | "xl\:align-text-top" | "xl\:align-top" | "xl\:all-small-caps" | "xl\:antialiased" | "xl\:appearance-none" | "xl\:bg-auto" | "xl\:bg-black" | "xl\:bg-blue-100" | "xl\:bg-blue-200" | "xl\:bg-blue-300" | "xl\:bg-blue-400" | "xl\:bg-blue-500" | "xl\:bg-blue-600" | "xl\:bg-blue-700" | "xl\:bg-blue-800" | "xl\:bg-blue-900" | "xl\:bg-border" | "xl\:bg-bottom" | "xl\:bg-center" | "xl\:bg-contain" | "xl\:bg-cover" | "xl\:bg-fixed" | "xl\:bg-gray-100" | "xl\:bg-gray-200" | "xl\:bg-gray-300" | "xl\:bg-gray-400" | "xl\:bg-gray-500" | "xl\:bg-gray-600" | "xl\:bg-gray-700" | "xl\:bg-gray-800" | "xl\:bg-gray-900" | "xl\:bg-left" | "xl\:bg-left-bottom" | "xl\:bg-left-top" | "xl\:bg-local" | "xl\:bg-no-repeat" | "xl\:bg-primary" | "xl\:bg-repeat" | "xl\:bg-repeat-round" | "xl\:bg-repeat-space" | "xl\:bg-repeat-x" | "xl\:bg-repeat-y" | "xl\:bg-right" | "xl\:bg-right-bottom" | "xl\:bg-right-top" | "xl\:bg-scroll" | "xl\:bg-top" | "xl\:bg-white" | "xl\:block" | "xl\:border" | "xl\:border-0" | "xl\:border-2" | "xl\:border-4" | "xl\:border-8" | "xl\:border-b" | "xl\:border-b-0" | "xl\:border-b-2" | "xl\:border-b-4" | "xl\:border-b-8" | "xl\:border-black" | "xl\:border-blue-100" | "xl\:border-blue-200" | "xl\:border-blue-300" | "xl\:border-blue-400" | "xl\:border-blue-500" | "xl\:border-blue-600" | "xl\:border-blue-700" | "xl\:border-blue-800" | "xl\:border-blue-900" | "xl\:border-border" | "xl\:border-collapse" | "xl\:border-dashed" | "xl\:border-dotted" | "xl\:border-double" | "xl\:border-gray-100" | "xl\:border-gray-200" | "xl\:border-gray-300" | "xl\:border-gray-400" | "xl\:border-gray-500" | "xl\:border-gray-600" | "xl\:border-gray-700" | "xl\:border-gray-800" | "xl\:border-gray-900" | "xl\:border-l" | "xl\:border-l-0" | "xl\:border-l-2" | "xl\:border-l-4" | "xl\:border-l-8" | "xl\:border-none" | "xl\:border-primary" | "xl\:border-r" | "xl\:border-r-0" | "xl\:border-r-2" | "xl\:border-r-4" | "xl\:border-r-8" | "xl\:border-separate" | "xl\:border-solid" | "xl\:border-t" | "xl\:border-t-0" | "xl\:border-t-2" | "xl\:border-t-4" | "xl\:border-t-8" | "xl\:border-white" | "xl\:bottom-0" | "xl\:bottom-auto" | "xl\:break-all" | "xl\:break-normal" | "xl\:break-words" | "xl\:capitalize" | "xl\:clearfix" | "xl\:common-ligatures" | "xl\:content-around" | "xl\:content-between" | "xl\:content-center" | "xl\:content-end" | "xl\:content-start" | "xl\:contextual-ligatures" | "xl\:cursor-auto" | "xl\:cursor-default" | "xl\:cursor-move" | "xl\:cursor-not-allowed" | "xl\:cursor-pointer" | "xl\:cursor-text" | "xl\:cursor-wait" | "xl\:diagonal-fractions" | "xl\:discretionary-ligatures" | "xl\:ellipsis" | "xl\:fill-current" | "xl\:fixed" | "xl\:flex" | "xl\:flex-1" | "xl\:flex-auto" | "xl\:flex-col" | "xl\:flex-col-reverse" | "xl\:flex-grow" | "xl\:flex-grow-0" | "xl\:flex-initial" | "xl\:flex-no-wrap" | "xl\:flex-none" | "xl\:flex-row" | "xl\:flex-row-reverse" | "xl\:flex-shrink" | "xl\:flex-shrink-0" | "xl\:flex-wrap" | "xl\:flex-wrap-reverse" | "xl\:float-left" | "xl\:float-none" | "xl\:float-right" | "xl\:focus\:bg-black" | "xl\:focus\:bg-blue-100" | "xl\:focus\:bg-blue-200" | "xl\:focus\:bg-blue-300" | "xl\:focus\:bg-blue-400" | "xl\:focus\:bg-blue-500" | "xl\:focus\:bg-blue-600" | "xl\:focus\:bg-blue-700" | "xl\:focus\:bg-blue-800" | "xl\:focus\:bg-blue-900" | "xl\:focus\:bg-border" | "xl\:focus\:bg-gray-100" | "xl\:focus\:bg-gray-200" | "xl\:focus\:bg-gray-300" | "xl\:focus\:bg-gray-400" | "xl\:focus\:bg-gray-500" | "xl\:focus\:bg-gray-600" | "xl\:focus\:bg-gray-700" | "xl\:focus\:bg-gray-800" | "xl\:focus\:bg-gray-900" | "xl\:focus\:bg-primary" | "xl\:focus\:bg-white" | "xl\:focus\:border" | "xl\:focus\:border-0" | "xl\:focus\:border-2" | "xl\:focus\:border-4" | "xl\:focus\:border-8" | "xl\:focus\:border-b" | "xl\:focus\:border-b-0" | "xl\:focus\:border-b-2" | "xl\:focus\:border-b-4" | "xl\:focus\:border-b-8" | "xl\:focus\:border-black" | "xl\:focus\:border-blue-100" | "xl\:focus\:border-blue-200" | "xl\:focus\:border-blue-300" | "xl\:focus\:border-blue-400" | "xl\:focus\:border-blue-500" | "xl\:focus\:border-blue-600" | "xl\:focus\:border-blue-700" | "xl\:focus\:border-blue-800" | "xl\:focus\:border-blue-900" | "xl\:focus\:border-border" | "xl\:focus\:border-gray-100" | "xl\:focus\:border-gray-200" | "xl\:focus\:border-gray-300" | "xl\:focus\:border-gray-400" | "xl\:focus\:border-gray-500" | "xl\:focus\:border-gray-600" | "xl\:focus\:border-gray-700" | "xl\:focus\:border-gray-800" | "xl\:focus\:border-gray-900" | "xl\:focus\:border-l" | "xl\:focus\:border-l-0" | "xl\:focus\:border-l-2" | "xl\:focus\:border-l-4" | "xl\:focus\:border-l-8" | "xl\:focus\:border-primary" | "xl\:focus\:border-r" | "xl\:focus\:border-r-0" | "xl\:focus\:border-r-2" | "xl\:focus\:border-r-4" | "xl\:focus\:border-r-8" | "xl\:focus\:border-t" | "xl\:focus\:border-t-0" | "xl\:focus\:border-t-2" | "xl\:focus\:border-t-4" | "xl\:focus\:border-t-8" | "xl\:focus\:border-white" | "xl\:focus\:font-black" | "xl\:focus\:font-bold" | "xl\:focus\:font-extrabold" | "xl\:focus\:font-hairline" | "xl\:focus\:font-light" | "xl\:focus\:font-medium" | "xl\:focus\:font-normal" | "xl\:focus\:font-semibold" | "xl\:focus\:font-thin" | "xl\:focus\:line-through" | "xl\:focus\:no-underline" | "xl\:focus\:not-sr-only" | "xl\:focus\:opacity-0" | "xl\:focus\:opacity-100" | "xl\:focus\:opacity-25" | "xl\:focus\:opacity-50" | "xl\:focus\:opacity-75" | "xl\:focus\:outline-none" | "xl\:focus\:placeholder-black" | "xl\:focus\:placeholder-blue-100" | "xl\:focus\:placeholder-blue-200" | "xl\:focus\:placeholder-blue-300" | "xl\:focus\:placeholder-blue-400" | "xl\:focus\:placeholder-blue-500" | "xl\:focus\:placeholder-blue-600" | "xl\:focus\:placeholder-blue-700" | "xl\:focus\:placeholder-blue-800" | "xl\:focus\:placeholder-blue-900" | "xl\:focus\:placeholder-border" | "xl\:focus\:placeholder-gray-100" | "xl\:focus\:placeholder-gray-200" | "xl\:focus\:placeholder-gray-300" | "xl\:focus\:placeholder-gray-400" | "xl\:focus\:placeholder-gray-500" | "xl\:focus\:placeholder-gray-600" | "xl\:focus\:placeholder-gray-700" | "xl\:focus\:placeholder-gray-800" | "xl\:focus\:placeholder-gray-900" | "xl\:focus\:placeholder-primary" | "xl\:focus\:placeholder-white" | "xl\:focus\:shadow" | "xl\:focus\:shadow-2xl" | "xl\:focus\:shadow-inner" | "xl\:focus\:shadow-lg" | "xl\:focus\:shadow-md" | "xl\:focus\:shadow-none" | "xl\:focus\:shadow-outline" | "xl\:focus\:shadow-xl" | "xl\:focus\:sr-only" | "xl\:focus\:text-black" | "xl\:focus\:text-blue-100" | "xl\:focus\:text-blue-200" | "xl\:focus\:text-blue-300" | "xl\:focus\:text-blue-400" | "xl\:focus\:text-blue-500" | "xl\:focus\:text-blue-600" | "xl\:focus\:text-blue-700" | "xl\:focus\:text-blue-800" | "xl\:focus\:text-blue-900" | "xl\:focus\:text-border" | "xl\:focus\:text-gray-100" | "xl\:focus\:text-gray-200" | "xl\:focus\:text-gray-300" | "xl\:focus\:text-gray-400" | "xl\:focus\:text-gray-500" | "xl\:focus\:text-gray-600" | "xl\:focus\:text-gray-700" | "xl\:focus\:text-gray-800" | "xl\:focus\:text-gray-900" | "xl\:focus\:text-primary" | "xl\:focus\:text-white" | "xl\:focus\:underline" | "xl\:font-black" | "xl\:font-bold" | "xl\:font-extrabold" | "xl\:font-family-unset" | "xl\:font-hairline" | "xl\:font-light" | "xl\:font-medium" | "xl\:font-mono" | "xl\:font-normal" | "xl\:font-sans" | "xl\:font-semibold" | "xl\:font-serif" | "xl\:font-style-unset" | "xl\:font-thin" | "xl\:font-weight-unset" | "xl\:h-0" | "xl\:h-1" | "xl\:h-10" | "xl\:h-12" | "xl\:h-16" | "xl\:h-2" | "xl\:h-20" | "xl\:h-24" | "xl\:h-3" | "xl\:h-32" | "xl\:h-4" | "xl\:h-40" | "xl\:h-48" | "xl\:h-5" | "xl\:h-56" | "xl\:h-6" | "xl\:h-64" | "xl\:h-8" | "xl\:h-auto" | "xl\:h-full" | "xl\:h-px" | "xl\:h-screen" | "xl\:hidden" | "xl\:historical-ligatures" | "xl\:hover\:bg-black" | "xl\:hover\:bg-blue-100" | "xl\:hover\:bg-blue-200" | "xl\:hover\:bg-blue-300" | "xl\:hover\:bg-blue-400" | "xl\:hover\:bg-blue-500" | "xl\:hover\:bg-blue-600" | "xl\:hover\:bg-blue-700" | "xl\:hover\:bg-blue-800" | "xl\:hover\:bg-blue-900" | "xl\:hover\:bg-border" | "xl\:hover\:bg-gray-100" | "xl\:hover\:bg-gray-200" | "xl\:hover\:bg-gray-300" | "xl\:hover\:bg-gray-400" | "xl\:hover\:bg-gray-500" | "xl\:hover\:bg-gray-600" | "xl\:hover\:bg-gray-700" | "xl\:hover\:bg-gray-800" | "xl\:hover\:bg-gray-900" | "xl\:hover\:bg-primary" | "xl\:hover\:bg-white" | "xl\:hover\:border" | "xl\:hover\:border-0" | "xl\:hover\:border-2" | "xl\:hover\:border-4" | "xl\:hover\:border-8" | "xl\:hover\:border-b" | "xl\:hover\:border-b-0" | "xl\:hover\:border-b-2" | "xl\:hover\:border-b-4" | "xl\:hover\:border-b-8" | "xl\:hover\:border-black" | "xl\:hover\:border-blue-100" | "xl\:hover\:border-blue-200" | "xl\:hover\:border-blue-300" | "xl\:hover\:border-blue-400" | "xl\:hover\:border-blue-500" | "xl\:hover\:border-blue-600" | "xl\:hover\:border-blue-700" | "xl\:hover\:border-blue-800" | "xl\:hover\:border-blue-900" | "xl\:hover\:border-border" | "xl\:hover\:border-gray-100" | "xl\:hover\:border-gray-200" | "xl\:hover\:border-gray-300" | "xl\:hover\:border-gray-400" | "xl\:hover\:border-gray-500" | "xl\:hover\:border-gray-600" | "xl\:hover\:border-gray-700" | "xl\:hover\:border-gray-800" | "xl\:hover\:border-gray-900" | "xl\:hover\:border-l" | "xl\:hover\:border-l-0" | "xl\:hover\:border-l-2" | "xl\:hover\:border-l-4" | "xl\:hover\:border-l-8" | "xl\:hover\:border-primary" | "xl\:hover\:border-r" | "xl\:hover\:border-r-0" | "xl\:hover\:border-r-2" | "xl\:hover\:border-r-4" | "xl\:hover\:border-r-8" | "xl\:hover\:border-t" | "xl\:hover\:border-t-0" | "xl\:hover\:border-t-2" | "xl\:hover\:border-t-4" | "xl\:hover\:border-t-8" | "xl\:hover\:border-white" | "xl\:hover\:font-black" | "xl\:hover\:font-bold" | "xl\:hover\:font-extrabold" | "xl\:hover\:font-hairline" | "xl\:hover\:font-light" | "xl\:hover\:font-medium" | "xl\:hover\:font-normal" | "xl\:hover\:font-semibold" | "xl\:hover\:font-thin" | "xl\:hover\:line-through" | "xl\:hover\:no-underline" | "xl\:hover\:opacity-0" | "xl\:hover\:opacity-100" | "xl\:hover\:opacity-25" | "xl\:hover\:opacity-50" | "xl\:hover\:opacity-75" | "xl\:hover\:shadow" | "xl\:hover\:shadow-2xl" | "xl\:hover\:shadow-inner" | "xl\:hover\:shadow-lg" | "xl\:hover\:shadow-md" | "xl\:hover\:shadow-none" | "xl\:hover\:shadow-outline" | "xl\:hover\:shadow-xl" | "xl\:hover\:text-black" | "xl\:hover\:text-blue-100" | "xl\:hover\:text-blue-200" | "xl\:hover\:text-blue-300" | "xl\:hover\:text-blue-400" | "xl\:hover\:text-blue-500" | "xl\:hover\:text-blue-600" | "xl\:hover\:text-blue-700" | "xl\:hover\:text-blue-800" | "xl\:hover\:text-blue-900" | "xl\:hover\:text-border" | "xl\:hover\:text-gray-100" | "xl\:hover\:text-gray-200" | "xl\:hover\:text-gray-300" | "xl\:hover\:text-gray-400" | "xl\:hover\:text-gray-500" | "xl\:hover\:text-gray-600" | "xl\:hover\:text-gray-700" | "xl\:hover\:text-gray-800" | "xl\:hover\:text-gray-900" | "xl\:hover\:text-primary" | "xl\:hover\:text-white" | "xl\:hover\:underline" | "xl\:hyphens-auto" | "xl\:hyphens-manual" | "xl\:hyphens-none" | "xl\:inline" | "xl\:inline-block" | "xl\:inline-flex" | "xl\:inset-0" | "xl\:inset-auto" | "xl\:inset-x-0" | "xl\:inset-x-auto" | "xl\:inset-y-0" | "xl\:inset-y-auto" | "xl\:invisible" | "xl\:italic" | "xl\:items-baseline" | "xl\:items-center" | "xl\:items-end" | "xl\:items-start" | "xl\:items-stretch" | "xl\:justify-around" | "xl\:justify-between" | "xl\:justify-center" | "xl\:justify-end" | "xl\:justify-start" | "xl\:last\:border" | "xl\:last\:border-0" | "xl\:last\:border-2" | "xl\:last\:border-4" | "xl\:last\:border-8" | "xl\:last\:border-b" | "xl\:last\:border-b-0" | "xl\:last\:border-b-2" | "xl\:last\:border-b-4" | "xl\:last\:border-b-8" | "xl\:last\:border-l" | "xl\:last\:border-l-0" | "xl\:last\:border-l-2" | "xl\:last\:border-l-4" | "xl\:last\:border-l-8" | "xl\:last\:border-r" | "xl\:last\:border-r-0" | "xl\:last\:border-r-2" | "xl\:last\:border-r-4" | "xl\:last\:border-r-8" | "xl\:last\:border-t" | "xl\:last\:border-t-0" | "xl\:last\:border-t-2" | "xl\:last\:border-t-4" | "xl\:last\:border-t-8" | "xl\:leading-loose" | "xl\:leading-none" | "xl\:leading-normal" | "xl\:leading-relaxed" | "xl\:leading-snug" | "xl\:leading-tight" | "xl\:leading-unset" | "xl\:left-0" | "xl\:left-auto" | "xl\:line-through" | "xl\:lining-nums" | "xl\:list-decimal" | "xl\:list-disc" | "xl\:list-inside" | "xl\:list-none" | "xl\:list-outside" | "xl\:lowercase" | "xl\:m-0" | "xl\:m-1" | "xl\:m-10" | "xl\:m-12" | "xl\:m-16" | "xl\:m-2" | "xl\:m-20" | "xl\:m-24" | "xl\:m-3" | "xl\:m-32" | "xl\:m-4" | "xl\:m-40" | "xl\:m-48" | "xl\:m-5" | "xl\:m-56" | "xl\:m-6" | "xl\:m-64" | "xl\:m-8" | "xl\:m-auto" | "xl\:m-px" | "xl\:max-h-full" | "xl\:max-h-screen" | "xl\:max-w-2xl" | "xl\:max-w-3xl" | "xl\:max-w-4xl" | "xl\:max-w-5xl" | "xl\:max-w-6xl" | "xl\:max-w-full" | "xl\:max-w-lg" | "xl\:max-w-md" | "xl\:max-w-sm" | "xl\:max-w-xl" | "xl\:max-w-xs" | "xl\:mb-0" | "xl\:mb-1" | "xl\:mb-10" | "xl\:mb-12" | "xl\:mb-16" | "xl\:mb-2" | "xl\:mb-20" | "xl\:mb-24" | "xl\:mb-3" | "xl\:mb-32" | "xl\:mb-4" | "xl\:mb-40" | "xl\:mb-48" | "xl\:mb-5" | "xl\:mb-56" | "xl\:mb-6" | "xl\:mb-64" | "xl\:mb-8" | "xl\:mb-auto" | "xl\:mb-px" | "xl\:min-h-0" | "xl\:min-h-full" | "xl\:min-h-screen" | "xl\:min-w-0" | "xl\:min-w-full" | "xl\:ml-0" | "xl\:ml-1" | "xl\:ml-10" | "xl\:ml-12" | "xl\:ml-16" | "xl\:ml-2" | "xl\:ml-20" | "xl\:ml-24" | "xl\:ml-3" | "xl\:ml-32" | "xl\:ml-4" | "xl\:ml-40" | "xl\:ml-48" | "xl\:ml-5" | "xl\:ml-56" | "xl\:ml-6" | "xl\:ml-64" | "xl\:ml-8" | "xl\:ml-auto" | "xl\:ml-px" | "xl\:mr-0" | "xl\:mr-1" | "xl\:mr-10" | "xl\:mr-12" | "xl\:mr-16" | "xl\:mr-2" | "xl\:mr-20" | "xl\:mr-24" | "xl\:mr-3" | "xl\:mr-32" | "xl\:mr-4" | "xl\:mr-40" | "xl\:mr-48" | "xl\:mr-5" | "xl\:mr-56" | "xl\:mr-6" | "xl\:mr-64" | "xl\:mr-8" | "xl\:mr-auto" | "xl\:mr-px" | "xl\:mt-0" | "xl\:mt-1" | "xl\:mt-10" | "xl\:mt-12" | "xl\:mt-16" | "xl\:mt-2" | "xl\:mt-20" | "xl\:mt-24" | "xl\:mt-3" | "xl\:mt-32" | "xl\:mt-4" | "xl\:mt-40" | "xl\:mt-48" | "xl\:mt-5" | "xl\:mt-56" | "xl\:mt-6" | "xl\:mt-64" | "xl\:mt-8" | "xl\:mt-auto" | "xl\:mt-px" | "xl\:mx-0" | "xl\:mx-1" | "xl\:mx-10" | "xl\:mx-12" | "xl\:mx-16" | "xl\:mx-2" | "xl\:mx-20" | "xl\:mx-24" | "xl\:mx-3" | "xl\:mx-32" | "xl\:mx-4" | "xl\:mx-40" | "xl\:mx-48" | "xl\:mx-5" | "xl\:mx-56" | "xl\:mx-6" | "xl\:mx-64" | "xl\:mx-8" | "xl\:mx-auto" | "xl\:mx-px" | "xl\:my-0" | "xl\:my-1" | "xl\:my-10" | "xl\:my-12" | "xl\:my-16" | "xl\:my-2" | "xl\:my-20" | "xl\:my-24" | "xl\:my-3" | "xl\:my-32" | "xl\:my-4" | "xl\:my-40" | "xl\:my-48" | "xl\:my-5" | "xl\:my-56" | "xl\:my-6" | "xl\:my-64" | "xl\:my-8" | "xl\:my-auto" | "xl\:my-px" | "xl\:no-common-ligatures" | "xl\:no-contextual-ligatures" | "xl\:no-discretionary-ligatures" | "xl\:no-ellipsis" | "xl\:no-historical-ligatures" | "xl\:no-ligatures" | "xl\:no-underline" | "xl\:normal-caps" | "xl\:normal-case" | "xl\:normal-ligatures" | "xl\:normal-nums" | "xl\:not-italic" | "xl\:not-sr-only" | "xl\:object-bottom" | "xl\:object-center" | "xl\:object-contain" | "xl\:object-cover" | "xl\:object-fill" | "xl\:object-left" | "xl\:object-left-bottom" | "xl\:object-left-top" | "xl\:object-none" | "xl\:object-right" | "xl\:object-right-bottom" | "xl\:object-right-top" | "xl\:object-scale-down" | "xl\:object-top" | "xl\:oldstyle-nums" | "xl\:opacity-0" | "xl\:opacity-100" | "xl\:opacity-25" | "xl\:opacity-50" | "xl\:opacity-75" | "xl\:order-1" | "xl\:order-10" | "xl\:order-11" | "xl\:order-12" | "xl\:order-2" | "xl\:order-3" | "xl\:order-4" | "xl\:order-5" | "xl\:order-6" | "xl\:order-7" | "xl\:order-8" | "xl\:order-9" | "xl\:order-first" | "xl\:order-last" | "xl\:order-none" | "xl\:ordinal-nums" | "xl\:outline-none" | "xl\:overflow-auto" | "xl\:overflow-hidden" | "xl\:overflow-scroll" | "xl\:overflow-visible" | "xl\:overflow-x-auto" | "xl\:overflow-x-hidden" | "xl\:overflow-x-scroll" | "xl\:overflow-x-visible" | "xl\:overflow-y-auto" | "xl\:overflow-y-hidden" | "xl\:overflow-y-scroll" | "xl\:overflow-y-visible" | "xl\:p-0" | "xl\:p-1" | "xl\:p-10" | "xl\:p-12" | "xl\:p-16" | "xl\:p-2" | "xl\:p-20" | "xl\:p-24" | "xl\:p-3" | "xl\:p-32" | "xl\:p-4" | "xl\:p-40" | "xl\:p-48" | "xl\:p-5" | "xl\:p-56" | "xl\:p-6" | "xl\:p-64" | "xl\:p-8" | "xl\:p-px" | "xl\:pb-0" | "xl\:pb-1" | "xl\:pb-10" | "xl\:pb-12" | "xl\:pb-16" | "xl\:pb-2" | "xl\:pb-20" | "xl\:pb-24" | "xl\:pb-3" | "xl\:pb-32" | "xl\:pb-4" | "xl\:pb-40" | "xl\:pb-48" | "xl\:pb-5" | "xl\:pb-56" | "xl\:pb-6" | "xl\:pb-64" | "xl\:pb-8" | "xl\:pb-px" | "xl\:petite-caps" | "xl\:pl-0" | "xl\:pl-1" | "xl\:pl-10" | "xl\:pl-12" | "xl\:pl-16" | "xl\:pl-2" | "xl\:pl-20" | "xl\:pl-24" | "xl\:pl-3" | "xl\:pl-32" | "xl\:pl-4" | "xl\:pl-40" | "xl\:pl-48" | "xl\:pl-5" | "xl\:pl-56" | "xl\:pl-6" | "xl\:pl-64" | "xl\:pl-8" | "xl\:pl-px" | "xl\:placeholder-black" | "xl\:placeholder-blue-100" | "xl\:placeholder-blue-200" | "xl\:placeholder-blue-300" | "xl\:placeholder-blue-400" | "xl\:placeholder-blue-500" | "xl\:placeholder-blue-600" | "xl\:placeholder-blue-700" | "xl\:placeholder-blue-800" | "xl\:placeholder-blue-900" | "xl\:placeholder-border" | "xl\:placeholder-gray-100" | "xl\:placeholder-gray-200" | "xl\:placeholder-gray-300" | "xl\:placeholder-gray-400" | "xl\:placeholder-gray-500" | "xl\:placeholder-gray-600" | "xl\:placeholder-gray-700" | "xl\:placeholder-gray-800" | "xl\:placeholder-gray-900" | "xl\:placeholder-primary" | "xl\:placeholder-white" | "xl\:pointer-events-auto" | "xl\:pointer-events-none" | "xl\:pr-0" | "xl\:pr-1" | "xl\:pr-10" | "xl\:pr-12" | "xl\:pr-16" | "xl\:pr-2" | "xl\:pr-20" | "xl\:pr-24" | "xl\:pr-3" | "xl\:pr-32" | "xl\:pr-4" | "xl\:pr-40" | "xl\:pr-48" | "xl\:pr-5" | "xl\:pr-56" | "xl\:pr-6" | "xl\:pr-64" | "xl\:pr-8" | "xl\:pr-px" | "xl\:proportional-nums" | "xl\:pt-0" | "xl\:pt-1" | "xl\:pt-10" | "xl\:pt-12" | "xl\:pt-16" | "xl\:pt-2" | "xl\:pt-20" | "xl\:pt-24" | "xl\:pt-3" | "xl\:pt-32" | "xl\:pt-4" | "xl\:pt-40" | "xl\:pt-48" | "xl\:pt-5" | "xl\:pt-56" | "xl\:pt-6" | "xl\:pt-64" | "xl\:pt-8" | "xl\:pt-px" | "xl\:px-0" | "xl\:px-1" | "xl\:px-10" | "xl\:px-12" | "xl\:px-16" | "xl\:px-2" | "xl\:px-20" | "xl\:px-24" | "xl\:px-3" | "xl\:px-32" | "xl\:px-4" | "xl\:px-40" | "xl\:px-48" | "xl\:px-5" | "xl\:px-56" | "xl\:px-6" | "xl\:px-64" | "xl\:px-8" | "xl\:px-px" | "xl\:py-0" | "xl\:py-1" | "xl\:py-10" | "xl\:py-12" | "xl\:py-16" | "xl\:py-2" | "xl\:py-20" | "xl\:py-24" | "xl\:py-3" | "xl\:py-32" | "xl\:py-4" | "xl\:py-40" | "xl\:py-48" | "xl\:py-5" | "xl\:py-56" | "xl\:py-6" | "xl\:py-64" | "xl\:py-8" | "xl\:py-px" | "xl\:relative" | "xl\:resize" | "xl\:resize-none" | "xl\:resize-x" | "xl\:resize-y" | "xl\:right-0" | "xl\:right-auto" | "xl\:rotate-180" | "xl\:rotate-270" | "xl\:rotate-90" | "xl\:rounded" | "xl\:rounded-b" | "xl\:rounded-b-full" | "xl\:rounded-b-lg" | "xl\:rounded-b-none" | "xl\:rounded-b-sm" | "xl\:rounded-bl" | "xl\:rounded-bl-full" | "xl\:rounded-bl-lg" | "xl\:rounded-bl-none" | "xl\:rounded-bl-sm" | "xl\:rounded-br" | "xl\:rounded-br-full" | "xl\:rounded-br-lg" | "xl\:rounded-br-none" | "xl\:rounded-br-sm" | "xl\:rounded-full" | "xl\:rounded-l" | "xl\:rounded-l-full" | "xl\:rounded-l-lg" | "xl\:rounded-l-none" | "xl\:rounded-l-sm" | "xl\:rounded-lg" | "xl\:rounded-none" | "xl\:rounded-r" | "xl\:rounded-r-full" | "xl\:rounded-r-lg" | "xl\:rounded-r-none" | "xl\:rounded-r-sm" | "xl\:rounded-sm" | "xl\:rounded-t" | "xl\:rounded-t-full" | "xl\:rounded-t-lg" | "xl\:rounded-t-none" | "xl\:rounded-t-sm" | "xl\:rounded-tl" | "xl\:rounded-tl-full" | "xl\:rounded-tl-lg" | "xl\:rounded-tl-none" | "xl\:rounded-tl-sm" | "xl\:rounded-tr" | "xl\:rounded-tr-full" | "xl\:rounded-tr-lg" | "xl\:rounded-tr-none" | "xl\:rounded-tr-sm" | "xl\:scrolling-auto" | "xl\:scrolling-touch" | "xl\:select-all" | "xl\:select-auto" | "xl\:select-none" | "xl\:select-text" | "xl\:self-auto" | "xl\:self-center" | "xl\:self-end" | "xl\:self-start" | "xl\:self-stretch" | "xl\:shadow" | "xl\:shadow-2xl" | "xl\:shadow-inner" | "xl\:shadow-lg" | "xl\:shadow-md" | "xl\:shadow-none" | "xl\:shadow-outline" | "xl\:shadow-xl" | "xl\:slashed-zeros" | "xl\:small-caps" | "xl\:sr-only" | "xl\:stacked-fractions" | "xl\:static" | "xl\:sticky" | "xl\:stroke-current" | "xl\:subpixel-antialiased" | "xl\:table" | "xl\:table-auto" | "xl\:table-cell" | "xl\:table-fixed" | "xl\:table-row" | "xl\:tabular-nums" | "xl\:text-2xl" | "xl\:text-3xl" | "xl\:text-4xl" | "xl\:text-5xl" | "xl\:text-6xl" | "xl\:text-align-unset" | "xl\:text-base" | "xl\:text-black" | "xl\:text-blue-100" | "xl\:text-blue-200" | "xl\:text-blue-300" | "xl\:text-blue-400" | "xl\:text-blue-500" | "xl\:text-blue-600" | "xl\:text-blue-700" | "xl\:text-blue-800" | "xl\:text-blue-900" | "xl\:text-border" | "xl\:text-center" | "xl\:text-color-unset" | "xl\:text-gray-100" | "xl\:text-gray-200" | "xl\:text-gray-300" | "xl\:text-gray-400" | "xl\:text-gray-500" | "xl\:text-gray-600" | "xl\:text-gray-700" | "xl\:text-gray-800" | "xl\:text-gray-900" | "xl\:text-justify" | "xl\:text-left" | "xl\:text-lg" | "xl\:text-primary" | "xl\:text-right" | "xl\:text-size-unset" | "xl\:text-sm" | "xl\:text-transform-unset" | "xl\:text-white" | "xl\:text-xl" | "xl\:text-xs" | "xl\:titling-caps" | "xl\:top-0" | "xl\:top-auto" | "xl\:tracking-normal" | "xl\:tracking-tight" | "xl\:tracking-tighter" | "xl\:tracking-unset" | "xl\:tracking-wide" | "xl\:tracking-wider" | "xl\:tracking-widest" | "xl\:transform-b" | "xl\:transform-bl" | "xl\:transform-border" | "xl\:transform-br" | "xl\:transform-fill" | "xl\:transform-l" | "xl\:transform-none" | "xl\:transform-r" | "xl\:transform-t" | "xl\:transform-tl" | "xl\:transform-tr" | "xl\:transform-view" | "xl\:transition-0" | "xl\:transition-100" | "xl\:transition-1000" | "xl\:transition-250" | "xl\:transition-500" | "xl\:transition-750" | "xl\:transition-all" | "xl\:transition-bg" | "xl\:transition-border" | "xl\:transition-color" | "xl\:transition-colors" | "xl\:transition-delay-0" | "xl\:transition-delay-100" | "xl\:transition-delay-1000" | "xl\:transition-delay-250" | "xl\:transition-delay-500" | "xl\:transition-delay-750" | "xl\:transition-ease" | "xl\:transition-ease-in" | "xl\:transition-ease-in-out" | "xl\:transition-ease-out" | "xl\:transition-linear" | "xl\:transition-none" | "xl\:transition-opacity" | "xl\:transition-transform" | "xl\:truncate" | "xl\:underline" | "xl\:unicase" | "xl\:uppercase" | "xl\:visible" | "xl\:w-0" | "xl\:w-1" | "xl\:w-10" | "xl\:w-10\/12" | "xl\:w-11\/12" | "xl\:w-12" | "xl\:w-16" | "xl\:w-1\/12" | "xl\:w-1\/2" | "xl\:w-1\/3" | "xl\:w-1\/4" | "xl\:w-1\/5" | "xl\:w-1\/6" | "xl\:w-2" | "xl\:w-20" | "xl\:w-24" | "xl\:w-2\/12" | "xl\:w-2\/3" | "xl\:w-2\/4" | "xl\:w-2\/5" | "xl\:w-2\/6" | "xl\:w-3" | "xl\:w-32" | "xl\:w-3\/12" | "xl\:w-3\/4" | "xl\:w-3\/5" | "xl\:w-3\/6" | "xl\:w-4" | "xl\:w-40" | "xl\:w-48" | "xl\:w-4\/12" | "xl\:w-4\/5" | "xl\:w-4\/6" | "xl\:w-5" | "xl\:w-56" | "xl\:w-5\/12" | "xl\:w-5\/6" | "xl\:w-6" | "xl\:w-64" | "xl\:w-6\/12" | "xl\:w-7\/12" | "xl\:w-8" | "xl\:w-8\/12" | "xl\:w-9\/12" | "xl\:w-auto" | "xl\:w-full" | "xl\:w-px" | "xl\:w-screen" | "xl\:whitespace-no-wrap" | "xl\:whitespace-normal" | "xl\:whitespace-pre" | "xl\:whitespace-pre-line" | "xl\:whitespace-pre-wrap" | "xl\:will-change-auto" | "xl\:will-change-contents" | "xl\:will-change-opacity" | "xl\:will-change-scroll" | "xl\:will-change-transform" | "xl\:z-0" | "xl\:z-10" | "xl\:z-20" | "xl\:z-30" | "xl\:z-40" | "xl\:z-50" | "xl\:z-auto" | "xs\:-m-1" | "xs\:-m-10" | "xs\:-m-12" | "xs\:-m-16" | "xs\:-m-2" | "xs\:-m-20" | "xs\:-m-24" | "xs\:-m-3" | "xs\:-m-32" | "xs\:-m-4" | "xs\:-m-40" | "xs\:-m-48" | "xs\:-m-5" | "xs\:-m-56" | "xs\:-m-6" | "xs\:-m-64" | "xs\:-m-8" | "xs\:-m-px" | "xs\:-mb-1" | "xs\:-mb-10" | "xs\:-mb-12" | "xs\:-mb-16" | "xs\:-mb-2" | "xs\:-mb-20" | "xs\:-mb-24" | "xs\:-mb-3" | "xs\:-mb-32" | "xs\:-mb-4" | "xs\:-mb-40" | "xs\:-mb-48" | "xs\:-mb-5" | "xs\:-mb-56" | "xs\:-mb-6" | "xs\:-mb-64" | "xs\:-mb-8" | "xs\:-mb-px" | "xs\:-ml-1" | "xs\:-ml-10" | "xs\:-ml-12" | "xs\:-ml-16" | "xs\:-ml-2" | "xs\:-ml-20" | "xs\:-ml-24" | "xs\:-ml-3" | "xs\:-ml-32" | "xs\:-ml-4" | "xs\:-ml-40" | "xs\:-ml-48" | "xs\:-ml-5" | "xs\:-ml-56" | "xs\:-ml-6" | "xs\:-ml-64" | "xs\:-ml-8" | "xs\:-ml-px" | "xs\:-mr-1" | "xs\:-mr-10" | "xs\:-mr-12" | "xs\:-mr-16" | "xs\:-mr-2" | "xs\:-mr-20" | "xs\:-mr-24" | "xs\:-mr-3" | "xs\:-mr-32" | "xs\:-mr-4" | "xs\:-mr-40" | "xs\:-mr-48" | "xs\:-mr-5" | "xs\:-mr-56" | "xs\:-mr-6" | "xs\:-mr-64" | "xs\:-mr-8" | "xs\:-mr-px" | "xs\:-mt-1" | "xs\:-mt-10" | "xs\:-mt-12" | "xs\:-mt-16" | "xs\:-mt-2" | "xs\:-mt-20" | "xs\:-mt-24" | "xs\:-mt-3" | "xs\:-mt-32" | "xs\:-mt-4" | "xs\:-mt-40" | "xs\:-mt-48" | "xs\:-mt-5" | "xs\:-mt-56" | "xs\:-mt-6" | "xs\:-mt-64" | "xs\:-mt-8" | "xs\:-mt-px" | "xs\:-mx-1" | "xs\:-mx-10" | "xs\:-mx-12" | "xs\:-mx-16" | "xs\:-mx-2" | "xs\:-mx-20" | "xs\:-mx-24" | "xs\:-mx-3" | "xs\:-mx-32" | "xs\:-mx-4" | "xs\:-mx-40" | "xs\:-mx-48" | "xs\:-mx-5" | "xs\:-mx-56" | "xs\:-mx-6" | "xs\:-mx-64" | "xs\:-mx-8" | "xs\:-mx-px" | "xs\:-my-1" | "xs\:-my-10" | "xs\:-my-12" | "xs\:-my-16" | "xs\:-my-2" | "xs\:-my-20" | "xs\:-my-24" | "xs\:-my-3" | "xs\:-my-32" | "xs\:-my-4" | "xs\:-my-40" | "xs\:-my-48" | "xs\:-my-5" | "xs\:-my-56" | "xs\:-my-6" | "xs\:-my-64" | "xs\:-my-8" | "xs\:-my-px" | "xs\:absolute" | "xs\:align-baseline" | "xs\:align-bottom" | "xs\:align-middle" | "xs\:align-text-bottom" | "xs\:align-text-top" | "xs\:align-top" | "xs\:all-small-caps" | "xs\:antialiased" | "xs\:appearance-none" | "xs\:bg-auto" | "xs\:bg-black" | "xs\:bg-blue-100" | "xs\:bg-blue-200" | "xs\:bg-blue-300" | "xs\:bg-blue-400" | "xs\:bg-blue-500" | "xs\:bg-blue-600" | "xs\:bg-blue-700" | "xs\:bg-blue-800" | "xs\:bg-blue-900" | "xs\:bg-border" | "xs\:bg-bottom" | "xs\:bg-center" | "xs\:bg-contain" | "xs\:bg-cover" | "xs\:bg-fixed" | "xs\:bg-gray-100" | "xs\:bg-gray-200" | "xs\:bg-gray-300" | "xs\:bg-gray-400" | "xs\:bg-gray-500" | "xs\:bg-gray-600" | "xs\:bg-gray-700" | "xs\:bg-gray-800" | "xs\:bg-gray-900" | "xs\:bg-left" | "xs\:bg-left-bottom" | "xs\:bg-left-top" | "xs\:bg-local" | "xs\:bg-no-repeat" | "xs\:bg-primary" | "xs\:bg-repeat" | "xs\:bg-repeat-round" | "xs\:bg-repeat-space" | "xs\:bg-repeat-x" | "xs\:bg-repeat-y" | "xs\:bg-right" | "xs\:bg-right-bottom" | "xs\:bg-right-top" | "xs\:bg-scroll" | "xs\:bg-top" | "xs\:bg-white" | "xs\:block" | "xs\:border" | "xs\:border-0" | "xs\:border-2" | "xs\:border-4" | "xs\:border-8" | "xs\:border-b" | "xs\:border-b-0" | "xs\:border-b-2" | "xs\:border-b-4" | "xs\:border-b-8" | "xs\:border-black" | "xs\:border-blue-100" | "xs\:border-blue-200" | "xs\:border-blue-300" | "xs\:border-blue-400" | "xs\:border-blue-500" | "xs\:border-blue-600" | "xs\:border-blue-700" | "xs\:border-blue-800" | "xs\:border-blue-900" | "xs\:border-border" | "xs\:border-collapse" | "xs\:border-dashed" | "xs\:border-dotted" | "xs\:border-double" | "xs\:border-gray-100" | "xs\:border-gray-200" | "xs\:border-gray-300" | "xs\:border-gray-400" | "xs\:border-gray-500" | "xs\:border-gray-600" | "xs\:border-gray-700" | "xs\:border-gray-800" | "xs\:border-gray-900" | "xs\:border-l" | "xs\:border-l-0" | "xs\:border-l-2" | "xs\:border-l-4" | "xs\:border-l-8" | "xs\:border-none" | "xs\:border-primary" | "xs\:border-r" | "xs\:border-r-0" | "xs\:border-r-2" | "xs\:border-r-4" | "xs\:border-r-8" | "xs\:border-separate" | "xs\:border-solid" | "xs\:border-t" | "xs\:border-t-0" | "xs\:border-t-2" | "xs\:border-t-4" | "xs\:border-t-8" | "xs\:border-white" | "xs\:bottom-0" | "xs\:bottom-auto" | "xs\:break-all" | "xs\:break-normal" | "xs\:break-words" | "xs\:capitalize" | "xs\:clearfix" | "xs\:common-ligatures" | "xs\:content-around" | "xs\:content-between" | "xs\:content-center" | "xs\:content-end" | "xs\:content-start" | "xs\:contextual-ligatures" | "xs\:cursor-auto" | "xs\:cursor-default" | "xs\:cursor-move" | "xs\:cursor-not-allowed" | "xs\:cursor-pointer" | "xs\:cursor-text" | "xs\:cursor-wait" | "xs\:diagonal-fractions" | "xs\:discretionary-ligatures" | "xs\:ellipsis" | "xs\:fill-current" | "xs\:fixed" | "xs\:flex" | "xs\:flex-1" | "xs\:flex-auto" | "xs\:flex-col" | "xs\:flex-col-reverse" | "xs\:flex-grow" | "xs\:flex-grow-0" | "xs\:flex-initial" | "xs\:flex-no-wrap" | "xs\:flex-none" | "xs\:flex-row" | "xs\:flex-row-reverse" | "xs\:flex-shrink" | "xs\:flex-shrink-0" | "xs\:flex-wrap" | "xs\:flex-wrap-reverse" | "xs\:float-left" | "xs\:float-none" | "xs\:float-right" | "xs\:focus\:bg-black" | "xs\:focus\:bg-blue-100" | "xs\:focus\:bg-blue-200" | "xs\:focus\:bg-blue-300" | "xs\:focus\:bg-blue-400" | "xs\:focus\:bg-blue-500" | "xs\:focus\:bg-blue-600" | "xs\:focus\:bg-blue-700" | "xs\:focus\:bg-blue-800" | "xs\:focus\:bg-blue-900" | "xs\:focus\:bg-border" | "xs\:focus\:bg-gray-100" | "xs\:focus\:bg-gray-200" | "xs\:focus\:bg-gray-300" | "xs\:focus\:bg-gray-400" | "xs\:focus\:bg-gray-500" | "xs\:focus\:bg-gray-600" | "xs\:focus\:bg-gray-700" | "xs\:focus\:bg-gray-800" | "xs\:focus\:bg-gray-900" | "xs\:focus\:bg-primary" | "xs\:focus\:bg-white" | "xs\:focus\:border" | "xs\:focus\:border-0" | "xs\:focus\:border-2" | "xs\:focus\:border-4" | "xs\:focus\:border-8" | "xs\:focus\:border-b" | "xs\:focus\:border-b-0" | "xs\:focus\:border-b-2" | "xs\:focus\:border-b-4" | "xs\:focus\:border-b-8" | "xs\:focus\:border-black" | "xs\:focus\:border-blue-100" | "xs\:focus\:border-blue-200" | "xs\:focus\:border-blue-300" | "xs\:focus\:border-blue-400" | "xs\:focus\:border-blue-500" | "xs\:focus\:border-blue-600" | "xs\:focus\:border-blue-700" | "xs\:focus\:border-blue-800" | "xs\:focus\:border-blue-900" | "xs\:focus\:border-border" | "xs\:focus\:border-gray-100" | "xs\:focus\:border-gray-200" | "xs\:focus\:border-gray-300" | "xs\:focus\:border-gray-400" | "xs\:focus\:border-gray-500" | "xs\:focus\:border-gray-600" | "xs\:focus\:border-gray-700" | "xs\:focus\:border-gray-800" | "xs\:focus\:border-gray-900" | "xs\:focus\:border-l" | "xs\:focus\:border-l-0" | "xs\:focus\:border-l-2" | "xs\:focus\:border-l-4" | "xs\:focus\:border-l-8" | "xs\:focus\:border-primary" | "xs\:focus\:border-r" | "xs\:focus\:border-r-0" | "xs\:focus\:border-r-2" | "xs\:focus\:border-r-4" | "xs\:focus\:border-r-8" | "xs\:focus\:border-t" | "xs\:focus\:border-t-0" | "xs\:focus\:border-t-2" | "xs\:focus\:border-t-4" | "xs\:focus\:border-t-8" | "xs\:focus\:border-white" | "xs\:focus\:font-black" | "xs\:focus\:font-bold" | "xs\:focus\:font-extrabold" | "xs\:focus\:font-hairline" | "xs\:focus\:font-light" | "xs\:focus\:font-medium" | "xs\:focus\:font-normal" | "xs\:focus\:font-semibold" | "xs\:focus\:font-thin" | "xs\:focus\:line-through" | "xs\:focus\:no-underline" | "xs\:focus\:not-sr-only" | "xs\:focus\:opacity-0" | "xs\:focus\:opacity-100" | "xs\:focus\:opacity-25" | "xs\:focus\:opacity-50" | "xs\:focus\:opacity-75" | "xs\:focus\:outline-none" | "xs\:focus\:placeholder-black" | "xs\:focus\:placeholder-blue-100" | "xs\:focus\:placeholder-blue-200" | "xs\:focus\:placeholder-blue-300" | "xs\:focus\:placeholder-blue-400" | "xs\:focus\:placeholder-blue-500" | "xs\:focus\:placeholder-blue-600" | "xs\:focus\:placeholder-blue-700" | "xs\:focus\:placeholder-blue-800" | "xs\:focus\:placeholder-blue-900" | "xs\:focus\:placeholder-border" | "xs\:focus\:placeholder-gray-100" | "xs\:focus\:placeholder-gray-200" | "xs\:focus\:placeholder-gray-300" | "xs\:focus\:placeholder-gray-400" | "xs\:focus\:placeholder-gray-500" | "xs\:focus\:placeholder-gray-600" | "xs\:focus\:placeholder-gray-700" | "xs\:focus\:placeholder-gray-800" | "xs\:focus\:placeholder-gray-900" | "xs\:focus\:placeholder-primary" | "xs\:focus\:placeholder-white" | "xs\:focus\:shadow" | "xs\:focus\:shadow-2xl" | "xs\:focus\:shadow-inner" | "xs\:focus\:shadow-lg" | "xs\:focus\:shadow-md" | "xs\:focus\:shadow-none" | "xs\:focus\:shadow-outline" | "xs\:focus\:shadow-xl" | "xs\:focus\:sr-only" | "xs\:focus\:text-black" | "xs\:focus\:text-blue-100" | "xs\:focus\:text-blue-200" | "xs\:focus\:text-blue-300" | "xs\:focus\:text-blue-400" | "xs\:focus\:text-blue-500" | "xs\:focus\:text-blue-600" | "xs\:focus\:text-blue-700" | "xs\:focus\:text-blue-800" | "xs\:focus\:text-blue-900" | "xs\:focus\:text-border" | "xs\:focus\:text-gray-100" | "xs\:focus\:text-gray-200" | "xs\:focus\:text-gray-300" | "xs\:focus\:text-gray-400" | "xs\:focus\:text-gray-500" | "xs\:focus\:text-gray-600" | "xs\:focus\:text-gray-700" | "xs\:focus\:text-gray-800" | "xs\:focus\:text-gray-900" | "xs\:focus\:text-primary" | "xs\:focus\:text-white" | "xs\:focus\:underline" | "xs\:font-black" | "xs\:font-bold" | "xs\:font-extrabold" | "xs\:font-family-unset" | "xs\:font-hairline" | "xs\:font-light" | "xs\:font-medium" | "xs\:font-mono" | "xs\:font-normal" | "xs\:font-sans" | "xs\:font-semibold" | "xs\:font-serif" | "xs\:font-style-unset" | "xs\:font-thin" | "xs\:font-weight-unset" | "xs\:h-0" | "xs\:h-1" | "xs\:h-10" | "xs\:h-12" | "xs\:h-16" | "xs\:h-2" | "xs\:h-20" | "xs\:h-24" | "xs\:h-3" | "xs\:h-32" | "xs\:h-4" | "xs\:h-40" | "xs\:h-48" | "xs\:h-5" | "xs\:h-56" | "xs\:h-6" | "xs\:h-64" | "xs\:h-8" | "xs\:h-auto" | "xs\:h-full" | "xs\:h-px" | "xs\:h-screen" | "xs\:hidden" | "xs\:historical-ligatures" | "xs\:hover\:bg-black" | "xs\:hover\:bg-blue-100" | "xs\:hover\:bg-blue-200" | "xs\:hover\:bg-blue-300" | "xs\:hover\:bg-blue-400" | "xs\:hover\:bg-blue-500" | "xs\:hover\:bg-blue-600" | "xs\:hover\:bg-blue-700" | "xs\:hover\:bg-blue-800" | "xs\:hover\:bg-blue-900" | "xs\:hover\:bg-border" | "xs\:hover\:bg-gray-100" | "xs\:hover\:bg-gray-200" | "xs\:hover\:bg-gray-300" | "xs\:hover\:bg-gray-400" | "xs\:hover\:bg-gray-500" | "xs\:hover\:bg-gray-600" | "xs\:hover\:bg-gray-700" | "xs\:hover\:bg-gray-800" | "xs\:hover\:bg-gray-900" | "xs\:hover\:bg-primary" | "xs\:hover\:bg-white" | "xs\:hover\:border" | "xs\:hover\:border-0" | "xs\:hover\:border-2" | "xs\:hover\:border-4" | "xs\:hover\:border-8" | "xs\:hover\:border-b" | "xs\:hover\:border-b-0" | "xs\:hover\:border-b-2" | "xs\:hover\:border-b-4" | "xs\:hover\:border-b-8" | "xs\:hover\:border-black" | "xs\:hover\:border-blue-100" | "xs\:hover\:border-blue-200" | "xs\:hover\:border-blue-300" | "xs\:hover\:border-blue-400" | "xs\:hover\:border-blue-500" | "xs\:hover\:border-blue-600" | "xs\:hover\:border-blue-700" | "xs\:hover\:border-blue-800" | "xs\:hover\:border-blue-900" | "xs\:hover\:border-border" | "xs\:hover\:border-gray-100" | "xs\:hover\:border-gray-200" | "xs\:hover\:border-gray-300" | "xs\:hover\:border-gray-400" | "xs\:hover\:border-gray-500" | "xs\:hover\:border-gray-600" | "xs\:hover\:border-gray-700" | "xs\:hover\:border-gray-800" | "xs\:hover\:border-gray-900" | "xs\:hover\:border-l" | "xs\:hover\:border-l-0" | "xs\:hover\:border-l-2" | "xs\:hover\:border-l-4" | "xs\:hover\:border-l-8" | "xs\:hover\:border-primary" | "xs\:hover\:border-r" | "xs\:hover\:border-r-0" | "xs\:hover\:border-r-2" | "xs\:hover\:border-r-4" | "xs\:hover\:border-r-8" | "xs\:hover\:border-t" | "xs\:hover\:border-t-0" | "xs\:hover\:border-t-2" | "xs\:hover\:border-t-4" | "xs\:hover\:border-t-8" | "xs\:hover\:border-white" | "xs\:hover\:font-black" | "xs\:hover\:font-bold" | "xs\:hover\:font-extrabold" | "xs\:hover\:font-hairline" | "xs\:hover\:font-light" | "xs\:hover\:font-medium" | "xs\:hover\:font-normal" | "xs\:hover\:font-semibold" | "xs\:hover\:font-thin" | "xs\:hover\:line-through" | "xs\:hover\:no-underline" | "xs\:hover\:opacity-0" | "xs\:hover\:opacity-100" | "xs\:hover\:opacity-25" | "xs\:hover\:opacity-50" | "xs\:hover\:opacity-75" | "xs\:hover\:shadow" | "xs\:hover\:shadow-2xl" | "xs\:hover\:shadow-inner" | "xs\:hover\:shadow-lg" | "xs\:hover\:shadow-md" | "xs\:hover\:shadow-none" | "xs\:hover\:shadow-outline" | "xs\:hover\:shadow-xl" | "xs\:hover\:text-black" | "xs\:hover\:text-blue-100" | "xs\:hover\:text-blue-200" | "xs\:hover\:text-blue-300" | "xs\:hover\:text-blue-400" | "xs\:hover\:text-blue-500" | "xs\:hover\:text-blue-600" | "xs\:hover\:text-blue-700" | "xs\:hover\:text-blue-800" | "xs\:hover\:text-blue-900" | "xs\:hover\:text-border" | "xs\:hover\:text-gray-100" | "xs\:hover\:text-gray-200" | "xs\:hover\:text-gray-300" | "xs\:hover\:text-gray-400" | "xs\:hover\:text-gray-500" | "xs\:hover\:text-gray-600" | "xs\:hover\:text-gray-700" | "xs\:hover\:text-gray-800" | "xs\:hover\:text-gray-900" | "xs\:hover\:text-primary" | "xs\:hover\:text-white" | "xs\:hover\:underline" | "xs\:hyphens-auto" | "xs\:hyphens-manual" | "xs\:hyphens-none" | "xs\:inline" | "xs\:inline-block" | "xs\:inline-flex" | "xs\:inset-0" | "xs\:inset-auto" | "xs\:inset-x-0" | "xs\:inset-x-auto" | "xs\:inset-y-0" | "xs\:inset-y-auto" | "xs\:invisible" | "xs\:italic" | "xs\:items-baseline" | "xs\:items-center" | "xs\:items-end" | "xs\:items-start" | "xs\:items-stretch" | "xs\:justify-around" | "xs\:justify-between" | "xs\:justify-center" | "xs\:justify-end" | "xs\:justify-start" | "xs\:last\:border" | "xs\:last\:border-0" | "xs\:last\:border-2" | "xs\:last\:border-4" | "xs\:last\:border-8" | "xs\:last\:border-b" | "xs\:last\:border-b-0" | "xs\:last\:border-b-2" | "xs\:last\:border-b-4" | "xs\:last\:border-b-8" | "xs\:last\:border-l" | "xs\:last\:border-l-0" | "xs\:last\:border-l-2" | "xs\:last\:border-l-4" | "xs\:last\:border-l-8" | "xs\:last\:border-r" | "xs\:last\:border-r-0" | "xs\:last\:border-r-2" | "xs\:last\:border-r-4" | "xs\:last\:border-r-8" | "xs\:last\:border-t" | "xs\:last\:border-t-0" | "xs\:last\:border-t-2" | "xs\:last\:border-t-4" | "xs\:last\:border-t-8" | "xs\:leading-loose" | "xs\:leading-none" | "xs\:leading-normal" | "xs\:leading-relaxed" | "xs\:leading-snug" | "xs\:leading-tight" | "xs\:leading-unset" | "xs\:left-0" | "xs\:left-auto" | "xs\:line-through" | "xs\:lining-nums" | "xs\:list-decimal" | "xs\:list-disc" | "xs\:list-inside" | "xs\:list-none" | "xs\:list-outside" | "xs\:lowercase" | "xs\:m-0" | "xs\:m-1" | "xs\:m-10" | "xs\:m-12" | "xs\:m-16" | "xs\:m-2" | "xs\:m-20" | "xs\:m-24" | "xs\:m-3" | "xs\:m-32" | "xs\:m-4" | "xs\:m-40" | "xs\:m-48" | "xs\:m-5" | "xs\:m-56" | "xs\:m-6" | "xs\:m-64" | "xs\:m-8" | "xs\:m-auto" | "xs\:m-px" | "xs\:max-h-full" | "xs\:max-h-screen" | "xs\:max-w-2xl" | "xs\:max-w-3xl" | "xs\:max-w-4xl" | "xs\:max-w-5xl" | "xs\:max-w-6xl" | "xs\:max-w-full" | "xs\:max-w-lg" | "xs\:max-w-md" | "xs\:max-w-sm" | "xs\:max-w-xl" | "xs\:max-w-xs" | "xs\:mb-0" | "xs\:mb-1" | "xs\:mb-10" | "xs\:mb-12" | "xs\:mb-16" | "xs\:mb-2" | "xs\:mb-20" | "xs\:mb-24" | "xs\:mb-3" | "xs\:mb-32" | "xs\:mb-4" | "xs\:mb-40" | "xs\:mb-48" | "xs\:mb-5" | "xs\:mb-56" | "xs\:mb-6" | "xs\:mb-64" | "xs\:mb-8" | "xs\:mb-auto" | "xs\:mb-px" | "xs\:min-h-0" | "xs\:min-h-full" | "xs\:min-h-screen" | "xs\:min-w-0" | "xs\:min-w-full" | "xs\:ml-0" | "xs\:ml-1" | "xs\:ml-10" | "xs\:ml-12" | "xs\:ml-16" | "xs\:ml-2" | "xs\:ml-20" | "xs\:ml-24" | "xs\:ml-3" | "xs\:ml-32" | "xs\:ml-4" | "xs\:ml-40" | "xs\:ml-48" | "xs\:ml-5" | "xs\:ml-56" | "xs\:ml-6" | "xs\:ml-64" | "xs\:ml-8" | "xs\:ml-auto" | "xs\:ml-px" | "xs\:mr-0" | "xs\:mr-1" | "xs\:mr-10" | "xs\:mr-12" | "xs\:mr-16" | "xs\:mr-2" | "xs\:mr-20" | "xs\:mr-24" | "xs\:mr-3" | "xs\:mr-32" | "xs\:mr-4" | "xs\:mr-40" | "xs\:mr-48" | "xs\:mr-5" | "xs\:mr-56" | "xs\:mr-6" | "xs\:mr-64" | "xs\:mr-8" | "xs\:mr-auto" | "xs\:mr-px" | "xs\:mt-0" | "xs\:mt-1" | "xs\:mt-10" | "xs\:mt-12" | "xs\:mt-16" | "xs\:mt-2" | "xs\:mt-20" | "xs\:mt-24" | "xs\:mt-3" | "xs\:mt-32" | "xs\:mt-4" | "xs\:mt-40" | "xs\:mt-48" | "xs\:mt-5" | "xs\:mt-56" | "xs\:mt-6" | "xs\:mt-64" | "xs\:mt-8" | "xs\:mt-auto" | "xs\:mt-px" | "xs\:mx-0" | "xs\:mx-1" | "xs\:mx-10" | "xs\:mx-12" | "xs\:mx-16" | "xs\:mx-2" | "xs\:mx-20" | "xs\:mx-24" | "xs\:mx-3" | "xs\:mx-32" | "xs\:mx-4" | "xs\:mx-40" | "xs\:mx-48" | "xs\:mx-5" | "xs\:mx-56" | "xs\:mx-6" | "xs\:mx-64" | "xs\:mx-8" | "xs\:mx-auto" | "xs\:mx-px" | "xs\:my-0" | "xs\:my-1" | "xs\:my-10" | "xs\:my-12" | "xs\:my-16" | "xs\:my-2" | "xs\:my-20" | "xs\:my-24" | "xs\:my-3" | "xs\:my-32" | "xs\:my-4" | "xs\:my-40" | "xs\:my-48" | "xs\:my-5" | "xs\:my-56" | "xs\:my-6" | "xs\:my-64" | "xs\:my-8" | "xs\:my-auto" | "xs\:my-px" | "xs\:no-common-ligatures" | "xs\:no-contextual-ligatures" | "xs\:no-discretionary-ligatures" | "xs\:no-ellipsis" | "xs\:no-historical-ligatures" | "xs\:no-ligatures" | "xs\:no-underline" | "xs\:normal-caps" | "xs\:normal-case" | "xs\:normal-ligatures" | "xs\:normal-nums" | "xs\:not-italic" | "xs\:not-sr-only" | "xs\:object-bottom" | "xs\:object-center" | "xs\:object-contain" | "xs\:object-cover" | "xs\:object-fill" | "xs\:object-left" | "xs\:object-left-bottom" | "xs\:object-left-top" | "xs\:object-none" | "xs\:object-right" | "xs\:object-right-bottom" | "xs\:object-right-top" | "xs\:object-scale-down" | "xs\:object-top" | "xs\:oldstyle-nums" | "xs\:opacity-0" | "xs\:opacity-100" | "xs\:opacity-25" | "xs\:opacity-50" | "xs\:opacity-75" | "xs\:order-1" | "xs\:order-10" | "xs\:order-11" | "xs\:order-12" | "xs\:order-2" | "xs\:order-3" | "xs\:order-4" | "xs\:order-5" | "xs\:order-6" | "xs\:order-7" | "xs\:order-8" | "xs\:order-9" | "xs\:order-first" | "xs\:order-last" | "xs\:order-none" | "xs\:ordinal-nums" | "xs\:outline-none" | "xs\:overflow-auto" | "xs\:overflow-hidden" | "xs\:overflow-scroll" | "xs\:overflow-visible" | "xs\:overflow-x-auto" | "xs\:overflow-x-hidden" | "xs\:overflow-x-scroll" | "xs\:overflow-x-visible" | "xs\:overflow-y-auto" | "xs\:overflow-y-hidden" | "xs\:overflow-y-scroll" | "xs\:overflow-y-visible" | "xs\:p-0" | "xs\:p-1" | "xs\:p-10" | "xs\:p-12" | "xs\:p-16" | "xs\:p-2" | "xs\:p-20" | "xs\:p-24" | "xs\:p-3" | "xs\:p-32" | "xs\:p-4" | "xs\:p-40" | "xs\:p-48" | "xs\:p-5" | "xs\:p-56" | "xs\:p-6" | "xs\:p-64" | "xs\:p-8" | "xs\:p-px" | "xs\:pb-0" | "xs\:pb-1" | "xs\:pb-10" | "xs\:pb-12" | "xs\:pb-16" | "xs\:pb-2" | "xs\:pb-20" | "xs\:pb-24" | "xs\:pb-3" | "xs\:pb-32" | "xs\:pb-4" | "xs\:pb-40" | "xs\:pb-48" | "xs\:pb-5" | "xs\:pb-56" | "xs\:pb-6" | "xs\:pb-64" | "xs\:pb-8" | "xs\:pb-px" | "xs\:petite-caps" | "xs\:pl-0" | "xs\:pl-1" | "xs\:pl-10" | "xs\:pl-12" | "xs\:pl-16" | "xs\:pl-2" | "xs\:pl-20" | "xs\:pl-24" | "xs\:pl-3" | "xs\:pl-32" | "xs\:pl-4" | "xs\:pl-40" | "xs\:pl-48" | "xs\:pl-5" | "xs\:pl-56" | "xs\:pl-6" | "xs\:pl-64" | "xs\:pl-8" | "xs\:pl-px" | "xs\:placeholder-black" | "xs\:placeholder-blue-100" | "xs\:placeholder-blue-200" | "xs\:placeholder-blue-300" | "xs\:placeholder-blue-400" | "xs\:placeholder-blue-500" | "xs\:placeholder-blue-600" | "xs\:placeholder-blue-700" | "xs\:placeholder-blue-800" | "xs\:placeholder-blue-900" | "xs\:placeholder-border" | "xs\:placeholder-gray-100" | "xs\:placeholder-gray-200" | "xs\:placeholder-gray-300" | "xs\:placeholder-gray-400" | "xs\:placeholder-gray-500" | "xs\:placeholder-gray-600" | "xs\:placeholder-gray-700" | "xs\:placeholder-gray-800" | "xs\:placeholder-gray-900" | "xs\:placeholder-primary" | "xs\:placeholder-white" | "xs\:pointer-events-auto" | "xs\:pointer-events-none" | "xs\:pr-0" | "xs\:pr-1" | "xs\:pr-10" | "xs\:pr-12" | "xs\:pr-16" | "xs\:pr-2" | "xs\:pr-20" | "xs\:pr-24" | "xs\:pr-3" | "xs\:pr-32" | "xs\:pr-4" | "xs\:pr-40" | "xs\:pr-48" | "xs\:pr-5" | "xs\:pr-56" | "xs\:pr-6" | "xs\:pr-64" | "xs\:pr-8" | "xs\:pr-px" | "xs\:proportional-nums" | "xs\:pt-0" | "xs\:pt-1" | "xs\:pt-10" | "xs\:pt-12" | "xs\:pt-16" | "xs\:pt-2" | "xs\:pt-20" | "xs\:pt-24" | "xs\:pt-3" | "xs\:pt-32" | "xs\:pt-4" | "xs\:pt-40" | "xs\:pt-48" | "xs\:pt-5" | "xs\:pt-56" | "xs\:pt-6" | "xs\:pt-64" | "xs\:pt-8" | "xs\:pt-px" | "xs\:px-0" | "xs\:px-1" | "xs\:px-10" | "xs\:px-12" | "xs\:px-16" | "xs\:px-2" | "xs\:px-20" | "xs\:px-24" | "xs\:px-3" | "xs\:px-32" | "xs\:px-4" | "xs\:px-40" | "xs\:px-48" | "xs\:px-5" | "xs\:px-56" | "xs\:px-6" | "xs\:px-64" | "xs\:px-8" | "xs\:px-px" | "xs\:py-0" | "xs\:py-1" | "xs\:py-10" | "xs\:py-12" | "xs\:py-16" | "xs\:py-2" | "xs\:py-20" | "xs\:py-24" | "xs\:py-3" | "xs\:py-32" | "xs\:py-4" | "xs\:py-40" | "xs\:py-48" | "xs\:py-5" | "xs\:py-56" | "xs\:py-6" | "xs\:py-64" | "xs\:py-8" | "xs\:py-px" | "xs\:relative" | "xs\:resize" | "xs\:resize-none" | "xs\:resize-x" | "xs\:resize-y" | "xs\:right-0" | "xs\:right-auto" | "xs\:rotate-180" | "xs\:rotate-270" | "xs\:rotate-90" | "xs\:rounded" | "xs\:rounded-b" | "xs\:rounded-b-full" | "xs\:rounded-b-lg" | "xs\:rounded-b-none" | "xs\:rounded-b-sm" | "xs\:rounded-bl" | "xs\:rounded-bl-full" | "xs\:rounded-bl-lg" | "xs\:rounded-bl-none" | "xs\:rounded-bl-sm" | "xs\:rounded-br" | "xs\:rounded-br-full" | "xs\:rounded-br-lg" | "xs\:rounded-br-none" | "xs\:rounded-br-sm" | "xs\:rounded-full" | "xs\:rounded-l" | "xs\:rounded-l-full" | "xs\:rounded-l-lg" | "xs\:rounded-l-none" | "xs\:rounded-l-sm" | "xs\:rounded-lg" | "xs\:rounded-none" | "xs\:rounded-r" | "xs\:rounded-r-full" | "xs\:rounded-r-lg" | "xs\:rounded-r-none" | "xs\:rounded-r-sm" | "xs\:rounded-sm" | "xs\:rounded-t" | "xs\:rounded-t-full" | "xs\:rounded-t-lg" | "xs\:rounded-t-none" | "xs\:rounded-t-sm" | "xs\:rounded-tl" | "xs\:rounded-tl-full" | "xs\:rounded-tl-lg" | "xs\:rounded-tl-none" | "xs\:rounded-tl-sm" | "xs\:rounded-tr" | "xs\:rounded-tr-full" | "xs\:rounded-tr-lg" | "xs\:rounded-tr-none" | "xs\:rounded-tr-sm" | "xs\:scrolling-auto" | "xs\:scrolling-touch" | "xs\:select-all" | "xs\:select-auto" | "xs\:select-none" | "xs\:select-text" | "xs\:self-auto" | "xs\:self-center" | "xs\:self-end" | "xs\:self-start" | "xs\:self-stretch" | "xs\:shadow" | "xs\:shadow-2xl" | "xs\:shadow-inner" | "xs\:shadow-lg" | "xs\:shadow-md" | "xs\:shadow-none" | "xs\:shadow-outline" | "xs\:shadow-xl" | "xs\:slashed-zeros" | "xs\:small-caps" | "xs\:sr-only" | "xs\:stacked-fractions" | "xs\:static" | "xs\:sticky" | "xs\:stroke-current" | "xs\:subpixel-antialiased" | "xs\:table" | "xs\:table-auto" | "xs\:table-cell" | "xs\:table-fixed" | "xs\:table-row" | "xs\:tabular-nums" | "xs\:text-2xl" | "xs\:text-3xl" | "xs\:text-4xl" | "xs\:text-5xl" | "xs\:text-6xl" | "xs\:text-align-unset" | "xs\:text-base" | "xs\:text-black" | "xs\:text-blue-100" | "xs\:text-blue-200" | "xs\:text-blue-300" | "xs\:text-blue-400" | "xs\:text-blue-500" | "xs\:text-blue-600" | "xs\:text-blue-700" | "xs\:text-blue-800" | "xs\:text-blue-900" | "xs\:text-border" | "xs\:text-center" | "xs\:text-color-unset" | "xs\:text-gray-100" | "xs\:text-gray-200" | "xs\:text-gray-300" | "xs\:text-gray-400" | "xs\:text-gray-500" | "xs\:text-gray-600" | "xs\:text-gray-700" | "xs\:text-gray-800" | "xs\:text-gray-900" | "xs\:text-justify" | "xs\:text-left" | "xs\:text-lg" | "xs\:text-primary" | "xs\:text-right" | "xs\:text-size-unset" | "xs\:text-sm" | "xs\:text-transform-unset" | "xs\:text-white" | "xs\:text-xl" | "xs\:text-xs" | "xs\:titling-caps" | "xs\:top-0" | "xs\:top-auto" | "xs\:tracking-normal" | "xs\:tracking-tight" | "xs\:tracking-tighter" | "xs\:tracking-unset" | "xs\:tracking-wide" | "xs\:tracking-wider" | "xs\:tracking-widest" | "xs\:transform-b" | "xs\:transform-bl" | "xs\:transform-border" | "xs\:transform-br" | "xs\:transform-fill" | "xs\:transform-l" | "xs\:transform-none" | "xs\:transform-r" | "xs\:transform-t" | "xs\:transform-tl" | "xs\:transform-tr" | "xs\:transform-view" | "xs\:transition-0" | "xs\:transition-100" | "xs\:transition-1000" | "xs\:transition-250" | "xs\:transition-500" | "xs\:transition-750" | "xs\:transition-all" | "xs\:transition-bg" | "xs\:transition-border" | "xs\:transition-color" | "xs\:transition-colors" | "xs\:transition-delay-0" | "xs\:transition-delay-100" | "xs\:transition-delay-1000" | "xs\:transition-delay-250" | "xs\:transition-delay-500" | "xs\:transition-delay-750" | "xs\:transition-ease" | "xs\:transition-ease-in" | "xs\:transition-ease-in-out" | "xs\:transition-ease-out" | "xs\:transition-linear" | "xs\:transition-none" | "xs\:transition-opacity" | "xs\:transition-transform" | "xs\:truncate" | "xs\:underline" | "xs\:unicase" | "xs\:uppercase" | "xs\:visible" | "xs\:w-0" | "xs\:w-1" | "xs\:w-10" | "xs\:w-10\/12" | "xs\:w-11\/12" | "xs\:w-12" | "xs\:w-16" | "xs\:w-1\/12" | "xs\:w-1\/2" | "xs\:w-1\/3" | "xs\:w-1\/4" | "xs\:w-1\/5" | "xs\:w-1\/6" | "xs\:w-2" | "xs\:w-20" | "xs\:w-24" | "xs\:w-2\/12" | "xs\:w-2\/3" | "xs\:w-2\/4" | "xs\:w-2\/5" | "xs\:w-2\/6" | "xs\:w-3" | "xs\:w-32" | "xs\:w-3\/12" | "xs\:w-3\/4" | "xs\:w-3\/5" | "xs\:w-3\/6" | "xs\:w-4" | "xs\:w-40" | "xs\:w-48" | "xs\:w-4\/12" | "xs\:w-4\/5" | "xs\:w-4\/6" | "xs\:w-5" | "xs\:w-56" | "xs\:w-5\/12" | "xs\:w-5\/6" | "xs\:w-6" | "xs\:w-64" | "xs\:w-6\/12" | "xs\:w-7\/12" | "xs\:w-8" | "xs\:w-8\/12" | "xs\:w-9\/12" | "xs\:w-auto" | "xs\:w-full" | "xs\:w-px" | "xs\:w-screen" | "xs\:whitespace-no-wrap" | "xs\:whitespace-normal" | "xs\:whitespace-pre" | "xs\:whitespace-pre-line" | "xs\:whitespace-pre-wrap" | "xs\:will-change-auto" | "xs\:will-change-contents" | "xs\:will-change-opacity" | "xs\:will-change-scroll" | "xs\:will-change-transform" | "xs\:z-0" | "xs\:z-10" | "xs\:z-20" | "xs\:z-30" | "xs\:z-40" | "xs\:z-50" | "xs\:z-auto" | "z-0" | "z-10" | "z-20" | "z-30" | "z-40" | "z-50" | "z-auto";
