+++ src/components/repos/repo.language.ts
@@ -1,19 +1,35 @@
-import styled from '@emotion/styled'
-import { prop } from 'fp-ts-ramda'
-import { flow } from 'fp-ts/lib/function'
+import { pipe } from 'fp-ts/lib/pipeable'
 import njsx from 'njsx'
+import { span } from 'njsx-react'
+import styled from 'styled-components'
 
 import { getColor } from 'src/data/languages'
 import { tw } from 'src/utils'
 
-interface Props {
-  children: string
-}
+const dot = pipe(
+  styled.span.attrs(
+    tw('inline-block', 'relative', 'mr-1', 'rounded-full')
+  )`
+    content: ' ';
+    bottom: -0.15em;
+    height: 1em;
+    width: 1em;
+  `,
+  njsx
+)
 
-const Language = njsx(styled('span')<Props>`
-  &::before {
-    background-color: ${flow(prop('children'), getColor)};
-  }
-`)(tw('language'))
+const Language = (language: string) =>
+  span(
+    tw(
+      'inline-flex',
+      'mr-4',
+      'whitespace-no-wrap',
+      'overflow-x-hidden',
+      'ellipsis'
+    )
+  )([
+    dot({ style: { backgroundColor: getColor(language) } }),
+    language,
+  ])
 
 export default Language
