--- src/styles/dot.scss
@@ -1,13 +1,0 @@
-@mixin dot {
-  &::before {
-    content: ' ';
-    display: inline-block;
-    position: relative;
-    margin-right: theme('spacing.1');
-    border-radius: theme('borderRadius.full');
-    vertical-align: baseline;
-    bottom: -0.15em;
-    height: 1em;
-    width: 1em;
-  }
-}
