--- src/styles/global.scss
@@ -1,19 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-@tailwind screens;
-@import 'dot.scss';
-@import 'select.scss';
-@import 'language.scss';
-
-* {
-  scrollbar-width: none;
-}
-
-html {
-  @apply antialiased;
-}
-
-body {
-  @apply font-sans text-white bg-gray-700 leading-normal;
-}
