--- src/styles/language.scss
@@ -1,6 +1,0 @@
-@import 'dot.scss';
-
-.language {
-  @include dot;
-  @apply inline-flex mr-4 whitespace-no-wrap overflow-x-hidden ellipsis;
-}
